Use Jack SDK reporter format for output
Use Jack SDK reporter for processing output. This is now using
JSON messages emitted by Jack, and they fit the standard
format that the plugin expects from the tools. At the same
time, this ensures that the message structure will not change
and that we can reliable pass them to Studio.

Bug: http://b.android.com/225063
Bug: http://b.android.com/220602
Change-Id: Ieb35e64caf30a61a91e55e7e1c95da625fbee9e4

Former-commit-id: 4ab8784efc16e6baaef8955cc0994c49a6055677