Do not delete lock files in ReadWriteProcessLock.
ReadWriteProcessLock uses lock files to synchronize execution across
processes. Previously, we delete the lock file after escaping a
process-synchronized block. This is not safe because of two reasons:

- A process might be deleting the lock file while another process is
  trying to grab the same lock file.

- If we delete the lock file before releasing it, the deletion will fail
  on Windows because Windows does not allow deleting a file while it is
  still in use. If we delete the lock file after releasing it, on Linux
  we might end up deleting the lock file after another process has
  grabbed it, which is also unsafe.

To fix the second issue, we may choose to have a system-dependent
implementation: On Windows, we delete the lock file after releasing it;
On Linux, we delete the lock file before releasing it. However, the
first issue remains.

To address both issues, ReadWriteProcessLock will no longer delete lock
files.

As a consequence of this change, the lock files will now co-exist with
the cache entries within the build cache directory. These lock files
and the cache entries will be deleted altogether when the user deletes
the entire cache directory.

Test: ReadWriteProcessLockTest, BuildCacheTest
Change-Id: I2566abf4ab998a714a03d05ef777735376660c87

Former-commit-id: 8c846ee6deb60962ebdf3806048cc39fd5f267ab