New internal dependency model. step 1
New classes. The idea is to separate what the dependency
is from its usage in order to allow more reusability (in
particular when we have the aar cache instead of exploding
in each consumming module).

First there is the Dependency class tree which represents
a dependency itself. Its properties are its own, and not
how the dependency is used. There is no transitive
dependency information since it can change depending on how
it's used (due to resolution strategy)

Then there's the DependencyNode that's a graph of all the
transitive dependencies. It's meant to be immutable once
it's built so information like skipped is not there since
that's computed later. Provided information is also not
there since 1. it's not always needed, 2. it's not specific
to a node but to a particular dependency that can be present
several times in the graph, and 3. not putting it here allow
for more reuse of the nodes through out multiple graphs
(main,tests, compile and package, ...)

Both provided and skipped are kept in a separate map.

Then we have a dependency graph that contains the
information, and it gets flattened into a
DependencyContainer with the flatten information.
For the package container we directly remove the skipped
packages since they are never needed.

Test: some unit tests for the new classes
      mostly based on the previous tests.

Change-Id: I9f8ea26f4edfbd8706356e19b5ae6caef9869ec9

Former-commit-id: 8fd10806167f50e0a2f3641afe23dacb1c862bc4