Fix additional SDK_INT comparison scenarios, and symbol resolution
On top of the recent fixes to make the API checker smarter about
recognizing SDK_INT version checks in the code, this CL adds a few
additional scenarios that weren't covered by bug reports:
(1) Also handle the ternary operator, e.g. this is safe:
    return SDK_INT > 21 ? requiresLollipop() : null;
(2) Also handle negation, e.g. this is safe
    if (!isLollipop) { return; }
(3) Also handle cases where the SDK_INT is on the right hand side of
    the comparison:
    return LOLLIPOP < SDK_INT ? null : requiresLollipop();
(4) Fix problem combining version checks in conditionals;
    https://code.google.com/p/android/issues/detail?id=199572

Now that we rely on symbol resolution heavily for the API detector I
ran lint on some large projects, like Nest, and found some additional
problems in the classpath computation which affected symbol
resolution, which was fixed. Also noticed we were generating a huge
number of warnings (deprecations etc) which are ignored, so we might
as well not compute them.

This CL also updates the ServiceCastDetector to be aware of the new
ShortcutManager service when enforcing casts. And changed the lookup
table for that from constructing a static map to just using Java 7
string switch tables. Made similar optimizations for the various other
static lookup maps that were used in various detectors.

Finally, fixed a bunch of warnings by IntelliJ (e.g. make eligible
fields final, remove explicit type args when not necessary.) And
removed the deprecated (and unused) getSpeed() and appliesTo(File)
detector methods.

Test: Added a number of regression tests, not just for the new
scenarios listed above (see VersionChecksTest) but also for some other
issues handled by the api detector but that didn't have unit tests
here: issues 172621, 78495, 56236 and 93158.

Change-Id: I0382236625da12f85de8ad7c3fbdab4d89ed00ec

Former-commit-id: e47fc2cee13f0648a3b1d90b05300dbf3503fbd6