Enable Jack scenarios for AntennaPod performance test
Jack scenarios (Jack in process, Jack out of process) were disabled for AntennaPod due to failing tests. This CL fixes and enables them back on.
The exception are Instant Run benchmarks as Jack does not support Instant Run, because it does not generate intermediate class files when compiling an app.

Change-Id: I43cc9cf7a2a9b72608d484047568da567c1bd9ce

Former-commit-id: ef1cf39cfc879bf8bcd312bec548f688fcf7f83b