Fix @RestrictTo checking to not consider inherited annotations
The @RestrictTo checking was incorrectly flagging as restricted APIs
methods on classes that inherit from a restricted class. This happens
because we normally pull in all annotations in the hierarchy; for
example if a class extends a class where an overridden method is
annotated as @UiThread, then the overridden method is also taken to be
@UiThread. However, this is the wrong semantics for @RestrictTo.

This CL also fixes a related problem: The @RestrictTo check depends on
being able to look up Gradle group id that a given method is contained
in. This requires mapping from a .jar file back to an Android
library. However, this broke recently (when Gradle's build cache was
turned on, because .jar files are no longer in a location where we can
guess the group id from the jar file location
(exploded-aar/$group-id/$artifact-id/$version/...).  This CL fixes
that by searching through the AndroidLibrary dependency hierarchy
instead.

Finally, the RestrictTo annotation recently added LIBRARY and
LIBRARY_GROUP to the set of available constants, so it now looks for
those as well.

Test: Included

Change-Id: I49f1f706192b97ac7a38292cd0ec51165bdfb347

Former-commit-id: f8d97aaed8fb3e82821cc61f50475bb29b3e30a1