227579: Disable registration detector for now
The registration detector has some false positives that
are now showing up when we're actually correctly modeling
the Gradle module hierarchy.

The real fix for this is to expose the merged manifest
to lint detectors, which is coming, but in the meantime,
turn off this check (in Gradle; it's still enabled in Studio)
to avoid false positives.

Test: Not applicable. (The registration detector is fully
covered by tests, but it's now disabled.)

Change-Id: I3d245a2ca56d8c8d7a463c79cff8b043e49f9907

Former-commit-id: 2bde7059b8cba455c18bd54df26d86d7e0ea226d