Fix flaky concurrency tests.
ReadWriteProcessLockTest and FileCacheTest sometimes fail with the
following message: "Actions are taking too long to start/finish".

The reason is that they use InterProcessConcurrencyTester and
ConcurrencyTester to test the synchronization of threads and processes,
and those testers currently set a time limit (2 seconds) to wait for a
thread/process to launch and another 2 seconds for it to exit.

However, on build servers, threads/processes might take longer to start
or finish (even though the tested dummy threads/processes do nothing
other than launching and exiting).

To fix this issue, we remove those timeouts and assume that the
threads/processes will eventually start and finish. Even if they don't,
the test itself will eventually fail as it takes too long to run, and
the issue will never go unnoticed.

Note that there are still other timeout values in the testers that do
not cause the failures in question, and yet are required for the
concurrency tests. These timeouts are used to prevent the tests from
waiting indefinitely in the case that the currently running thread or
process is blocking other threads/processes from running (e.g., if they
are not able to run concurrently, whether by design or by error). Refer
to the javadoc of these timeouts for the rationale behind their values
and how the values are chosen to prevent any possible future flakiness.

Test: Rerun ReadWriteProcessLockTest and FileCacheTest unit tests
Change-Id: I0563d0c804059619f176d6d5b35929fbb34ba7cd

Former-commit-id: 8b4df90620c7bbbca4142139a1a9d665862beb2e