Added support for symbol table merging.
Removed the old SymbolLoader and SymbolWriter as they are
no longer used.

The current behavior while handling R.java is exactly the
same, with the difference that an intermediate in-memory
representation of the symbol table is now used and the
import / export is now clearly separted from the
in-memory representation.

Test: Included as JUnit tests.

Change-Id: Ib4fa44a8dd13a31c8e18d1f16da6d7ce0fb7a784

Former-commit-id: 49e19ae43a4bb976a665c6e0be033b041ebd957f