Fix issues 228570 and 228481 along with infrastructure
This CL fixes these two bugs:
    228570: StringFormatInvalid fixed in Lint but not in AS inspections
    228481: Don't flag API violations in plain java modules

In order to fix the second one it also improves the new testing
infrastructure a bit in order to be able to provide plain-java
test projects.

It also switches the AccessibilityDetectorTest from JUnit4 to JUnit3
since there are some problems in the test infrastructure running this
test as JUnit4. And it bumps the known stable version of Guava from
19 to 20.

Test: New unit tests included
Change-Id: Ib99c2e7b0857de3410d063df53dcb9e45ea9832d

Former-commit-id: 2066bb190718b8e47c119ab50f885426e9714dea