Do not cache exploded aars of snapshot Maven artifacts.
Snapshot Maven artifacts (whose versions ending with "-SNAPSHOT") are
not yet stable and therefore should not be stored in the build cache.

As a more stable fix to this type of issue, we now include a hash of the
artifact's contents when computing the inputs to the cache so that if
the artifact is still in development (e.g., if it is an "X.Y-dev"
version), the cache won't reuse outdated (incorrect) copies.

Although using hashes guarantees that the cache will produce correct
results, it is possible that the cache may create more and more entries
over time without ever using the old entries anymore.

In the future, we could allow the cache to delete unused old entries
automatically, or we could let the user decide whether to use hashes
behind a flag. In the worst case, the user can always delete the entire
cache if it grows too large.

Bug: http://b.android.com/228623
Test: New test in PrepareLibraryTaskTest
Change-Id: I4e1eedbc8253a43eba38da990070a8a8d8f51e71

Former-commit-id: 73a010f09d3fe0b596b06ee90f73de1a589432cc