Explode AARs during sync.
Previously, AARs were not exploded during sync. This was not a problem
before as we would later explode them in the project's build directory.
However, recently we began to explode AARs in the user-level build
cache. Because of this, the file watcher in the IDE does not know when
these files are exploded, and as a side effect, the IDE does not refresh
symbols, resulting in unresolved symbols.

Besides, we have code in the IDE that expects annotations.jar to exist
before attaching them to a library. This happens during sync.

Therefore, we need to explode AARs at sync time in order for the IDE to
work properly.

Test: Updated LocalJarInAarInModelTest
Change-Id: Ide73d75f2d247d9da7f9663912945f6205db9a04

Former-commit-id: a845ef3e19bd9b4c177fb3804c23010ee7ba7d41