Fix more problems.
Fix issues with non android sub projects by adding new streams with
their content which sometimes requires doing substraction with other
collections. This is because even though the library module publish
classes and java resources separately, java projects don't. (for
external aar we can separate them via the getFiles() filter).

Also add support for test component, by including the artifacts of the
tested aar into the FileCollections, or manually adding it for tasks
using ArtifactCollection. We need to find a better mechanism here.

This is handled by having the publishArtifact method now record what
gets published. This is handled by each variant (rather than being a
static method in AndroidArtifacts). When the test variant queries for
collection, it automatically look for a tested variant, get the matching
artifact (by type) and inserts it in the collection.

Test: fixing them!
Change-Id: If5e4316b5233d5dc8dfce324ceec0b08fba6c774

Former-commit-id: d7d501f2f9263a009e0266a96c48cdaa01383ac9