Adding sub-progress indicators
Right now, if you have some operation that takes a ProgressIndicator and
uses it to indicate progress, there's no way to incorporate that into a
larger sequence of operations. That is, if you're downloading 10 files,
the downloader will set the progress to 0 at the beginning and 100 at
the end of each download: there's no way to know that once you've
downloaded five files you're halfway done.
Adding the concept of a sub-ProgressIndicator that just delegates to the
current one, but maps the sub-indicators progress range onto only a
fraction of the parents.

Test: ProgressIndicatorTest
Change-Id: I019324184019c3e29e6c13fa7ede113d57797e48

Former-commit-id: d54c88f8f4006338aedb5cae06d3c4b7851d62d3