Do not continue after build cache failure.
Currently, if an exception occurs during the creation of the build cache
or during pre-dexing, we disable the build cache and continue the build.

Although we do print out a warning message in those cases, the user
might potentially miss the messages and might wrongly assume that the
build cache is being used. Also, we might keep pre-dexing a library
twice (we first try to pre-dex using the cache and fail, and then
silently try again without the cache), and that pattern is repeated for
every build.

To address this issue, we should simply fail the build and ask the user
to fix the underlying cause or file a bug, or let them disable the build
cache manually.

Test: BuildCacheUtilsTest, new tests in DexTransformTest and
PrepareLibraryTaskTest
Bug: http://b.android.com/231645
Change-Id: I76bac108e7e9961b2694f33201e85f0c282bc2d2

Former-commit-id: 8aa76863304e9c901ffd13e7651d24c9f07b4e30