force hasCode attribute on application to be true.
some applications only have native code and will set hasCode="false"
which is not true any more in IR mode where we add the IR service code.
Ultimately the IDE should probably not turn on IR mode in such cases
but changing the attribute to true will allow the application and
IR machinery to work correctly.

Bug : b.android.com/231842
Test: unit test

Change-Id: I097a0de3d5468670afd3d5987da92d1169e9bbb2

Former-commit-id: 009ac38dda1d13defd7af1d952b527db20eee875