Create FileCache without throwing an exception.
Currently FileCache construtor normalizes the cache directory's path
(via File.getCanonicalFile()), and throws an exception if it fails to
do so.

However, it is no longer necessary as the recently revised locking
mechanism already normalizes the file paths.

With this commit, creating a FileCache will now be fast and exception
free.

Test: Rerun FileCacheTest
Change-Id: Ie5f16401d770d76571fd5ff766223daf19442127

Former-commit-id: 65b43ced052ee001cd386327071de9e5fc9427bb