Fix InstantRun deploying split APK for the wrong ABI.
TaskManager, ApplicationTaskManager and many other places use
variantData.getOutputs().get(0) to get a single output among the
possible outputs, with the assumption that there is only one output.

In the case of multiple APKs, this is no longer true. This causes
Instant Run to pick up the wrong ABI to deploy (when the given ABI is
not the first output, which is usually the universal APK). Also due to
this wrong assumption, there might be other latent issues not yet
manifested as bugs.

This commit does two things:

- Replace all variantData.getOutputs().get(0) with
  variantData.getMainOutput(), where the main output is defined by the
  variant data when it is constructed.

- If an ABI is specified, set the main output of the variant data as the
  one corresponding to the given ABI. Otherwise, the main output is
  simply the first of the outputs.

Bug: http://b.android.com/227610
Test: Updated InstantRunFullSplitApkTest
Change-Id: Icdb8dc5fe91512d498af98a202089abc5b223dd6

Former-commit-id: 578ae20c13ed1d0b16818a0df3256f4a8d29302e