Fix jni packaging.
First, only strip library's own so and not all the
depenencies one.

Then, fix a problem with so files coming through folders. The issue
is due to java-library project now being processed as jars. Their
folders does include the lib/ part of a valid so path, but before we
considered that only jar dependencies would have lib/ while folders
(which was all AARs) didn't (because AAR don't actually have the
lib/ folder).

So now the distinction is not based on jar vs folders but on external
vs sub-projects. This requires also changing intermediate library project
outputs to contain the lib folder.

Change-Id: I57618547a6094f0aecae98e58c916c6b664caeb7

Former-commit-id: 872935a06d7d00ed44ab390cd38ee16f2283d9dd