Move API database from platform-tools into platforms/<name>/data
In an upcoming release the platform will start shipping the
lint API databases with the SDK components instead of the
platform-tools. This CL updates lint to look in both places
such that we also handle the case where users update platform
tools but don't yet compile with the preview SDK.

It also changes the resource lookup for these databases to
just use the base name such that the fallback logic can be
placed in a single place: the LintClient#findResource call.

Test: This code is already covered by tests (and these have
been updated.)

Change-Id: Iae88255899a44696615088ff13e4a43d48cd5de3

Former-commit-id: fcb9560e552026fb7148d9ad0acc9b070bd2ac01