Fix dependency Order.
Change ResolutionStrategy.Order from DEPENDENT_FIRST to
CONSUMER_FIRST. It makes more sense for most tasks.

This means the 2 tasks that needs the other order need to be adjusted
so that we reverse the order when processing the artifacts. This
is the resources and assets/res/jni mergers.

This fixes other tasks (like the manifest merger) that were using
the result of the FileCollection as-is when it actually needed to
be reversed. Now it can use the normal order.

Test: fixing them.

Change-Id: Ief335961a8e12c56328cf1400c34cb522da3f28f

Former-commit-id: 3903193988cbe1ef09b966fdd6d33cfc3d286bd6