Fix unit test support.
Also migrate part of it to FileCollection outputs in order to
embed task dependencies directly.

In order to handle the mockable Jar as a file collection,
this changeset also refactor the handling of task output from
VariantScope(Impl) to its own interface/class so that we can
both have a global version in GlobalScope and a per-variant
version in VariantScope(Impl).

Test: Fixing them.
Change-Id: Ib6ee28d3885645a63203937df6ec1d0f34a0d915

Former-commit-id: 3a7d7f041891cbd946e3070f679473f1f0717c2c