New templates for new project structure
This CL is an overhaul of how new projects and modules with
Instant App support are generated enabling projects that
meet the new structure described in http://b/34667965

Over all template changes, user-facing parameters are now
documented in the appropriate template.xml file which
will also allows future TemplateTest extensions to
automatically cover the template variations introduced by
InstantApp support. Template variables (including derived
values) are now documented in globals.xml making it easier
for UI updates to know what parameters are expected /
required by the templates and providing default values
in case the UI does not provide a value in some cases.

The new project templates now use an aia.properties file
more in keeping with projects generated by hand in the
existing demos. This file is a temporary solution to the
non-packaging of the SDK and non-published state of the
plugin / support libraries. Once these other tasks have been
completed this functionality can be removed

The new module template has new parameters added:
hasSplitWrapper which adds an empty split to wrap a new
library project and hasInstantAppWrapper which adds a new
instant app module and a base split / library pair to the
project. This allows us to output the entire multi-module
setup we want based on the single module the user has
requested (which in this case is the feature library).

Note that the ultimate aim here is to generate "feature
modules" and the current dual split / library
implementation is a temporary work-around. Updating to
feature modules should mainly be a case of simply removing
the unnecessary files from the templates / recipes.

Finally there is an update to the Activity templates.
Previously only an activity added to a new project was
given an intent-filter and only the domain could be
customised. This CL adds parameters to all activity
templates to allow users to set the URL domain, route
type and route. The recipe is also updated to inject the
intent filter into the correct location (currently the
library associated with the base-split).

Note due to a bug, activities containing constraint layouts
will not compile if added to the feature-library of a
project. This CL temporarily disables the generation of
constraint layouts in InstantApp enabled projects so that
functioning projects are generated.

Test: TemplateTests and NewProjectTest GUI Test confirm that
normal templates are not broken and InstantApps GUI test
confirm that new parameters are correctly producing a
working project.

Change-Id: Id66fd0d6ca62fcb15c72afbd63ce8a5ea29815e6

Former-commit-id: f6099d6ff0f30ef2db6de13d136726b411fffeac