The .dex reader (incrementally creating .dex IR)
This is the first part of the .dex manipulation pipeline: the
reader parses an in-memory .dex image and offers a low-level interface
for iterating the main .dex indexes, as well as the ability to
create .dex IR incrementally (once class at a time). Adding the .dex IR
for a class will automatically include the trasitive closure of all the
declarations referenced from that class so we always have all the .dex IR
that describes a partial but valid .dex file.

Also included a custom driver for end-to-end testing based on comparing
the expected ('golden') output with the actual output. Right now this is
fairly simple but we could make it more generic if we want to use this
for other components.

Test: added end-to-end tests (bazel only)

Change-Id: I36b26ffe872d89745ca532b2daf8e8890d34b94c

Former-commit-id: bfbbcd56e941e9efe2363449b8a12fc80018f360