[AVD Mgr] Create AVD user data file earlier
When an AVD is created or "wiped," create the modifiable
copy of the user data at that time. Incuring this cost
at creation or at wipe-data is less objectionable than
incurring it at the subsequent AVD start-up.

AVDs keep both a static and a modifiable copy of the user
data image. Currently, new and wiped AVDs have only the
static file--the modifiable file is created at launch if
it does not exist.

This change makes the copy happen early so that processing
does not happen at launch.

BUG: b.android.com/233123

Test: Extended testCreateAvdWithoutSnapshot(), testRenameAvd(),
      and testDuplicateAvd() in AvdManagerTest.java. Created
      AvdManagerConnectionTest.java with  testWipeAvd().

Change-Id: I7060906b7e5f1560b551131b0f229155c6865dd0

Former-commit-id: ee00808d95042f9c459530e2e991a270f7041d71