Wait for dex merging completion on the main thread only
There are now two executors used for dex merging, and the
transform will wait for tasks scheduled on both of them.

First executor is used to execute DexMergerTransformCallable
that schedules dex merging on the other executor. This
guarantees that all dex writing tasks will be scheduled by
the time DexMergerTransformCallables are done. At the same
time, worker threads will not block any more, which avoids
deadlock.

Because we are adding callables to the executor from
multiple threads, support has been added to the
WaitableExecutor.

Test: DexMergerTransformTest
Change-Id: I9c61948d9e01a8e6eb43d729bef5fa0264633bfb

Former-commit-id: 4cac1f3a4ef11d30fa8115906f5e88bff1b294bb