devicelib: Add ability to execute adb host commands
This CL adds infrastructure to execute adb commands, and adds a
couple of simple host commands (get version and kill server).

Each command is implemented as an instance of AdbCommand. When executed,
it receives a connection to the adb server and is responsible for writing
the command, and then parsing the server response.

The connection class (see StreamConnection) makes it simple to read and
write data in the format as expected by the adb server.

A command is executed against an instance of the adb server service,
which is responsible for setting up a connection, executing the command
on its executor, and then tearing down the connection. The result of
executing the command is returned via a future.

Test: unit tests included
Change-Id: I4e1ef84ce8e5208bdb4e7ddfa45488a9be58e042

Former-commit-id: 84e34f353654e1b5b66835acf9f98cd694f79dbc