Make lint test infrastructure work outside the tools tree
Some recent changes to make more tests stable in our environment
are causing problems for third party lint checks.

In particular, there are calls to find the "workspace root" (e.g.
the tools tree, by looking for the WORKSPACE file at the top) and
then looking up source files relative to it. This ensures that the
tests are always running against the current sources for things
like typo databases and API annotations.

However, when lint is running outside of our tree, the workspace
root isn't found and lint calls (such as attempts to open the
external annotations database) fails.

Test: Yes - this CL is updating tests to work correctly

Change-Id: I004cde75a27ba49968c7ac5d99618472561db1f9

Former-commit-id: 75ffe968c8c7558b5249a01ba0e3131354bf3bcd