Code IR lowering (to .dex bytecode)
This is the conversion from code IR to .dex bytecode, including debug information and EH (try/catch blocks)

With this we can do a full roundtrip :
.dex -> dex IR -> code IR -> .dex

In order to allow code IR modifications we need a few more pieces (allocating .dex IR nodes for example)
and helpers to make it easier than manipulating IR data structures directly. This will be added in a follow-up CL.

Also included is a new dexter option for running experiments. The first such experiment is "full_rewrite", which
demonstrates the full roundtrip through code IR.

Test: bazel test :dexter_tests

Change-Id: I34db27b07aa827592c618ddb01ddb816ff8cf590

Former-commit-id: ac3ea64967838d323a802548899ef4745c3c3312