Explode AARs during sync only when the cache is used.
An AAR may be exploded into either the user cache, the project cache,
or the project's intermediates directory.

The bug in question occurs when during sync, the static method
ResolveDependenciesTask.extractAarInParallel used by ModelBuilder
wrongly explodes an AAR into the project cache instead of the
intermediates directory.

We could fix the bug by letting that method explode the AARs into the
intermediates directory (for AARs that have that directory as the
exploded directory). However, it would introduce a new subtle bug: It
might interfere with the UP-TO-DATE check of the PrepareLibraryTask if
that task is executed at a later time. For example, PrepareLibraryTask
might later try to explode an AAR to the same location under the
intermediates directory which has already been exploded into earlier,
and Gradle may mistakenly assume that the output of the
PrepareLibraryTask is empty.

Therefore, we decide to not explode those AARs during sync (i.e., AARs
that are exploded without a cache). Since they will be exploded into the
project's intermediates directory, it will not cause problems to the IDE.
(The problems on the IDE only occur if AARs that will be exploded into
the user cache are not exploded during sync, which was the motivation for
the earlier CL that tries to explode all AARs during sync.)

Bug: http://b.android.com/237547
Test: New test in LocalJarInAarInModelTest
Change-Id: I4e8a1750692238196ecd4d59d4b5547ba089084a

Former-commit-id: da321fafe276697afac256339e1e092566091567