Split part of Supportlib into Agentlib
This changes sets up the modules for post-O support, where an
agent will be loading in the Agentlib at runtime. We now output
two transform jars. One that include both supportlib and agentlib
(profilers-transform.jar) and one that only includes supportlib.
(profilers-transform-slim.jar). An additional command is also run
to dx the agentlib which will be used for runtime loading.

Separated out the Network instrumentation code as an example.
Supportlib now contains only the HttpWrapper classes and the
rest resides in Agentlib.

Minor clean up to the gradle build files to use the new compileOnly
commands and removed configurations no longer needed.

Bug: 36189101
Test: N/A
Change-Id: I9f3193ffef8c54a9f9943c1810e78e55e730a659

Former-commit-id: b276138782801b1a428eb6194b23a7ca78069b5f