Moving event initialization to a thread.
Events don't work on the butterknife sample due to the event system,
attempting to grab the Apps application class before the application class
is initialized. Moving the event system init to a thread, allows us to try until
we are able to grab the Application and register events on that applicaiton.

Test: None
Change-Id: Iac3b392c73db3120e2740953a3a3a847fde00f47

Former-commit-id: 93fb80ed823e806875b109f93710d3bb2755b209