Passing Fragment Events to Studio via Activity RPC.
This change adds basic support for fragment added/removed events.
The events are stored in the cache, and passed to studio via
GetActivityData RPC. A flag has been added to the RPC to determine
if the event is a fragment or an activity. The cache had to be updated,
to store the list of active events.

Test: None
Change-Id: I889f317877480ff96f8bf68e350f2b77d7548a76

Former-commit-id: 8c5e2b6fbe4a6846dbf51ee5fe774f186c434cd4