Java 8 lang support messages
Log deprecation warning for Jack. If Desugar is
disabled (via flag), log error if 1.8 is used without
external plugins that support it. If external plugin is
used inform users that we have our built-in support.

Fix error reporter to output messages in json format if
plugin is invoked from the IDE.

Test: JackDeprecatedTest
Bug: http://b.android.com/234217
Change-Id: I954a1f154a81b5c55bacd3b9db5d2ab842ef9490

Former-commit-id: 5ce3b5dad46ae81dacd6e97c498b73e2e149a2ce