Stress exit hooks instrumentation
For every method in the .dex image, insert an "exit hook" call
to a fictitious method: Tracer.OnExit(<return value...>).
OnExit() is called right before returning from the instrumented
method (on the non-exceptional path) and it will be passed the return
value, if any. For non-void return types, the return value from OnExit()
will also be used as the return value of the instrumented method.

This is similar to the stress_entry_hook test case, in that it's just
an example of how the instrumentation can be done. For real world usage
we'll build higher level helpers.

Test: bazel test :dexter_tests

Change-Id: I297c44b8cf6df5af77b475b79a4516aad506073f

Former-commit-id: 33c91955c0f70a134afa5dd49ee58fc7b26416a4