Renames perfa to support jni to make room for the real perfa
perfd and perfa are our daemons and agents, however
with time perfa became the jni library inside supportlib.

Now that we have the real agent coming, this CL makes only
a rename. The object shared by both perfa and the jni of
support lib is called Agent and so is the perfd service.

Now the jni library of supportlib is called supportjni.

And perfa will be the real name of our agent

Test: existing tests.
Change-Id: I908adb3e9e93ee8adca63a187d5a423880805d84

Former-commit-id: 11eb954ff6c8ca85995854fff30ffed1ab80877a