Prepare for new postprocessing DSL
Deprecated BuildType::isMinifyEnabled. As the decision tree for what
shrinking tool should run in what mode gets more complicated (and there
are more DSL options to control it), this simple flag should no longer
be read directly, if possible.

I also renamed the transform for the built-in (soon to be known as
'android-gradle') shrinker.

Test: pure refactoring, existing tests.
Change-Id: I5b2170a4e27ccd013b386fbd3f706813a5ceff1d

Former-commit-id: e6ce692b81e9b1f8bcba7cc4d715a6da1d5d102a