Don't fail the build if platform-tools is missing
Now that the API database for lint is shipped in the platforms/,
platform-tools is not required for a successful CI build. So we will
print a warning (and show it in the IDE with a quickfix), but carry on
building.

We will still download the package if the build is not run with
`--offline`, since tests cannot pass without it.

Test: SdkAutoDownloadTest
Change-Id: I892edd74bef365b581be114a0813ad6a0dac4824

Former-commit-id: 864e000ed4b80c2691b543bdcbbb5d073a1407a5