Refactor DSL tests
Migrate the tests to JUnit4, and try to simplify
the tests.

This CL removes the base BaseDslTest class, and injects
all necessary information in the test classes. Logic
for examining variants has been pushed to the
fixture, and creation of android test projects is
now reusable.

The tests shared between atom and app plugin are now
in a PluginDslTest class, that is parameterized.

Test: existing DSL tests
Change-Id: I531d450487f23e406c087925236f47dca2e50dd1

Former-commit-id: 85722b8775c1ffeb81bd597cba733c73523e896e