Initial commit for QA profiler app
This is an initial checkin of a new Android project created from Studio.
It will serve as a foundation for an app which will be used by QA to
verify that profilers are working correctly. However, in this first
checkin, there is absolutely no new code except for boilerplate
generated by Studio.

Test: NA
Change-Id: Ibfd912f53a1d842a575a9bbbd025d2e9b830d528

Former-commit-id: ab92311a833e4388215a9358ac486345eab4563d