Set test_summary=detailed for bazel on go/ab
This is to make bazel list failed test cases at the end of a run. Right
now, users have to scroll up to find which tests failed. Gradle tests
are very noisy in that they generate lots of stdout/err and users have to
sift through TONS of lines just to get to the test class. Purely
cosmetic change.

Test: tested with --test_summary=detailed in local bazel workspace.
Change-Id: Ie3be046b83d27f92c99c71afae30cdd0e88878a3

Former-commit-id: 40f7b2cadf0c872d70a0879a5c2fe1daa0ddc87e