quick fix for databinding in libary androidTest.
When databinding is applied to a library project, the annotation
processor needs to be applied automatically to the androidTest
dependencies as well.

It used to be automatic because a library project test dependencies
extended from the main component but that's not the case anymore.

This way we actually only include the databinding annotation
processor and not every other processor that are added which is
better.

Bug: 38170505
Test: update DataBindingTest to test the output for androidTest
Change-Id: I63aca8206746f6e6f52cc8736398a7725ee21a8f

Former-commit-id: 865361a235da5a8dde2a5b4b64a32c59a5744d4e