Encoding callstack data in agent before transmission
Added simple, special-purpose prefix tree structure to encode
callstacks for allocation events.

Stats:
For DisplayingBitmaps, out of ~122k allocations with callstack,
a total of 2140 unique methods were encountered. The stacks generated
from those methods created ~4.1k unique leaf nodes in the trie, with
a total of ~9.3k nodes. The callstack depths were measured to be ~25
on average per allocation.

Previously, we send the entire list of method ids (longs) with each
allocation:
122k x 25 = ~3M (longs) = ~24Mb of data

After this change, we send the encoded stack id (ints) with each
allocation, and we only send the list of method ids for each unique stack
id the trie encounters:
122k ints + (4.1k x 25 (longs)) = ~1.3Mb of data

Bug: 37175866
Test: unit test for prefix tree
Change-Id: Ie278d2c08c4c43734147c4b0c3cc4a10c4e23e00

Former-commit-id: 213dcea2b9b2df760b025e1159ee5c3928123f28