Use org.jacoco.agent:x.y:runtime directly.
we use to unzip the agent to get the runtime and then package it in the aar.
However the latter does not work anymore: we don't package local jars inside
aar during project to project publishing.

This was always kind of broken anyway because this local jar could cause
conflict if 2+ libraries used code coverage. So using a proper dependency
is better.

To make this happen we make the runtimeClasspath extend from the agent
configuration where the dependency is set. This automatically fills
the streams for the apps and fill the runtimeElements for libraries.

Bug: 38411800
Change-Id: I3d05aa1ddcb2322bd1c0a03feb628e2622851c9f
Test: New JacocoInLibTest

Former-commit-id: 3b93a6b50552de62dfecf76564f407b33db23744