Fix Out-of-Memory error in IncrementalFileMergerOutputs.
The current file merging algorithm works by merging file input streams
into a ByteArrayOutputStream, copying this output stream into a
ByteArrayInputStream, and then writing this input stream to a file.

This is inefficient. Also, if the input file is large, loading the
entire file into a byte array in memory could result in an
OutOfMemoryError, as happened in http://b/38345547.

To fix this, we remove the intermediate copying step and write the input
streams to the file directly (i.e., the output of stream merging in
StreamMergeAlgorithm is not an OutputStream but is still an
InputStream).

Bug: 38345547
Test: Updated StreamMergeAlgorithmsTests,
      IncrementalFileMergerOutputsTest.
Change-Id: Ib89589e8838b86731e07bd63474cee59a03cee66

Former-commit-id: 6ef22073ecc54799877cde556ba16da5eb532090