Prevent silent error on missing variant.
Disable the default configuration as consumable so that
if the resolution fails to find a matching variant it does not
fall back on default which is a match with no artifact.

However we don't want sync to fail so we only do this if sync
is not on. Lenient mode should handle this but does not.

Bug: 37136720
Change-Id: I21a5f74e20b6759665112f0eb4721160f773c773
Test: new integ test: BuildTypeMismatchTest

Former-commit-id: d36db1e2b27ec0d60528db45010f4f5b6ac73a3b