Add fallback logic to Interceptor#intercept handling
When running through a sample app, I was able to occasionally
get a stream reset exception. When that happens, our Proxy
invocation handler would return null for the OkHttp response,
which would crash the user's app.

Now, we add a fallback path, which makes sure that even in the
worst cast, the app continues to work.

Test: To do later with integration testing
Change-Id: Ibac10c20820b8bfbc57262f242649a286e3ef832

Former-commit-id: b20412020692c3dfd003e56b3f69d98e3cb8c842