devicelib: use channels instead of streams
The connection between the devicelib to the adb server was implemented
using (input & output) streams. Streams however cannot be cancelled,
i.e. if a process is blocked waiting to read from a stream, it cannot be
cancelled by closing the stream. However, we need support for canceling
in flight operations, so this CL changes the implementation to use NIO
channels instead.

Test: existing tests updated
Change-Id: Id727945ab5ef16a84217f1e0a80346cfed694510

Former-commit-id: caaa0c82b31130de5866d58e1b60cef855f25766