Remove :app:lint dependency on :app:assemble.
I'm not sure why the dependency on :app:assemble exists for :app:lint,
and removing it only broke one test. Investigating the test, I'm also
not quite sure what the test exists for. The change required to make it
pass is fairly trivial, and remains ~semantically equivalent to what it
was originally.

Bug: 62060471
Change-Id: Ie706fbceee951e5e9dcd7e4ed7742799505f36f9
Test: bazel test //tools/base/...

Former-commit-id: 051d773b52eabfecd2773c440d6a48254d9d7d65