37140703: lint will report all lint warnings and errors from dependencies
Lint deliberately analyzes all dependencies for a project along with
the project. This is necessary to for example correctly identify
unused resources in a codebase; you have to consider both the
declaration of the resource in a library and potential usages in
clients of the library.

However, this is a bit counter to how Gradle normally works (where
you get one report per project), and it makes things quite a bit
slower (since libraries can be analyzed multiple times).

This CL adds an option for users to deliberately turn off analyzing
dependencies:

android.lintOptions.checkDependencies = false

Test: Integration tests included
Change-Id: Ic4c6a9ddde714ef6f6267959fc7b10fc99033d90

Former-commit-id: 41fc089c8e40874cae9777b7c3022e8793fba655