Ensure Tools PSQ profile is a subset of Tools IDE local.
This is a one-off sync to the profiles. There is a hook
being added by ag/2394341 to ensure profiles are sync'd
on every change to them so manual sync wouldn't be required.

Note: This includes attributes reordering as this is how python’s
etree dumps xml. This impacts diff, but the actual difference in
inspections is about 30 of them.

Test: N/A
Change-Id: I8f75dc0a661108588b357eebd8b99f5a00fbe8a1

Former-commit-id: cf4700431e0919ee7438cfd0995a0cdc25fedca2