Move ide model classes to sdk-common.
Move classes to new package com.android.ide.common.builder.model.
Move stub classes and unit test classes as well.

This is part of the effort to switch to DependencyGraphs in IDE. Move Ide*
classes to sdk-common so they are available to modules in tools/base.

Made the following necessary changes to original files:
1. In IdeModel.java, replace Computable with Supplier, remove annotator
   Contract.
2. In ModelCahce.java, change @TestOnly to @VisibleForTesting.
3. In IdeBaseArtifactImpl.java, replace intellij.contains with List.contains.
4. In IdeLevel2DependenciesFactory.java, remove dependency on NewGradleSync,
   check new model with dependencies field.
5. Change jetbrains.annotations to android.annotations.

Test: Unit tests.
Change-Id: I97439030d450410fbf725706ff764289bda62dce

Former-commit-id: 89a7c2693595b00bafcbd841592b89ca98cbd027