Add D8 dex builder and merger
This CL adds options to use D8 dexer
and merger. Feature is behind a flag, that is
off by default.

Transform tasks read these flags, and choose
the appropriate implementation of dexer and
merger. Relevant unit tests have been updated
to run with both dx and d8.

R8 is built, and packaged as part of builder
module. R8 iml module is created as well.

BUG=62861605
Test: DexArchivesTest, DexArchiveMergerTest,
DexArchiveBuilderTest, existing tests with flag on
Change-Id: Iba6f065b089c0f589885f2389a9923f50ba7bdba

Former-commit-id: 38edacb7b3ad978e80e2e8ad8e248d7a5f08d88d