Fix AIA module template
Dependencies on support libs should be handled in recipe.xml, because
that mechanism allows us to resolve the latest version and not put "+"
in the generated project. The dependency on support-v7 is already in
recipe.xml, so ended up creating base/build.gradle which included both
an "api" and "implementation" dependency on support-v7, with the
"implementation" one always using "+".

Bug: 63147281
Change-Id: Ic8dab1dbe6bf8b83001ef27d1f9d3cff7044b224

Former-commit-id: a12a3cafed086fa9a676667872b5dc6fc9d8ec38