Adaptive two-level trace representation
Seperates trace data into two parts:
1. trace_info: where meta/header data supposed to be located,
   such as traceId, startTime, endTime, threads.
2. data: where we keep the binary trace file

So api looks like:
* getTraceInfo(range) -> list of trace_info
* getTrace(trace_id) -> (trace_data, profilertype)

Some advantages:
1. Faster and cheaper L2 in cpu. Previously the threads view to be
able to color thread states under captured area, it was quering for
all captures. If it doesn't exist in memory it was parsing, so bringing
them back to the memory. With this change it doesn't need to parse
trace file. The same behaviour for duration renderers
in Cpu L2.
2. Ability to keep only the current capture object or part of it,
   so leaving much room for memory improvement.

Bug: 62692407
Bug: 63289171

Test: coming

Change-Id: If27275c96434aa8a7c1e116a8458912485f6be6e

Former-commit-id: e3414ad9d00b8e985b235fd0448e2022ea7c294d