Fix publishing of AARs.
Publishing an AAR requires it to be in 'archives', but
this is not enough as its consumers would not find it during their
own publication. So putting it to 'default' solves it (it also
gets added to 'archives' automatically). It does not interfer
with the normal inter-project publishing as Gradle will favor
direct intermediate artifacts vs the AAR being used via
transforms.

Bug: 63321533
Change-Id: I91350f67369d53bb31570492a60110a72e7ac93a
Test: none for now.

Former-commit-id: e2eb89c49e84c601fc4cf615badcfe25720ef6a8