Dont clean up too much after tasks.
running with --rerun-tasks makes an @InputFile annotated
getter be called after the task is run. While we figure out
what's going on, we stop doing the clean up.

However, this change also changes things a bit. The previous
mechanism built a more than needed data structure for the input
and the processing. Now this uses a simpler collection of files
for the input and delays creating the more complex data structure
for task execution. This mitigate the lack of clean up.

Bug: 63096043
Change-Id: I827420473e9167897f480f5607eb031aec8612ee
Test: existing ones.

Former-commit-id: 9283ad95e641e32cba81407702bdeb6b0a856b1f