Bump autovalue to the latest version: 1.4.1.
Recently, the tools tests have been failing sporadically while trying to compile the gradle plugin. gavra@ added some flags to aid in debugging, which pointed to an IOException being thrown in the autovalue library.

We tracked this down to the following issue https://github.com/google/auto/issues/365, which is present in 1.3 but the CL that fixes it (https://github.com/hzsweers/auto/commit/4070289243cbb40c9cde48b96a764c56fd564d1b) is not present until 1.4.

This should fix the sporadic build failures, but given that I was unable to reproduce the problem locally I can't be 100% sure. If it turns out that this CL causes other problems, it should be relatively safe to roll it back.

Test: ¯\_(ツ)_/¯
Change-Id: I2dfb889fb5069144db479dcd650516565a50acfe

Former-commit-id: a6d435d90988696faec681003d62c94c1563748c