Fix the "targetSdkVersion > compileSdkVersion" quick fix
Fix "The targetSdkVersion should not be higher than the
compileSdkVersion" quick fix that was associated with a wrong file
location.

Modify the order of visiting of AST nodes in GroovyGradleDetector and in
GradleDetectorTest to make it more natural and to make sure that
GradleDetectorTest would be able to catch the bug in GradleDetector.

Test: GradleDetectorTest
Bug: 63869544
Change-Id: Id4b96c2c7e746b9b88d5581a1323a7c91ea4eb02

Former-commit-id: 311ed1b268d898b60fd919308514a68b16ffd80c