Increase O+ perfd->perfa connector timeout
The "wait for debugger" workflow can vary quite a bit in time. In this
bug's case, Studio ended up waiting for LLDB to initialize which took
20+ seconds - over the limit of our existing attach agent timeout.

There can potentially be other scenarios where waiting can take a long
time. e.g. if the user starts the app outside of Studio with "-D", and
go through the "Attach debugger to Android Process" workflow at a later
time.

This change increases the retry timeframe to 5 minutes which should
largely mitigate the issue. We might need a better solution when
profiler's process discovery logic is moved onto device.

Bug: 64003406
Test: N/A
Change-Id: I16d04f9c2b3c6f57d2b5b3cbabd2aeda517eb776

Former-commit-id: 4d3527ed1b850fb687a0dd8eadcd91c023c90858