Remove equals methods from DSL classes.
The only place these classes are used is in the DSL where they are
modified by the user, and we then use the information they contain to
configure the build.

Each one has a unique name within a project, so two in a project should
never be equal, and comparing them between projects doesn't make sense.

They are stored in NamedDomainObjectContainer which maps from a string
name to an instance.

Test: existing tests.
Change-Id: I529312f501fdd98e3f4f7694fae913273adfe234

Former-commit-id: 6a780d0016e27f1424eef3561ddaa67ef89e3edf