Use main resources as import to split apk builder.
when building split APKs in InstantRun mode, it is possible
that the small manifest generated for each split reference
a resource. Pass the main resources bundle as an import
but not a dependency (as the dependency are set through
the transfrom input parameters) to avoid rebuilding all
split APKs each time the resource bundle changes.

Also pass the resources bundle containing the main
manifest to trigger full rebuild in case the main
manifest has changed (which might cause the
embedded resource ID to change).

Bug=37137338
Test: added integration-test

Change-Id: I5d469b7a5501ced2f4e20f7a55ccc34c2e66bb01

Former-commit-id: 6770faff98752bc29aeb9983dc732c84492d99a8