Return false for MESSAGE_PING if app has crashed
Server was returning true for MESSAGE_PING when app has crashed which
leads to studio trying to do a HOTSWAP to update the application. Such
update would fail since the activity is not actually in the foreground.

When an app has crashed, the activity's paused flag is false while the
crash dialog is visible. To detect if an application has crash, check
getProcessesInErrorState and check the app's packageName is not in the
list of ProcessErrorStateInfo.

Now the activity won't be in the foreground if app has crashed. So
Studio will do a COLDSWAP for update which restarts the application.

Fixes: b/37133823
Test: n/a
Change-Id: I67ccb86b2b117aae5ea59eb92189060e3a7a6c40

Former-commit-id: d487467e197d33c2bacd5c0a52a0d11344584b16