[IAPP] App with GoogleMaps activity fails to build
Instant apps, need all resources referenced by the Feature module
manifest to be on the Base module. In the case of GoogleMaps activity
this means all content of strings.xml and google_maps_api.xml
- Update recipe.xml.ftl to either copy resource as usual, or copy them
to the instant app module root, if we have an instant app project.
- Removed from globals.xml.ftl all variables already defined by
"../common/common_globals.xml.ftl"

Bug: 64375704
Test: Run TemplateTest. Manual test, as per bug, for the iapp path.
Change-Id: I2420df720444ff8667859fe43d379bebb02b017a

Former-commit-id: 9fcea6863f8a709b572da55fa2f5b83df3e58593