Add manifest merging feature to automatically handle conflicts
This CL adds a new feature flag to the manifest merger:
HANDLE_VALUE_CONFLICTS_AUTOMATICALLY

When this flag is set, and the manifest merger comes across a case
where an attribute is set by both the higher and lower priority
manifests but to *different* values, and there's no tools:replace to
guide it, it will now just stick with the first value.

This feature has only one intended use case: the manifest merger
invoked to insert a new template into the current manifest.

For example, in bug 64552606 there's a scenario where the manifest
already contains the attribute definition standalone="true", and then
you try to insert a new template which attempts to set that attribute
to standalone="false". Currently, this makes the manifest merger abort
with an error, which makes the project break -- we just show the old
and previous manifests to the user and make them pick. This isn't
really helpful. With the new merging flag, set only for this case
(merging a previous version of a manifest with the manifest fragments
specified for a new template), we'll merge in the new changes for the
template and reject the values that are incompatible with the current
manifest.

Bug: 64552606
Test: ManifestMerger2SmallTest#testAutomaticallyHandlingAttributeConflicts
Change-Id: Idb8a84b6ca3ea37fb25ce147457ee7f1e8db4879

Former-commit-id: cfb4a7812c3489e9d8a0b12c757f6396aac3fe60