Updates wear templates to use new wear widgets for 26.
Also, removes deprecated swipe to dismiss classes.

Important Note: This CL was already +2 and submitted (original
CL/2774915) but
reverted because the test CL associated with it required the sandboxed
Android Studio test instance to have the latest version of the
support library (which it did not). It will have it in there before
submit this time.

Bug:64723675
Bug:62281168
Test: test covered with unit tests in TemplateTest.java methods: testNewProjectWithGoogleMapsWearActivity(),
testNewProjectWithBlankWearActivity(),
testNewProjectWithBlankWearActivityWithKotlin(), and
testNewProjectWithGoogleMapsWearActivityWithKotlin().

Change-Id: I0b891dd84d82fb9fac7fc358d25bdd41ce075999

Former-commit-id: dbe5e4ca41686122629fc5fc87a4379c5e38f0cf