Adds port forwarding support to fake ADB Server.
This CL adds fake port forwarding support. Since it is "fake" (as in
no actual forwards are occurring), it supports Unix domain sockets
as well, even on non-Unix-based hosts.

However, since no real forwards are being made, one side of the test
needs to know it is in a test mode. This usually means the services
that the client or test code is attempting to connect to will know
about the fake. Therefore, the services will need to be listening on
the host port that the client is attempting to forward from.

Test: Added test for port forwarding.

Change-Id: If1185cc13eea184eb289f9c16c875d9acfb5dcbf

Former-commit-id: 272685d9525998f25d6cc51c6456ea2e6878f5bd