Reintroduced getManifestOutputFile API to manifest tasks.
The API was removed during the split/feature work as the manifest
tasks are now using an output folder which may contain 1 to many
manifest files.

This caused some user's scripts to fail without much information
on what to do next.

With this change, an exception will be raised providing a link
to the migration guide.

Bug=64021986
Test: added unit test

Change-Id: If8df6c1e762acee49b6b2a1451ec76bc5b68139d

Former-commit-id: b65ccae9cced965f4b84f33253bbba952df39f38