Revamp tools presubmit Java profile basing on errorprone mapping.
See the commit message for ag/2487569 for details. This is its full
counterpart for tools, and as far as PSQ is concerned, there were no
differences based on what I saw was enabled in the IDE (there were
some differences like 'Access level can be reduced' but those checks
aren't part of PSQ anyway).

Test: N/A
Change-Id: I97b7ef919f2960ddc116f911c869d4c2ec83f9a2

Former-commit-id: 9733146e1fb814a76f15ef16de373361eeac8060