Coerce the IDE profile XML to IntelliJ attribute ordering.
Now that the pre-upload hook is going to be fixed not to
interfere with attributes ordering, it makes sense to reset
the current IDE profile to IntelliJ format. After that, and
once the hook fix is submitted, any changes to the profiles
should yield a representative diff instead of one bloated with
irrelevant line changes due to attributes reordering.

Test: N/A
Change-Id: I3e91d513d90957ccaf0586cd0c667a42c39d2ff8

Former-commit-id: 2a794d8a45df04bd87f37ae35e0789d8172a1026