Clean up AAPT1 & AAPT2 daemon modes code
Clean up the code for AAPT1 and AAPT2 daemon mode code - get rid of
VERBOSE_LOGGING constant as it was always set to false and we have no
need for it anymore. I left some log messages for log.verbose that can
still be useful to us in the future.

Test: gwt
Change-Id: Ia0316199606f627f98e667da9016b8a3ce934869

Former-commit-id: 306c7a5e5c81a349170a4884361d5c174d128292