[AVD Mgr] Modify the "Cold Boot Now" behavior
The "Cold Boot Now" behavior was the same as if the general "Cold
Boot" option was selected. This option does not save a snapshot
on exit, so the subsequent launch would not be Quick.

For "Cold Boot Now," give a different option to the Emulator so it
does a cold boot, but saves its state on exit so the next invocation
can be Quick.

(The general cold boot option is "-no-snapstorage"; the option for
Cold Boot Now is "-no-snapshot-load".)

Bug: 64851650

Test: Expanded AvdManagerConnectionTest::testAddParameters()

Change-Id: Id9459c0485c8a13915e884855b7d31017da032d5

Former-commit-id: 1757528be34c4a0aa3c8085c5de6a0f3f2000808