Remove custom resolution strategy.
This strategy was put in place to reject dependencies that
can't be part of the apps. These includes the broken android.jar
published to maven by third parties as well as open source artifacts
that are part of the Andoid API like httpclient.

This is not working because the rejection cause is not displayed
anywhere so it looks like a broken/unresolved dependencies instead.

We are moving this to a lint error instead.

Bug: 65050576
Test: existing integ tests.
Change-Id: I489ab48f96d1bd0b074ac9062350439aad4c1895

Former-commit-id: 7700459df431d98b57d21efb45ca5ac1bdc26879