Kotlin command line lint support
This CL adds support for analyzing Kotlin from the command line, and
in particular from Gradle.

First, to run lint from outside the IDE, it now has access to a
complete kotlin-compiler environment, which includes the compiler
itself, along with some additional code (such as
KotlinLintAnalyzerFacade) to invoke the compiler and generate a
binding context.

Second, in order to run this compilation environment from Gradle, lint
is run in its own class loader.

(This is necessary because otherwise the Kotlin compiler above would
conflict with the Gradle Kotlin plugin's embedded compiler which
contains many of the same classes, but many of them partial thanks to
ProGuard, and we cannot repackage/jarjar lint's compiler copy since it
needs to also access the same APIs when running in the IDE and
accessing the IDE's PSI directly.)

To achieve this, all dependencies on lint and lint's APIs were removed
from the Gradle plugin -- this wasn't just lint itself, but seemingly
unrelated parts like the annotations extraction code, and the resource
shrinker. Then, there is a new clean invocation API (which has no
other dependencies) which lets Gradle ask lint to run itself. This
module then generates a suitable class loader containing lint's
dependencies, and calls into a new library which performs the
Gradle-specific integration tasks for lint. This CL therefore moves a
bunch of code out of the Gradle plugin and into a new lint execution
library. It also moves some code which was sitting in lint utility
classes out to sdk-common where they can be used without invoking
the class loader machinery.

This CL also turns on a number of lint unit tests that exercise
the lint checks on Kotlin source files. It also fixes
67712955: Kotlin when statement fails if subject is Build.VERSION.SDK_INT

Test: Included
Inspect: OFF

Change-Id: I6a699bc98f021fe78e459d7cd6350dbe9327dfc9

Former-commit-id: 849ebbc5b286bef031639e9521964eb00d95c733