Update Jacoco Transform to copy .kotlin_module files
Updated the Jacoco's Transform to blindly copy
the .kotlin_module files to the transform output,
according to the suggested fix for b/64435479

Also updated LibraryBaseTransform.mergeInputsToLocation() to
ensure META-INF files are merged when dealing with an
intermediate jar.

Updated TransformInModuleWithKotlinTest to also run
the jacoco transform (by enabling test coverage
for the library module).

Bug: 65345148
Test: TransformInModuleWithKotlin, LibraryIntermediateJarsTransformTest,
	JacocoTransformTest
Change-Id: Id760e7989c8195461d7aabe3501d16983ff66627

Former-commit-id: 59565424bef8319eb7d3d42a6f40e902cf280290