Fix lint false positives in Kotlin files
I'm working on enabling lint handling for Kotlin files outside
of the IDE. Those changes are too large to be included in 3.0,
but in doing so I've started running lint on sample projects and
discovering bugs in the AST analysis that is unique to Kotlin.

Three examples:
- Lint often looks for a return statement in the code to find
  the exit point or return value, but in Kotlin methods can
  be defined using expression body statements.
- Lint often looks for the type element, but this may sometimes
  not be there (because of type inference)
- Lint looks for UCallExpression statements, but with property
  syntax those may look like reference statements.

Test: There are regression tests included for these, which pass
in the 3.1 codebase where lint can run Kotlin from the command
line, but for this 3.0 backport the tests are disabled.
Inspect: OFF

Change-Id: I41900b78e481fdc2ceb5ea7cc7260ae1971ccef2

Former-commit-id: 775446e9906edcccbfbc148562b1c92075bd98b8