Clean up cache keys for users of FileCache.
This commit:

1. Removes obsolete cache keys for users of FileCache.

2. Provides cleaner API to define the properties of a file as the cache
   key, using either the file's hash (recommended), or the file's
   path+size+timestamp.

Since this change touches the cache keys, old cache entries in the build
cache created in prior plugin versions will not be reused.

Therefore, it is better for this change to go into the 3.0 release
rather than 3.1. (If introduced in 3.1, it would make the build cache
grow larger one more time unnecessarily.)

Test: Updated FileCacheTest
Change-Id: Iba7f7e21b8fbf7b01b58448636ae1261e989432a

Former-commit-id: 46f6cb8601e973b9bcf95133440cc96d8417ba44