Tidy up ActdProfileUploader and add tests for it.
We can't rely on the buildbot build number as an act-d build ID, because
there's no guarantee that the buildbot build number is monotonic.
Instead, we generate a build ID by counting the number of commits in
the git repo we're uploading stats for.

The tests added are quite nominal, in that we can't reasonably test
the act-d API. Instead, we just do some basic sanity tests to make sure
we haven't completely broken anything.

Change-Id: I17c1fff7bc3355904f1c80bdac362d929b4b4fb7
Test: ./gradlew --parallel :b:i-t:te -D:base:integration-test:test.single=ActdProfileUploaderTest

Former-commit-id: 9d834a483b585f2e8946222a8a25dcfe49a64c1d