Show CMake/Gradle error messages
If and when errors occur during Gradle/CMake operations, we currently
throw a runtime error (in vanilla-cmake) and return.
Instead, fix the error messages (correct the file paths) and return the
configured messages to Android Studio so it can parse it out correctly
and report errors (create a link to the file with the errors etc).

Test: 1.Unit test
2. run a sample app with errors and verify manually

Change-Id: Id3a799adcca262624bbf7644e8addd9ae0262251

Former-commit-id: 1514b942a64730a423edee2e220b5de7c9a7799c