Fix StableApiTest.apiElements on Windows.
The expected API list is read froma a text file with LF line
separator. The splitter should be done on the LF character ("\n")
rather than the system native line separator.
In addition, the line separator for this specific file is enforced as
LF via .gitattributes to ensure consistency across all platforms.

Test: StableApiTest.apiElements
Change-Id: I297a12f7587da611d3aa36615b20d6b00a115f07

Former-commit-id: 060795199b44ac972c69cda5c00deb2ddec81533