Rename IO "SessionDetails" and "SpeedDetails"
Work is currently underway to introduce a general purpose
concept of profiling sessions. However, C++ has detected a
naming collision between that work's "SessionDetails" and
IO profiler's "SessionDetails". Since the IO version is really
encompassing the idea of an IO session, we rename it
accordingly.

"SpeedDetails" is also pre-emptively renamed, even though there
aren't currently any problems with it. The change is included
just for consistency and clarity.

Test: N/A, refactoring only
Change-Id: Ia5438f10ef4aa31508debcc7d26e89aacb1c364b

Former-commit-id: b75b2b266788ebdbe97d878021004811eba17d1d