add interfaces hierarchy for default methods list.
when calculating the default methods list, we used only
the immediate interfaces implemented by classes, ignoring
the interface's hierarchy.

Instead of passing an ASM ClassNode for the visited type
in ASM, and a list of ASM ClassNodes for the parent types,
we now have a mini model of classes, superclasses, interfaces
and super interfaces.

Added a few functional methods to traverse the mini model.

Test: Added unit tests.
Change-Id: I1cc027321fab56e3779eaa86fa028d7f6d797c56

Former-commit-id: c3fed23723a54405c9144a477d1dadfbcb8bb17f