Use PathSensitivity.ABSOLUTE for transform inputs.
As at least the DEX archive transform depends on more than the relative path

In detail, if an AAR is updated, the input switches from one classes.jar
in an exploded AAR to another. When using PathSensitivity.RELATIVE, this
leads to the file being outOfDate, which leads to the TransformTask
marking it as CHANGED. Using PathSensitivity.ABSOLUTE, this is a removed
and outOfDate, leading to one REMOVED input and another ADDED input, as
the transform expects.

We can re-add transform path sensitivity relative, but it should be
opt-in for transforms.

Issue: https://issuetracker.google.com/68144982
Test: DexArchivesTest.testIncrementalDexingRemoteDependency
Change-Id: I9893f2c6e42e760d9553bde3d9c53a99f3b60862

Former-commit-id: 84878a6e6fd81efb425e5a779c098954420a315e