Add more info for the user in case of a sideload failure.
The current side loader only extracts and presents to the user the line
from the log that contains the Install token. Unfortunately, this often
may not include the message thrown with the InstantAppBundleException or
its direct cause. Novice Android developers may not realize that more
information about the failure is available in the Logcat.

Bug: 68286114
Test: View message in dialog.
Change-Id: Ie8cee6e74b873e5f0e0252da4f36c006bf132c4c

Former-commit-id: e34958385b38681ecb19b6d62c90bbaefe784adc