Adds support for module runtime cycles.
When generating build files we no longer treat runtime dependencies
as compile dependencies, but now we move them to top level removing
them from the compilation graph and ensuring no bazel cycles.

We do this as follows:

 1.- All runtime dependencies are removed from build rules.
 2.- We compute the transitive closures of
     a. The compile+runtime dependencies of each module
     b. The compile dependencies of each module.
 3.- If the module's compile dependencies is < its runtime dependencies
     we add runtime dependencies to the test rule until all the
     needed dependencies are covered.

Because the runtime dependencies are all added in the test->build
direction, this ensures no cycles.

Test: existing tests
Change-Id: I4ef79072c43a75d13b2f827d3b4ff69bd20b60d1

Former-commit-id: 4bae0ee098eb0da929792e35c93a38ca77b474d4