Add disabling of the retry logic to the BaseGradleExecutor
Something is causing very frequent DaemonDisappearedException, which
cause retrying to build the test project, which then in turn cause the
tasks to be marked as 'UP-TO-DATE' (because they ran already before the
retry).
This CL adds the option to disable the retry logic and updates the tests
to opt-out of retrying.

Test: Updated DataBindingIncrementalTest
Bug: 65284926
Change-Id: I62206ff80396809961fcabc87f28e5f94297020d

Former-commit-id: 839d563a9f7ad8411bd962cc4a493e5db427c2a0