Android Studio can't load info of older Add-ons
Recent SDK add-ons are described by a package.xml file. For old add-ons
that file is automatically generated based on old add-on format files.
When converting a local SDK add-on, we scan source.properties for
"Addon.NameDisplay" and then for "Addon.Name", and if no value are
found, we assume the "Unknown" value.
This CL adds code to extend search to manifest.ini too, and
to re-generate package.xml for package.xml with "Unknown" display names.

Fixes: 69835275
Test: Updated LegacyLocalRepoTest
Change-Id: Ife6c1b69d865544290c466ffbd0437a7a5211e20

Former-commit-id: bc15634bdea1d0f99cde0c7c2236b71206b65b08