Upload profiles in the background.
This is quite a large structural change to allow uploading our profiling
information in the background while tests are running.

The big changes are the relationship that BenchmarkRecorder and
ProfileCapturer have with each other and the rest of the code base.
They're now mostly stand-alone components and don't interact with
GradleTestProject.

Change-Id: Ic62eb81c25b0f154cda609c37ae3975eb1df52f1
Test: added ProfileCapturerTest and BenchmarkRecorderTest.

Former-commit-id: e9a5104420486fcf483cf15c213ac97dfbc5fad2