70036345: Lint doesn't understand Kotlin standard functions
This introduced some additional cases to check for both
the shared prefs editor and for the fragment transactions,
and there was a lot of redundancy which I then cleaned up.

It also fixes these bugs:
69534659: False positive on forEach
37140903: Reactivate failing tests after SDK update

Finally I discovered that in some cases UAST returns null
for UCallExpression.getMethodName(); added a workaround and
used it in the various places getMethodName() is called.

Bug: 70036345
Bug: 69534659
Bug: 37140903
Test: Unit tests added
Change-Id: Ib6281a9e48b995ee8d0143a52ca6c2627465b1ed

Former-commit-id: 58f687a646daecd5483151ea8d2f4bd402ae9759