tools.apksig depends on apksig-all
Because of the way //tools/base/build-system:tools.apksig target is
used, it needs to directly depend on all classes of apksig library,
including private API and impl detail classes. As a result,
tools.apksig needs to depend on //tools/apksig:apksig-all rather than
//tools/apksig:apksig. The latter will soon be switched to export just
the public API of the apksig library.

Test: bazel test $(<tools/base/bazel/targets)
Change-Id: Ie1fb5eb0e539eff179b22bcc526749514db1ef52

Former-commit-id: fbfe9f94cbfe1388d80d7d6d016f1559e1eb7784