Improve new DSL to work on test project.
This is the bare minimum to evaluate the
perf-android-large project.

- make more classes instantiated via objectFactory
- Fix non open DSL classes that are decorated by Gradle
- Also ensure that their constructor is annotated with @Inject.
- Implement some missing options classes.
- Implement some missing fun from the SealedContainer.

This also does some clean-up:

New scope object for things that all the DSL objects need,
including the issue reporter, the deprecation reporter and the
new instantiator.
This has 2 effects: 1. we can add new thing to it without changing
a LOT of constructor, 2. less storage for all the classes that needed
2+ of these items.

Update OptionalSupplier to not use a lambda but instead receive
the class to instantiate and its arguments,
and to know about the parent sealable to make get() easier to use

Change-Id: I0359b6110c622613d8e9bff1d17c3ff66af1a092
Test: existing.

Former-commit-id: 43e456ed699b8c5a4e2ceea68185e77921601bd2