Proper support for composite builds in the builder-model.
The dependency information in the model did not include build
information in the case of composite builds. Because different
included build can have colliding project path, it's important
that any reference to a project does not solely use the
project path as a unique identifier.

The other identifier is the build ID, however what's available
inside Gradle vs in the model is different. To fix this, we
use the root dir of each included build.

This CL updates the level 1 and level 2 models to include
the build id in the dependency information.

This also updates the API for querying model for the
integration tests. The previous model Map went from
project path to model instance and as we mentioned above
this is not enough anymore. The old method is still
there but deprecated and will fail in case of composite
builds.

This change also adds a new module to debug models. It's
just the basic foundation for it. The end goal is to
have it print out the model in a structured format for
debugging, but also potentially for testing. There are
a lot of tests that query the model to validate it
and doing a string comparison may be better. TBD.

Change-Id: I13ea45ec3cbb3c557bea13d0c14dbd2e25b45395
Test: existing tests.

Former-commit-id: ad1fe10f3206cfc79137b8c012e5edcdab16988b