Don't delete .o when CMakeLists.txt is touched
Bug: https://b.corp.google.com/issues/69408798

It is typically safe to just regenerate the generated ninja project and to
leave .o files intact. In most cases this will cause the subsequent build to 
be incremental rather than full.

A few things:
(1) There is a case where an edit to CMakeLists.txt would remove a 
target (and corresponding .so). We shouldn't just delete all .so files 
because link time can be very long. 
(2) Instead, during sync, I delete all .so files that aren't referenced in the 
JSON after sync.
(3) This requires a reading JSON immediately after sync which we never 
needed to do before. To ammortize the cost of doing this, I now persist 
the miniconfig in a new build file called android_gradle_build_mini.json.
(4) The behavior also affects ndk-build based projects in the same way.

Test: Modified NativeModelTest to cover new cases of touching 
CMakeLists.txt not causing removal of intermediate .o files.

69408798-cmake-refresh

Change-Id: Ic921a6272586948ee436cd62d5d0facfeae003df
Former-commit-id: 5abf264c02389ff0865b89feed30ddf7c5cbdd71