Add namespace dimension to ResourceResolver
Pass all resource values to the resolver in a single data structure, to
hide any special handling of framework or other namespaces. It's the
responsibility of the instantiating thread to efficiently come up with a
map that contains the right resources. To that end, this change adds two
implementations of Map that can be used to efficiently store resources
in non-namespaced projects, as well as "merge" project resources with
resources coming from outside (like framework). This basically moves the
if statements (if framework then this, else that) from ResourceResolver
into the map.

For now this ignores style resolution, which requires changes to
StyleResourceValue to correctly store all namespaces.

Test: ResourceResolverNamespacesTest, ResourceResolverNoNamespacesTest
Change-Id: I005bbf7fc454187a76577337eb4d0bf9c5991820

Former-commit-id: 3e399ae639285e964bf710da9a686e0986ecf664