Fix composite build model for dependencies inside same build.
The current code misunderstood the BuildIdentifier.isCurrentBuild
as "is it the root build" vs "is it the same build for both the
dependency and the consumer". It's the latter instead of the former.

However BuildIdentifier.getName() return ":" for the root
build so we can use that in the map directly and not have to use
isCurrentBuild.

Added a new multi-builds composite test to validate the composite
build support.

Fixes: 71518470
Change-Id: I7989b3bb6fb7b40c2e7a0a6e26147617d82c4113
Test: new test.

Former-commit-id: ef0acdfe1324f50dcb80e815521a4d1aa943c276