Fix inconsistent UI when app crashes during CPU capture
If the app crashes during a CPU capture (any technology), we were losing
the information that the capture stopped (or should have stopped), and
the next time the user opened CPU profiler the UI would respond in a way
that gave the impression the recording was still in progress.

This also applies if the user switch to profile another process, which
terminates the previous profiler session. Since tracing, as an action,
is part of a session, it should stop, too. If the user swtich to profile
the old process again, the profiler should be ready to record.

For simpleperf-based captures, the problem was even worse, because we
were keeping simpleperf always running and some temporary files (raw
trace, log, etc) stored on the user's device. Now, were making sure to
communicate with SimpleperfManager, so it can kill simpleperf and cleanup
the temporary files.

Fix: 71600616
Fix: 71701240
Fix: 71806927
Fix: 71815030
Test: unit tests for SimpleperfManager::StopSimpleper
Change-Id: I8c941a836b589708b8f5aa234d8f78afd9ff557d

Former-commit-id: 23d2ef97a481c5c70ff0f884c369d797a39879b9