Do not duplicate try-with-resources in test apk
Runtime classes for try-with-resources should not
be packaged in the test apk, if the tested one
already contains them. Doing so might cause
class loading issues.

This CL removes try-with-resources runtime classes
from test apk, except for the library test apk as
it must contain both test and producation classes.

BUG: 71737186
Test: DesugarAppTest, DesugarLibraryTest
Change-Id: I048491382332fa0de149b62697c2c79b2781c573

Former-commit-id: 26841209429771cab7e0f5bfc6e7082ed22c92b6