do not allow densities in resConfig when splits are requested.
Whether pure splits or full splits are requested, do not allow
densities to be placed on the resConfig(s) config setting. This
is not supported by aapt2 and is inconsistent.

Allow for languages to still be placed on the resConfig even if
language based config splits are requested. These languages will
be stored in the main APK.

this will allow for configuration like :
   defaultConfig {
           resConfigs "en"
   }
   generatePureSplits true

   splits {
        density {
            enable true
	}
	language {
	    enable true
            include "de", "fr", "it"
        }
   }

which will create three pure splits and package "en" strings in the main
apk.

When resConfig(s) is not specified, all remaining localized strings
after the resources merge will be packaged in the main APK which is
undesirable. File b/71819665 so AAPT2 provide a flag to only package
"default" in the main APK. Such flag will be used when resConfig(s)
is not specified.

Bug=70181913
Test: adapted some integration tests and added new ones.

Change-Id: I86d90c5b788b7e051693619583265f72f30e35a4

Former-commit-id: 8bf55bc018328fad0920fe928a2fbba5c2836816