Update to Gradle 4.5
Fix PackagingOptionsFilteringTest: The first part of the Java
resource processing use the built-in Sync task which used to
not filter out folder named after source control system. However
it looks like in 4.5 it now does filter them out so the test
had to be updated.
This is not super awesome because the test already checked whether
our own filters worked when applied or not. There's definitively
some mess in there anyway that was done to make it pass on windows
already. we should fix this at some point.

Change-Id: I30a1955e1c732b636a7f6c4b92e18416c14e9cd7
Test: existing tests.

Former-commit-id: 56a5c51203d8b5b71e72264d534d1f0d7d793257