Make package tasks fail on underconfigured SigningConfig
IncrementalPackageManager was silently performing no signing when given
a SigningConfig that was missing a keyPassword. The same did not happen
if the SigningConfig was missing one of the other required parameters
because those parameters were set as @Inputs to the packaging task.

Also removed storePassword from CoreSigningConfig so it would no longer
be an input to the packaging tasks, in order to keep it out of Gradle's
input cache.

Fixes: b/64502600
Test: Added test to SigningTest
Change-Id: Iae2ff6d09d9e4bdef29db191ce419561430ddf6a

Former-commit-id: 47805208df7f450522df308cc378f767839e8899