Enable data binding  in instant app features
This CL enables data binding in instant app features.
There are 2 requirements:
data binding v2 must be enabled.
base feature must have data binding enabled.

When compiling a base feature, we pass the list of all manifest
ids for feature packages. This allows the annotation processor
to generate code that will load them dynamically.

When compiling a feature, we pass the list of its runtime dependencies
that are not included in compile and also its ID offset.
Data binding annotation processor uses this ID to offset BR ids
and uses the runtime dependencies as the constraint to decide
for which dependencies it should generate a BR id.

Bug: 63814741
Test: DataBindingWithFeatures*
Change-Id: I0f6acf03c0a855afbdfc5af6f059967fb67a7630

Former-commit-id: e2275d387440195d65eb8a5201418619deba303a