Add the ability to run benchmarks N times.
The theory being that running a benchmark multiple times and graphing
only the lowest scores will yield much smoother graphs, and much more
reliable data.

The way this is going to work is that all results will get uploaded in
to GCS and thus Perfgate, and Perfgate has the ability to show
aggregates like mean, min, percentiles, etc.

This will work now that we key on the commit time instead of the build
time. These runs will be grouped together.

While I was here, I removed the LocalProtoProfileUploader and its tests.
They're not used anymore.

Test: bazel/bazel test -- //tools/base/build-system/... -//tools/base/build-system/integration-test/...
Change-Id: I84a717b3a2de7c0b5d1dd2c50ca0b6d05b6ab095

Former-commit-id: 0b67def2c707518e7d56db9b975df97a741f3339