Use java.nio.file.Path for file deletion.
We have had a few bugs where the error message only says "Failed to
delete [some file/directory]" without further details.

On Windows post submit build bot, starting from build 1494, we also
started seeing these failures occur more often in tests.

java.nio allows file deletion to throw an exception with a specific
cause that will help debug the issue better.

In terms of performance, we did the following experiment:

On Windows:
 - Task: Delete a directory that contains 3500 files, 3500 folders,
     35 MB, 3 sub-directory levels
 - Results: Averaging over 10 times,
     + java.io file deletion takes 13244 ms +- 17% stdev
     + java.nio path deletion takes 8234 ms +- 13% stdev

On Linux:
 - Task: Delete the same directory as above (3500 files, 3500 folders,
     35 MB, 3 sub-directory levels)
 - Results: Averaging over 10 times,
     + java.io file deletion takes 177 ms +- 34% stdev
     + java.nio path deletion takes 162 ms +- 17% stdev

On Linux with a larger directory:
 - Task: Delete a directory that contains 50000 files, 50000 folders,
   750 MB, 4 sub-directory levels
 - Results: Averaging over 10 times,
     + java.io file deletion takes 2666 ms +- 8.8% stdev
     + java.nio path deletion takes 2636 ms +- 8.6% stdev

These results show that java.nio is better than java.io: It is slightly
faster on Linux and much faster on Windows, it is also more stable as
the stdev is also smaller.

Bug: 70267359, 67484963, 64592700, 62291921, 37103657
Test: Refactoring change, no new tests (performance tests were done
      separately)
Change-Id: Ibb73414c2b71ec8413c2918f0eb0ad3f0a646884

Former-commit-id: bee782691fa069c0178d0457f8d43239e2cac174