Properly set up ddmlib logging.
The logger wrapper for ddmlib was set incorrectly and not actually
logging anything from ddmlib except for error messages.
Calls to DeviceProvider.init() are expected to be followed by a call
to DeviceProvider.terminate() but this was not the case in many
instances. These have been fixed. In addition, a failed call to
DeviceProvider.init() is now guaranteed not to leave any straggling
resources.

Test: N/A
Change-Id: I3453c6bf0ea6899a5ff293e9317d65f237280d7d

Former-commit-id: 0e4bd1c95271f817cf0f224e7081d7bdfc637f98