Introduce the concept of a BenchmarkEnvironment.
At the moment, Benchmark objects reach all the way up to the top level
class, BenchmarkTest, for a bunch of static variables that it needs in
order to work.

This isn't ideal, so instead I seek to formalise the concept of a
"benchmark environment" by making an explicit class to hold that data
and passing it down in to Benchmark objects.

I'm still not a fan of how the BenchmarkEnvironment gets to the
Benchmark object. I'd like it to be a lot more opaque to benchmark
writers, in that they shouldn't have to worry about passing a
BenchmarkEnvironment around, but this is a start.

Bug: none
Test: presubmit
Change-Id: I9a2b234e31ca337a7d1a9f7b31a82aaf549f39c6

Former-commit-id: b194e868adcc983bd7cb3e830f1ed59d903e4998