Move from VariantType enum to Variant type interface
There was a lot of code that tested for specific values of VariantType. This
changeset moves all that logic to testing the capabilities of the variant via
the VariantType interface. This allows code handling features and app and dynamic
modules to be less repetitive.

The enum is now VariantTypeImpl. Ideally it would become private but it's still
used in multiple places.

This change also unifies all the logic that was FEATURE specific to now apply
to dynamic feature as well. There are new enums to split the base and non base values

FEATURE -> BASE_FEATURE, FEATURE
APK -> BASE_APK, OPTIONAL_APK

and a new TEST_APK value for test-only modules.

In order to properly handle the separate test module, the existing
isForTesting property in VariantType has been renamed isTestComponent.
This change also adds a new isForTesting property that now is true for
both TEST_APK, ANDROID_TEST and UNIT_TEST. It is used in much fewer
places than isTestComponent.

Change-Id: I4a22fef3730fa722057c087459b62b575dec4fe5
Test: existing. New bundle test coming soon after more wiring.
Bug: 73356281

Former-commit-id: a21aba8ee3b8c5fdf42d89309ea92fc5198abfa2