Fix HttpTest flaky FakeAndroid Address in use
Test print tells the FakeAndroid communication port is in used when
serverSocket is initiated, which means the port that was available
has been occupied. We need avoid the time gap between getting an
available port and start the ServerSocket. So make the driver read
log to get FakeAndroid port and remove communication port param.

Bug: 71865664
Test: HttpTest
Change-Id: I88b3c46dbeb9cd4b3d439b2b09d76d30bf5404b6

Former-commit-id: 9e21d97a67afd31178fd3b0e32acfcc0800caddf