Disable the original set of failing targets on foundry.
This is based on results on the bazel foundry bot. There are a lot of failures
but these are the most common:
go/upsalite-prod/c1760076-e66a-4854-92e8-5e4e69907455
This needs to land with ag/3673117 and then since this landed ag/3665907, the
foundry bot will keep going after failures so we can get a more complete list
of what we need to disable and file bugs for.

Bug: 73948555
Test: No PSQ tests yet. Currently, we will need to check the bazel foundry
post-submit bot to see if the tests are correctly not being run on the bazel
foundry bot.

Change-Id: I2878fc695942ee405b60bdd25e61ef354da02b0f

Former-commit-id: ba0d7c2b81e1df7638440b56bfc85fa2d67bfd53