Split out the translatable=false warnings from ExtraTranslations
The translation check for specifying translatable=false on a string
which is translated in a different locale was using the same issue
as the ExtraTranslation issue. That's unfortunate since the latter
is a fatal problem (can lead to crashes), and the former is not. This
CL creates a separate, warning-severity isssue for the translated
issue (but continues to look up suppress markers for ExtraTranslation
such that any manually suppressed recordings of this issue continues
to not fire.)

It also demotes MissingTranslation from fatal to error.

Change-Id: Ie04a6234b79c2011f5222a7894fbd3d0290b2114

Former-commit-id: 3f34819fd3fdde709442804dd7600fcf0392ddef