Handle the general case for Instant Run not finding byte code.
In the current implementation for instant run instrumentation, there are
at least a couple of situations that would cause as to be unable to find
a class's bytecode:

1) The class was introduced in to Android at an API level higher than
the one we are currently running at.
2) The class was included as a compileOnly dependency in a library and
we've inherited from one of its classes.

We solved case 1 by searching for uses of RequiresApi and TargetApi on
classes and not instrumenting those classes, under the assumption
they'll never be used and if they are the app is broken anyway. We have
no story for case 2 and currently we show users an exception at build
time.

This CL introduces a solution that solves case 1 and case 2, as well as
slightly simplifying the implementation: if we can't find a class's
bytecode we throw a ByteCodeNotFoundException, and we catch that higher
up the stack and use it to skip instrumenting the class.

Bug: 72811718
Test: added a test in IncrementalVisitorTest
Change-Id: I9957f5e5239b76c8aa7b5808c52f8e74a75d40d5

Former-commit-id: 22212a9bf49883368890ed89ff78403a9da42fc3