Extract all parsing from ViewNode
Take all the parsing logic currently inside ViewNode's parse and
constructor method and move it into a ViewNodeParser class.

This prepares us to add a ViewNodeParser for the v2 protocol and have both
parsers return a ViewNode object.

There are still some some ugly code and one use of lateinit for displayInfo
that's to be cleaned up in follow up CL.

Fixes: b/74242992
Test: ViewNodeTest
Change-Id: I9bcf405a061b9d5f6d496e57ba834a6484f5d934

Former-commit-id: b2d68e6982127439c99c4a30ef37b23233e6dadb