Exploratory change to Mockable JAR process.
This change removes the Mockable Android JAR creation task from the Task
Manager and replaces it with a generic blank task for indexing purposes.
In order to provide the Mockable Android JAR, we attempt to use an
ArtifactTransform to do the same operation that the task previously did,
using a new configuration in the gradle file.

Bug: 66345712
Test: PluginDslTest and UnitTestingDefaultValuesTest
Change-Id: I0436066ca9e205157bccac4019a98eba4d48e27e

Former-commit-id: 77caec2b5b755cad4c9e5f01c2b9070c7e508694