Reformat lint source code
This CL reformats the lint source code. It's never a great
time to do this, but given that it's early in the 3.2
cycle it seems as good a time as any. Much of the Kotlin
code was written before there was a Kotlin style guide
so this brings the source code in line with the current
recommendations.

First, reformats the source code using the Kotlin plugin's
new formatter (as of 1.2.30) which implements the recommended
Kotlin style guide. I then ran ktlint which applies some
additional cleanup that the default formatter does not (such
as removing accidental semicolons), and then formatted again
to change back a few things ktlint handles differently than
the default formatter.

It also reformats all the Java files with the google3
formatter; it doesn't do a great job, particularly with the
test sources, but since this is done as a repo hook for the
rest of tools/base I don't want to keep fighting it.
(And it's a diminishing problem anyway as more and more of
the codebase is replaced by Kotlin.)

Finally, it updates the IDE Kotlin code style settings files
to use the official Kotlin style guide.

Test: This is just a source formatting change; everything is
covered by existing tests.

Inspect: Off
Change-Id: I46b15382ca307af78deab3cfc58c7739a90f7f5e

Former-commit-id: ff37cc094223d66343f0529f67c31f3909017895