Baselines: Record variant and version and give better diagnostics
This CL updates the baseline creation code to record extra metadata such as
which variant the baseline was created under, which version of the tool was
recording it, etc.

Then, when lint is running against the baseline, it emits extra warnings
if the current variant doesn't match the creation variant, and it also
warns if it looks like the baseline was created with a newer version
of lint than the one currently running, etc.

Test: Covered by existing tests
Bug: N/A
Change-Id: I28502f4cb52fda6d9d49b27a4fd4a5282c5a5059

Former-commit-id: d089d4b9c3b398013910a064a51eb05d9c1ddbf9