Add integration test running proguard/R8 for multi-APK project
This CL is the 3rd in a series to provide proguard/R8
support for multi-APK projects.

This CL creates a skeleton integration test which runs
proguard and R8 on a multi-APK project.

As support is added for proguard and R8 in multi-APK projects,
code will be added to this test to test it.

Bug: 74433581
Test: MinifyFeaturesTest
Change-Id: I91b7d8d18cf25960678840a7bef1aec9f6928e77

Former-commit-id: 184d9a94ae01bc3cd8cac810689a901f148b2f43