Debug DirectoryNotEmptyException in GradleTestProject(cont.)
Other people on the Internet have seen this issue too.

Theory: There seems to be a timing/visibility issue on Windows
filesystem, such that even if we delete the contents of the directory
before deleting the directory itself (and no other thread/process is
accessing the directory), the directory is still being seen as
non-empty. Let's wait a little bit and hope that the directory will be
seen as empty shortly.

If the theory is confirmed, we will need to figure out a better way to
delete directories on Windows.

Bug: 69271554
Test: Fixing them
Change-Id: Ifa372533ab08feab47ded0192b0cf0d4ec3f7051

Former-commit-id: 2a6deb679c1611d346d85f87af3e3528069605db