Move overly verbose log to "fine" level
The purpose of the warning was to highlight that a mapping for a package
that we expect to have an AndroidxMapping didn't have one. The log
showed up no matter if were using that name so reducing to fine so it
helps debugging issues but it does not randomly popup for everyone.

Bug: N/A
Test: N/A
Change-Id: I3ff813f3b1c89684114bc0802077ad808fcb9165

Former-commit-id: 0f41aa6191e0c462860b2d119b640098e156622c