AndroidX support: Use correct multidex dep. for AndroidX.
This commit also adds a flag, android.useAndroidX, for users to indicate
whether they want to use AndroidX instead of the old support libraries.

This flag is currently disabled by default.

Note that this flag is different from an existing flag called
android.enableJetifier, which specifies whether the users want to jetify
dependencies automatically after setting android.useAndroidX to true.
Having the two flags provides the users with the option to use AndroidX
but disable Jetifier (e.g., for debugging).

Bug: 76099977
Test: Pending publishing androidx.multidex:multidex:1.0.0-alpha1 to
      Gmaven
Change-Id: I224758231602dd4b0064f2a7485cfc187388d515

Former-commit-id: 0d3e47410784592a487e1567d1dfefc1dea5b7d9