Add DexSplitterTransform
This CL is the 4th in a series to provide proguard/R8
support for multi-APK projects.

This CL creates a new DexSplitterTransform which is
run after Proguard/R8 for base modules in order to
split the resulting dex file(s) based on their source
feature.

This CL wires the new feature-classes and
feature-java-res streams to the transform pipeline.

This CL also adds java code, android manifests, etc.
to the features in MinifyFeatureTest.

This CL also disables resource shrinking for
multi-apk applications.

This CL also disables instant run for multi-apk
applications.

Bug: 74433581
Bug: 77685496
Test: DexSplitterTransformTest, MinifyFeatureTest
Change-Id: I5d6597b83e3090b79b4a518204600a5818367b20

Former-commit-id: caaba04f5dd3557904fa2f65d5da8cbe1e2a9de3