Revert "Jetifier support: Transform libraries to use Androidx."
This reverts commit 46d2fc3d85726e121abb6580f26db083520478ce [formerly c21e40cad3b26e0f1ef4d48683a0033245122148].

Reason for revert: The jetifier-core library is not yet available on
remote repositories (only in prebuilts). AGP cannot use the library
until it is available on jcenter() or google(), or until we repackage
the library as part of AGP.

Only this commit needs to be reverted. The other commits of the same
topic can stay without any harm.

Change-Id: I9b160ff8e9e6daded144be628c1eb21613ce82b4

Former-commit-id: d27ac9e0b579efe0d92e11d9842754433c79e123