Moves more responsibilities to the Daemon
The main profiler service is only the link with grpc.

This change highligted the fact that the heartbeat maps
are accessed from multiple threads, and their behaviour
is undefined if race conditions occur. Will be cleaned
up further in follow up CLs, but this is now starting
to move the ownership of centralized objects, to the
Daemon.

Test: existing
Bug: 73538507

Change-Id: If69ab43193e7c102571902b4a7e296b1945b3404

Former-commit-id: 0a3b6caa91f132c74065192452a5a98f26d26932