Show launcher checkbox even when creating a new project
The isLauncher flag is hardcoded to be enabled and hidden when creating
a new project. As a result, the checkbox is hidden when creating a new
project, and shown only when adding more activites to an existing
project. In Android Things, the developer should have the option to
create a project that won't launch automatically. So instead of relying
on the default isLauncher flag, we are introducing our own
isThingsLauncher.

Test: TemplateTest tests both activities
Bug: 63867883
Change-Id: If30ae457f55dc93d5f8854d2e9925b8abea2a501

Former-commit-id: 2546f795208297948b436f2a86217be7d7d7f73d