Use Guava for recursive file deletion in GradleTestProject.
As decribed at bug https://issuetracker.google.com/77478235, there is a
timing/visibility issue on Windows filesystem, such that even if we
delete the contents of a directory before deleting the directory itself
(and no other thread/process is accessing the directory), the directory
is still being seen as non-empty. If we retry deleting the directory
(not its contents) one or a few more times, the directory will be seen
as empty and deletion will succeed.

We now use Guava for recursive file deletion in GradleTestProject, to
see if Guava can handle this issue. If this works, we will use Guava for
file deletion in the plugin.

Bug: 77478235
Test: Fixing them
Change-Id: I60e426e189e023893d0f7aa1b6238f07bbf60103

Former-commit-id: 5ebd2f2105a4ddf1e4ad9423023a7c987b4b98f7