Remove platform specific bazel build/test filters from bazel wrapper script.
This allows developers to more easily try to fix these targets. The change
requires that platform specific build/test filters be explicitly specified
as command-line arguments to bazel calls. This change ag/3869571, which is
in this topic explicitly adds the command-line args to the bots.

Bug: 77319498
Test: PSQ will test the new wrapper, but the topic will not pick up the recipe
change, since that is part of the infra code. Expect some of the disabled tests
that are filtered to fail on the PSQ. As long as only the disabled ones are the
ones that fail for this CL, the change can be added with a PSQ -2, because it
will be added with the recipe change which will then explicitly add the filters.

Change-Id: I5977fbcde27a10e962a9f2a8c535624a605b610f

Former-commit-id: 4f33dc9e5ff2c5f80aa3e9c21202e4eff4d20735