Treat fake Android system as having nothing installed
When the package manager returns a string of the form "package:<some
path to an installed APK>" to the adb shell command "pm path <package
ID>", then the the adb client's assumption is that the package with ID
<package ID> is installed on the Android system. When the sanity UI
tests run with this test fake's behavior, the IDE will try to uninstall
the package. This behavior does not occur normally on the UI tests
because each of the UI tests start with a newly created AVD
configuration for each test.

In order to retain the original behavior of the UI test, the package
manager command handler should instead behave as if no package is
installed on the system. There is currently no known test case that
relies on checking for the existence of system packages such as Chrome
or the Play Store.

Test: Test fake prepared by future UI tests.
Bug: 77870331
Change-Id: If18f12be22e018571d19356563f30f2457f5acbd

Former-commit-id: 5b959450c3606d41d955637ed59ba580c21fb3ce