Stop tracing before deallocating CPU cache
In perfd, much of ongoing traceing data is managed by CPU
cache. When the monitoring of an app ends including When
the app exits, we should stop tracing first and then
deallocate the cache associated with the app. If the
order is wrong, the profiler would incorrectly think there
is no ongooing tracing.

Fixes: b/77218929

Test: Will be added (openned b/77973736)
The test will requires making CpuService testable, which
implies making a number sub-components testable.

Change-Id: I70122ef35cfa91b4a9d9406f4467f73aab571837

Former-commit-id: 599f536e4a35c93478523467c1d70e97a2c198d2