change BuildableArtifact output strategy
When dealing with a single task producing an artifact,
there is no need to add the task name to the output folders.

Final version of an artifact also have the ability to be in
different location while intermediate version of that artifact
will be stored in intermediates/ no matter what.

Task should not use appendArtifact that return File but
the new version that use RegularFileProperty. That guarantees
the file name to be adapted until execution time depending
on the presence of other tasks producing the same artifact
type or not.

In a susequent CL, I will replace all usage of the appendArtifact
returning a File with the one returning RegualarFileProperty.

Test: unit tests added, adapted integration tests.
Bug: 77971282
Change-Id: I777a162dd5f1a7d6b66434b0607c1c52e95c118c

Former-commit-id: 5a4475384a41ac4d16fd6e11a925f18cc8af9bec