Require explicit opt-in for legacy multidex.
As the application class must call MultiDex.install(), we sadly can't
make legacy multidex work automatically for the bundle, so switch to
respecting the multidex flag.

This means that generating the monolithic APKs for versions of Android
before Lollipiop (i.e. API level <= 21) might fail if the app doesn't
fit in to one fused dex file and the developer has not enabled multidex.
This is preferable to failing at runtime due to a lack of
MultiDex.install()

Bug: 75321654
Test: BundleTest
Change-Id: I630b9547106eff8b6c494aaf21232edf670727f1

Former-commit-id: cc74c236e52054fbf0e684afaf0027b0c627df48