Add Dynamic App test for multi-apk code shrinking
This CL is the 8th in a series to provide proguard/R8
support for multi-APK projects.

This CL changes MinifyFeaturesTest to also test
multi-apk code shrinking for a Dynamic App
(we only tested it for an Instant App before)

This CL also adds java resources to the features'
library dependencies. I installed the resulting APKs
on my phone and verified that the dynamic app finds
the java resources at runtime.

This CL also adds some external lib dependencies.

Test: MinifyFeaturesTest
Bug: 74433581
Change-Id: I96cb67b600215fd0067f9d39decf1f61da33c537

Former-commit-id: 9ee057270b51883e322350a8f6ae280196d50508