Bundle: Include features in multidex calculation.
This is only done if there are any features that would need to be fused.

Therefore, the bundle will no longer contain a main dex list if it has
no dynamic features, as the bundle tool will never need to remerge the
dex files in that case.

Note that in either case, the base module classes are still processed
for legacy multidex, as if none of the features are fused to the base
the dex files will be delivered as is, and that is currently not known
at configuration time.

Bug: 78511020
Test: DynamicAppTest.
Change-Id: Iaec7c03e0754ab7fa579c38744ecce35b3f87828

Former-commit-id: 01ce6bc5972db65114e782aa6ba02b82c714c8f5