Mark useAndroidX and enableJetifier as stable options.
If the user uses the AndroidX refactoring tool in the IDE, the IDE will
add these 2 flags for them and they don't need to be aware of the flags.

Therefore, showing the warnings that the flags are experimental is
confusing to users.

This commit marks the 2 flags as STABLE so these warnings won't show up.

Bug: 78573742
Test: Manually verified that the warnings do not show up
Change-Id: I947d457e47471fde3df6da06126358cc03323a55

Former-commit-id: 233db93f71dce2970600af6a46c8fba5e1a35a1d