DataSet should save and restore the ignore patterns used.
During a incremental build, MergeSourceSetFolders rebuilds the
AssetMerger from the .../incremental/mergeDebugAssets/merger.xml.

But when it loads the DataSets from the xml, it ends up using
the default for aapt (which excludes directories starting with
underscore for example), since it doesn't have the information
of which set of pattern to use instead.

Bug: 64846237
Test: AssetPackagingTest
Change-Id: I252c37fb8e155fc983c06197610e9e622922428d

Former-commit-id: 1493cf2845d4a75e542045646cbb7917cee9dfaa