Don't inject splitName in dynamic features.
It is only needed for instant apps.

The presence of splitName makes entry points merged from features in
the fused standalone APK on O+, as it causes the platform to try to
install a non existent feature split, when everything is already present
in the base.

Still inject featureSplit for both.

Fixes: 78696166
Fixes: 78918941
Fixes: 78646345
Test: DynamicAppTest, ManifestMerger2SmallTest
Change-Id: I293af334859a8d6beb79f0360f97b3fcd8a4db15

Former-commit-id: da820790d32ada4e4b6698d361d8a2058769d012