Change wording around experimental flags
This CL adds an option property to allow
for additional information regarding a flag.
When introducing new features (which have
experimental status), this additional information
is reported. It can be used to supply user with
advice how a feature should be used.

This CL uses this new property to add a  warning message
when using R8.

BUG: 79148738
Test: existing
Change-Id: I96abad0faf0c7b7eec4ca0d3356f1580a121af6c

Former-commit-id: dfdb138a692c95e01503f92e35a0288703d78501