Use crash reporter in the plugin
This CL adds a crash reporter that can be
used in AGP for automated exception reporting.
If users have opted into analytics (in the
Studio UI), reports will be uploaded.

Reports will be available on go/crash, with
productId "AndroidGradlePlugin".

For start, only specific kinds of exceptions
will be reported, in order to reduce false
positives to minimum.

BUG: 73644139
Test: PluginCrashReporterTest, PluginExceptionReportTest
Change-Id: I399e314ba0fde002b345696238165a1b8a4ea536

Former-commit-id: d2af9a83a3acbecf50275256b73bad93ba121b29