Log the application ID from projects.
This only has effect if the user is opted-in to logs.

As calculating the application ID might involve parsing
AndroidManifest.xml files, the profile submission is no longer done
directly in the build completion callback, but in the profile executor
service, and the calculation of the application ID is deferred.

Test: ProfileContentTest
Bug: 79201094
Change-Id: I4b8114b177786b917819526deab72a6bf7c12d4a

Former-commit-id: 50375a57bf64db8701b7dc602b7cbb1171a6f635