Bump ThreadPoolScheduler timeout from 60m to 90m to fix flaky tests.
RunnerScheduler.finished() is called after all child tests are
scheduled and the documentation suggests it could contain logic to wait
until all tests are finished.

My hypothesis is that CheckAll tests grew organically and now takes more
than [schedule] + 60m to run everything, so we are going on even when there
are still tests that didn't finished.

Also adding extra logging to check, if the problem persists, what is the
state of ThreadPoolScheduler during these runs.

Bug: N/A
Test: gw :base:build-system:integration-test:application:automaticTest
Change-Id: Ib3f538b4a1662c4b812e096e1482e75fd6aa132f

Former-commit-id: a739f3c156dbf38ea55aaed0b9565cbafee45524