Modify CheckInstantAppLibrariesTask for dynamic apps
This CL renames CheckInstantAppLibrariesTask to
CheckMultiApkLibrariesTask and changes it to work for
dynamic apps as well as instant apps.

The new task is run from the base module instead of the
instant app module.

Also, the merge java resources transform task is set to
have a dependency on this task so that it's guaranteed to
run before any errors from java resource merging or
code shrinking.

Fixes: 80083154
Test: CheckMultiApkLibrariesTaskTest
Change-Id: If2c09c0b5b25f9b57221472c9c60020a8236b275

Former-commit-id: 53a7d066a3aae632661a1a19ba72db3f2fdf978c