replaced all default glob pattern implementation.
The PackageOptions uses glob: patterns to merge, exclude files.
By default these patterns are implemented by the
FileSystem.getPathMatcher which treats the pattern as a regular
expression. This is an overkill as glob patterns are not as
sophisticated as regular expressions and the matching takes a
lot more time than necessary.

replaced all default exclusions using dedicated PathMatcher.

reduce the task exectution for some open source android apps by
2/3, overall build time by 15~20%.

Test: added unit tests.
Bug: 79880023
Change-Id: I7083f69c3a1942af37ccce037a2a47282e04c105

Former-commit-id: e9ac4003fa92948cc506fc02c083586018a6f335