Make data binding work with Gradle remote caching.
(Note: This doesn't work when Kapt is present yet.)

In a previous CL, we made data binding work with Gradle caching given
the same project location. This CL makes data binding work with Gradle
caching across different project locations. That is, a second project
can reuse the build cache that was populated earlier by the first
project.

We do this by removing the unnessary hack for Kapt when Kapt is not
used. This hack (see code for details) is only needed when Kapt is used,
so that Kapt can read annotation processor arguments. Note that the
limitation of the hack is that it will break remote caching. Therefore,
after this CL, data binding will work with remote caching if and only if
Kapt is not used.

Bug: 69243050
Test: Updated DataBindingCachingTest
Change-Id: Ibe60c7ce9142bf2fe44c9b4689b4490905dd1d1a

Former-commit-id: 0638635001b7c30a8d14405f072caf758803a4b5