Let GradleTestProject::getSubProject accept project path
Previously GradleTestProject::getSubProject accepted arguments
of the form ":foo" but not ":foo:bar". This CL adds the ability
to create sub projects which aren't right inside the root project
directory.

This functionality is necessary to properly test splitting the
dex files in a multi-apk app, because the split dex file publication
is directed by the full gradle project path.

Bug: 74433581
Test: existing, MinifyFeaturesTest
Change-Id: I8d813455744dc232cb9812cb1991911167ea1c95

Former-commit-id: a50fab5c27d7d3d5408770a2760fa0f0c07063fe