[SVS] Multiple variants in single-variant sync.
With single variant sync, the variant of dependency modules
were picked based on the dependent modules. In the case of
selection conflicts, we fetch both variants.
So that sync is not needed when user swtiches back and forth
from Build Variants Window.
This change makes the constructor of IdeAndroidProject accept
a list of Variatns to inject.

Test: SyncExecutorIntegrationTest.
Bug: 80003481
Change-Id: Iae9119cb94cf87901edc58bd17745783f9a78fb4

Former-commit-id: 17e2dc6ef3866152cafcbe8d1d93d57ff4794f86