Add native code for CPU
Currently, there's only C++ code for memory. In order to test
CPU native profiling, though, we should have some native code
in CPU as well.

In this CL, we introduce a new library (to keep CPU and memory code
separate) and rearrange the code to accomodate it. We also add a task
which calls it.

Bug: 80135731
Test: N/A
Change-Id: I976b48212b74b416a21da2673c71d22e5db0fdbd

Former-commit-id: 7422f3dc93f933d73a715a30689fec6f5ca590e8