Suppress targetSdkVersion expiration checks in intg-tests
As of June 1 lint will start to enforce targetSdkVersion
expiration (as a heads up notice to developers that
this will be enforced in the play store in August.)

Various integration tests already have a targetSdkVersion
lower than the required minimum. This CL adds some
suppress statements for these checks; this prevents
the tests from start failing in a few days.

Test: Covered by existing (and updated) tests
Bug: N/A
Change-Id: Icc52cf5c8357b99c8bb350070dfcedbe3069e3bb

Former-commit-id: 8ce819196008f93a9087a96384414c21c1fa4e40