Publish baseFeature proguard rules to app module
This CL changes how proguardFiles and consumerProguardFiles
are handled for features and dynamic-features.

Previously, only consumerProguardFiles were included when
publishing proguard rules for features and
dynamic-features. With this CL, the listed proguardFiles will
also be merged and published for features (for both
library and feature variants) and dynamic-features.

The user can now use proguardFiles and/or consumerProguardFiles
when specifying proguard rules for features and dynamic-features;
they both have the same effect.

In the case of the baseFeature of an Instant App, the proguardFiles
AND consumerProguardFiles specified in the baseFeature are both included
when proguard or R8 is run in the baseFeature, and they are both published
to the app module.

Fixes: 80257425
Test: MinifyFeaturesTest
Change-Id: I5570c73c58a716f36a05dc0a8bd4c840e4c611ba

Former-commit-id: 7f9c014159f284dcbb344eb3657064a58622c03e