Disable unstable inspection
The most recent 2018.2 builds introduced (or widened?) an
"unstable API" inspection which is flagging any API in a
class annotated with @Beta. Unfortunately, this includes
a lot of Guava APIs, that have been "@Beta" for years
and years. I've alerted the Guava team to this but in
the meantime, this inspection is lighting up a huge
number of files with misleading warnings.

Test: Not a code change affecting production
Bug: N/A

Change-Id: I03fddf7e90f1eb592d8608027434d6e1acaf8f05

Former-commit-id: 97674173570cbce434054d93847567f888d5e222