Clean-up task storage.
We have 2 different location where we store tasks, with tons of
duplicates between the 2. This moves everything to a new container.
This is written in Kotlin to not have to manually manage setters/getters.

There was also duplication between direct task field and task in
a map by type/kind, and this remove the map one as there was only a
few usage and the API isn't great for typed getters.

This also introduces a read-only version of the container for the
API objects since they should not set tasks. Every task not
part of the reader should be eliminated from the container in the
near future as we should not deal with tasks directly (and instead
link them via FileCollections). This will happen later.

Bug: Clean-up
Test: existing tests.

Change-Id: I4648d74e13882cd3a6d05721fc9d6012bfd21a44

Former-commit-id: f88d2962df4c14bd1034c0c36f8484257bb8aa61