Refactor cxx process spawning logging
Separates logging concerns and converts to idiomatic Kotlin.
The underlying purpose of this change is to make it more straight-
forward to execute processes in the future. For example, for
Windows sync performance we will want to execute a probing call
to CMake to discover information about ABI that we can cache.

Bug: NA
Test: Refactor, no new behavior. Well I could test the broken out
classes.

Change-Id: I486b2fb64cb9a03a18951f5c5eefb99dc27b2283

Former-commit-id: 08f37242b506d3e5c1adce41f091443192f35b00