Avoid null properties in Gradle invoker bazel rule
In Gradle Tooling API 4.8, when using Bazel rule to run Gradle builds,
we get an exception:

DaemonConnectionException: Could not dispatch a message to the daemon
caused by IllegalArgumentException: Cannot encode a null string.

This happens because of system properties null in Linux / Mac.

This CL adds the given properties only if not null.

This CL also sets gradle home as a system property as a workaround for
issue https://github.com/gradle/gradle/issues/5188 of Gradle 4.7.

Test: existing should pass
Bug: N/A
Change-Id: Ia191d6ab8311e1e01ce8b568fd4622209ec193b6

Former-commit-id: 1b2856701224cd743c7f081fc322c83e3164226c