Pack main dex completely for dx and legacy multidex
This CL passes information if variant is debuggable to
the DX dex merging. Flag is used to determine if
all available space should be used in the main dex.
For debuggable builds, minimal main dex will be
produced, while for the release builds, all available
space in the main dex will be used.

BUG: 110028496
Test: DexMergerTransformTest
Change-Id: I7f020942d18802dbff293cc71f60abeb2011fad0

Former-commit-id: dbac5898541b78f34143f4774e46b2ce11182ad9