[6/N] Structure Plugin: Add SourceFilesCollector.
Collects the usage of Java and Kotlin files. This is based on a few
assumptions:
1) Each file correspond to one class.
2) The directory corresponds to the class package.
3) The source files will be under src/main/java of the project root
dir.

So for now we put a hardcoded value on the methods per class, as I'm
also not sure how that actually impacts the ASPoet generation... in the
worst case we can crank it up.

In the future we might move to run ASM through the bytecode generated,
although in the spirit of ASPoet I'd prefer to catch most of the stuff
as raw as possible, to avoid reflecting changes that our plugins makes
in the project (and missing the whole points that is getting the
project's original structure).

It also fetches the info from Gradle's Project API directly in order to
work with non-android/pure java/kotlin library modules.

Bug: N/A
Test: manually with test project
Change-Id: I4ae64a553b465780f95875eb6954d9854a407c3e

Former-commit-id: 7e6bd4621d3567a9d4b92f0b67a7df579943d7d9