added change information for folder based APK.
when creating an APK as a folder, record all added/changed files as well
as deleted files. This will be used by a "deployer" facility to figure
out what needs to be updated on the phone.

in incremental mode, the change list will only contain the delta from
the previous build.

in full mode, the change list will contain all the files in the folder
APK.

Test: added unit tests, integration tests.
Bug: 109751732
Change-Id: Ib65c7cda1b550fe4e3445faba64df18635dfa101

Former-commit-id: 18dc26e3cbde0830e78ded8fb9c0f8d512bba0fb