Implement PathHashMap
PathHashMap is a read-only implementation of the PathMap interface,
which wraps a given Map<PathString, T>. Although it can wrap any kind
of map, the name PathHashMap derives from the fact that the most common
use-case is to wrap a HashMap.

This implementation is fairly efficient for most operations, but some
operations require an exhaustive search of the map entries.

Although we currently only pass one one parameter to it, the PathMapTest
is parameterized because it can (and should) be applied to all
implementations of PathMap interface. Another such implementation
is coming in the next CL.

Bug: 110264755
Test: PathMapTest
Change-Id: Ib0436474a49e8d3120e8cbf433e3999a94b345f3

Former-commit-id: 2d5ecc151939ee8f31d34b2aaa1be7c3da19394c