Implement PathTreeMap
PathTreeMap is a mutable PathMap implementation in which all methods
run in constant time with respect to the map size. Although some methods
on PathTreeMap are asymptotically faster than the implementation in
PathHashMap.kt, the latter has a lower constant amount of overhead so
should still be preferred in most situations.

Bug: 110264755
Test: PathMapTest, MutablePathMapTest, PathTreeMapTest
Change-Id: Ia92e156c2ff0fd6c4b7ef7e814ab16808791ef6c

Former-commit-id: 688087f3bd4282d55f4c057a4884010a23cb75a5