Preserve actual resource names in RDef.
RDef is the file listing the resources defined in an AAR for the
purposes of auto-namespacing.

Normalizing makes it impossible to tell if the original name was foo.bar
or foo_bar, which is problematic for automatically making all resources
public.

We still use the canonical name as the internal key in SymbolTable, and
in all externally visible files, for example the AAR R.txt.

Test: NamespaceRewriterTest, LibraryDefinedSymbolTableTransformTest
Test: SymbolIoTest, SymbolTableTest, SymbolTest
Bug: 73626400
Change-Id: I9ddb0f00a13e44fccaadab1390ee5114605ffd77

Former-commit-id: 66c36fa112a3d5a18dd52f6ed8f8ccebd7d981bf