Add a SourceProviderAdapter type for converting SourceSets
The normal way to obtain the adapter would be via the new extension
function SourceSet.toSourceProvider("mysourceprovidername").

The adapter accepts an optional "fileResolver" argument. If present,
it allows a custom conversion from PathString to File, which allows
for the possibility of creating temp files for files that aren't
present on the local filesystem.

Bug: 109669023
Test: SourceProviderAdapterTest
Change-Id: Ic785f80dc28d8e53ded80edea98dd39bb3a6cd60

Former-commit-id: 9f3d45b7266818649f2a5e05c04b1e5eee4c8a35