provide ability to have change list even when producing APK.
folder based deployment cannot handle folders yet (!) but
need to have the list of changed files even for APK for
prototyping.

Adding -Pandroid.deployment.provideListOfChanges=true to the
list of invocation paramters to get the change list.

the change list file will be the name of the APK without extension
followed by __adt_change_list__.json, so if the apk is named
base.apk, the change list is base__adt_change_list__.json

Test: added unit and integration tests.
Bug: N/A
Change-Id: I2b59552f5aa575dfa53f3da8689ed70d286eac30

Former-commit-id: e875bd31f5615682b4abdccf514cb2554f0f23f3