Initial Check-in for Code Swapper
The core of the CL is setting up an agent to call RedefineClasses and
have the art host test verifying a class is actually redefined.

I did have to modify FakeAndroid a bit. In particular, I am setting up
the context class loader to be the loader of the dex file (see comments).

Test: ./tools/base/bazel/bazel test tools/base/deploy/swapper/... tools/base/profiler/...
Bug: 109892884
Change-Id: Ib47a396af3f2505f51c9e545df1146a1fbb6b2e3

Former-commit-id: a6d62bd61c91425eaabc2879e606de5b7e0cae32