Add a generateVariants method to ConfigTable
The method generates all the variants and artifacts described by the
ConfigTable.

This will be useful in conversion code (where the generated variants
can be used as default values, and customized if necessary using .map).
It will also be useful in test fixture construction, since the fixture
can just describe the ConfigTable and then use this new method to
generate valid Variants based on all valid permutations of the
ConfigTable.

Bug: 109669023
Test: ConfigTableTest.testGenerateVariants
Change-Id: I6fe595efce8242382b12bf8f41bb6624df1a6e87

Former-commit-id: 442c2625d1612d2fe28f7638d84b833d503160ba