Dexing artifact transform
This CL adds support for using the artifact transforms
for dexing external libraries and library projects. Also,
the dex merging pipeline has been changed (described below).

This feature is behind a flag, and it can be enabled by
specifying:

android.enableDexingArtifactTransform=true

For variants that are both debuggable and native multidex,
there are 3 dex merging tasks created:
- a task to merge external libraries' dex files
- another task to merge library projects' dex files
- and third task to merge project dex files

Outputs of these 3 tasks will be packaged in the APK/bundle.

For other types of variants, there is a separate task
to merge external dependencies first, and then the result
of that merge is merged with the project's and libraries'
dex files.

BUG: 112087084
Test: added new tests for new artifact transform,
new task delegate, and intergration tests

Change-Id: I3313a5d84a07fa0489b7effed9d8e3651e2ab32e

Former-commit-id: e6f8e29e449bc11a4417c53e54a774e5d4153c8c