Handle missing symbol file in IdeLibraryFactory
getSymbolFile is a @NonNull property that does not exist in models from
old versions of our Gradle plugin. Since one of the points of the IDE
model classes is to be safe to use, the current behavior of delaying the
error is not very useful for the consumers. Instead let's guess where
symbol file is most likely to be and use that path instead (the caller
needs to handle the file not existing anyway).

Bug: 77801019
Test: GradleSyncWithOlderPluginTest when StudioFlags.IN_MEMORY_R_CLASSES
      is enabled.

Change-Id: I24a830e4eeae03719baba49b4075eac96491c697

Former-commit-id: fcb4502b79b7c95d1cec549a7c74e23b95720479