[CS] Support Compound Sync from AGP model builder
A new field is introduced in ModelBuilderParameter to indicate if tasks
for setting up the IDE should be scheduled to run after model is
fetched.

Since we don't know what is the selected Variant when SVS is not used,
it only makes sense to support CS when SVS is used.

IDE setup tasks for all artifacts of the Variant are shceduled to run.

Test: CompoundSyncModelBuilderTest
Bug: 110973453
Change-Id: I4d68204176cc6cb339d0b9d8360742f18318d0b9

Former-commit-id: 957d76ddf320798e80f39c60dfcc729afe708f9a