Resource Shrinker for bundles.
Create a parallel resource shrinking pipeline for resources in the
bundle.

Add a new format parameter to the resource shrinker, to write stub
resources in proto-apk format for the bundle.

This also removes the special case for res/raw/keep.xml, as the bundle
tool validates that resources in the resource table have corresponding
files.

Following this, the existing resource shrinking pipeline should be
migrated to a task too (b/111491977).

Test: ShrinkResourcesTest
Fixes: 111376695
Change-Id: Ic2cd58fd8280e78a4a553acc7915fd2b400790e3

Former-commit-id: 1e2e01e34e94ebe9904650a3f50a08e95c5b45bb