Fix merging of android:hasCode="false"
This CL disallows merging android:hasCode attributes
from libraries, features, or dynamic-features.

This CL adds a method to the MergingPolicy interface to
whitelist lower priority manifest types which attributes
can be merged from.

Originally, I considered just setting a NO_MERGING_POLICY,
but that would have potentially broken existing projects in which
none of the variants had code but the android:hasCode attribute
was only set in the main manifest.

Bug: 110417580
Test: ManifestMerger2SmallTest, manually checked repro project
Change-Id: Iae74263418df22f0333e9e0cec4aba82e870af0e

Former-commit-id: 5b10f0b36d70d983b7d6de1ecf8fa0ebe6aa517f