Throw exception when duplicate java resource in feature
Previously, if the base module and a (dynamic)-feature module had a
java resource with the same name, and mininyEnabled was set to true,
we would just use the base module's java resource, instead of
throwing an exception.

This CL changes the behavior to throw an exception instead.

An exception will also be thrown if there are duplicate java
resources in multiple dynamic-features or multiple non-base
features.

Change-Id: I77fb31442dd193a29b135eb2934fa5b1c9583bf3
Fixes: 111520742
Test: MergeJavaResourcesTransformTest, DynamicFeatureJavaResTest

Former-commit-id: af3ce4cf752575d2168a1a849e436b07cd6f09ea