Introduce a DexArchiveDatabase
I was getting carried away with this CL initially, trying to optimize the information needed to
be stored in the cache. Since then I extracted most of the core functionality only to this CL.
It does, at the very least, not store duplicates classesXX.dex checksums for different APKs.

With this CL we can call the DexArchiveComparator (much like in the DexArchiveComparatorTest)
and get a Result object that contains a map that we can feed directly into one of ClassRedefiner.

Here are the list of things that is change / added:

-Added two representation of an APK should it be on disk or on cache
-Added a database to store the APK cache and a very simple implmentation that is in memory.
-Changed the DexArchiveComparator to deal with the new DexArhive objects instead.
-Updated the DexArchiveComparatorTest which behaves like a hotswap of two APKs.

Test: ./tools/base/bazel/bazel test tools/base/deploy/swapper/...
Bug: 109892884

Change-Id: If08e10bcee5324247a05a4c000c2022a67e98926

Former-commit-id: 928d569638926365e9d13d7374a3540567649916