Revert "Fetch remote repositories in parallel."
This reverts commit 3fecb240730e03e5756947b1d64bf28a2a2cb861 [formerly 4f304bf0c598de5fc8daddf8d6e9879ef9a30bf5].

Reason for revert: Greedy approach to parallelize XML parsing
in addition to downloads per se didn't work because post-submit
testing indicates that the XML processing impl is not thread-safe.
Instead of adding a quick follow-up with synchronized() blocks
around XML parsing, I think it's reasonable to revert now in
order to: (1) isolate only download part to be executed in the
thread pool; (2) restore determinate progress indication for
non-download part as we'll execute it sequentially so can use
regular progress indicator.

Change-Id: I14b376971f4b79eafd4fc513a0d34bf4f9a744c1

Former-commit-id: 554a02af12982044dfdce3b1ea512781fae53fc0