Fix progress indication consistency warning in RepoManager.
Progress indicator that is passed to RepoManager appears to be
in the indeterminate state in practice, so reporting precise
fractions on it leads to warning about visual inconsistency.

This CL fixes that by explicitly setting the indicator to
the determinate state, and restoring it to the previous state
when SDK repo loading is done.

Bug: N/A - observation aside when working on b/85839361.
Test: N/A
Change-Id: I9231cd92a7480e5b9cab93af0a5b1ac62681af2c

Former-commit-id: 97a9d22c8b8fdb827df3d04ed2532c21ea415870