Run simpleperf report-sample step on the host
When running simpleperf report-sample on the host, we need to make sure that
we don't run the same step on the device. Instead, we should make sure the
raw trace (or intermediate format) is copied to the CPU service response.

This change is done under a flag, synchronized via AgentConfig with Studio's
flag StudioFlags.PROFILER_SIMPLEPERF_HOST.

Bug: 117758275
Test: Added test to simpleperf_manager_test.cc

Change-Id: I76be07dd0518724d6ddd11b600430228ea1ba50b

Former-commit-id: c5d5187d603dad9c51576e901335ab734bc8e012