studio stats: clean up AnalyticsSettings API
AnalyticsSettings API used to be instance based, this was exposing an
implementation detail of the GSON object being serialized.

This caused issues as people kept cached instances around.

This change makes all public API static on AnalyticsSettings
and creats a separate AnalyticsSettingsData object that
is only used for serialization and testing.

Change-Id: I8fdd88fa5ecb403c97bae3029466337618b0d137
Test: updated
Bug: 110560339

Former-commit-id: 25277d1d3e43b9f9ebedf0d27ae8a45f0b99cc5b