Auto-namespacing: use correct resource visibility
Use the public.txt to filter out which resources need to be placed in
the public.xml file (for namespaced app/libs).

Auto-namespaced references will use the private resource access modifier
(the "*" symbol) in order to reference resources that might have not
been declared in the public.txt files.

These maybe-private references work in the non-namespaced world, so we
need to keep it this way, but using the public.xml file will mean that
the namespace aware app/lib depending on these auto-namespaced
libraries will not have access to the private resources.

Change-Id: Ib86b8cd97e4bef2b7fca93ec9d9e5a8cf9bc3e89
Fixes: 111545408
Test: AutoNamespaceTest and NamespaceRewriterTest

Former-commit-id: d0294ab672bfb83048575a686b56368bee805ae3