[Clean up] Builder style for adding test projects' contents.
This CL adds the following builders to construct the contents of
different types of files in a test project:
  - BuildFileBuilder
  - JavaSourceFileBuilder
  - LayoutFileBuilder
  - ManifestFileBuilder

The goal is to hide syntatic sugar in the builders and only keep
relevant parts in the tests, making the tests easier to read and
maintain.

An example usage is the new AnnotationProcessingAsSeparateTaskTest
(which will be expanded in later CLs).

This CL also cleans up the AndroidTestModule interface to simplify its
usages.

Bug: N/A (clean up)
Test: Existing tests + New AnnotationProcessingAsSeparateTaskTest
Change-Id: Iac8de99629f8c762dc7699a3ba486bc2d754ac3c

Former-commit-id: ad4546f79bff32eb6f47552f59ac764b8d066930