Base TaskCreationAction with some logic.
New VariantTaskCreationAction that receives a VariantScope
and does some basic logic:
- sets dependency on preBuild Task
- sets variant name if applicable
- sets android builder if applicable.

Also move some DefaultTask or other Gradle-provided
task types to use custom tasks that implements
VariantAwareTask (or extend AndroidVariantTask) in order
to ensure we have variant names for all tasks (for
analytics purpose).

The main goal is to slowly remove logic sprinkled everywhere
to handle the preBuild task dependency. This changes makes all
tasks depend on preBuild which is fine, and actually reduces
the risk of having a task not depend on it.

Test: existing tests.
Bug: N/A
Change-Id: I062edeb4cb3ead0a284ad0987394bdea90e4c2d8

Former-commit-id: 217a4bcac0a5521d09c8f7cbb8b2c5b306468710