Auto-namespacing: introduce AttributeSymbol
Introduce AttributeSymbol to be used instead of NormalSymbol during
handling of R-def.txt files. AttributeSymbol contains additional
information on whether the attr was defined under a declare styleable
element (if it was a 'maybe' definition).
This will be used during auto-namespacing when finding the best
candidate for attribute package resolution.

Test: updated
Bug: 110088465
Bug: 73626400
Change-Id: I0afc360c5aedf93b1bc831b8813f4b963e3c9f09

Former-commit-id: 845286886a770ea943589a95ba29a77381226547