Create anchor tasks after the variants are created.
The only task we need as the variant task are created is the
assemble<Variant> one as we have some manual dependencies right
now.

This change moves all other anchor tasks to after so that they
can be created in a single pass rather than keep trying to create
them for every variant.

Some additional clean up related to the feature plugin. There was
some issue around the name of the variant specific task and the build
type ones (when no flavors) when there is both debug and a debugFeature
variants.
In this case, there is now always a variant specific task (assembleDebugAar,
assembleDebugFeature) and a build-type assemble task (assembleDebug) that builds
both.

Bug: N/A (clean-up)
Test: existing tests.
Change-Id: I2dbc412c2bf99ee51c950bf6f58eb0093252a6fb

Former-commit-id: 9ffedcfef19e9656843b1dd82f5e393972115994