Suppport metadata and description fields in Benchmark
This data is Benchmark-specific (as opposed to Run-specific), so there
is no logical way to include it in the JSON files generated by the
Metric class. So the PerfData class was added, which writes Benchmark
data to a new file, 'perf_data.json', which will be picked up by
Buildbot and sent to GCS along with the <metric>.json files.

Bug: b/112777956
Test: PerfDataTest.java
Change-Id: Idb21a28833e80a1c248f398163495c4fcecb83e6

Former-commit-id: 7b83c845828068631682a86ef5d82ce5b2ec234b