Update XML reporter to (optionally) include fix diffs
This CL updates the XML reports to contain descriptions of quickfixes,
along with editing operations, describing how to perform these fixes
mechanically when possible.

Here's a sample snippet from the unit tests:

  <issue ...
    summary="Spelling error"
    explanation="This check looks through ...
    errorLine1="    &lt;string name=&quot;...
    errorLine2="                          ^">
    <fix
        description="Replace with &quot;used&quot;"
        auto="false">
        <edit
            file="res/values/strings.xml"
            offset="81"
            after="message&amp;quot;>%d "
            before="unsed resour"
            delete="unsed"
            insert="used"/>
    </fix>
    <fix
        description="Replace with &quot;unused&quot;"
        auto="false">
        ...

As you can see, under the <issue> node, there can now be one or more
<fix> elements which describes fixes offered to the user. The "auto"
attribute indicates whether this fix can be applied without human
intervention; this is true for fixes that are always a good thing
(such as removing unused imports).

A fix can contain one or more <edit> elements.  Each edit lists a file
and an offset, and it also includes "before" and "after" attributes
which shows what the text is directly before and after that offset in
the document; this can be used for extra sanity checking. Note that
the offset refers to the character offset in the file after newlines
have been converted to \n (from \r\n on Windows for example).

Finally, the edit lists an optional string to delete, and an optional
string to insert. If both are specified, the deletion should happen
first; this is basically a change operation.

(Including XML fix diffs is needed by google3 tooling.)

To enable this from the command line driver, add the flag
--describe-suggestions.

This CL also adds new lint testing infrastructure methods to test the
XML and HTML reporters, in addition to the existing text report
facilities.  Also ported the test result class to Kotlin such that it
can benefit from default parameters (since the new testing methods
supply many parameters that tests shouldn't have to pass in.)

Test: New unit tests included
Bug: N/A
Change-Id: I11b02feb307d8221f98485e63f55bbc43c127678

Former-commit-id: caf08687a247c4abce8a118c2cc08760cd7283ae