Update the AarLibrary struct to reflect how its actually used.
Some of the fields were not present in certain situations. Currently,
they are being filled in with bogus data in these situations, which is
likely to cause problems for API consumers that rely on the fields to
be filled in as documented.

This change makes a bunch of fields optional and updates the javadoc
to explain the circumstances in which the fields will be omitted and
what API consumers are expected to do about it.

This change also adds the "representativeManifestFile" property.
Previously, the "manifestFile" attribute served a dual purpose: it
may have pointed to a manifest bundled with the library intended
to be merged into consuming apps, or it may have pointed to a
representative manifest whose only purpose was to provide metadata
to help understand the resources. It may have also been omitted entirely,
in which case it had been filled in with a bogus nonexistent path.

With this new property, API consumers can disambiguate between the
three possibilities. If they just want to obtain the package name, they
can look at the representativeManifestFile attribute. If they want to
compute a merged manifest (for example, when opening the manifest editor
for a dependent app), they would look at the manifestFile attribute.

This change also adds default values to all the properties, for the
convenience of construction code.

Add helper methods to simplify AarLibrary construction from Java

Bug: 112598316
Test: AarLibraryTest
Change-Id: I14df877e9116d2a1e2207fa2038793325b0900c0

Former-commit-id: e18143d0b290453e67626abb1f41bb1da8c1e3e1