[AVD Mgr] Don't update hw.ramSize when the hardware definition changes
When a hardware definition changes, update the configuration of each
AVD that is based on the definition, but do not update hw.ramSize.

The AVD's RAM size is already set to the recommended value or to the
value the user explicitly set. Even if the hardware definition changes,
we'd like to keep the existing hw.ramSize setting.

Fixes: http://b/113050265
Test: Updated AvdManagerTest.testUpdateDeviceChanged
Test: AvdManagerCliTest still passes
Change-Id: I22692cd1ae806cceef571f803d34b86f0be32d3c

Former-commit-id: ebab7db7294819fc34f1880b0257941c097e9e8e