Sanitized BuildableArtifactHolder APIs.
replace bunch of unclear APIs with saner APIs based on the
return type expected :

createArtifactFile will create a BuildableArtifact and return a
Provider<RegularFile> to be used a single file task output.

createDirectory will create a BuildableArtifact and return a
Provider<Directory> to be used as a directory task output.

createBuildableArtifact will create a BuildableArtifact based
on existing files or folders.

Each API takes an OperationType parameter to indicate whether this
is the initial provider of the BuildableArtifact or it appends or
replace the existing one.

modified all tests to use new APIs.

subsequent CL will :
1. add new tests for INITIAL/TRANSFORM/APPEND collisions.
2. wire new APIs to annotations like @Initial, @Replace
3. remove all remaining deprecated appendArtifacts APIs from code.

Test: updated existing tests
Bug: 110890815
Change-Id: I17f1aa42ebd191f7e85100855dc1a1657f665385

Former-commit-id: 68b2733258bb709645575e54150cdcea611809fc