Allows min sdk to be greater than base for dynamic-features
Both normal features and dynamic-features output a new feature metadata
manifest that will be consumed by the base.
For dynamic-features, we strip the min sdk from this manifest for
merging as dynamic-features can have a min sdk greater than the base.
Updated optional feature names to be clearer.
Removed separately merging the bundletool manifest, as that is
duplicated work now that we have a separate manifest we can use for
merging feature specific manifest information to the base module.

bug:112878394
Test: Added unit and integration tests
Change-Id: I33d2cf05b50d2c367bd92d997a7738552a834f66

Former-commit-id: 5001125219d987558de1c59185607f7e3961b7bc