Allow the agent to communicate with the installer.
+ Adds an agent server, which is started by the installer using run-as.
Agents can communicate with the server via socket; the server
communicates with the installer via pipe.

+ Adds a 'common' directory, which contains utilities for communicating
between deploy components (agent, agent server, installer).

+ Adds server/client code to the agent and agent server. Passes the swap
request to the agent via the server; removes the copy-request-to-disk
step of agent setup

+ Removes the AgentConfig proto (no longer needed)

Messages from agents are forwarded via the server to the installer and
vice versa; the server is a proxy.

The communication from the agent to the installer is currently a simple
debug message. This will change in following commits.

Bug: 112310583
Test: common/socket_tests.cc, common/message_pipe_wrapper_tests.cc
Change-Id: I5182eb0e5cdd25e42986a8d07eb5326169b076ec

Former-commit-id: 54531dbfcd4a317617afd37001a64ec07de5fd52