Change installer swap to use stat instead of test
The installer swap command checks if the agent has already been copied
to the target package's data directory by using the 'test' command.

However, "run-as test" results in a 'permission denied' error, which
causes the swap command to always assume it needs to copy the agent.

Since the agent binaries are not given write permissions, overwriting
the agent fails, and swapping fails if the agent already exists.

Bug: 112378860
Test: Attempt to swap via installer twice in a row. This may succeed on
some emulators and devices, depending on the behavior of test() on the
system.

Change-Id: I7d4b94c1acb69764ee2c6a52932423b00de71a37

Former-commit-id: 987094b5d0509a58c0df0166736680d619f90b1f