Fix DirectoryNotEmptyException when deleting Windows files.
We have confirmed through a number of experiments (see the bug below)
that there is a timing issue on the Windows filesystem, such that if we
delete a directory immediately after deleting its contents, the deletion
might occasionally fail. If we wait for a little after deleting its
contents, then it has a higher chance to succeed.

Note that in our experiments, we did not see any other thread/process
accessing the same directory at the time of deletion, so this is not a
concurrency issue.

We have also tried using Guava
(com.google.common.io.MoreFiles.deleteRecursively) for file deletion,
but (surprisingly) Guava also faces this same problem.

To work around the issue, we now keep deleting the directory until it
succeeds or hits a reasonable limit.

Change-Id: I5b0bf3ad15c125858c5b4f6ae57f6d29f6e7f9b5
Fixes: 77478235
Test: Existing tests

Former-commit-id: 1fe91e9581ce412be20e134c7759b871b707a28c