Res merger: fix isFileBased when scanning for ids in layouts etc.
ResourceSet has a mode where "id generating" non-values XML files are
scanned looking for "@+id" declarations (currently this mode is only
used by the IDE). When that happens, we create a 1+ ResourceMergerItems
from a single ResourceFile, but the information that the "main" resource
represents the whole file was lost. This change overrides the
isFileBased method and fixes loading from the merger blob to preserve
this information.

Test: ResourceMergerTest
Bug: No bug, but this was discovered when trying to switch the IDE to
  resolve all references using ResourceRepository and needs to be fixed
  for the namespaces refactoring to work correctly.
Change-Id: If803d31cbe3eff5c17e54195d64477f6490dee3b

Former-commit-id: e2935835541b8779ba714ba5da33ea606d9c5cd2