ProjectOption: enableUncompressedNativeLibsInBundle
This is a flag for a feature we don't expect developers to generally
want to disable, but adding it as an option to make it easier to debug
issues and control the risk of deploying the feature.

This is already enabled by default in this change, and connected to the
bundle tool invocation.

Bug: 113022642
Test: TODO
Change-Id: I18f979199dc4dc4b3210973f711bac53eab33c43

Former-commit-id: d8a444d545bad6a1134736074acbc0bc70e41989