Expose TaskProvider via the Variant API.
Now that the plugin is lazily creating the tasks, we do not
want other plugins to force configure them when calling our API.

This change adds a new getter for the TaskProvider wherever there
was a task getter. Calling the task getter now outputs a deprecation
warning that the API will disappear end of 2019 (which may be at
the same time as we remove the old variant api anyway).
It also displays a URL for more information (address TBD).

Test: existing tests updated.
Bug: N/A (clean-up)
Change-Id: I6939647192d827e4349820ffc61143430ad66cbd

Former-commit-id: 8d9060333ea113bb6efcccb6a7b7426dfd9ba7fe