Use new MIN_ANALYZER for noisy metrics
Use new MIN_ANALYZER for DEX_MERGER and EXTERNAL_LIBS_MERGER metrics.

This is a temporary fix to stop the email spam until there's a better
way to configure the analyzers per metric/benchmark.

MIN_ANALYZER aggregates run data by taking the minimum instead of the
median. The minimum is much less noisy for DEX_MERGER and
EXTERNAL_LIBS_MERGER because both of these metrics are fairly consistent
when the tasks run fast, but when they are slow, they tend to be very
inconsistently slow.

Bug: N/A
Test: bazel test //prebuilts/studio/buildbenchmarks:Plaid_cleanBuild
Change-Id: Ia6d0e9998216f5af3e6f602d239c3981f86412ed

Former-commit-id: 45771f1add548977162bc733ddefc8a73eceea90