Pass feature info to dex splitter via file name instead of manifest
Previously, the feature source of each merged-classed jar was passed to
the dex splitter via the manifest in the merged-classes jar.

While designing a similar mechanism for splitting java resources, I
rejected the idea of using the manifest because the manifest files
are merged and/or filtered out during java resource merging.

My plan is to convey the feature sources of the java resource jars to the
dex splitter via the name of the java resource jars (e.g,
feature-foo.jar)

It's better to have the same mechanism for conveying the feature source
information for classes and java resources, so this CL changes the
mechanism for classes.

Bug:111082605
Test: existing
Change-Id: Icd84f13eb619d7961bd8579c4512295fe1381a99

Former-commit-id: 7fd14c7d0db4fd156c613f8ee43d924924feb850