Replace "representative manifest" with packageName
Previously, we allowed the build system to attach a "representative
manifest" to External Library, which would be a valid manifest not
associated with the libary but containing the correct package name.

It is much simpler for the build system to just determine the package
name and attach it explicitly in such situations, so this CL replaces
the representativeManifest attribute with an optional explicit package
name.

Bug: 112667464
Test: ExternalLibraryTest
Change-Id: I215b0d8fdc45dbb151b473e3bcc57f099999b7df

Former-commit-id: bb5209be51c606788736e69d3a873c3b80512651