Add dependency reader to GoogleMavenRepository.
This CL is a preparation for the implemention of logic to compute
a compatible version of a new dependency for a module with the new
rules for semantic version checks introduced for AndroidX.

In this CL we are adding a method to give us the dependencies with
version numbers of a library from the GoogleMavenRepository. We
need the dependency information of libraries that may not exist on
the users harddrive. Therefore we are adding it to the Google
Maven index.

The format chosen is: the original pom files distributed with the
actual library. The pom files should be added to the google index.
This information is lazily read from the server since the info for
the older version are seldomly needed.

Also in this CL add a GradleVersionRange as a preparation for the
same logic.

Bug: 111656600
Test: Unit test added.
Change-Id: Iebc9993522fb8dd7b58e5b5ad735f64168613ad3

Former-commit-id: cbbf531fe9119db0f56005c40c36b2805724fff6