Set initial Java heap size to 10GB for baseline memory test.
The memory baseline test for go/adt-perfgate shows high variance due to Java
garbage collection. While the max heap size is set to 10GB, and the test
allocates ~6GB, the print of free space shows that the heap size starts at
about 2GB. Thus, a lot of GC time is probably spent up front getting to the
heap size required to allocate the 6GB needed by the test. This CL changes
the initial heap size to 10GB, and running a few tests shows that this
greatly reduces the GC time (by about half).

Bug: N/A
Test: bazel test --config=remote --cache_test_results=no --test_output=all
                 foundry:perfgate-baseline-tests

Change-Id: I8da2bcaa7914b26916310abcedb06b38df496334

Former-commit-id: 22712d89c7ec89425d9556778445610b0b6d2f1c