Create a task to write the signing config to a file.
We need this task so that dynamic-features can read the signing config
from the base module. The signing config should not be read for each
dynamic-feature, but shared with the base module.

Future cl's will read the signing config and move any dynamic-features
that read the signing config from the variant configuration to read it
based on a task. This will also configure the instant run packaging too
later.

bug:114626324
Test: added unit test
Change-Id: I946eec4a67b9c25e00eea1acc2aab89ec7f9d107

Former-commit-id: 825bb616b495629080fcb980ab667aa07e32241c