use allocation policy rather than module's minSdk.
We used the module minSdk version to determine whether or not
package id below 0x7f are allowed.
This prevents creating optional features that support newer
platforms than the main application.

From now on, we will trust the allocation from the base module
and set the flag for aapt2 depending on the offset value.

fixes: 113857937
Test: added integration test.

Change-Id: I5b832eda974cf60a5304427181c2b873a49ebe52

Former-commit-id: 4587b9c2f25a5ee037025c73c78e5b227afaae37