Fix dex artifact transform for library androidTest
Build dex for the main variant, to be consumed when
building the library android test variant. Dex is
built from LIBRARY_CLASSES i.e. the classes that
are published in a library project.

Once b/115334911 is fixed, this can go away, as
dex for the tested variant will be built using
artifact transforms.

BUG: 112087084
Test: DexingArtifactTransformMultiModuleTest
Change-Id: I0b1d9544f72b0cbd156f957e34e005b87550319c

Former-commit-id: 8ea760c6f7266a5e06279e843d040f60867c01f6