Improve error message for enforcing minimum plugin versions.
A plugin may come as a transitive dependency that is not visible from
the users' code, so when we show a violating plugin, it's not clear to
users how to fix it.

To address that, this CL traverses the dependency graph to display the
dependency path from a project to the plugin that violates the minimum
version requirement.

All related information is also passed to Studio, so Studio can perform
a quick fix later.

Test: Updated PluginVersionCheckTest
Bug: 115893234
Change-Id: I5c8b1798796dd38eeb3632f1f846ae25a2619e22

Former-commit-id: ad5a217b48573d344c857e5acd775f20f2f9bf25