Make LibraryJarTransform do less work.
Register compile-only R-classes as PROVIDED_ONLY, so they don't need to
be processed.

Unless shrinking is used, in which case they are it is needed to support the case
where all program classes are shrunk away.

This is to reduce the amount of processing the library jar transform has
to do, but we still need an eventual fix for b/115974418.

Test: existing tests.
Bug: 115974418
Change-Id: I0ff7578aaebca8d273bf31c8e6918777ac63a2ac

Former-commit-id: 5d5f31e8922065487e95f7244a55cf824f212a7a