remove automatic injection of @InputFiles
tracking the issue has been challenging, there is an issue
with setting a private field using the kotlin library but it is
intermittent and not reproduceable for now.

I suspect an issue with the Kotlin runtime library so turning
off the feature until I rewrite the injector to use plain old
JDK reflection APIs.

Bug:114361898
Test: existing tests + some unit tests trying to repro this issue.
Change-Id: I589822dc453596b36d4b2f006b95f989dc50f6d3

Former-commit-id: 01c92f93d85421a6f3e6acb8cbacce3d5cf7e9df