Rename all namespaces to 'deploy'.
We had a scattering of different namespaces that only served to add
confusion; the agent was 'swapper', the installer was 'deployer', and
the common libraries were 'deploy'.

This renames all namespaces to 'deploy', since that is the shared
purpose of all the namespaces (C++ namespace != java package)

This commit also has applied clang-format --style=Google to all the
modified files. If that is an issue, I can re-do it.

Bug: None
Test: Existing tests still pass.
Change-Id: I0489e00b6034ebfa4b6e6402890348037afd8539

Former-commit-id: e265e74c1b40932c6f1cdd1eb0bd1cedabec505a