Removes config file parsing logic from the agent.
Removes the ability to pass a SwapRequest to the agent, now that all
tests have been migrated to communicate on a socket.

The agent now requires a socket connection to communicate
success/failure messages, so we can remove the file parsing logic.

The functionality in the "config.h" header will be folded into agent
logic in a future CL, since it's not really neccessary to have it in a
distinct header.

Bug: None
Test: Existing tests still pass.
Change-Id: I6ee8e7d3f34600f6b59a2d1a5b63962a6e40ac00

Former-commit-id: df1357aec20c68e9b616e180580d4869a42bfb97