Create the SubmodulePath type
The projectmodel uses some paths that can contain wildcards
and some paths in which wildcards are forbidden. Previously,
we had been using the ConfigPath type for both sorts of paths,
which forced methods that received a non-wildcard path to
deal with extra corner cases that could not occur and didn't
offer any strong typing to protect against the types being
mismatched.

This new type makes it clear where wildcards and adds
strong type protection against the types being misused.

Bug: 116617697
Test: SubmodulePathTest
Change-Id: I04c40dedb01381a884d034e89c75770dc687afcb

Former-commit-id: d7e902ace591c1b177e120cba9ac8cfb085fdf4b