Refactor projectmodel to simplify modification
It would be easier to construct AndroidSubmodule instances
if the Variants just held metadata and the Artifacts were
contained in the AndroidSubmodule objects directly. Test
cases that wanted to modify the set of Artifacts would then
no longer need to modify any of the Variant objects.

This change moves the Artifacts up one level in the model to
be associated directly with the AndroidSubmodule rather than
the Variant. The Artifacts are associated with their variant
by using their SubmodulePath.

This change also introduces the notion of a SubmodulePath.
A SubmodulePath identifies a specific artifact or variant
within an AndroidSubmodule. SubmodulePaths cannot contain
wildcards. This distinguishes them from ConfigPaths which
can contain wildcards, identify a Config, and may match
multiple Artifacts or Variants.

This is useful since it creates a distinction between methods
that accept wildcards (which continue to take ConfigPath)
and methods that do not accept or return wildcards (which now
take SubmodulePaths.

Bug: 116617697
Test: AndroidSubmoduleTest, ConfigTableTest, VariantTest
Change-Id: Ic8162970623e5518ce492a06a9a375c100080b23

Former-commit-id: a21312de1b43f2a9c858e0d470423a4b767231ac