Protect logic for retrieving class tag during ObjectAlloc
To fix a potential nullptr that can happen when multiple threads
are loading classes and writing to the class tag map simultaneously,
which can cause the map to be rehashed if the number of classes are
large enough. The old code pre-size the map to avoid re-hashing but
we should not make assumptions about the max number of classes
given multi-dexing.

Bug: 110924477
Test: Failed to repro the bug but found the most likely culprit.
(See description for details)

Change-Id: I54385647bfc777b67f3be1a9be6c6cc79eba4192

Former-commit-id: 61b36fbe2d3e187a3e2735468e35475a5a51c005