Accept ids defined in values file used in barrier.
We used to suppress reference problems if an id wasn't found
anywhere. However if an id is defined as a values item we don't
know if it is defined before all files have been analyzed.
The reference check happens earlier.

In this CL always make the reference check.

Also pulls in id's from <include> tags, if necessary.

Fixes: 79407972
Test: Add lint tests for barrier references to ids in values file.

Change-Id: Id10d999d286c67754399a82f728a4c24f3f1984a

Former-commit-id: bba10c7dd17463fb1ef766a2b430b8254602fa92