Collect sources to compile at execution time.
Users/third-party plugins may add sources to compile through the DSL, so
sources that are collected during the configuration of the Java compile
task may be incomplete.

So far we have worked around this issue by adding sources to the Java
compile task also when the following DSL calls are made to register
custom sources:
  - BaseVariantData.registerJavaGeneratingTask(...)
  - BaseVariantData.registerExternalAptJavaOutput(...)

The current approach has several issues:
  - It forces the Java compile task to be created and configured.
  - The custom sources are not added to the new ProcessAnnotationsTask.
  - Sources may be added twice (which may slightly decrease
    performance).

To solve all of the issues above, this commit collects sources to
compile at execution time instead.

Bug: 117161463
Bug: 115774858
Test: Existing tests (Refactoring change)
Change-Id: Ief66555be7a74af50c896d11071c081c5b29acd1

Former-commit-id: a3825b797f5be363bb47aa51085957eab136f5e0