Installation is aborted when swap fails
Installation now happens on device and it will be
committed iff swap succeeded.

There was an issue with the old approach where
in order to know if an application could be installed
we needed to know if the swap was successful, but by
swapping during the reload app info loop it's already
too late to do that.

This CL moves the swapping to OnAgentAttach, which
increases the chance of code running between that and
the app info message is received, but to mitigate this
we block the main thread after swap until the installer
issues the reload appinfo command.

Test: to follow up.
Bug: 112430219
Change-Id: I35beac6864553ea43893b056b4098836cb45e876

Former-commit-id: ea4c94c709d323c8b7464f567288dca2973b3221