Fix performance problem in toString
The toString method on large models was creating a performance problem.
Two cases were causing the worst problems:
- Long lists were causing the toString runtime to grow without bound.
- ArtifactDependency graphs could recurse to arbitrary depth, which
  also grows without bound.

This change puts a cap of 10 elements when printing the contents of
any collection in the model (it appends a string like "...300 more").
It also removes the content of the recursive dependency list from
the toString method, and now just displays the number of dependencies
rather than their content. This latter change prevents unbounded growth
when there is deep recursion between dependencies.

Test: Covered by the *toString tests for test cases in the
      com.android.projectmodel package.
Change-Id: I2707f4ca20ba75bf32a28bc80e36e53e39096035

Former-commit-id: 7c990b8652266f23578824de0c754886543d996c