AAPT2 Daemon: Better idle daemon quit error message.
Check for the end of the stderr stream rather than standard out for the
idle daemon listener, to avoid a race between the closure of the two
streams. This caused the daemon quitting to sometimes be reported twice,
once during compile, which returns an error and again through the idle
daemon listener as a spurious log message.

Bug: 117486770
Test: Aapt2DaemonImplInternalErrorHandlingTest is not flaky 0/5000,
      with the previous flakiness mitigation removed.

Change-Id: Ie03489ad6970ab26f26007dcd762523d1e8372d7

Former-commit-id: 89b08bd999ec6da6aa2e46dad4076d3880fcfbcd