Adding retry when Atrace fails to allocate requested buffer.
This change should address most of the "small traces" issues.
What happens is when we request a buffer from Atrace sometimes it fails,
and then allocates a 1 KB buffer. This buffer is not enough for any signifcant
captures. With this change we check the allocated buffer size if it is
not enough we reduce the amount we are requesting and try again.
The reason the buffer may fail is because the memory of the device is
to fragmented to be allocated from a contigous block.

Test: Updated AtraceManagerTest
Bug: 117845741
Change-Id: I129b6a29fe513640b11bb6232a3c617b7ce1cb17

Former-commit-id: 59aba537e49a105a63c0f6d0b97fa6568052da17