Toggle network profiler's samplers based on flag
- Refactored the sampler, session code slightly to allow
better support for the added tests. The refactoring also
fixes a flaky sampler test, because previously, the sample
worker thread starts immediately in the constructor and
can inadvertently hold on an unitialized mutex. By moving
the worker thread to an explicit start, we ensure that the
mutex will be initialized.

Bug: 117826867
Fixes: 117965764
Test: Added

Change-Id: I6824a7d24b2055b432188a9c8facd400fb021b77

Former-commit-id: c13de4b2234afc0281649f02300db2bee49aa716