Incap: Simplify ProcessAnnotationsTask's logic.
So far, when the separateAnnotationProcessing flag is enabled, if all
annotation processors are incremental, ProcessAnnotationsTask will be
skipped so that AndroidJavaCompile can perform both annotation
processing and compilation incrementally in the same task.

Because this decision requires resolving annotation processors, it
needs to be made at execution time. At configuration time, we don't know
how the decision will be made, and need to prepare the two tasks for
both cases:
  - ProcessAnnotationsTask always has the annotation processor generated
    sources directory as an output.
  - AndroidJavaCompile always has the AP directory as an output, in case
    it needs to perform annotation processing.
  - AndroidJavaCompile always has the AP directory as an input, in case
    ProcessAnnotationsTask does annotation processing.

This leads to two issues:
  - Both ProcessAnnotationsTask and AndroidJavaCompile share the same
    output directory. It confuses Gradle and breaks the build cache.
  - The AP directory is both an input and output to AndroidJavaCompile.
    This breaks incrementality when running with Gradle 5.0 (the first
    build after a clean build is not incremental; however, subsequent
    builds are still incremental as expected).

(Changing the tasks' inputs and outputs dynamically at execution time
won't work because at the time of modification, the task graph will have
already been resolved and can't be udpated.)

To fix these issues, we now always create and run ProcessAnnotationsTask
when the separateAnnotationProcessing flag is enabled, regardless of
whether annotation processors are incremental or not (i.e., the decision
to run the task will now be made entirely at configuration time).

Bug: 115774858
Bug: 117293097
Test: Updated + Re-enabled IncrementalJavaCompileWithAPsTest
Change-Id: Icafd28c2e8dadce44cb27cd27b253d480bd27c04

Former-commit-id: e28a8d62cbbfa079e9ff2e0ab62291d30efaf210