Add option to fetch ProjectBuildOutput in GradleTaskExecutor
Lots of tests use GradleTestProject::executeAndReturnOutputModel to run
tasks, then receive the output model. This CL adds the option to
GradleTaskExector to fetch the ProjectBuildOutput object during the same
Gradle invocation that executes a set of tasks. This allows
executeAndReturnOutputModel to return the output model without making a
separate Gradle invocation via model().fetch().

This CL also adds a new variant of executeAndReturnMultiModel named
executeAndReturnOutputMultiModel, which mimics the behavior of
executeAndReturnOutputModel to return the models for a multi-module
build.

Bug: b/67909922
Test: Test change
Change-Id: I1155d6882ce9fe4a998d4cdbf07ebcb0d1c5c57d

Former-commit-id: 3767dee35fdce14add15f9b46bdedf76b19cd744