IDE model: Expose compileJarFile instead of apiJarFile.
In a recent change to support api.jar in AARs, we exposed the
getApiJarFile() method in the model. This is not good design because the
IDE should not have to know about api.jar. All it cares about is the jar
file that will be used for compilation.

The jar file used for compilation is api.jar if that file exists;
otherwise, the regular jar file used for packaging will also be used for
compilation. This logic should be hidden from the IDE.

Therefore, this CL exposes the getCompileJarFile() method instead of the
getApiJarFile() method in the model and hides the computation of the
compile jar file in the plugin.

For old plugins where getCompileJarFile is not available, fall back to
use regular jar file.

Bug: 117145267
Test: To be included from the IDE side
Change-Id: If6c5bebea3aa9fd918f0c92d3d28b9194f27ad42

Former-commit-id: a589fb6985095779eb14835966b8e197c7b87004