Allow a few more system ticks between file reads.
This will slow down the test but should help us see if the flakiness is
coming from race on reading file metadata or if the file is actually
being changed but the low clock resolution thinks it's the same.

If still flaky after this, we might need to find a better way to assert
if an output artifact has changed or not.

Bug: 113323972
Test: existing
Change-Id: Ib480ab5a77dfe8f5dcb478d217d8685e2fb43500

Former-commit-id: b19d67149906cf1f2be0193d45b68d558fac5f2b