Adding field to CpuProfilingAppResponse to contain atrace buffer size.
Atrace has a requested buffer size it attempts to start profiling with.
However this buffer cannot always be allocated as such we reduce the
size requested and try again with a smaller size buffer.

We want to inform users we are doing this so we pass back the buffer size
in the response message to allow studio to verify and inform users of
buffer size being used by atrace.

Test: Updated AtraceManagerTest
Bug: 118707788
Change-Id: I1d4abe659ae785b9fc85b84ce98f99d7be115681

Former-commit-id: a6c53f28c2f1205799e0098ecf39d897c099b4e9