Compiler settings cache: Plumb ABI config
This is part one of N CLs to implement caching of compiler settings
for CMake projects.

go/cmake-compiler-settings-cache

The purpose of this CL is to extend the data class
JsonGenerationAbiConfiguration to hold information about compiler
settings caching and to plumb this value to the locations it will
be needed. There should be no behavior change due to this CL.

Bug: 119116588
Test: Added unittests

plumb-abi-settings

Change-Id: I31ae8a18d769a67db0c784bc88f9ec7267a6429b

Former-commit-id: ff77e1f9da408b1bf43e7599ede843c828c6f615