116789824:  Lint rule to avoid scrolling glitch in RecyclerView/Paging
Also ports RecyclerViewDetector & test to Kotlin. I did
this first because I had planned to update the recycler view
detector to add the new diff util checks, but after doing that
I decided it's better to do that in a separate detector.

Note that there are no semantic changes to these two files; they
are direct ports.

Fixes: 116789824
Test: Unit tests included
Change-Id: I41db087bc03210a30f13bc0be9ec8c64871023de

Former-commit-id: 5efabbf2beb226021ebfa71cc121083f26d1c05c