Sampler expects ns however we were passing ms. Changed to accept ms.
One of the performance regressions was due to the fact that we
were sampling at 500ns instead of 500ms. This means the unified pipeline
was sending events almost as fast as it could recieve them flooding the
database and the UI with data. This fixes that and makes all other
performance issues seem minor.

Test: N/A
Bug: 119134037
Change-Id: I20b5fdbddc144f892ebc04efab73ac85d313eb16

Former-commit-id: 8c2a594ddc767fda4b534ee6b520475ec4213639