Compiler settings cache: Cache implementation
This is part one of N CLs to implement caching of compiler settings
for CMake projects.

go/cmake-compiler-settings-cache

This CL implements the physical caching logic. The files in this
structure can be read and written from multiple processes at once
so the logic needs to be "thread safe". It does this by employing
a checksum file on a per hash basis.

Bug: 119116588
Test: Unittested
Change-Id: I7418656557e7a917a5b3e215753e40d580cd842e

Former-commit-id: c08e69a96f7d13f0fb0f417c3bc9244908999853