Compiler settings cache: Cache implementation
This is part one of N CLs to implement caching of compiler settings
for CMake projects.

go/cmake-compiler-settings-cache

This CL implements the physical caching logic. The files in this
structure can be read and written from multiple processes at once
so the logic needs to be "thread safe". It does this by employing
a checksum file on a per hash basis.

Bug: 119116588
Test: Unittested
Change-Id: I7ee311182def99204accd503e618777428e00d4e

Former-commit-id: abd7e6eae31b41c19d26e7db86d640b37c3d3987