Compiler settings cache: functions for reading NDK/source.properties
This is part one of N CLs to implement caching of compiler settings
for CMake projects.

go/cmake-compiler-settings-cache

Compiler settings cache needs to include information about the version
of the NDK that produced the settings. For SDK packages (like NDK)
the version information is stored in a file called source.properties.
This CL adds a class for reading that file.

Bug: 119116588
Test: Unittests
Change-Id: Ia593387945d31a7121ee3361d421ace04b4570f8

Former-commit-id: 357582316e274766011a0faf15420cdd8d4906f4