Compiler settings cache: Error, warn, and logging support
This is part one of N CLs to implement caching of compiler settings
for CMake projects.

go/cmake-compiler-settings-cache

Compiler settings cache needs logging and especially the ability to
issue warnings to the user that will show up in Android Studio at
sync time. CmakeLocator already had infrastructure to do this so
I generalized it to work in both places by registering the loggers
in a thread-local.

Bug: 119116588
Test: Unittests

Change-Id: Ibb3cfcb98932d833122ec2c917022f2ea5570e0e

Former-commit-id: 66969b38e3f9b4fc36553d0bdc923e5f4a48f6f6