Fixed PackageSplitAbi and PackageSplitRes not serializable params
After change ag/5533042 the signing config was loaded from a file passing a
FileCollection to the worker. However Gradle's FileCollection is not
serializable.
This change replaces it with Java's File instead and removes
CoreSigningConfigImpl as it was only created as a serializable way to pass
signing config to workers which is not needed now.

Bug: N/A
Test: N/A
Change-Id: I2c2c16b4ba81b9e80a99f9774f56177eb233cc58

Former-commit-id: 96c81e178b8f5d4da539abbf41e6425fc22573af