Compute provided classpath using component identifiers
This CL changes how we compute provided classpath. When computing
it, component identifier is used to find artifacts that are
in the compile classpath, and not in the runtime one. We consider
those artifacts to form the provided classpath. They are used by
e.g. R8 and Proguard when adding "-libraryjar" keep rules.

Previously, provided classpath was computed by subtracting
compile classpath from the runtime classpath. Because library
projects now publish two separate artifacts (files) for the
compile and the runtime classpath, provided classpath contained
the compile library artifact. This CL fixes that.

BUG: 119758914
Test: MinifyLibAndAppKeepRules
Change-Id: Ia567f0f877454be198259f33cd8e867795276c9a

Former-commit-id: 535b2ab225782cd5758c9b66320bc21e7c9f33df