Dexing artifact transform with desugaring
This CL adds support for running dexing and desugaring in the same
artifact transform. In order to avoid re-processing inputs when
dependencies change for build that should not be desugared, there are
now two transforms: DexingNoDesugarTransform which does not support
desugaring, and DexingWithDesugarTransform one that does support it.

Desugaring as implemented here will likely break with file dependencies,
(b/128599004) as they will not have any classpath dependencies. Those
should be processed by a separate task using all the module (i.e. not
the project dependencies) as the classpath.

It is not enabled by default yet.

Bug: 120592547
Bug: 120592547
Test: DexingArtifactTransformTest, Desugaring tests
Test: Added DesugarFileDependencyTest
Change-Id: Id198397c80b6d8f4f9aad3c91a8ad1d611570386

Former-commit-id: fc5d13fc5a790956ae7171d0e008d9bb759f1637