Fix memory leak in GradleDetector
Some static caches in GradleDetector reference anonymous classes
which capture references to the LintClient instance. This permanently
leaks the first LintClient instance used in the IDE or Gradle daemon,
along with all associated project data (such as PSI). This CL converts
these caches to be per-Lint-run in order to avoid memory leaks.

If it's important for these caches to survive across Lint runs, the
alternative fix would be to change the LintClient API so that you
don't need to retain a LintClient instance to call methods like
LintClient#log or LintClient#openConnection.

I've tested manually by profiling with YourKit and ensuring that
the LintCoreProjectEnvironment instance (for example) is no longer
strongly reachable after Lint finishes.

I tried to write an automated test for this, but it's tricky. We
might normally use VirtualMachine.attach(...) to get access to heap
information for the current VM, but in our lint-gradle integration
tests, Lint appears to run in a separate process from the test.

Bug: 119833503
Test: manual (see above)
Change-Id: Ia17206f5331d373652f5d8120c64928e7eae7fa2

Former-commit-id: b09d849086cd1ee07194079244efae05d3120637