IRR: Preinstall & DeltaPreinstall to replace "push"
Problem: Pushing apks to the device has many issues.

1. Some devices don't even have enough storage space to store
the apks.
2. It is a wasteful process since apks are written twice (once
during push, and once when feeding them to the package manager).
3. Preinstalling is more efficient than pushing and later install
since while preinstalling we can analyze the dex content.

Solution:
Preinstall creates a session and write the content of all apks to
the Package Manager but does not commit or abandon. This decision
is left to the swap request to make.

DeltaPreinstall does the same thing as Preinstall except it
attemps to minimize the amount of data transfered by detecting zip
dirty and clean areas.

Test: Follow-up CL
Bug: N/A
Change-Id: I622268e8c6fec79cb55c4913e05b7f85cbe48f6c

Former-commit-id: f3e6f5f7dfac8d6ec6089e4e08ce5c72b4da23f0