Revamp CacheabilityTest to ensure cacheability of tasks.
The current implementation of RelocationTest doesn't actually check
cacheability (perhaps by mistake): It checks that all tasks that are
known to be not yet cacheable but need to be made cacheable are
currently not cacheable. This is not a useful check. As a result, this
test passes even when the build cache is not used.

This CL updates this test to make sure the checks are effective and
exhaustive. Also rename it to CacheabilityTest to make it sound more
familiar.

Bug: 69668176
Test: CacheabilityTest
Change-Id: I605d8027138c99cf13d1f834ae8f7fd6e7c91603

Former-commit-id: 0de8ad8d75c59a205d36004cbbf3ed9d24e645f6