Remove deprecated NdkCompile codepaths
NdkCompile was an older codepath used to support C++ in AndroidStudio.
It has long been deprecated and the user has received a warning about
this. In 3.3 we introduced a 'lease' flag that the user would have to
periodically reset to continue using. It's time to get rid of this
old code.

Bug: NA
Test: Code removal
Change-Id: I3eb04858e1d42c68c577909daf226d79aebe47dd

Former-commit-id: bbb5fdd10c4a0179433509dfbd2f43dc58b9ba34