Merge legacy multidex in a single task
Merge legacy multidex in a single task, in order
for synthesized classes from main dex list classes to
be packaged in the primary dex. Previously, the external
libraries would be merged together, which means that any
synthesized class from an external library main dex list
class would not be packaged correctly.

At the moment it was not possible to hit this bug as
dexing artifact transforms are not enabled when desugaring.
However, that code path will be enabled soon.

BUG: 120039166
Test: DesugarAppTest
Change-Id: I01866982118bdafe1a044612887830506ec9b106

Former-commit-id: 574b5b2618081d98fc8b1c9de6304cb79546fbf0