120087311: Enforce RestrictTo(LIBRARY) within local projects
This is required by the support library.

Normally, the restrict library and library group calls
need to know the maven coordinates, which we don't have
(there's a tracking bug on the Android Gradle plugin
to provide those from the Maven publishing plugin),
but for library we know they also need to be in the
same lint project, so we can simply check that they
are.

Test: Unit test included
Change-Id: Id7fc577b686cee7284eb838a78cd0db0ed9628b9
Fixes: 120087311

Former-commit-id: f40b383cbfa4575e396fb64edfd49ef0ac28e552