Workaround for JDK Bug.
Workaround for https://bugs.openjdk.java.net/browse/JDK-8054213
which only seems to appear on Unix derived platforms.

The new incubating-api.txt is correct (as the previous version)
was contaning the wrong package name.

The workaround is ugly but necessary until we upgrade to JDK 9 or
later where it was verified that the bug is not present any longer.

Test: Fixed a test on Windows
Bug: N/A
Change-Id: I57fcdbc1338cdcafe31b1421990a70381f69bc16

Former-commit-id: a8c3a7694be5a85a9061fdf677c3beb055d34e8a