Provide emulator as a Bazel target for sanity tests
The emulator and system images are extremely large files that they are
not checked into the prebuilts directory. In order to enable the tests
to run with the emulator within a Bazel environment, this change
instead allows Bazel to pull in the emulator and system images from an
installation outside of the Bazel workspace. This is done by creating a
custom repository rule that searches an external Android SDK directory.
The external SDK path can be specified using the 'path' attribute when
invoking the repository rule or setting the QA_ANDROID_HOME environment
variable when starting Bazel.

Bug: 112553978
Test: DebugOnEmulatorTest will use this functionality once the test is
migrated into Bazel
Change-Id: I626ab5dc64a9c392686d3215d00feffa8e0d4e17

Former-commit-id: 7a52deae5b838c731fba08855421275ec1ef822e