Don't fail if R8 is enabled using DSL with desugaring
If R8 was enabled with "useProguard false" in the DSL,
and Java 8 language features were used, configuration would
fail. Reason is that AGP expected both android.enableR8 and
android.enableR8.desugaring to be true. Because it is possible
to specify R8 usage through DSL, this assertion does not
hold any more.

This CL allows R8 desugaring (on by default) to be used if
the R8 shinker is used (either set using the flag, or DSL).

BUG: 120196378
Test: AppPluginDslTest
Change-Id: Icf3ada2f04aede06383186a8c5ef627b02baebfd

Former-commit-id: a225204b4dbf0675a61e635b5cb53765f3adbe8d