Remove Event.type field and friends
To avoid confusion between Event.kind and Event.type, this change
eliminated the type concept. Event.kind serves as a high level data
grouping, and the type categorization is now embedded within the
data payload associated with the event kind.

To support determining if an event group has ended, an "ended" field
has been added to Event to allow users to specify if a particular
event marks an event group as ended. This mechanism replaces the previous
"end" kind in the GetEventGroupsRequest.

Bug: 73538507
Test: Updated
Change-Id: Id95f48670bdfe88a6ca4dfc4757a20032c7e3f9d

Former-commit-id: 2fb1eacba57c3941516afc0f0a1f7c1c497843a2