Changes the task runner to have explicit run and runAsync
Tasks are not automatically started, they are created and queued.
When run is called, tasks start being executed and the runner
waits until all tasks are done.

There is a new runAsync method that will do this on a separate thread
raising any exception as a runtime exception all the way to the top.

Test: added new tests
Bug: 120506342
Change-Id: I234f10ee1447f176bf387e8682ddae14cd70e767

Former-commit-id: 6edd687c4bda074abb97ed512b00d4b26def431e