Prevent emulator test targets from building in PSQ.
Some tests, in particular some QA tests, depend on the emulator to be
available. The emulator is found in an installation of the Android SDK.
However, the prebuilts Android SDK does not provide the emulator
packages. Therefore, such tests normally pull in the emulator files
from an external installation of the Android SDK. However, an external
installation of the Android SDK is not available on the bots running
PSQ. Builds on PSQ will fail due to this missing dependency.

This change disables targets that depend on the emulator from building
on the PSQ instead.

Bug: 112553978
Test: DebugOnEmulatorTest
Change-Id: I244ccf35690586060bdd20ccb837355748b7827b

Former-commit-id: 9ae3f8230ca962e4808a91a155d7d7abfbeecebe