Fix crash when progress indicator goes out of range.
When loading repos, progress indicator could go out
of range due to an arithmetic error. This could
cause crashes in each case when a fallback repo source
is loaded and a relevant subprogress indicator is
created, at which point it asserts the correct range.

Test: Added an explicit assertion to FakeProgressIndicator
- this should ensure all the tests using FakeProgress
actually verify that the rest of the code is free from
this issue.
Bug: b/116361869

Change-Id: Ib0212be986325ee70ca9502a5ad315b767ef1694

Former-commit-id: 2b84c11fef804bc17986aa59e2f86339d4bdcd84