Make --jobs=N config arg not common, so that we can adjust for resources.
The --jobs bazel configuration option "specifies a limit on the number of jobs
that should be executed concurrently during the execution phase of the build".
This will vary depending on the pool of resources running, so remove this as
a common arg, so that we can adjust it for the different bazel invocations
that we support.

Bug: N/A
Test: N/A
Change-Id: I5804aa28ca211fe3b31e44a34acb37a250fe67b9

Former-commit-id: 1a039ec29b3d457dd8dc0de41c906bd79c1fc39f