Moving daemon, and other core components to transport.
In addition to moving components from profilers to transport
perfd.cc is no longer the main entry point for the profiler service.
The entry point has been moved to transport.cc where it then calls
perfd::initialize to setup the required perfd services. This allows
a clean break between the pipeline/utils and profiler specific
code.

The seperation is as follows,
--Daemon Lib---
- Daemon
- EventBuffer
- EventWriter
- Connector

--Utils Lib---
- Utils

-- Main --
~~~Depends on
~~~~Daemon Lib
~~~~Perfd
~~~~Utils Lib
- Transport (entry point)

-- Perfd --
~~~Depnds on
~~~~Daemon Lib
~~~~Utils Lib
- Profiler specific code

-- Transport Protos --
-- Contains all protos previously defined in profilers.

Test: Existing test (this change moves files to a new TLD)
Bug: 122659360
Change-Id: Ib2740f287d834c6eda178a5376046ddfc293ee8e

Former-commit-id: 44a6a8bc7420193c8185f4a1e8a3daf6a09780d9