Don't run cleanup tasks if there aren't any.
Running gradle with no tasks runs the :help task by default
which requires configuring the project. On the large android
test project this can take 1-2 seconds. Multiply this by 6 warmups
and 10 iterations and it can take some amount of time (especially
with 20+ benchmarks).

So just run the cleanup tasks if there are any tasks.

Bug: N/A
Change-Id: I777ad8dbf8990ceceb663eb80778a760f5da40b7
Test: N/A

Former-commit-id: 035ad385863531c0731111ca07a3dd1fae79d1f9