Add option to enforce unique package name.
Adds a feature on ManifestMerger2 to enforce that all libraries being
merged, as part of the application manifest merging, have unique package
names.

Also add a BooleanOption, android.uniquePackageNames, to turn that on.
This option is currently false by default but will switch to be on in a
subsequent release and then removed so the enforcing will be permanent.

We need the option because since we regressed the check (somewhere
between 3.1.x and 3.2.0), usage of this become widespread including in
android support libraries, for example:
Package name 'android.support.graphics.drawable' used in: com.android.support:animated-vector-drawable:28.0.0, com.android.support:support-vector-drawable:28.0.0.

We need to at least clean up first the official libraries before
enabling this flag by default.

Bug: 116109681
Test: existing + local + ManifestMerger2Test:95 and 96
Change-Id: I4d099cb44c9a701f234d5307f1667d8f468e5079

Former-commit-id: b75bbb17e99ad9d293cb14e41853d1b9b389d22c