Make AndroidUnitTest fully cacheable (tools/base).
When includeAndroidResources=false, the AndroidUnitTest task is already
cacheable (done in a previous CL).

When includeAndroidResources=true, the AGP generates a test config file
containing absolute paths, which breaks cache relocatability. Changing
these paths to relative, however, would break existing users of this
property including Robolectric.

Therefore, we introduce a flag called
    android.testConfig.useRelativePath,
which currently defaults to false. When the user sets it to true, the
test config file will contain all relative paths, which allows the
AndroidUnitTest task to be fully cacheable.

Once Robolectric and others have migrated to use relative paths, we will
default this flag to true, and eventually remove it. We might enforce a
minimum version of Robolectric at that point.

As part of this work, we've also made the GenerateTestConfig and
PackageForUnitTest tasks cacheable.

Change-Id: I4b7089af0b1ebe66fca1d7aa0d190730d6c718da
Fixes: 115873047
Fixes: 123343127
Fixes: 123578979
Test: Existing tests + Updated CacheabilityTest

Former-commit-id: e1d267e226d7c69cebe8f53a4df519ab658a1a67