Automatically rename Benchmarks with host-specific names
This reverts commit faed664e46cd6fc1c477aecfa6ed4a5734bd45c5 [formerly 520e390886acc8f0f546d3818906834122297618].

This will keep Benchmarks run on Mac and Windows from colliding with the
"default" Benchmarks, which are run on Linux. Changing the name of the
Benchmark object is sufficient to force data generated on Mac or Windows
hosts from appearing in the charts which already contain Linux
performance data.

In addition, this change updates Metric.java and MetricTest.java to
close some input/output streams that were previously being left open,
causing MetricTest to fail on Windows.

Bug: b/122838052
Test: Updated MetricTest and PerfDataTest
Test: Ran test locally and looked at perf_data.json
Change-Id: I7dc6646befcb6552cf3fa8c0f8618cb3a785326e

Former-commit-id: 1a30e315eafdd76349e54428629f210ceeb3e6ce