IRR: Allow device commands retry with InputStream
Problem: Device command invocation have a built-in retry mecanism
to accomodate the installer auto-update system. Upon running a
command on the device, if the versions expected and the version
installed do not match, the lastest version of the installer is
pushed to the device and the command is retried once.

This system worked well when all parameters to a command are on
the command-line but breaks when an inputstream such a protobuffer
is used. With the inputstream exhausted, the command cannot be re-
run.

Solution: Make the invoker take a resetable ByteArrayInputStream
instead of an InputStream. Upon retry, reset the stream.

Bug: None
Test: None
Change-Id: Ic816a6ee3185161c4a93b6f78257c4e73f8be020

Former-commit-id: b508a0bdd401d6674c452986d3be646151438248