Close input stream in MetricTest and rename confusing variable
Renamed outputFile to jsonFile because "InputStream outputStream" was
a little confusing. Also closed an input stream I forgot to close in a
previous cleanup CL, copying the syntax used in the other tests already
in the file.

I also ran google-java-format manually, which caught some stuff that the
presubmit hook wasn't catching.

Bug: N/A
Test: MetricTest still passes after changes
Change-Id: I087e875f89b1b1c5def796dd3533824ace4ae296

Former-commit-id: 872804e0ed0d0313aeacf58924ba59d2b1cda262