Normalize before hash table insert/lookup.
So far we have not encountered a need to normalize before
inserting into the hash table, because the compilation database seems to
always provide normalized paths. However, this CL normalizes before
inserting anyways, just to be safe.

This CL also makes sure we don't get empty string after normalization,
and uses non-normalized strings if that's the case.

Bug: 123123307
Test: Added new unit test to cover paths with ../ in it.

Change-Id: Ib6fb48e106a4e1cdb86171411c4f132a33dd795d
Signed-off-by: Emre Kultursay <emrekultursay@google.com>

Former-commit-id: c9afd4db23acf1a3a7983a0176d62b907fa38e36