Optimize disk space utilisation during SDK package installation.
This CL changes the install operations preparation
logic to download and unzip packages into SDK_ROOT/.temp
as opposed to java.io.tmpdir. This will optimize disk
space utilisation whenever system temp partition and
SDK root partition are different (which is the case
on most Unix systems, for instance), since instead
of requiring the space for unzipped package to be
available both on temp and SDK partitions, we will
only require it to be available on the SDK one.

This makes sense, since ultimately the only place
we really require that space is the SDK root, as
this is where the package gets installed anyway.

Apart from that, this CL ensures that the directories
containing incompletely unzipped package contents
are removed as part of the operation cleanup along with
other things considered temporary/inconsistent.

Bug: 37092608 et al.
Test: Added relevant checks to BasicInstallerTest & adjusted
AbstractInstallerTest.
Change-Id: Id02e09d78875089e9ba02a888ce704d85db5faec

Former-commit-id: d734f9db6de4742222e48770ffff4a13e5c3277a