Record UsbDevice serial numbers on Mac
The serial numbers were already present in the data being passed to
MacParser. This patch adds them to the UsbDevice struct. The Mac
appears to also fill in the serial number with all-zeroes if it isn't
present. This patch checks for that case and replaces it with a null
in the UsbDevice struct.

Test: MacParserTest
Change-Id: Id55738dc1a4d6e82d9f4de5f22d7767b05ca122a
Fixes: 123886676

Former-commit-id: 94116586a92b35a1650a88aeb96e13149e411006