Fix potential exception on windows
This patch ports the fix for bug 122955511 to windows. If the windows
command reports any unexpected output on the first few lines, it would
cause an exception to be thrown and the connection assistant to fail.

This makes the error handling more robust and will cause lines to be
ignored until the parser finds the markers it is looking for.

Fixes: 124070991
Test: WindowsParserTest
Change-Id: Ic2b62a93431583f6e5d6c991b95e982b5f21d1d7

Former-commit-id: d39958e6f30029f0ac55f247c58c5e76f4214795