Add config for perfetto protos
Perfetto has a set process to combine protos into a
single proto file. We use this proto file to build
the perfetto config initializing a capture. As well
as deserializing the trace after a capture is
completed.

The change to proto.bzl allows the perfetto repo to
access the proto helper functions that live outside
the perfetto workspace.

Test: proto only.
Bug: 124038138
Change-Id: Ia77a543ed0f90f09ae0959236823b0d1bf15513b

Former-commit-id: ba5c82943128264aa923038cb9170fec35ea06cc