Adds first end to end tests for deployment
Using fake adb, this change sets up the first steps
to a virtual library of devices. For now the devices
only support the minimum necessary for a basic adb install
support.
It is now possible to fake all the installation errors that
can occur when installing. For example this CL exercises the
case where on API 24, install-commit does not return
"Success" but just nothing.

We now run the deployment pipeline on the basic case on them.

Bug: 122475948
Bug: 122907314
Test: this is it.
Change-Id: I2bbf0f98ab7a562082b3057efeaec8a287fed636

Former-commit-id: 79e0823558fe53993f9375ae20801330b2af7858