Include process abi information in Session's metadata
This information is needed by the studio profilers to fetch
the correct symbols from the project based on the currently
selected session. This is need by both CPU recording code
navaigation and JNI ref callstack resolution.

Previously, there was a subtle bug where CPU code navigation
uses the ABI of the currently selected process. If the user
selects an older session, then the selected process information
is invalid.

This change is also needed so on Studio side, we can completely
remove dependency on the project in the ProfilerService.

Bug: 79772836
Bug: 124526769
Test: Studio-side tests added

Change-Id: I6babd969cc64979ebc1c578eef54e2f0b7de0e7a

Former-commit-id: cac6968b4d84475cef1560760a73abcab4c3590d