Remove usage of aapt2 from BuildTools.
We should use aapt2 from maven (or from a custom override,
see 'android.aapt2FromMavenOverride'), since we can have a faster update
cycle with it than with the one bundled with the BuildTools release.

Some of the code was already using from maven only, this CL just change
the rest to also use it only and don't consider the buildtool only as
fallback.

This CL also change the tasks inputs so the aapt2 from Maven is not
optional anymore.

Bug: n/a
Test: existing
Change-Id: I769b5124c45c9d9bb5434d5b23e54ef67eeed1f0

Former-commit-id: f99f21a2df887f201499b5505db6421f8f38eada