IRR: Only return NO_CHANGES when nothing changed
Problem: The current implementation can return a partial patch when
only a few splits have changed in an application. This results in
only the changed splits to be installed and the unchanged to be
completely discarded.

Solution: Change the patch generator to return NO_CHANGES only when
none of the apks in an app have changed.

Test: assert
Bug: None
Change-Id: I44a52099b1a68bce84e9d56865f87fd4c5608640

Former-commit-id: 83ea022297606bbdd4e91f539300d7f23e0640b8