Jetifier: Remove signatures from signed libraries.
Jetifier cannot rewrite a signed library (having files ending with *.SF,
*.DSA, *.RSA, etc.), since the purpose of the signatures is to prevent a
library from being rewritten.

Therefore, we need to strip the signatures from a library (if it
contains references to old support libraries). If a library doesn't
contain references to old support libraries, then Jetifier will not
modify their contents including their signatures.

Stripping the signatures is harmless because the signatures will not
contribute to the build's output, and if developers want to verify the
libraries, they should do so independently of the AGP.

----------
This CL also:
  - Moves the dependency substition part in JetifyTransform to the new
    AndroidXDependencySubstitution class for clarity since they are
    conceptually different.
  - Sets the Jetifier blacklist option as an input of JetifyTransform
    (by passing it to the constructor) so that the transform can
    rerun when the blacklist changes (this fixes Bug 122717256).

Bug: 115556774
Bug: 118785891
Bug: 122717256
Test: Updated JetifierTest
Change-Id: Ic940f017a8bf933d012dae5d5e6259ef94c61ebd

Former-commit-id: 0ace07355ae565fa0fdca3c1606445cac3ce1587