Only add javac output as MergeJavaResourceTask input if there are APs
This CL is a repeat of previously reverted http://ag/6306741

We only need to add javac output as input to MergeJavaResourceTask if
javac generates java resources, and that currently only happens if
there are annotation processors.

We check for annotation processors by checking for dependencies of the
annotationProcessor configuration, and add javac output to the
projectJavaRes input if we find such a dependency. This must be done via
a lazy file collection in case a dependency is added via
Configuration.defaultDependencies(), for example.

Bug: NA
Test: MergeJavaResourceTaskTest
Change-Id: I117ddfa4dc651f63904309c068fa2217a095f4dd

Former-commit-id: 07368f1e5e4e2d9554d5c2277f348ed610a67d4e