140154274: Allow suppressing warnings on catch clauses
As described in 140154274, you couldn't suppress the API
warning on exception class types.

This was broken in Kotlin try/catch blocks; here are the tracking
UAST bugs:
https://youtrack.jetbrains.com/issue/KT-35804
https://youtrack.jetbrains.com/issue/KT-35803

This CL adds a workaround for this, as well as new UAST tests.
It also adds various additional other UAST regression tests,
and a few other regression tests as well such as for issue
122928037.

Fixes: 140154274
Bug: 126439418
Test: Unit test included

Change-Id: I0cc160f5353eb6ce486b2847f5ffb7d5ca19e698

Former-commit-id: b4b2041cde212c15a086e7a5505c127f9bf2b778