IRR: Report deltaInstall issue error -> info
Problem: Issue occuring during the delta install phase of install
is reported as error which is then bubbled up all the way to our
crash reports.

Solution: Log error as info(). This is not a problem to do so since
the issue is recoverable and these kind of error are tracked by our
metric system.

Bug: 126760749
Test: None
Change-Id: I6325d1b162ec67364de70fa5f5149191e078d9a3

Former-commit-id: f87c8fc320dcf0afc53afe5e7c9278e3dac46a90