NDK side-by-side: Move NdkHandler under SdkComponents
This moves NdkHandler under SDK components and makes it
delay-constructed.

After this CL, a non-C++ project that doesn't have
shaders and doesn't need to strip any external .so
files will not construct an NdkHandler.

Creates one source of truth about the location of the
user's selected NDK.

see go/ndk-sxs

Bug: 80478236
Test: Refactor
Change-Id: Ib4a3802b5b1b048aaecbd7aef5cfd65eca32b795

Former-commit-id: ba95688e51407c005eb3b80eedeeb4e746fcf0b2