Fixed an issue in MergingLog where it reported incorrect file
Fixed an issue in MergingLog where it pointed to an incorrect file
although the contents of the map were correct, this was due to some
times getting the ceiling entry returns a range that starts after the
query. This change adds a check that the candidate file position is not
starting after the query and the query is completely within the
candidate range.

Test: Updated MergingLogTest
Fixes: 128525861
Change-Id: I7dda36217e897f839682529f6cb4ed19a7e4880d

Former-commit-id: 702be35b8f1de8ad511c49f42d41d1054ee75586