Add stricter regression analyzer for TOTAL_BUILD_TIME
TOTAL_BUILD_TIME is much less volatile than some of the task times, but
previously we used the same regression analyzer for most tasks and the
total build time. This CL creates a stricter regression analyzer which
flags 5% regressions in TOTAL_BUILD_TIME.

This CL also increases the window sizes for all AGP analyzers from 25
runs to 50 runs, in an attempt to reduce the false positive rate.

Bug: N/A
Test: Existing
Change-Id: I0f294e3a0be18626733481bd04709bc2b7678e4c

Former-commit-id: 7a310a50a7ed7a5eb7a6222c6c0c8680f739668d