Attempt to run emulator-dependent tests on AB again
The last attempt to run sanity UI tests that use the emulator on
the continuous integration system known as AB led to builds failing
because Bazel exits with status code 4 if Bazel fails to find any tests
when a Bazel test command is given. This was not a case considered
because the only cases encountered were when Bazel exited with 0 or 3.

For reference, exit code 3 indicates that some tests failed or timed out
when Bazel was running them. We do not want to fail builds when the
sanity tests fail or time out, since reviews of test results are carried
out manually. We also do not want to have our build target disabled by
showing too many failing builds.

The previous commit that attempted to get the emulator-dependent tests
running was change I7af729cdd3dabe2913646f4f378fea212fe13090.

This commit covers cases where the exit code can be 0, 3, or 4. If all
Bazel test commands exited with one of these 3 exit codes, then we exit
with exit code 0 in the script. Otherewise, we exit with the same exit
code that Bazel gave us

Bug: 123900444
Test: Ran script locally
Change-Id: Ifd1c35967bee6e857a5df533953816127faf52a3

Former-commit-id: aab4e8bcca7d829c147e4d4968b0682b2887e5a3