Delete perfgate ZIP files after adding to archive
Windows builders are currently encountering an issue when calling the
bazel clean command. Because this command is failing, we are seeing
stale data get zipped into our perfgate uploads. Until the bug is fixed,
we will delete the perfgate data once added to the archive to ensure it
is properly removed.

Bug tracking the Bazel issue it at: b/128869827
Test: PSQ
Bug: NA
Change-Id: I5721f5d35853a24e63d70fa8d1ce314035085ae7

Former-commit-id: cf07b08e68d6f9c90fe65028add70598bd172588