Migrate packaging to use new gradle API.
This avoids the need for the packaging task to track which input the
changed file comes from which input, as Gradle gives us that information
directly.

RelativeFile is updated not to need the 'base' file, as all use cases
constructed the actual file from the base and relative path anyway.

This also changes the split under studio injection semantic to not keep
the other apks (see update to InjectedAbiTest).

Still to do:
 (1) The merge java transform can also be migrated.
 (2) Finish cleanup and removal of infrastructure after (1).

Fixes: 74595859
Bug: 128912423
Bug: 72085541
Test: Existing gradle integation tests.
Change-Id: Ib469b2562358622d66e30df0f5700997dc67ac64

Former-commit-id: 664cc96d49703e3caa242628f641cd8a7fa15188