Do not create lintVital task for any variants in test modules
Currently, if a buildType is added to a test module with
debuggable=false, the TaskManager attempts to create a lintVital task
for that variant. This fails because lintVital requires a
MANIFEST_MERGE_REPORT artifact, which is not created for test modules.
No other Lint tasks are created by test modules, so it seems right to
just stop trying to create this task as well.

I also changed the implementation of TaskManager.isLintVariant() to
check VariantType.isForTesting() rather than
VariantType.isTestComponent() because that seems more semantic.

Bug: b/127843764
Test: Added test to SeparateTestModuleTest
Change-Id: I16e4abf1ca68ea3507ae4b5854f0739051aa2c4c

Former-commit-id: 4a46a4ef611389b63447409d8c12c308bace869a