Enable exporting an aspects agent baseline file
Add the possibility of exporting a baseline file through a JVM flag.
aspects.baseline.export.path can be used to determine the output file
resulting from running UI tests (or Android Studio) with the agent
enabled. When doing that, instead of throwing an exception when a rule
is violated, we add it to the baseline. When the JVM shuts down, we
export the baseline to the path provided using the flag.

Bug: 128497963
Test: BaselineTest and AssertionsTest
Change-Id: I1072c6836ebc160b1b82da7b2171071f94de151f

Former-commit-id: dd75b5a8899e73ec14014a1f77e4c14c9a57bdd1