Revert to outputting absolute paths in MergeResources task.
When data binding is used, it used to write absolute paths to a log
file so the IDE can read those paths and navigate to the files (e.g.,
for error reporting).

In a previous CL, we tried to pass relative paths to the log file, so
the MergeResources task can support cache relocatability (bug
121288180).

However, we ran into 2 issues:
  1. The paths to some resources can't be relativized (bug 128579779).
  2. The relative paths need to be relative to the root project
     directory, not the current subproject directory, as the IDE
     searches relative paths from the root project directory (bug
     128643036).

To temporarily fix the above 2 issues, we need to revert to outputting
absolute paths in this CL, and will come up with a proper solution
later (tracked at bug 129276057).

Fixes: 128579779
Fixes: 128643036
Bug: 121288180 (re-introduced)
Bug: 129276057 (to track the proper fix)
Test: See other CL in the same topic (in tools/data-binding)
Change-Id: I8790fc6de84d6f8fb250530b67442435063883e5

Former-commit-id: 1fb5e983555a64860accacb3fe303227af0d99ed