Rename Enum::getName to Enum::getTag to avoid confusion from Kotlin
Kotlin has it's own synthetic 'name' that it adds to Enums. When
called from Kotlin you naturally expect Enum::name to map to getName
but it doesn't, it map's to Kotlin's synthetic function which returns
the enum name like "ARMEABI" rather than the value that is supposed
to be returned, "armeabi".

Bug: NA
Test: NA
Change-Id: Iaf13750e5322ff6c1d44c0c7b8050e1ddce5cd07

Former-commit-id: 17c5c45aa2ecb94404c49ae801a609ddee4f6876