Annotate all method on com.intellij.openapi.application.Application
This was done using the "annotate method" intention, which keeps the XML
file in certain order (that I don't completely understand).

To use "annotate method", you need to first enable "use external
annotations" in the project's code style. Then choose the annotate the
method externally with some other annotation and then edit the XML to
use our annotation. This will be the flow for now, until we can add a
custom AnnotationProvider to Studio Dev Support plugin.

Bug: None
Test: N/A
Change-Id: I4e61fe8dc2af759da8d124876c5a63309a594175

Former-commit-id: 036dc98a207d44f5ca70bb15f65dff474594bdb8