Use nanosec timestamp as CPU trace Id
Instead of having perfd responsible for generating an unique id
for each trace, simply use the current time. The goal is to simplify
the communication needed between the daemon and agent, which previously
requires perfd to send back the generated trace id to the app. In the
new pipeline, that direction of communciation is not readily available
anymore.

Bug: 119261457
Test: Existing

Change-Id: I9483ea2b94b6af92b8b4266e4074557d92e43092

Former-commit-id: 75fe226a469c0e4a83988c2c0d8d88cab41a2d05