Don't compress bundle modules.
The .zip files produced by PerModuleBundleTask for each module
are only read by BundleTool and in order to do that it needs to
decompress which is a waste of CPU time.

We use setLevel(0) instead of doing setMethod(STORED), for two reasons:
1) setMethod only make the entries which don't specify a method use
STORED.
2) Using STORED method requires that entries contains file size and
CRC32. To compute them, we would require an additional pass over the
files, since we need to have those values before passing the entry to
the backing [Jar|Zip]OutputStream.

Bug: 118601166
Test: JarMergerTest
Change-Id: If20f8bd6bfcc808134b48895035618262de3195b

Former-commit-id: 2aee43146cfe939e8c1d90ee8da76b95b4b5b7ae