Produce sync error when useAndroidX property is missing.
The android.useAndroidX property indicates whether the user wants to use
AndroidX dependencies or not. It allows the Android Gradle plugin, data
binding, and several IDE tools to add dependencies automatically and
provide proper support.

Typically, this value is set automatically when the user creates a new
project with AndroidX, or migrates an existing project using the Migrate
to AndroidX tool in the IDE.

However, if the user adds AndroidX dependencies manually to an existing
project, they may forget to set the property or may simply be unaware of
the property.

If that happens, we'll need to issue a sync error for the user to fix it
manually or preferably using the IDE quick fix.

This CL should be accompanied by another CL that provides the IDE quick
fix for better user experience.

Bug: 130286699
Test: AndroidXPropertySyncIssueTest
Change-Id: Ida643d7d225ebf1a065541a0656e36703fb8d0df

Former-commit-id: 160148399d3dacf98c14f305d4ae1228fe8a94da