Add more UAST resolve tests
The first covers resolution for Kotlin -> Java source references,
which required adding a test fixture correctly registering the
top-level directory as a Java source root (otherwise the Kotlin
compiler will not resolve those references). This can be useful
for writing tests mixing Java and Kotlin sources.

The second test covers property access, which is currently left
unresolved in UAST (even though kotlin-reflect is included on the
compilation classpath and the binding trace contains the resolved
calls).

Bug: N/A
Test: this

Change-Id: I965a547ce5bcc5d108732101b8cde3e417c6ad52

Former-commit-id: b0173ba757763a5807558927e5933d664cde4fdb