Backport support for callable references to ThreadDetector
We made it work in the internal check used for IntelliJ threading, but
not in the "upstream" version. This change brings the two closer
together.

I also rename the internal check to IntellijThreadDetector, to avoid
going insane when switching between the two.

I didn't try to extract a common superclass for them for two reasons.
One is that such a superclass would contain big parts of the checks'
semantics, and lint-api is probably not a good place for opinionated
helpers like this. The other reason is that we may want to change the
behavior of the internal check without worrying about breaking the real
lint.

Bug: None, follow-up from change Ib390847573b31ab7bdb623c3ea08fd0824f5e3c3
Test: Existing tests for both checks, added new cases in ThreadDetectorTest
Change-Id: I920733bd03a519020c686f87ed48984fcf128057

Former-commit-id: 477c4ea5437908ef37b906f4db385df111a80d74