Consolidate directory deleting logic.
We used two methods to delete directory but the straight forward
way is not good enough as some files can be held by other processes
or threads a little longer than expected.

Adapted the retry logic to add a small Thread.sleep which seems to force
JDK to either flush or re-read state I/O data. With this adapted retry
logic, I am able to consistently run all integration tests with Gradle or
with Bazel.

Test: Existing
Fixes: 131623810

Change-Id: Ic239e57def683e9621aa43163ba755c04c92c8ce
Change-IdChange-Id: I61dd384ea694e0960ff423072d8eb54ef9e392b4

Former-commit-id: eb77ce64f7eb43b69bd0f97d269698be5583a8bd