Dalvik error ReflectiveOperationException
This CL modifies the API check such that when you have
a catch clause exception which requires a higher API
level, and the minSdkVersion is less than Kitkat (e.g.
there's a possibility the code will run on Dalvik),
then complain about the potential crash even if the
try/catch clause is within a version check or in a
method annotated with @TargetApi or @RequiresApi.

This code will crash on Dalvik, even if the relevant
code is not run at all (the crash happens during
classloading). This CL also adds an error message to
explain this.

Fixes: 131843501: @TargetApi on a method should not suppress NewApi in a catch
See: 131349148: Dalvik: java.lang.VerifyError
Test: Unit test included
Change-Id: I9b228f5018c8b755a9f50d373da6e4e67729d5b8

Former-commit-id: b43de9331a5eaeeb726ab67bb42b926f46714857