Fix NPE in case of multiple classloaders
This change remove dependency on BuildSession, which uses reflection
magic to create a singleton across different classloaders. This is
problematic because we rely on the state of this JVM-scoped singleton
when writing to classloader-scoped static variables.

This change instead registers classloader-scoped listeners directly on
the root Gradle project. That is, if different classloaders are
involved, there would be different CxxBuildModel objects, which is fine
for our purpose.

Bug: 131419191
Test: manual
Change-Id: I5e7b92c2441bf0af1dd8e01d1b40fdb862d5831b

Former-commit-id: 51087dc206dcfac40ca700d5367e4b8eac0f4a27