Use caller-defined paths and standard byte cache to store cpu trace contents
- Caller is now responsible for providing the path that trace files should be
written to. This prevents a special path that startup profiling had to take
to obtain the file path from the daemon and pass that in to the launch argument.
- After stopping a trace, the content stored in the path is copied to the generic
pipeline's byte cache (referenced  by the trace's id), which provides a standard way
for the CPU profiler to fetch the content when needed.

Bug: 119261457
Test: Existing

Change-Id: Id53ff6b39c5aceb3394210571d74a770f9282722

Former-commit-id: d581dd47f1968f42f42d7b503b816f50e59de039