Enable android.testConfig.useRelativePath=true by default.
Three months ago, we introduced a flag called
    android.testConfig.useRelativePath,
disabled by default. When the user enables it, the test config file that
the AGP generates for unit tests will contain all relative paths, which
will allow the AndroidUnitTest task to be fully cacheable.

This feature seems safe and we haven't received any bug reports for
this, so let's enable it by default now.

Bug: 115873047
Test: Existing integration test + Updated CacheabilityTest
Change-Id: I2402239840753b0624bfbd4847c3c2b3f254af95

Former-commit-id: e48c3831b9a9dd4829d63ba80fddc46ba7e67e61