Fix MergeJavaResourceTask regression
Use PathSensitivity.ABSOLUTE for the MergeJavaResourceTask and
MergeNativeLibsTask project inputs, until we get custom snapshotting
from gradle.

This CL is a better solution than http://ag/7246394 because it doesn't
add time to the UP-TO-DATE checks and it still allows incremental builds
when files are added or removed from the project java resources or
native libs.

Fixes: 132521634
Test: JavaResPackagingTest, NativeSoPackagingTest added in http://ag/7246394
Change-Id: I029055ae123003e901571c9e9460ba4379bd37ab

Former-commit-id: 918331f9dc93fa06870ee6a416b27b85ebc9c297