Update agent to always close the socket
This fix causes the socket always be closed by the agent at the end of a
swap. Currently, the socket may remain open if the swap fails due to an
instrumentation failure, or if the swap request fails to be read/parsed.

This does NOT add error codes/responses for these cases.

The test case will time out indefinitely on the unfixed version. Note
that we only test the "unparseable swap request" case, but the failure
path is identical for all three cases.

Bug: None
Test: AgentBasedClassRedefinerFailureTest#testUnparseableSwapRequest
Change-Id: Ia23cdc1dc39c36c4b3153ab1889f750d8a6bb136

Former-commit-id: f96cb641b862ae4fa11eab092799dc3deb060bbd