Handle library project dex files from jars when merging
If there are a few library project dex files, instead of merging
them, we just copy them to the APK. This is to help incremental
builds. However, when artifact transforms are not used, these dex
files are in jars. Previously, we would end up copying jars to APK,
which resulted in ClassNotFoundException at runtime.

This CL fixes that, and we will extract these files from the jar.

BUG: 130965921
Test: DexingArtifactTransformMultiModuleTest
Change-Id: I235c5acdde816a46e1ea429a6b98ad43bd2f8b6d

Former-commit-id: 98298bc45deeb8e18c82ae187c28056b3f5918d5