Add incremental desugaring for nest-based access
Starting with javac in JDK 11, there are no synthetic methods
in the outer class to be used by nested class, but "NestMembers"
and "NestHost" attributes from the class file are used. This means
that D8 has to desugar these, whenever outer or nested class change.

This CL adds support to incremental desugaring to handle
outer-nested relationship, and to create bi-directional dependency
between outer and nested type.

BUG: 130578767
Test: DesugaringClassAnalyzerTest
Change-Id: I701417e89e7001c76b63dc40b12bc464b2e1c20a

Former-commit-id: 73c814982d77c617184d393274189986a1132d10