Force perfetto launching to wait for each process before starting next
With startup profiling lots of things are happening at the same time.
As such the timing of starting the perfetto processes is often out of
order. This change blocks the start up of each perfetto process until
the process updates the /proc/[pid]/cmdline. This is used as a way to
validate the process is running. If this fails an error is returned.

Test: No detailed test for specific process launching code
Bug: 131412010
Change-Id: I32a0156fd8afb8356a2b847a981563085d263c16

Former-commit-id: 7d97d1f9414b365c06b57a0b3aa8e4ac4ccfa559