Throw InvalidUserDataException instead of errorln for missing NDK
This exception seems to be required for Android Studio to respond with
error handlinger which are needed to supply hyperlinks to update the
ndkVersion in build.gradle.

In a followup CL I need to evaluate whether all messages from NdkLocator
should throw the same exception.

Also, I found that many unittests have been disabled. I didn't reenable
them in this CL since I need the cherry-pick to be as smooth as possible.

Bug: 133186158
Test: Manual-tested only using studio-3.6-release branch.
Change-Id: I53d336eb430a28cdc564b576bb139dbd1a07e253

Former-commit-id: 5d43379a9628330d5ebd5853611f3877337c0ce6