Log baseline stack traces that are currently being called
Write the stack traces from the aspects agent baseline that are
currently being called when running the UI tests to a log file. That's
useful to compare the content of this file to the actual baseline after
running all the UI tests. Stack traces that were already fixed should be
removed, so we end up with a shorter whitelist whenever is possible.

Test: BaselineFileTest#logActiveBaselineEntries
Bug: 133301342
Change-Id: Ia5ffe26e8c9c7cb3f2ce8c49676b260880019094

Former-commit-id: e11af71851b20876954dc97386e2811425ea92cd