Conver dex builder transform to a task
This CL converts dex builder transform to a task. This task
still needs to consume artifacts from the Transform API, but it
does not output them.

This also means that dex merging task can now rely on providers
for consumption of correct scopes it needs to process (project,
subprojects, or external Maven libs).

BUG: 136244270
Test: updated existing tests
Change-Id: Ib87791af1082f188e109c893e70937c2d1f7b153

Former-commit-id: 5d42ac0db2ba163272ea8dc51f17f711dd52a7a8