Remove explicit checkAppProfilingState rpc
The information about whether a trace is in progress can be derived
directly from the list of trace infos. Removing the redundant rpc
to avoid having to maintain extra states and code.

The CpuTraceInfo has been modified to return the entire configuration
message that was used to capture a trace (which was previously returned
via the ProfilingStateResponse).

Bug: 119261457
Test: Updated studio tests
Change-Id: I21dba61fec13867723f17a2b6676e8e978baf186

Former-commit-id: a0cfd4355f52f2de1113899d73641116da2a333d