Added jar artifacts as inputs to Lint
There is a deadlock that happens from time to time running lint on a
multi module project, :lib:lint tries to resolve jar artifacts and
:app:lint waits for :lib:lint to finish to enter a synchronized block
which doesn't happen.
This change adds those artifacts as inputs to make sure that they are
resolved before starting the task to avoid this issue.

Fixes: 132783036
Test: Verified manually. Ran LibMinifiyJarDepTest 2000 time successfully
while it usually takes only 100 times to produce a timeout.

Change-Id: I7c83d77fbaf7e3cefdf8615bc204598a97a84c90

Former-commit-id: 9da4dbdef1418c3c370db00f853264c9248adf84