Disregard class annotations if they conflict with method's
When the agent intercepts a method annotated with an annotation
belonging to the same group as one of the class-level annotations, we
should disregard the one from the class.

This is useful for scenarios where we want to generalize all the methods
of a class apart from a few specific ones. For instance, we could
annotate a UI component with @UiThread, but annotate certain lightweight
methods with @AnyThread. The agent shouldn't resolve both annotations
for that method, specially because they can trigger contradictory
methods (e.g. assertEdt and assertNotEdt).

Bug: 133300743
Test: InstrumentTest#testConflictingAnnotations
Change-Id: Iac032951cfb20c0332018708e93d794ede1eea1d

Former-commit-id: 96ca4b617a69f10e613a197720e49fbf808d80f1