[aapt kotlin] add basic parsing
Added classes that are used for Resource types (in ResourceValues.kt).
These are used inside Resource Table to be flattened into the end result
of the compilation process. Added Source, ResourceName, and ResourceFile
which are used for both the flattening and Attribution of Resource
Types.

General parsing has been added in ResourceUtils.kt for Boolean, Null,
Empty, and Integer Types.

Bug: 132800341
Test: Tests added for both ResourceTypes (parsing Integer values) and
ResourceUtils (parsing bool, Null, and Empty)

Change-Id: Id021a400331d3b4e0a002c8cfbae3d26b854cac8

Former-commit-id: 4dfb139a72054f5ba073cd13d0072c1322371018