Create new debug agent
This agent will be used to implement async stacktraces
as well as Kotlin coroutine debugging support.

This is dead code for now, and is not yet bundled with Studio.

Some of the code is modeled after the agents already created for
profilers and Apply Changes. This includes most of the code in
log.cc and util.cc, as well as some of the code in agent.cc.
At the moment, code sharing would probably not be worth the added
dependency complexity.

Bug: 111444154
Test: manual for now (agent attaches successfully)
Change-Id: I1940fa25f2b27e3feecdb5fa1c211fd7485fe3f5

Former-commit-id: 01c676d42ad3fa2725cbb38333ad0b8ebc441c74