Add async stack instrumentation
This is dead code for now, but it's a good checkpoint before
making more progress.

A couple corner cases are still unimplemented, but those are
marked with 'TODO' comments, and they cause the instrumentation
to fail gracefully.

The instrumentation works successfully for the
built-in java.lang.Thread capture point.

Test: stress test added in a pending child commit.
Also tested manually by attaching the agent to a small application
and verifying that the correct hooks get called from java.lang.Thread,
even when an exception is thrown.

Bug: 111444154
Change-Id: If15ff925b24a390f08ab1f94083c3808da4710cf

Former-commit-id: f5fef2790ecace723ac8fb9d6e90beba2496183c