Task that analyzes the dependencies of a variant
The task finds all the direct dependencies that are not used in the variant and builds a dependency graph to find all transitive dependencies that need to be added when removing the unused ones.
It also looks at the API dependencies that can be configured as implementation (those that are only used in private fields/methods and will not be available to the library's consumers).

Bug: N/A
Test: -
Change-Id: I4e4fdf8fd710a7ff550c2d1f1c1012de041ffebd

Former-commit-id: 51c5b91ae7ee656a87ece40ff0da0d9783871119