Add navigation to BasicActivity by default
This adds Navigation Architecture Components to BasicActivity by
default with two Fragments.

The first Fragment has a button that navigates to the second Fragment.
The second Fragment has a button that navigates to the first Fragment.

The original CL was reverted because of b/136632498.
This is the same change as the original one but needs to be after the
bug is fixed.

Bug: 134539629
Test: TemplateTest
Change-Id: I19c32c4d8a857cce038ec914779e42a60f73777e

Former-commit-id: 388afe94a623e2532cd19ad03ae9d5d527c3d02d