Ignore lambdas visited by the aspects agent
Lambdas are internally converted into static methods named like
"lambda$containingMethod$1". Because of this conversion, we intercept
lambdas as if they're regular methods and trigger the aspects defined in
the rules file when we're not actually supposed to.

This CL changes our InterceptVisitor to ignore methods that are actually
a lambda conversion.

Test: InstrumentTest
Bug: N/A
Change-Id: I79e0538d6d00f43ad60d5485113e88c69aad308e

Former-commit-id: 6174e575c9eec0b5d740125978221cb0bb44ade8