Remove references to non-existant files and folders.
While moving tasks from BuilableArtifacts to Provider<>, I
used a non existent EMPTY_FILE, EMPTY_FOLDER to avoid returning
non set providers.

With this CL, we still return a provider, but it is not set to anything.

Changed the getFinalProductAsFileCollection to return a Provider<>
rather than the FileCollection directly so the FileCollection is not
resolved before execution time (unless get() is invoked earlier of
course).

Fixed coupld of issues in the process, espcially around main dex list
that were not published/consummed correctly.

Test: Existing.
Bug: N/A
Change-Id: Ica18ae56cea966d4a8549048fc05c03d5da310ff

Former-commit-id: 18db7498c4dd6a195603fc91423efe480af775f2