Change FeatureSplitTransitiveDepsWriterTask to not depend on CLASSES.
Using the CLASSES artifact to get the dependency graph adds an
unnecessary dependency on the task that generate this artifact.

Also, this works well right now because runtimeElements in APK
modules does not include transitive information, but this is
something we need to change to allow for multiple levels
of features, as well as supporting unit tests inside features.
Consuming the CLASSES artifact would lead to cyclic dependencies
because of the way these packaged dependencies file gets
consumed.

Bug: N/A
Test: existing tests.
Change-Id: I9457441ab39259a12a05ee154ab9bfdc72f7c32a

Former-commit-id: 8b01e580618c4242e6e36b011b05832661573a56