[aapt Kotlin] Added Xml Flattener.
In order to process all of the resource types from the xml table,
we need to have the ability of flattening xml resources in a kind
of easy to consume format.

That is, escaped characters are handled properly:

"Hello \\t World" -> "Hello \t World"

Whitespace is reduced similiar to html if not contained in quotes:

"Hello \t World" -> "Hello World"

Whitespace is preserved if in quotes ->

" \"Hello \t World\" " => "Hello \t World"

Spans and untranslatable sections (for psuedo localization) are
collected.

Bug: 132800341
Test: Added StringBuilderTest Class

Change-Id: Ie25bda20fb771621b230adff6e77871f1b13b142

Former-commit-id: 3c132d166747ba9a0dc1d9114a59c731034142e6