Use delayed computation of Java sources
Java compile task can be configured before Variant API
is invoked to add the generated sources. KAPT 1.3.40 behaves
in such way.

This CL wraps the source computation in a Callable, makes sure
sources are evaluated as late as possible, and fixes
regression introduced in ag/8023538. A regression test is
added as well.

BUG: 135780031
Test: RegisterExternalAptJavaTest
Change-Id: I4f55402230cead50e8e370846e1a59f7d6cea916

Former-commit-id: e98440d5d2d9040535922de565e064ed397eabc2