Pass android.aaptOptions.noCompress to other tasks
Previously, MergeJavaResourceTask and ProcessAndroidAppResourcesTask
ignored android.aaptOptions.noCompress. Those tasks would compress the
files, but they would be uncompressed in the packaging task. This works
fine for apkzlib, but zipflinger cannot create an uncompressed entry
from a compressed entry, so this CL changes the behavior of the tasks to
create uncompressed entries for the files corresponding to
android.aaptOptions.noCompress.

Bug: 135275558, 134585392
Test: NoCompressTest, IncrementalFileMergerTests,
MergeJavaResourcesDelegateTest
Change-Id: I8fae2da5e87d3300223471c537f94cc34641c561

Former-commit-id: eae56de5569ce08e1de1f82f61f6c8a76a05f769