[aapt Kotlin] Add Resource Table
The Resource Table is the primary class for aggregating resources in an
app. This groups resources in order of importance by:

package -> resource type -> resource name -> config -> product

It also manages the visibility, the <add-resource> overlay, the
overlayability of the resource, as well as whether or not the resource
was a weak reference (Ex. using an attr in a declare-styleable).

Test: Added a ResourceTableTest class with tests on visibility,
allowNew, Overlayablility, general resource collision, etc.
Bug: 132800341

Change-Id: I899b248cda4dee8a4a60a5d1e546c97ad6a52534

Former-commit-id: 6efe1bf07131002848a2185015a40a6d32bf080c