Fix commandline SDK problems
On windows the commandline was too long when many jars are on the
classpath. Switching to generating a jar with the classpath set in the
manifest and referencing only that from the wrapper script.
Also fixing the logic to find the sdk root in screenshot2 and
avdmanager.
Also on windows we need to not have "" around multiple arguments.

Test: QA testing :-(
Fixes: 135575661, 135575661
Change-Id: I9cd1ffef53bd11e28932cc0d36bbc36de35aa787

Former-commit-id: 219a0e9ec931fa4f343793640e26c554516b8205