Modify sanity run script to remove empty files
This change modifies the qa sanity script to remove the empty files
under directory testlogs. That way the target won't be disabled by AB.
AB builder would fail with 0 bytes files.

This change will also limit the number of tests build on the builder to
a total of 4 jobs at the same time.

Bug: N/A
Test: Manual Testing
Change-Id: Ia37714ea75700ef66197c00de7c89d1345862476

Former-commit-id: 4bfc199f97a9152cf7ad9077b381605012388344