Prevent calling initWith with the receiver.
this.initWith(this) can end up with calling
this.set(this) on Property<*> objects which leads
to stack overflow. This CL bails early in initWith if the arg
is this as a safety measure.

The real fix is making the property get the current value
of the other rather than dynamic link itself with the other
property. initWith is meant as a copy of values, not a
synchronization between build types.

Change-Id: Id6804f177b830b1b7bd1b53e855984d0f39f7d16
Fixes: 130102823
Test: added a new integ test: InitWithThisTest

Former-commit-id: 200165a0a324c38e359747ef4bcf9e7bebe36f91