Make AppPreBuildTask not depend on MANIFEST
We cannot depend on the Configuration objects directly
as we need to query for the manifest to find android
dependencies.
So instead, generate an input from the ArtifactCollection
and use that as an input.
This still make this task not depend on the task outputting
these manifest files which is a win.

Bug: N/A
Test: existing tests.
Change-Id: I18c0e006d9f199bf67483ce71610dbb930d1ebb1

Former-commit-id: 3c5d4dfacd3f39c94dbfd81166176a01106f522d