Zipflinger: Add semi-paranoid mode
Problem: A corrupted zip could potentially trigger zipflinger to
generate "mashed potatoes" in output.

Solution: Improve the mapper by retrieving not only the local
extra size but also the name size and compare it to the central
directory name. This gives a fairly good certaincy that the CD
offset is valid.

Since zipflinger was already retrieving two bytes for the extra,
there is no performance impact to retrieve the two bytes of the
name size (there are locate next to each others).

Bug: None
Test: None
Change-Id: Ia6e0ecedf860ee2c112d6387ac979864f6c135f2

Former-commit-id: 6c19f3da4fed4411f2da6c6370b9fe995f40f0b3