136492453: Remove FloatMath detector
Issue 136492453 discovered that when the unit tests are running on
the CI server the normal code path is not executed, because we don't
have an old compileSdkVersion there.

However, fixing this is pointless. FloatMath was deleted from the
SDK a long time ago (API level 22), so code that refers to it doesn't
compile at all unless compileSdkVersion < 23.

Instead, we'll just delete this lint check now.

Fixes: 136492453
Test: Existing unit tests
Change-Id: I581bac000dd51214f78f32d27698b5f042cc4102

Former-commit-id: 1e70758e1fb41f954ce8e1bb13471caa764a6128