Add new compile classpath R class flow.
This makes a few related source-level breaking changes behind the Gradle
project property `android.useCompileClasspathLibraryRClasses`, which is
currently off by default.

These changes are designed to reduce the amount of
quadratic-in-the-depth-of-the-dependency-graph work done when generating
library R classes and results in library R classes that are smaller for
typical projects and generating the R class depends on fewer things.

The changes have the following semantic changes:

 1. The Library R classes are no longer generated with unique IDs within
    that library, instead all IDs at compilation are 0.
 2. The non-namespaced R class in libraries contains resources from the
    compile classpath, rather than the runtime classpath.

Allowing these changes means that in the new flow, R classes are
generated by the producing side and exported as part of the API jar of
AARs and Android Library subprojects.

For AARs the compilation R class is generated directly from the R.txt in
the AarToClassTransform and for library subprojects the compilation R
class is included as part of the API jar for compilation published for
other projects to consume.

While it might be conceptually simpler to have library projects export
two different artifacts, one without the compilation R class for
consumers that generate constant IDs, this is in practice made difficult
by the inbuilt assumption of the AGP publishing code that there are not
multiple artifacts of the same type published from the same
configuration. Instead we rely on the compilation classpath ordering,
and  CompileRClassFlowTest includes a regression test for this.

Bug: 136244851
Test: Included new AGP integration test: CompileRClassFlowTest
Change-Id: I5dd972cf0e9ffffe3112cfe3b728c8f3d77cb611

Former-commit-id: 534e5bc09328e2655bd8481d9a3415671781b04c