Add command handling and event generation for allocation tracking
In the old pipeline, the daemon handles the AllocationsInfo generation for
both the pre-O (studio proxy) and O+ (perfa) tracking. Given that the app-agent
is the one handling the command(s), it is cleaner to have the app-agent generate
the AllocationsInfo messages instead. Therefore no daemon-side command handler
is added. The legacy allocation tracking pipeline in Studio proxy will be handled in
a follow-up.

Bug: 121392346
Test: new pipeline tests to be added.

Change-Id: I5081c8449b8c7c2195201b09728124bc1ae9aced

Former-commit-id: e96cbce22816a779710367575000389d0dad4b0e