[Aapt Kotlin] Add parsing Bag types.
Added the ability of TableResource to parse resources with child
resources (colloquially refered to as bag types). These refer to
plurals, arrays, declare-styleables, styles, etc.

In addition to adding the different parsing methods, which are unique to
each "bag" type. A few changes were added to ResourceTable to fix some
bugs that surfaced in testing, as well as added tests to cover said
bugs.

Test: Added testing methods to TableExtractorTest, as well as testing
methods to ResourceTableTest to fix bugs.
Bug: 132800341

Change-Id: I52e944b16d19aec7d387ca50be14b238c42dab2d

Former-commit-id: bbf74d2330cc70e0c95b4afa4dda4d756c549780