Fix task property warnings with SigningConfig.
In SigningConfigWriterTask, we write the entire SigningConfig object to
a json file, so all properties of the object should be marked as @Input.

When consuming the SigningConfig json file, tasks should annotate the
file with PathSensitivity.RELATIVE for cache relocability (and it can't
be PathSensitivity.NONE because the file name matters).

Bug: 135900510
Bug: 135509623
Test: ValidateTaskPropertiesTest
Change-Id: I672f6895e23328f71a0d3f8ad0f2d747af90115b

Former-commit-id: 0f82c1ebcd6c73a727630c369fd39612f88f7782