IRR: Fix PM crash caused by early pipe closure
Problem: Upon issuing a cmd command, the forked process STDERR
and STDIN pipe were returned to the caller. If they were not
needed (a.k.a the caller was not interested in the child
outputs) these fds were closed. These closed pipe resulted in
a SIGPIPE in the child process if it tried to write().

Solution: If we know that the child process stdout and stderr
will not be inspected by the parent, we can redirect the
child stdout and stderr fds to "/dev/null" which will act as
a sink for any write(2) from the child process.

Bug: 137795791
Test: None
Change-Id: I5d7469cedacc9124191f549e054c70e47b5cb00c

Former-commit-id: d12b3bcb95478b49b1fc91086ace929318b9e192