Make JavaCompile and tasks using ProcessExecutor serializable
JavaCompile configuration was capturing entire config action
causing it not to be serializable. This CL fixes that.

Also, tasks using GradleProcessExecutor are now creating it
only when necessary and they are not storing it in their fields.

Additionally, methods from GlobalScope that provided executors
have been removed, to prevent further regressions.

BUG: 137370122
Test: incoming
Change-Id: Iac922379f6a1048c24ff4050b63da628dc7458e2

Former-commit-id: cac7ac8d2dd8184c8bbb68f7c669d52ed7f61dfc