Fix some correctness issues in packaging.
Using the absolute paths of files as inputs is likely to have issues
with classpath normalization, as the absolute path is not considered.

This change uses the fact that classpath normalization preserves order
of roots, and so the cache keys are derived from the input and the index
of the root file.

ApkFlinger/ZipFlinger is stricter than apkzlib, so some cases that used
to work in a very brittle way now fail, as with zipflinger trying to add
an entry which already exists results in an error.

Fixes: 138348852
Test: expanded MinifyFeaturesTest, NoMappingTest to also test with ApkFlinger enabled.
Change-Id: If2791937bf14751796dc6a5e7a5d60449e51741c

Former-commit-id: 5c92fb4c09d64cfa84ae28754c22e78a7e087d1c