allow registration of Provider<> of artifacts.
Up to now, we only allowed Property<> of RegularFile
or Directory to be registered as Producers. This is
not always practical with Tasks inherited from other
plugins or stock Gradle tasks.

Allow registering a Provider<> will have the same
functionalities as with a Property<> except the
location of the file or directory will never be set
by the system but instead will need to be set by
the Task configuration block.

Adapted InternalArtifactType.AAR to rightfully be
a FILE and use the new API inherited from the Gradle Task
to register the provider.

Test: added unit tests
Bug: N/A
Change-Id: Iaf1fbf45dfc7f38e34d7f00feae26bc3ea831cf9

Former-commit-id: ba7e141c266e20b5935346467a877e7246336191