Don't write signing config info to disk in IDE builds.
We currently write the signing config to disk, originally to pass this
info from app to dynamic-feature (see bug 114626324), and later on also
between tasks within the app to make use of this mechansim.

However, the signing config contains sensitive data that shouldn't be
written to disk or at least not in plain text.

With this CL, during an IDE build, tasks can get this information
directly from the IDE-injected project properties, so we can avoid
writing it to disk.

For a command-line build, because the information is specific to a
variant, and global properties do not exist, we still need to write this
info to disk per variant. We will address this in a follow-up CL
(probably by encrypting the data).

Bug: 137210434
Test: Updated SigningTest
Change-Id: I9157e4640e52d5fd84f58330cbb86c387d8ed1c4

Former-commit-id: 4b9b87645771df919a8c867520df828b93bd2dcf