Add multi-threading to ApkFlinger
This CL changes ApkFlinger::writeFile to use multi-threading in order to
do compression in parallel.

Currently, SignalAndroid_apkFlinger_cleanBuild's PACKAGE_APPLICATION
time is ~500ms longer than SignalAndroid_cleanBuild:
https://screenshot.googleplex.com/RAgXW6ggLSU.png

Fabien profiled it and found that most of the extra time was because
compression is done serially in zipflinger while it's done in
parallel in apkzlib.

Testing this CL on my local machine,
SignalAndroid_apkFliner_cleanBuild's PACKAGE_APPLICATION time is about
~200ms *shorter* than SignalAndroid_cleanBuild.

Bug: 134585392
Test: existing
Change-Id: I1a27256c8283985740ffec06ee85c978092620e1

Former-commit-id: a88a61b9e606b4afc0a8c2779f256096a7728b12