Add testBuildType = "release" to Benchmark template
This change updates the template to set the testBuildType to "release"
which prevents benchmarks from pulling in debug artifacts that could
affect results.

The default build type is also set to "release" to make it easier to run
benchmarks out of the box from within Studio itself.

These two changes are hidden behind an AGP version check since isDefault
is only supported in AGP versions >= 3.6

Fixes: 139150022
Test: AddBenchmarkModuleTest
Change-Id: Ic5f53dbf3032fdaaccd040b64a7c53725f84208f

Former-commit-id: 463f38bb6ad0a5a45aa60508bc19aa24a552e924