New public plugin and move existing to internal.
All current plugin classes are considered public API
because of how Gradle allows finding plugins. Therefore
we need these classes to not change.

However, we also want to have plugin authors target gradle-api
instead of the 'gradle' artifact. This change forks the current
plugin classes into a new set of public class (name unchanged)
and the actual implementations as private, internal classes.

The new public plugins delegate to the internal plugins
by applying them as separate "internal" plugins. For now
the public plugins stay in gradle-core but we'll move them
to gradle-api at some point. This is currently limited by
the presence of getExtension on BasePlugin, both of which are
now deprecated.

Because our classes have no other public API this should not
break anything.

Bug: N/A
Change-Id: Iedda24f8e363e014a71a93a4fc471ac10a932e64
Test: existing tests have been updated.

Former-commit-id: ecdfaee5fbdfa69e82bb9266b6742d9c3db27880