Exit when gRPC server fails to start (cannot bind a port)
gRPC may failed to bind a port number to the grpc::Server
for the corresponding endpoint. Existing implementation
proceeded blindly, and this CL checks the error, and in
case of an error, it prints a better informed error
message and exits the process.

When the transport process exits, Studio is expected
to automatically re-launch the transport daemon.

Bug: 139150224
Test: N/A
Change-Id: I43cf74e6db41d070f8dff8771ee0424328569fc1

Former-commit-id: e2027134d3e5a6d1476a03461710e8b24ebbd6cc