Make the resource name of the navigation graph XML unique.
The resource name of the navigation needs to be unique across the app
even if the flavors are different.
b/137524683 isn't going to be fixed only by this CL. The package name
of the synthetic properties need to flavors into account.
E.g. If a new BasicActivity is created in a differenct flavor (as flavor1),
the package name should be kotlinx.android.synthetic.flavor1.activity_main
instead of kotlinx.android.synthetic.main.activity_main.
But that should be fixed fixed by a follow up CL.

Also fixes:
- The name of the XXXXArgs class in the SecondFragment.kt.ftl
- fragment_simple.xml.ftl should include <fragment /> even if it's not
  a new module

Bug: 137524683
Test: TemplateTest, UniqueParameterTest, ParameterTest
Change-Id: I8dd11f341f127ffb747be83440c0bdb64eee5762

Former-commit-id: b2c940f2e0bef009d63185719839ab1df45857f0