Don't create .json files for build benchmarks with no data
Currently, when a build benchmark has 0 iterations, we still create the
perf_data.json file, even though there's no data to upload to perfgate.
The creation of this file causes an (inactive) benchmark to be created
in perfgate.

This CL adds a check that iterations is > 0 before we create any .json
files for perfgate.

Fixes: 139495572
Test: bazel test //prebuilts/studio/buildbenchmarks:Base100_Java8_noop_singleWarmupTest
Change-Id: Ic286315db81bf600efc6736af297ccf3896ce4c0

Former-commit-id: 286d3035856eb7eac8f2d1d9bf6ff813b9d1a56c