139011783: Ability to suppress lint warnings on dependency artifacts
This CL attempts to include the Maven library name of
a dependency which references invalid packages. This makes
it possible to suppress these on the group and artifact ids
rather than the paths in the Gradle cache.

Since this changes an error message, it would potentially
make any existing baseline items no longer match and show
up as new errors. Therefore, this CL also updates the
baseline mechanism with an ability to match older and newer
error messages. This is then used to also help with issue
68802305 where a single error message included multiple
file names which meant that the message wasn't stable.

Fixes: 139011783
Fixes: 68802305
Test: Unit tests included
Change-Id: I23cdb79d2bc09d5b355987bc2a8b6a74ddbc8a93

Former-commit-id: aebb43312c5cc361508c451f8d8aa0659ce93ff1