Make STOP_CPU_TRACE command return immediately
When handling STOP_CPU_TRACE command, the daemon may
need to wait several seconds so the trace file is complete
and ready for use. This may block the UI thread in Studio
unnecessarily.

In the old pipeline, a CompletableFuture is used to
unblock the UI thread.

This CL is part of the effort that in the unified pipeline
all commands should return immediately.

Fixes: 139502114
Test: N/A
Change-Id: I387cff5824ca0bdbff75980ae7647fa99e555e55

Former-commit-id: 54c3f21b4986459d8e8f3d83fcad7d46c61accd5