Improve error message for missing manifest file.
The manifest file may or may not be required to exist depending on the
variant type.

Previously, the manifest parser always assumed that the manifest file
was not required, which was not correct.

This CL makes it explicit whether the manifest file is required or not,
and fails the build if the manifest file is required but does not exist.

Fixes: 139928154
Test: New test in DefaultManifestParserTest
Change-Id: I79e6ce9120f2c07c6f314169ec0b569f391e883b

Former-commit-id: 58abb2ccc2b71e37dde9ae2a2b1adbececd7ddd9