Warn when building AAR with local .aar file dependency
Change-ID Ib332aa5a1bb107341ed1c70806a303223a9ca0b8
caused builds to fail when building an AAR with a local .aar
file dependency. This CL changes the build failure to a
warning so that we don't completely break users without warning when
they update AGP. The plan is to turn this back into a build failure for
AGP 3.7.

Bug: 139948018
Test: LibWithLocalDepsTest
Change-Id: I07119cff87946fb6dd6ba7b622127028d67039c8

Former-commit-id: 7f61ff6d9d80a22322c9281da755a788fe854923