Allow spaces in merged manifest paths issues
When there are issues in nerged manifest files, the original path is
extracted from the manifest merge blame, but if the path contains spaces
it is reported incorrectly. The cause is that parsing this type of
patterns:

<line>-->[description] <path>:<position>

for example:

13-->[androidx.core:core:1.0.1] /usr/.gradle/caches/transforms-2/files-2.1/cb5e0295e6631df8cf1172ae152a4ad4/AndroidManifest.xml:22:18-22:86

is done by spliting by white spaces and taking the last element. This
works fine if the file path does not contain whitespaces but can result
in incorrect paths when it does.

After this CL, the case of brackets is handled by looking for the
closing bracket followed by an space and then looking for the ".xlm"
extension, allowing withe spaces between those elements.

Bug: 140525396
Test: Added testFindingOriginalManifestFilePositionWithSpace
Change-Id: Ida08ea144395d303815909ce8c7aaccf173fa1ab

Former-commit-id: 526875e7597be60899f8bdaa9cd49e013a8f3cdf