Convert path to Unix-style to check if it is class file
Convert paths to Unix-style path when checking if path is
a valid class file that should be processed when e.g. dexing,
computing main dex list etc. Previously, users of this method
were responsible for converting path to Unix-style one, but
that is error prone. Now, we always convert path before examining
it.

BUG: 140290149
Test: ClassFileInputTest
Change-Id: I38270852910836ffcff50ca8d85a17a34c80adb3

Former-commit-id: 7ade7ac16a1b17f20d761026d1c4eb1987b1b97f