Removes ApkDiffer dependency on Dump in our code.
This CL is a refactoring that does not have functional changes.
In the Deployer, Diff still depends on Dump to finish.

It is just the ApkDiffer now knows nothing about dump. This will
facilitates ApkDiffer to run on some cached result which will
be part of speculative dex diffing should we want to go that route.

If not, I still think this makes code / test code cleaner.

Bug: 134706433
Test: ApkDifferTest has them. This is just a refactor.
Change-Id: I3d8264e57bdf21e55c48220ced8e7ae0df519a9c

Former-commit-id: 4431c02e61e5e0ba35f1570eca2937e3754f67b5