IRR: Improve deltapush by reducing # dirty areas
apkzlib and zipflinger never change the information contained in
the local file header extra field. Therefore, if an entry is
deemed "unchanged", there is no point in creating a dirty area
for its extra field.

Morever, since we don't have access to the Local File Header extra
size, we cannot accuratly know where is the end of an entry.
Therefore, we use an extra LFH value of zero which result in a zip
entry location where the start is accurate but the end is only
"equal or before" the actual end of the zip entry. This cover most
of the entry filespace, while the reminder (if any) will be marked
as "dirty" by deltapush.

Test: None
Bug: None
Change-Id: I05a6464d71c921b9610940eb2e1b2379eeccde8e

Former-commit-id: a495543e08eafcd5725be9fa8e0f63f55953de50