Add memory integration tests for unified pipeline
- Add back tests that mimics the legacy MemoryTest and JniTest
- Switch the integration tests to use an async stub for TransportService.
This allows the tests to support cancelling a test so we can more easily
test different conditions in sequence within a single test. Without cancel,
events arriving at the daemon would keep streaming events to the first
registered listener.
- Switch the live allocation performance test to use the new pipeline since
that's the one being maintained going forward.
- Fixed an issue in the daemon's EventBuffer where if a stream fails when
being written to, the event would be missed even if a new stream listener
is registered.

Bug: 121392346
Test: This is it
Change-Id: Ib59923454b6a14829254033884a4c5c93713d51f

Former-commit-id: 5a3037731662860790655b3e52c3aa8f80b16552