Do not generate un-needed exceptions.
When dealing with asynchronous notifications, or just jvm memory
model visibility across threads, do not generate an exception
when start time has not been set correctly.

Also protected thread visibility with Synchronization.

When such a situation arise, will approximate the worker time
to zero and keep the build healthy instead of generating an
exception while profiling may not even be turned on.

The worker profiling is an approximation to start with.

Bug: 141350484
Test: adapted unit tests.
Change-Id: I3c5ade1ec425403aaac00cb5c5fa09432f517ed7

Former-commit-id: ebbe812a47e1f69040c75deab33ad88f0f29b7a4