Remove <nav-graph> elements in aapt friendly manifest
In app modules, <nav-graph> elements in the source manifest are
removed and replaced with corresponding <intent-filter> elements
in the app's merged manifest.

In library modules, these <nav-graph> elements remain unconverted
in the library's merged manifest. This causes a problem during the
VerifyLibraryResourcesTask because AAPT2 fails when the manifest
contains <nav-graph> elements.

This CL removes all <nav-graph> elements in the aapt friendly merged
manifest.

Fixes: 140856013
Bug: 112513722
Test: LibWithNavigationTest, ManifestMerger2SmallTest
Change-Id: Id7ee76df1d20f8a1d3d49b8d95c96dbbb65efea3

Former-commit-id: c0ae7e23221a14fcba5727e0aa2da70497871533