Bye-bye feature and instant-app plugins.
These were deprecated in 3.6 for removal by the end of 2019.
By the time 3.7 is out it'll be early 2020.

This is a first pass that simply removed the META-INF entry
for these plugins and updates tests to pass.

A 2nd pass will require removing the actual code, which starts
from the plugin but then goes deep, deep, deep into AGP. Care
must be taken to ensure we do not end up with a lot of unused
code.

Bug: N/A
Change-Id: Ie0d0bb7424d2dff8a7ce62274412deff4a5050a9
Test: updated parameterized test and removed feature only tests.

Former-commit-id: c41947884d0ce8d1ae79134026adb58b41ebfce2