Remove feature plugin code.
This remove the plugin, extension, variant factory, etc...

This also cleans up the VariantType and tweak the API sightly to
make it less confusing now that we only have one type of feature.

This removes AndroidTypeAttr as we do not need to disambiguate
between the feature and the aar part of the same module.

However this was also used to disambiguate building vs
publishing configurations for libraries. So we still
use an attribute for this: AndroidUsageAttr.

Special handling for other plugins which may require that
FeatureExtension and/or FeaturePlugins classes to exist to work
properly. The classes are therefore not totally gone, but are
now mostly empty.

Bug: N/A (clean up)
Test: removed some obsolete tests, but this is only code removal so existing tests are all the coverage we need.
Change-Id: I5c1fc13cb41ad61bdd3fcc9636f1341511d1d2fe

Former-commit-id: 7bb16a5aacb837dff1b1d8521df37482a8ecdd7e