Don't remove native libs when external libs are updated
Previously, when updating an external lib with native libs, sometimes
some of the native libs would be removed from the APK. The reason for
this is because the IncrementalFileMerger assumed that the absolute
paths of input directories wouldn't change between builds, but the
absolute paths of the input directories do change when the version
changes.

This CL changes the IncrementalFileMerger logic such that it no longer
relies on the absolute paths of the input directories being constant.

Fixes: 141536986
Test: IncrementalFileMergerTests, NativeSoPackagingFromRemoteAarTest
Change-Id: I2f07b868b25d19527d25a551e5cb3a9fe5cc82d9

Former-commit-id: ffb9c68ec32ef3f7b082c226013e298984de8a44