Provide better reporting when SDK path is not set.
When the SDK path is not set, we want to report this as a sync/build
error. The error reporting mechanism was already in place before this
CL; however, it was a bit late in the execution and some execution paths
may fail with NPE before reaching the code that performs error
reporting.

This CL moves that error reporting earlier to immediately when the SDK
location is resolved (in SdkLocator#getSdkLocation).

Fixes: 141735050
Test: Updated SdkLocatorTest
Change-Id: I857f90c70c6f5f8ac0b7407a7fd290cd4c643e0b

Former-commit-id: 18bf3bb6956e9c5627318ad50ff28ad2645f5d7f