Uses BES to copy outputs.zip
Currently, performance data that is generated as part of running
tests is saved in bazel-out, and a shell script copies over all
files matching a pattern into the DIST_DIR. This is error prone,
as it doesn't take into account what tests were actually run by
bazel.

This CL fixes it by parsing the bazel build event stream to look
at the tests that were run, finding their completed targets and
their output files.

Ultimately, this will go away once the backend knows how to
extract this data directly from the BES (as all the artifacts are
available on resultstore).

Test: Not included
Bug: 141926350

Change-Id: I90300cc529663b25e24adef806f1cef067e43e86

Former-commit-id: 31b4feb07068d528397810a91c098762ad7e1ec6