Switched BuildableArtifactHolder to use new Variant API.
Switch BuildableArtifactsHolder to use VariantAPI underneath.
BuildableArtifactsHolder is not an empty shell and will probably
be removed in a subsequent CL.

Moved some private artifact type to a new type MultipleArtifactType
when cardinality was more than one.

All internal ArtifactType now implements Single as a lot of
internal fields or parameters are used to store things like
task output type and we cannot cast to two types before calling
Operations methods.

Removed reporting feature from the BuildableArtifactHolder as it
is not very useful anymore (we don't keep track of intermediate
producers any longer). Will either reimplement or remove the
feature in a subsequent CL.

Bug: 139752844
Test: Adapted Existing

Change-Id: Ia90edc1e68a71471de9de93f50dd7646d8fdc0f2

Former-commit-id: 0960593f8384f148f63cf43fc39d65adf6cf037b