Tweak API and implementation of Feature Set data.
VariantScope.getFeatureName() cannot receive a Function<String, String>
anymore. Instead we expect the call to use Provider.map() to transform
the string into another string (or something else).

Internally, use a Provider<FeatureSetMetadata> to avoid code duplication.
Sadly, this does not prevent the file from being parsed twice at the moment.

In fact, it's going to be parsed for every Property that reads the value.
We need to find a solution for this.

Bug: N/A clean-up
Change-Id: I27e4521126c3e3a8791d6a30a1f229d2f092aa6c
Test: existing tests.

Former-commit-id: 32fc68bcbd35a91a585260d08f4e83cefcaec98e