java_proto_library: Make grpc version explicit
The java_proto_library macro used a hard coded version of
the grpc proto compiler plugin, and always added a particular version
of the grpc runtime dependency. This CL changes the version to be a
parameter, and users have to explcitly add the grpc libraries in
their dependencies. All of this allows other targets to use a different
version of grpc if needed.

Bug: 141926350
Test: existing
Change-Id: I5b49b0d6fa9d26fa8c75ed08f12d5dc49f3b33c4

Former-commit-id: 599307999f48657c6769b9209ac005f413f8b0f3