Refactor ApkEntry -> Apk relationship
This begins the process of paying down some technical debt before implementation of install-without-install begins.

Refactoring the ApkEntry -> Apk relationship such that an Apk contains a
list of ApkEntries allows for more natural APIs and easier access to
basic pieces of information, such as the package name.

Immutability of the Apk and ApkEntry objects are still maintained
outside of the model package, and a bi-directional relationship from
Apk -> ApkEntry and ApkEntry -> Apk exits.

Bug: None
Test: Existing tests pass
Test: ApplicationDumperTest
Test: ApkDifferTest
Change-Id: I8168af86a440ea3c67096f485c385ae16c80ab57

Former-commit-id: 4c042ed101003028b23579fbc804beb4021c891f