IRR: Fix off by one error in zip parsing
Change I05a6464d71c921b9610940eb2e1b2379eeccde8e introducted an
"off by one" error where clean portion of a zip are one byte past
the end of a zip entry. This cause Patcher to fail in the case
where a zip entry is added at the end of a v2 signed apk.

In that case, the first byte of the v2 signature is marked as
"clean" and subsequent v2 verification will fail.

Fixes: 142883453
Change-Id: I06a5cd708515703199c9ea81964ed78b30a4ada9
Test: Unit Tests

Former-commit-id: a376c8b20c92f97c8064925d58ca67a31416015f