Zipflinger: Add debugging tools
To investigate problems with zip or apks, this CL introduces tools
to verify and map.

Verify: Check that top-down parsing and bottom-up parsing return
the same list of files. This is a good way to diagnose if a zip
file contains "holes" since top-down parsing will stop at the
first hole.

Map: Returns the sorted list of entries (bottom-up built).

Test: N/A
Bug: N/A
Change-Id: I2dd008d5c13df85a3cc6729923c91294935cf2ae

Former-commit-id: 7de785203d380e6fbb76794f04cd2a14a2eaebb2