Document that CheckManifest task is unused.
So that when the new variant API is created, we don't create a public
artifact representing the output of that task.

Note that we are not removing the task yet for compatibility reasons.
Once the new variant API (which removes all task getters) is in place,
we will remove this task together with the other task getters in the
old API.

Bug: 139855995
Test: Existing tests should pass
Change-Id: I6ecab7f8dcfaf42f6aef2553ad6ef7cd2ede00a0

Former-commit-id: 0de5f98ccbe08a0e1605e8bc6767f5d6570d77eb