Change PrecompileRemoteResourcesTest published lib name
The test is inspecting gradle's artifact transform cache looking for the
output of the published lib. The name was used in another test so
sometimes the test would grab the output of the other test instead.
This change renames the published lib package name so it contains the
test name.

Bug: 143200215
Test: fixing the test
Change-Id: Idd245e53aa1ef748c94cd82835c99661197a956f

Former-commit-id: 13bb807ad62ccc60e9dc9636169f83120d13b131