(Hopefully) fixing native on heavy loads
RenderscriptNdkTest has been failing intermittently ever since it was re-enabled.

My theory is that when doing native compilation with Renderscript in Ndk mode,
ExternalNativeBuildJsonTask fails to wait for Renderscript to generate
the source files, causing a compilation error when the machine is under
extreme load.

The test has not failed when running after this change, so hopefully
this will fix it for good.

Bug: 141373709
Test: RenderscriptNdkTest
Change-Id: I74d000a3dabb7c16a2ae476dfa411bebafae1e3a

Former-commit-id: 2dfcc9afa0a1e94a1e06fcc3f294ed5fd9ed41b3