Skip direct load if an add-on SDK is being used.
We currently cannot predict where add-on SDKs will be installed,
so we cannot use the direct loading mechanism to try to load them
up.

If we manage to find another way to detect them without a full scan
of the SDK directory, we should change this to also parse the
package.xml from the add-on SDK, do a parseAdditionalLibraries() on
where it is installed and merge the result of both additional
libraries.

Bug: 141573404
Test: SdkDirectLoadingStrategyTest.load_addOnSdk + existing
Change-Id: Icf113a2af6bec44575d41012955345e248868926

Former-commit-id: 063a1e2ddbaeecdfa23c4cfe4c317cbba972dcca