Warn about java res collisions between app and lib
Currently, when an app or dynamic-feature module has a java
resource or .so file with the same name as a java resource or .so file
from one of its dependencies,
AGP automatically chooses the version of the file from the app or
dynamic-feature module and ignores the file from the dependency.
This can lead to unexpected behavior or errors at runtime, so eventually
this should result in an error.

This CL adds a warning for this case, which will be converted to an
error in a future version of AGP.

Bug: 141758241
Test: JavaResPackagingTest, NativeSoPackagingTest
Change-Id: Id23de956747593e6c5a9c6d6f0175a612d0f53b5

Former-commit-id: 6a979561420d0d64f96e2de0dc3b6acd62eb120b