Merge PackageManager into CmdCommand
Removes the single method from package_manager.h and folds it into
cmd_command.h

Uses this as an opportunity to begin refactoring cmd_command.h to act as
a single point of abstraction between Android command line utilities and
the rest of the installer:

- Removes global static state in the form of PM_EXEC and CMD_EXEC in
favor of additional workspace configuration

- Moves per-API APK retrieval logic from DumpCommand to CmdCommand; code
that calls CmdCommand should just need to specify a package to dump
without worrying about which method is best to call

- Reduces the API surface of CmdCommand - only exposes a single
GetApks() method for the reasons described above

Note on Workspace changes:

My goal is for the Workspace to be just a collection of information;
essentially a glorified dictionary with compile-time safety. This is the
motivation behind moving pm_exec and cmd_exec to the workspace, as
opposed to storing them in global state.

Bug: None
Test: Existing, Manual
Change-Id: Ifc49662147e367da946f8cc23753aceb16f8fe7e

Former-commit-id: 3e6954a8c0ea106cad3f96509afbc0f1ab60eae3