Split java resources back to feature APKs (like feature dex)
Previously, all of the dynamic-feature java resources were packaged with
the base APK (or into the base folder of the .aab file). This CL changes
this behavior to instead package each dynamic-feature's java resources
into the corresponding dynamic-feature APK (or into the corresponding
dynamic-feature folder in the .aab file)

This CL adds java resource splitting only when code shrinking is done
with R8, not Proguard.

Bug: 111082605
Test: R8ToolTest, R8Test, MinifyFeaturesTest,
Test: R8TaskWithDynamicFeaturesTest, DynamicFeatureJavaResTest
Change-Id: I33541cd414ed10e2c3e7a640c3e1691246ef782c

Former-commit-id: 65bc57feddabe3289228329e041d424a016988d7