Update AGP to Gradle 6.0
Update tests and make sure we use correct APIs from
Gradle 6.0. This is also a new minimum version.

Task validation is now done in an intergration test because
it was substantially changed in Gradle 6.0.

When running test with AGP 3.3.2 and 3.0.0 we need to
specify Gradle version, as 6.0 is not binary compatible.

Also, because of changes in 6.0 regarding the dependency
management, Maven repos that we are using in our tests need
to specify artifact() metadata source. More information can
be found at https://docs.gradle.org/current/userguide/repository_types.html.

BUG: 142630205
Test: existing
Change-Id: I0a1232993807a1e875e1814f2d07c239a05b79c0

Former-commit-id: 708d9be0dbb33101d2f84031e3408cb71f87d00b