Add an install server and client for IWI
IWI overlays can only be modified by a process running as the
application uid. Thus, there must be an installer component that runs as
the shell user, and a component that runs as the app uid.

This CL adds server and client objects that encapsulate this logic, as
well as methods to start the server optimistically, only copying it if
the binary is missing.

It also adds an additional binary, called "install_server", that can be
unpacked from matryoshka and run as the app uid.

Bug: 144062091
Test: install_server_test
Test: Modify swap.cc to start the server and check for logs.
Change-Id: Idbc93aa867be4af8f747003ff6022142db4ec293

Former-commit-id: 19c90839370d9377ff595704d66e7e04f3fcfacd