Remove separate annotation processing feature.
This feature is no longer needed as the most popular annotation
processors have become incremental (see bug 110890707#comment10).

The feature has never gone stable so we can remove it directly without a
deprecation phase.

For users that are still using the feature (~1500 projects on 3.5.0,
~1700 projects on 3.5.1, and very few on other stable and alpha
releases), we ask them to use incremental annotation processors via a
warning.

Bug: 142515190
Test: Existing tests to verify that the removal doesn't break stuff
Change-Id: I13b2c4186e4584bf2ef48f3f15a0aa8267ce2aa8

Former-commit-id: 609d9cdba53e5dd88a4c0c4e869afde0c09263e7