Split Variant object into 3 and generify extension
Split Variant interface into 3 interfaces :
1. VariantConfiguration will contains properties that can never
be changed through the variant API like the variant name, build
type, etc...
2. Variant will contain properties that can change the build flow
like setting the minSdkVersion or disabling minification. These
objects will be available through the onVariants() APIs.
3. VariantProperties will contain all properties that are input
to Tasks. These objects will be available throug the
onVariantsProperties() APIs.

In a subsequent CL, Variant and VariantProperties will be subclassed
for all types of modules.

Generify the CommonExtension interface to parameterized those on the
right Variant and VariantProperties. There is only so far but this
will change in the next CL.

Test: adapted existing, some unit tests.
Bug: 139752844
Change-Id: I528c16a3e8334238bcd68933c4d24c260161f45a

Former-commit-id: aa796ecdf713c44c95afd93745a85cff3ad2e0e3