Implement "test_excluded_packages" attribute for iml_module rule.
This replaces the hard-coded list of packages names in JarTestSuite,
which was specific to android.core test target.

This new attribute allows for a more generic approach. It uses a
"black list" semantics (i.e. list of packages to exclude from tests)
as opposed to a "white list" semantics from the previous hard-coded
implementation.

Bug: 141868211
Test: This is it
Change-Id: I25a86111805bb71a208dd2617e4c64dc00f20073

Former-commit-id: d513dc36cd5f4ca724d023fe96f1eed8a21a7505