IRR: Fix session leak
Problem: It is possible for Apply Changes pipeline to create a
session on the device and neither commit nor abandon it (e.g: when
an Exception is raised while AC runs). In this case, a session is
leaked on device. This leak may saturate the device package manager
overtime. Example of scenarios where a session is created and
leaked:

- APK not recognized (APK not in local DB). Session creation
  and diffing are done in parallel on the host. Diffing will throw
  an exception but the session was already created (and leaked).
- App is not running. Dump, session creation, and diffing succeed.
  However since the app is not running, the swap task will throw an
  exception and the session is leaked.
- Resource in the APK were changed. Session is created. Diffing happens
  in parallel and resource modification is detected. Exception is
  raised and session is leaked.

Solution: Introduce the concept of Task Fallback. If a Task did not
complete execution (threw Exception) or was not run at all due to a
failure of one of its dependencies, it is eligible for fallback.
If a fallback callback was provided, it is called to let the Task
perform clean up operations.  To abandon a session, ApkSwapper provides
an error method which checks if preinstall ran and abandon the session
if needed.

Test: Unit Tests
Fixes: 141867712
Change-Id: I2fa24ca1eeb36f1a3776e609ebdeda518bbfc140

Former-commit-id: 2dcc232f4617bb58ad8d67be47b84c9525221221