Added Variant and VariantPropeties sub types.
This is a reworking of I74620d18971ee230945912fd5062ee44d6135794 but
without the breaking change of making the pseudo-api
DynamicFeatureExtension not extend pseudo-api AppExtension. To do this,
this change splits the class heirachy of the new api from the existing
pseudo-api. Pseudo-api BaseExtension no longer implements api
CommonExtension, and thus does not need to be parameterized.

Added Variant and VariantProperties for application, dynamic feature,
library or test modules.

Changed relevant extensions to be parameterized by the new sub type

Test variants are not yet handled.

Test: Existing (mostly just interfaces added).
Test: PseudoApiTest and StableApiTest.
Test: TODO: Land regression test for kotlin in features before landing this ideally.
Test: TODO: Add integration tests for this new API.
Bug: 139752844
Bug: 144403807
Change-Id: I75a6da9b9a886ede74e277055e37fe778b70c499

Former-commit-id: 1bbb45a937b0f68918092e6d92bf303545820497