Make the variant API compatible with buildFeatures
It is now possible to disable aid, renderscript
via BuildFeatures. This impacts the variant API
as queried tasks can now be missing.

This makes the getters throw an exception
(or return null during Sync) since this breaks
the previous API. The action on the devs will be
re-enable the feature.

Bug: N/A
Test: N/A
Change-Id: Ief0d3146342f67e2cf7b48ea53ea8a428d406cee

Former-commit-id: 8a6d35adfcc0440e7f2fcfdeb60ed25c58e5d12e