Organize Boolean options by their status.
This CL only sorts the lines in BooleanOption.kt into groups having the
same status. More changes will be made in a separate CL (to make it
easier to review and prevent accidental semantic changes).

Also add a test to prevent adding a new Boolean option in the wrong
place.

Bug: 144917849
Test: New BooleanOptionTest
Change-Id: Ic5b788f4629e411a0d33fb03d87a59f0bb9d0a49

Former-commit-id: 48c6455b8a44c166c6eb4d0874c6e9f33e09f1e1