Split sourcesets out of VariantDslInfo.
They are now in a new object called VariantSources.

This will allow us to have a single VarianntDslInfo per
combination of flavors/biuld type, meaning that this can
be the same object for prod and androidTest.
This will be done in the next step.

One main benefit to this is that VariantDslInfo is
only meant to feed the default values into VariantProperties.
Nothing else should access it, and certainly not the tasks.
(this is Work in progress).

Sources, however, should still be available to the tasks,
so this separation will ensure that we don't give access
to tasks to both VariantProps and VariantDslInfo.

Bug: Clean-up
Test: existing tests
Change-Id: I7e55be61c1e97eac023e7b2d1583aeffd5f50b96

Former-commit-id: b55446712b34c72eeb18d32d15f2376c5fd16012