Provide post build information through a file.
Instead of relying on post build sync to get list of APKs,
all information necessary for deployment will now be also
provided though a json file. The file path (which is per
variant) will be provided through the initial project sync.

This CL only wires the file for the assemble task for APKs,
Bundle and APKsFromBundle will come in a subsequent CL.

The post build sync is left untouched for now but will be removed
once the IDE stopped using it.

Test: adapted some integration tests, added unit tests
Bug: 141647392
Change-Id: Iaf0d90b18425854b028ce12c1f50f76193082ddb

Former-commit-id: 5f78b6dacc3b14336c7840748cc4f11736e96adc