Look for lcov files in bazel-bin instead of bazel-genfiles
This usage of bazel-genfiles was missed in the migration to bazel-bin

Since the local version of the script is not checked by CI at all this
was only caught when someone attempted to generate a local report. It
could be worth considering adding a quick sanity check of this script to
the coverage bot by invoking it for a very small custom report such as
the example report.

Test: Checked locally
Bug: 146011558
Change-Id: I914dc54ede9d872c0db54256965975706423ab44

Former-commit-id: 811cf71402a0d71fc465d5ace96d10f1400951ab