Signal the initation of CPU trace stopping
A STOP_CPU_TRACE usually generates a CPU_TRACE_STATUS
and a CPU_TRACE events. They were generated at the
same time before. This CL generates the first one
right after the command is issued (before the stopping
initiates) and the second one after the stopping has
finished.

The purpose is to let the first event to signal
the profiler is in the stopping stage, which is
persistent even if the user navigate away from
CPU profiler.

Also fixed a minor bug in stop_cpu_trace.cc.

Bug: 139218459
Test: manual
Change-Id: I5087d1133340cd5628d05cc1bbe536410e8a65d4

Former-commit-id: 4c7db36dc351a2c8bc4ffc7d4ecdd1433500e802