Fix ~5% failure when establishing JDWP connection
This CL fixes the 5% failures of failing to establish a JDWP
connection. The problem arises from the fact that if a
connection is present, the app can send an APNM chunk to the
host before the host establishes a HELO. This is not expected
by the host Client state machine, and it bails + reopens the
connection.

This CL fixes the issue by detecting if there's an APNM packet
arriving before the HELO, and discards that packet.

Fixes: N/A
Test: manual
Change-Id: Ib49b6af152ba0164f6b4e12d4a8cd5a21bdb2821

Former-commit-id: 462997c9b3da7da20d601697d44c51718d606516