Rename android-classes artifact type to android-classes-jar.
To support incremental dexing for subprojects, we need to distinguish
between a jar and a directory containing classes.

The current android-classes represents a jar so we need to rename it to
android-classes-jar.

In another CL, we will introduce android-classes-dir. We will also add
back android-classes, but it will have a new meaning: it will represent
either android-classes-jar or android-classes-dir.

Bug: 132615827
Test: Existing tests (refactoring only)
Change-Id: I69ea17012ba1ca0c264cec5c5bdfdb16643d0fff

Former-commit-id: be8b75f36fd876b783614092d279c6425ac2e3b5