Fix kotlin script ExternalNativeBuild
ag/I59ef09801d45fef01d915dd961f839edee10a5b4 was insufficient, as
although this fixes cases that compile against the API class, it does
not fix the cases that compile against the runtime class.  Before this
change, the runtime class was implemented in java, so it lacked the
kotlin metadata on the implementation of those methods.

This worked before the work to add the kotlin interface as the methods
generated by gradle had the correct metadata attached to them, but as
soon as they were added explicitly to the kotlin interface and
implemented in java, they were no longer generated by Gradle with the
appropriate metadata.

Bug: 140406102
Test: Expanded funcitonal test: KotlinDslTest
Change-Id: I72b8307049d674944899f30ff47f26425df6f78c

Former-commit-id: e44b6ef8517c47e865893067a5515ddee88767c0