Remove NavHostFragment check
Derived classes of NavHostFragment can also reference
navigation resources. Removing this check means that
they will not be flagged as errors.

If a non-NavHostFragment fragment tag has a graph attribute
that references this graph it will not be flagged as an error,
but this should be rare. The alternative would be to test whether
the class name is a derived class of NavHostFragment, but this
would be more expensive.

Change-Id: Ibc5a5a5ef93e0ab22587f015336eecd4ce1a4702
Fixes: 144871832
Test: None

Former-commit-id: 354e8f0a2cdb7bee093e830d3104de1b20e34608