Refactor VariantManager
... which has become too big and does too much.

Moved out 3 components:

1. The part that handles the DSL callbacks (when adding
build types, productflavors, and signing configs).
This is now in a specific class that implements VariantInputModel,
which used to be VariantModel. VariantManager does not implement
this anymore. There is still a VariantModel but it now
contains the actual variants, and is used by the ModelBuilder
rather than have it use VariantManager.

2. The computation of the default variant which is
only needed to during the model. This is now part of the new
VariantModel which returns the list of variants, the default variant
name and the inputs of the variants (as a VariantInputModel).
This is what gets passed to the ModelBuilder rather than a VariantManager
instance.

3. The part that configures DependencyHandler setting up all the
   artifact transforms.

The goal is to make VariantManger (and the new VariantInputModel)
more transient and allow more unit tests on these steps.
Right now these objects are still held on to (and accessible via
the Plugin class which is not good), but this is a step in the
right direction.

This change had to do some other misc changes:
- Some Kotlin code now deal with the isDefault mess on BuildType and
  Productflavor so I had to add a temp method to deal with the property.
- Changed VariantModelFixtures to create normal instances of build types
  and product flavors instead of mocks. This allows more tests to change
  any property they need. Like...
- DefaultVariantTest is changed to not run the actual Plugin. This is not
  necessary anymore.
- Did a bit of clean up on the IssueReporter classes but more need to
  happen.

Bug: N/A (clean-up)
Test: existing tests
Change-Id: I3b1a2c430ca884aa4ff246cfe7b488f3fac18ce3

Former-commit-id: 10f389d9b130c427ae49fd64fc4c200821a28daa