Create new variant objects when they should be.
They used to be created via BaseVariantData after VariantManager
was done doing its things, which includes create all variants
and tasks.

Variant must be created first as it allows filtering out variant
(actually it needs to run after the old VariantApi since we want
the new one to be more granular).

VariantProperties must be created (and actions done) before
tasks are created.

This changes quite a bit how VariantData and VariantScope
are handled. This is a bit of clean up but we really need
to invert everything and make the new objects
1. contain the old ones
2. be used throughout the code base (with getter on the old
   ones as needed until we remove them)

Remove the old Transformer from BaseVariantData to new
variant object, and reworked the FilteredVariantOperation/Action

Bug: N/A
Test: existing tests
Change-Id: I3f052e49ab23b7756aa4f690d2d0badd8cd9ed9a

Former-commit-id: cd6d67ecca11b9c7b25278762dde234b0135f293