Introduce AndroidTest and UnitTest interfaces.
This change rework the inheritance hierarchy. In order
to reduce confusion, Variant is now the base only
for the actual build output (apk, aar). Variants
are components and tests are components too.

The new tree looks like this:
- Component
   - Variant
      - ApplicationVariant
      - LibraryVariant
      - ....
   - UnitTest
   - AndroidTest

The Properties tree looks the same.

This means the onVariants API only acts on
variants but not components. We will introduce
later a onComponents that requires Type filtering,
but onVariants can be specialized in each Extension
to default to the actual Variant that most users
will want to touch.
To offset the lack of tests in onVariants we give
access to them via the variant:

onVariants().withName("foo") {
    onProperties {
      // variant props here
    }
    unitTest {
        enabled = false
        onProperties {
          // test props here
        }
    }

    androidTest {
        enabled = false
    }
}

Because the root type is now Component[Properties],
this change also renames some classes for *Variant*
to *Component*

Bug: N/A
Test: existing tests
Change-Id: Ie56cf03779b414a8d869d0db9a3268c54268a522

Former-commit-id: 4a6c7fd99652a7d8ea3ca2d726b78a87c9348abe