moved metadatafile loading/writing to sdk common
studio needs to be able to read this file so made Studio specific
APIs that do not depend on Gradle APIs and have some specialized
persistence methods to read/write.

factored out common code in sdk common and specialized further in
gradle-core to match the gradle-api types provided to end users.

Test: Added unit tests
Bug: N/A
Change-Id: Ia7b0b7cc0f8df469bb6be3a0f4d6f386f195966f

Former-commit-id: 2511adc9c9ac8fa8baec5ae22121bb828be22ffc