Migrate remaining lint unit tests to the new Test DSL
Lint has had a new test DSL for several years now, but a lot of the
older detectors hadn't been migrated. This CL does that.
In some cases, when the whole test case was using the old mechanism,
I also ported the class to Kotlin. In other cases, particularly
large test cases where there was a mixture, I just left it in Java
and switched to the new DSL.

The new test DSL performs additional checks automatically, for example
for the UInjectionHost handdling, but with the upcoming support for
provisional reporting, that gap will widen significantly, and it's
important for all the lint unit tests to be going through the
additional checking in the new test DSL.

(Also applied google-java-format and ktlint on the modified files.)

Test: This is it
Bug: None

Change-Id: I5664cc8e8de8f4a77f5af7d7c8a1b9aa0edf81d6

Former-commit-id: 26ac4442331a1caec6c88b053861102455cb9ab2