Fix per-variant external native build types.
By using the more-specific type, and update some tests that were using
mocking to use the more-specific type too.

While this didn't matter for groovy, which always used the
implementation class, it matters for kotlin script where, before this
change, an explicit cast was required to set values in these DSL
elements.

Fixes: 146488072
Test: Updated pseudo-api
Test: demonstrated use in KotlinDslTest
Change-Id: I2ed387415d12d4d9535cc1d5b65d877cbc0f5494

Former-commit-id: de7dfafb7440e670afd3c27167b67ae9634d2779