Added Workers friendly API to BuiltArtifacts
A number of internal tasks and potentially future external
tasks can use BuiltArtifacts.transform() style of methods
to move an artifact type into a new one.

Added an external API that is friendly to Gradle workers
for external tasks.

Added an internal API that we can intercept to be able to
continue measure work items execution time (as well as
start and stop time). Wiring this API to the actual profiling
objects will be done in a subsequent CL.

Bug: N/A
Test: Added unit tests and integration test for public API.
Change-Id: I84fa0d30ad7a646eeff48e3a4f1b22564a6d8fb3

Former-commit-id: 13562c894326edca11231689b9664468ab5dc3a0