Limit week-based-year check to combinations with days/months
Limit the date format check for week-based years to only
complain if the week-based-year is referenced in the same
string as days or months, which is where it's particularly
likely to be wrong. This makes false positives less likely,
which is why this CL also turns on the check by defaut.

Bug: N/A
Test: Unit tests updated
Change-Id: Ia9a88c5785b2744f1c0147304a9fb9f319331eae

Former-commit-id: 77015739cdd8c71b3436d99416805521f4fe3db7