Move some task creation methods around
Methods that contained test on VariantType are
now moved to their own proper custom TaskManager.

It's an important steps as future changes
will make the TaskAction receive a specialized
CreationConfig and we need to ensure these
methods act on the right types. Right now
they all act on the base ComponentPropertiesImpl
so it does not matter but when we move to
specialized types where they are located
will matter.

Bug: N/A
Test: N/A
Change-Id: I711d60b0e79239a34b778f300c7c261b2a758c99

Former-commit-id: 6c8531b79711cd7f3a4d8ef1b2747828f49a79db