Include the lint implementation checks in non-Android lint
Lint can be used to analyze non-Android code as well, for example
using the standalone lint Gradle plugin.

This CL moves the lint checks out of the Studio registry and
into the standard lint registry, but marks all the issues as
not applying to the Android platform, so the checks will not be
run in Android Studio, but will be included when you run lint
elsewhere.

Test: Unit tests included
Bug: None
Change-Id: Id735add62bb004094c10d096c27858345bfb0514

Former-commit-id: 165ab6af9512f1d97944e435068221ece3dafc47