Migrate to official Googletest build rules.
This CL drops tools/base custom gtest and gmock custom BUILD in favour
of the official one already present in the repository. This also cleans
the bindings that are not needed anymore.

Besides the cleanup aspect, this should proper set-up gtest/gmock to be
used in windows environment, as the official build rules properly export
the native symbols required.

Bug: 147797511
Test: bazel test //tools/base/...
Change-Id: I61d29560a4678e4cf7f2556cb3ff976ce36f4232

Former-commit-id: efc8e239e56e4fc7fdd9c2e1b35e725d27d60975