Introduce a type parameter for Component
Previously we added a type param for ComponentProperties.
However this is not enough as the task manager needs to
query Component/Variant to figure out which task to
create.

Therefore this change adds a new type param to VariantManager,
*TaskManager, *Plugin.

Doing this, we came to the realization that AppVariantImpl
extends VariantImpl<AppVariantProperties> rather than
using AppVariantPropertiesImpl.

This is because the public facing interface is
ApplicationVariant which is declared as
Variant<AppVariantProperties>. This creates a lot of
trouble internally as we want to use the Impl for both
the Variant and the properties objects.

To fix this, the public AppVariant interface still has
a type parameter for the properties, but everything
will always ever use it with AppVariantProperties only.

We will write some samples and see how the API feels
with this changes and tweak as needed.

This also does a tiny bit of refactoring of TaskManager
for the tasks that must be created after the Variant API
(This is mostly compose)

Test: existing (some tweaked)
Bug: N/A
Change-Id: I9d40929f016ff6bdd12a827fbac72ce075f05314

Former-commit-id: e69931af81210bafa45ad66498256cb9e6bc4988