Variant-specific dependenciesInfo
Introduce appVariant.dependencyInfo to allow
per-variant updates of the dependency info attributes,
and a read-only AppVariantProperty.dependenciesInfo
to read the values if needed during the 2nd stage of
the API.

The Variant interface is the same as the DSL one since
we do not want to use Property there. The implementation
is currently different but it would be made the same
if we were to tweak the flag to enable/disable the setter
and getter.

The variant properties interface is different as it's read
only. (note that this is a case where the value is a stage 1
property and therefore is not exposed as a Property in
stage 2).

AppVariant.dependenciesInfo use laziness to avoid
creating separate objects if the value is not changed
during the variant actions.
AppVariantProperties just re-use the DSL or Stage 1
objects since they implement the right interface.

The Variant implementation uses a new Value wrapper
that works similarly to Property but only allows
setting from the actual type (no provider/callback).
It also allows turning setters and gettes on and off
separately.
This object is not meant to be exposed, but instead just
used as the backing property for the public setter/getter.

The change also introduces 2 new scope objects, one
for Variant and one for VariantProperties. Instead of
providing direct access to all of Gradle's instantiators
they offer very specific features that are needed for the
narrow usage they are designed for.
Adding new features to them should not be taken lightly.

The VariantProperties scope offers a way to create
lockable properties. This is not currently used by
VariantProperties expect for the versionCode/Name
in VariantOutput, though special care must be taken
when dealing with these since they are accessed
via old APIs. A further change will clean this up.

Because we now have several scope objects (and more to come
as the task creation will need its own scope), this change
also introduce a new ProjectScope that backs most scope
and owns various objects like issue reporter, factories,
etc...

Note that DslScope, currently available on GlobalScope
will be removed so that it's transient and only available
on the DSL objects.

Test: existing tests
Bug: N/A
Change-Id: I243d0a4d4cd4f1e8e984ef82912ec37a5043244e

Former-commit-id: fd1f42453d6e10058593f233a75936927eedcf55