Zipflinger: Add StableArchive
Problem: In Release mode, it is preferable for the APK to be
stable. Zipflinger provides no such guarantee since it carries
operations as soon as they are requested.

Solution: Buffers all operations until close() is called. The
trade-off for this feature is an increased memory footprint as
well as increased zipping time.

Test: Unit tests
Bug: 148414726
Change-Id: I0969d14484f0dd59b6f333dfae99d6297126a9e1

Former-commit-id: 56fb29dfa2e9de87e3efcb0fa827afc407e06e6f