VariantInputManager now owns the containers.
The goal is to extract logic from BasePlugin
into reusable objects so that we can more easily
test the DSL outside of running plugins.

This change is quite disruptive since the containers
are using type parameters for DSL objects who's
base interface is changing in the new DSL.

BaseAppModuleExtensionTest already shows that it's
much easier to instantiate an extension object
with minimum mocking required.
Next step is a full fixture to create/configure
DSL and look up the resulting variant[Properties]
objects.

This also fixes a timing issue where changing
testBuildType would not impact already created
androidTest source sets for the previous build type.
To fix this without breaking anyone
we now create an AndroidTest source set for
each build Type. Later we'll have to actually
delay creation till the end, but it's a breaking
change. This is tracked via b/149489432

Bug: N/A
Test: existing
Change-Id: Ib0da9a2566f03c6a164b0094dd76495ccdbe0bb3

Former-commit-id: 0308eebcf8b3b36f5545faf14b7e95d1b6f61dcf