Add hardcoded case for "findInstances(Class.class)"
Currently, calling findInstances for the Class class returns a bunch of
extra classes that would normally be hidden to the user, as they are
dummy classes used as placeholders during startup in some cases. The
appropriate way to find classes with jvmti is to call getLoadedClasses,
but that's not something we want to expose through our
InspectorEnvironment to all users, so instead, we patch "findInstances"
to behave as expected in that case.

Bug: N/A
Test: Covered by existing tests
Change-Id: Ie93d2f49b3a1d638f2751f242c2d885cdd22f2b3

Former-commit-id: 7582dc752e2dbfbc93f062f98a42a1e8d5715a70