Added APIs to use workers with BuiltArtifact with Profiling
Some public APIs were added to allow to use BuiltArtifacts with
workers but these cannot make use of the profiling information
for workers.

For internal AGP tasks that will use such APIs, we can wire up
automatically the profiling information as long as the Task
implementation is loaded by the same class loader as the AGP.
This is necessary as we need to be able to serialize and
deserialize generically target parameter and work item types.

Added a second API that can be used by Tasks not using the same
classloader that require passing a concrete WorkAction implementation
that will provide the right context classloader to Gradle.

Test: Added unit tests and integration tests.
Bug: N/A
Change-Id: Ibd7e491c0889cfab57cb4eaa0fe8e3bdfd5a7378

Former-commit-id: cdef78e48df8a748f65a763da932018fc3930486