Use native singlejar for build speedup
In the past, we wrote custom Java code for merging jar files.
Meanwhile, Bazel has written an optimized version of singlejar
in C++ that is much faster. It even works on Windows now; see
https://github.com/bazelbuild/bazel/issues/2241

Our previous jar merging tool takes about 10 seconds to produce
intellij.android.core.jar, whereas the native tool takes about
200 ms (50x speedup).

Jar merging tends to take around 5% of the critical path build
time, so hopefully this change will improve presubmit latency.
It's hard to say, though, because singlejar is not as fast on
RBE as it is locally.

I've verified that singlejar is deterministic when we pass
the --normalize flag, just like our previous jar merging tool.

Note that while singlejar supports a --no_duplicates flag, we
cannot use it for most targets because we do in fact pass in
jars containing duplicate entries. This is worth investigating,
but is not tackled here.

Migrating to an official singlejar implementation also brings some
correctness benefits. For example, it properly merges META-INF/services
resources.

Change-Id: Ib87292d4e0ba626f28caffd95ac474eaf2102606

Former-commit-id: f2bedac245b06234244979a579269d22fe147f6a