removed hasFinalArtifact(s) methods.
this change removes these fragile APIs and replaced the
implementation of getFinalArtifactAsFileCollection to not
need to test for presence.

Unfortunately, a number of users of that API were calling
a get() on the returned provider which was ok with the
previous implementation (since we relied on artifacts
being created and queried in the right order) but is
not ok with the new implementation as we would loose
the Task dependency (which is embedded in the provider).

Test: Existing
Bug: N/A
Change-Id: I4865b8e69ac1302662cc989c8638539c7b69e2d1

Former-commit-id: be0d96aad792ef8e4aa678b8f84ab8e687a2a209