Fix to sending wrong OID to agent.
It should have been getSha() instead of getRepresentation().

Sending getPrepresentation() was designed to work in place of getSha() as well since
it was designed to be use debugging purpose.

However, only getSha() had to workaround for the unimplemented empty OID fresh install
case so using getPrepresentation() always create a mismatch after fresh install.

Test: none yet
Bug: 146086924
Bug: 146085649
Change-Id: I3ebb26b11dfd9426c0c027060b257ead2f56e25e

Former-commit-id: fafa60d1b2404d1f931c54f3311e91e4899c9630