Exit successfully if tests fail for post-submit builds.
studio_rbe performs both a build and test, but ATP doesn't understand
this concept very well. If we return a failure code, the test pipeline
will think the build failed and contains no test results. We return
success to indicate the build was successful, and once the test results
are processed by ATP the build will receive the "test error" label.

Bug: 152755167
Test: n/a
Change-Id: I0514bc18c59f88a9d6832e1cec78b5847d4c11b4

Former-commit-id: cc5803bf9d810b4ceb90a5f7a775aea4a7b41b92