Update test data for JarContentsTest
...to be compatible with the native singlejar tool,
which differs very slightly from the previous jar
merging tool in that it will not automatically create
parent directory entries for resources under
META-INF/services.

Omitting the META-INF/services directory entry appears
to be harmless. If for some reason we need it, we can
track down the input jar that contains
META-INF/services/com.android.tools.r8, and change its
build rule so that it includes directory entries.
Probably there is a Bazel glob somewhere that just
needs to have exclude_directories set to false.

Test: existing JarContentsTest
Change-Id: I1633e74684511d11cf0bee451dc3ed35781ec0ad

Former-commit-id: f9889af43da7daef598fdda6ddf418abe0dff12c