Coverage support for split tests
Extends coverage results processing to support test splitting.
Assumes a test A with split B generates a target A__B, and assumes that
all split targets are lumped into a test_suite with the same name as an
unsplit test target. Custom coverage reports do not require any changes
when (un)splitting a test.

Each test/split pair is treated as an indpendent test for lcov
aggregation purposes, which can result in extremely large lcov files.
This may merit a follow up to compress that lcov for report generation
purposes (storage quota should not be an issue once the Placer migration
is complete). This also ups the parallelization requirements for results
processing as each test/split pair requires its own deploy jar + xml
conversion pipe (test_suites do not have deploy jars so we need to go
down a level). This does not seem to be an issue in testing.

Bug: 148149634
Test: Verified locally with split android-core
Change-Id: Ia5c49c213572986cec0ac37515df08a88ed6778a

Former-commit-id: a90d405e355765c3c9b588d5c18586f0fdf4239e