Per-variant BuildFeatures.
While we don't want (yet?) to allow per-variant
configuration of build features, internally we do
want to have a per-variant instance of BuildFeatureValues.

This is because some of its properties can be automatically
computed to change depending on the variant. e.g databinding
for test is tied to another DSL value in some cases.

So this changes moves BuildFeatureValues from GlobalScope
to BaseCreationConfig (and ComponentPropertiesImpl as it's
the impl class, but not the public ComponentProperties)

This change also moves data binding from BuildFeatures base
to App/Lib/Dynanimc-Feature as it does not apply to test
plugin.

Bug: N/A
Test: existing tests.
Change-Id: I3b3d261091c3eed3bfef744a6ce3c3eca49a0c55

Former-commit-id: 7e7985f55d14e02e2ae034cd14c7e4c0b704396b