Enable dynamic scheduling for studio_win.
Dynamic scheduling has shown to improve the build time for Javac and
kotlinc actions by 50-60%. After http://ag/10923854, our bazel binary
can now successfully make use of persistent workers with sandboxing.
This blocked dynamic scheduling on Windows machines, since workers
must be run in a sandbox.

Other strategies for Javac and kotlinc are removed. The 'sandboxed'
strategy is not supported on Windows, and we don't want to use the
'local' strategy if 'worker' is not available. These only pertain to
when the dynamic strategy is set, i.e., if Javac=dynamic then the
local strategy should only be worker.

Bug: 147615404
Test: Checking presubmit
Change-Id: I2024eb804729b0cc8f24af71c24281136e1be4eb

Former-commit-id: fe288704a16c3bf38a8ad68dd363118dc9ae182c