Create either .so.dbg or .so.sym files, not both
Currently, if packaging the native debug metadata in the bundle, AGP
creates both the .so.dbg and the .so.sym files, even though only 1 type
will be packaged in the bundle.

This CL changes this behavior by creating 2 tasks and 2 artifact types
instead of a single task and single artifact type.

Bug: 143862514
Test: Modified existing ExtractNativeDebugMetadataTask tests
Change-Id: Id441ba0c75dcf1e6c708ae179f5f5d53ffa3b18b

Former-commit-id: 5690dc51bd6e3ddb2c1942f23b144cf3431363a0