Pass 30 to the NDK for R, not 29.
I'm not sure if this is the right fix or not, because it's not clear
why this problem wasn't present in Q. Currently R is
`AndroidVersion(29, "R")`. This code expected `AndroidVersion(1, "R")`
or `AndroidVersion(30, "R")`. This change alters the behavior to handle
the case where it's `AndroidVersion(29, "R")` by just assuming the code
name is right when the code name and the API level disagree.

Test: existing
Bug: http://b/149605123
Change-Id: I2ac9c52b1485c9aa025922eb0d94232e307c7a45

Former-commit-id: 2c83b6284cd73b5629a7ccb4a0d29d8d959f0e2f