Change GradleTestProject.getAar
Now receives a callback for the Aar instance. This
guarantees that the zip file is closed after the callback.

There's also a testAar that automatically creates a
AarSubject.

This is a first step, the goal is to move to this
kind of API for get(Test)Apk, as well as the
ZipFileSubject.assertThat methods.

Bug: N/A
Test: fixing tests
Change-Id: I977274950ee460d1eb24500eddda3bdd91de6202

Former-commit-id: e51f56525e57281f47e1e8f688b5f2137624f6d9