Remove ABSOLUTE path sensitivity in a couple tasks
MergeJavaResourceTask and MergeNativeLibsTask were using ABSOLUTE path
sensitivity instead of RELATIVE, which resulted in cache misses for some
users.

In both cases, we couldn't use RELATIVE because we were filtering the
corresponding FileTree objects in a way that caused the relative paths
of the files to be the file names instead of the relative paths.

This CL uses a different mechanism to filter the corresponding FileTree
objects which preserves the files' relative paths.

One of the MergeJavaResourceTaskTest tests needed to be updated because
it was adding an .so file to the java resources source directory,
expecting the MergeJavaResourceTask to be up-to-date. After my changes,
the task was no longer up-to-date in this case because of the new
directories... we can't ignore added directories because they can serve
as java resources. This case isn't something that we need to support
though, because .so files belong in the jniLibs directory; I changed the
test to do this instead, and it works as desired.

Bug: N/A
Test: existing
Change-Id: I94980d16f84849f39a7f4822083305c49c573689

Former-commit-id: 9944d48a323ad2c606bed1e20e5bc90bc520e9fb