Use Gradle 6.3 nightly for instant exec tests
This CL uses Gradle 6.3 nightly that supports
detection of usage of getProject() in task action.

Also, three tasks are fixed that were using getProject()
during task execution. In addition to that, when setting
the location for generated sources for JavaCompile task,
File is used instead of Property. This is because property
that was previously used is also a task output. During
configuration, while evaluating CompileOptions for JavaCompile
this property is evaluated, which causes Gradle to detect issue
with querying output property before task that produces it is
finished.

Bug: 138843659
Test: existing instant exec test
Change-Id: I4685f317132da78ae2ed4d6b121ee46c91ad22be

Former-commit-id: 62be4184e7ae72ea227041b01ab23cf585315290