Tweak compatibility mode flag
New Property instances of the new API should
be made to be finalizeOnRead() and should disallow
calling get() during configuration (feature pending in
Gradle).

However some of these properties back the old variant
and variantOutput APIs which does not use Property
instances.

Because of this we want to not turn on finalizeOnRead()
or disable get() because it would break projects
that potentially do get(), or projects that do set
after get (which is unsafe but can happen in the old
implementation).

A new flag allows turning on and off these features. If
the old API is used and the compatibility mode is not
turned on, this will fail the build.

Bug: N/A
Test: new tests
Change-Id: I4037a38058912eb9f811c34407a5b615e9875e9d

Former-commit-id: 91c11450190359277aa009428a70a8766d5e73de