LiveStore library and inspector
The LiveStore API essentially allows for the creation of a LiveStore
(with a name) and then the ability to add ValueEntry instances to it,
via its add methods:

 val store = LiveStore("user")
 val name = store.addString("name", "Jane")
 val age = store.addInt("age", 34)

These values can later be queried via their value properties:

 label = Label(name.value)
 numberChooser = NumberChooser(age.value)

The library itself is a glorified wrapper around primitive values;
however, the power comes from the inspector, which can read these
values and send them to the IDE, where they can be presented to the
user in a simple UI.

Note: There was already the beginnings of a library and inspector here
before, and their diffs show up as part of the review, but they were
updated in response to changing requirements as our team worked on the
demo. The demo app will be uploaded in a followup CL in this CL chain.

Bug: 148613577
Bug: 148613499
Bug: 148613343
Test: Added for the library. A test for the inspector will be done in a
followup CL. It's not urgent for now because this is only demo code;
however, we will want to add it later since it is hoped that future
inspectors will copy the test patterns into their own code.

Change-Id: Ic61e0bf6ac1a663927a76d0d69c42c18b6b68898

Former-commit-id: 9b4e66e14f44aa925a553d73f5361e1745266b55