move BUNDLE_MANIFEST out of the processManifest
To be able to make merged manifest a public artifact ID,
we need to move its generation in a separate task so it can
be replaced by end users.

Moved bundle manifest generation in its own task that uses
the merged manifest, perform bundle related processing and
generates it.

This has the potential to create regression as we occur a
extra reading of the merged manifest. On the other hand, we
will only write out the bundle manifest when needed so there
should be some improvements as well.

Test: unit tests + existing.
Bug: N/A
Change-Id: I13bf9a28fe4b744f88e511c69b24c4cbbc1eac36

Former-commit-id: 983fe0f22019ca975c2e1efc722bbc6b5f2a8ade