Add AaptOptions interface
This CL creates an AaptOptions interface which users will be able to use
to set aapt options per variant. The values of AaptOptions' properties
will be initialized from the corresponding DSL values. Tasks will be
modified to use these values instead of accessing the DSL objects
directly (in subsequent CLs).

Bug: 151103668
Test: existing
Change-Id: I0cfca6ffcf7083f4be1c13d5b81cb6a52c5606f3

Former-commit-id: 624392b7818f3a3f1e74e56758cd87565b668d83