Disable InvalidPackage check and tweak Cordova check
This CL disables the InvalidPackageDetector by default; it's observed
to be expensive in some scenarios and not adding a lot of value
anymore.

This CL also tweaks the CordovaVersionDetector to (a) iterate the
assets folder using a file walker (which is potentially more efficient,
especially on Windows) and (b)to only do pattern matching on lines
that look likely to contain the version number.

It also bumps the required Cordova version, after looking up some
recent security bulletins.

Test: Existing
Change-Id: Iddec6a38b82b16cb9fbc319c21ceb4f9b9a2d7d2

Former-commit-id: 266e3288b46c1cf05d85dd2ccbc8fca9f8388f70