Adds instrumentation for resources to IWI
Two new instrumentation hooks are added:
 * ResourcesManager applyNewResourceDirsLocked()
 * AssetManager$Builder build()

The ResourcesManager hook is called during update-appinfo, and inserts
the modified/added resources over the app resources.

The ResourcesManager hook is also called during app startup, and handles
persistence of changes between app starts.

The AssetManager$Builder hook is used to ensure that changes to resource
ids are appropriately propagated throughout the app (methods such as
Resources#getIdentifier)

In addition, three methods in AssetManager are instrumented such that
their first argument - a "cookie" parameter - is set to zero. This
is the parameter that determines where the framework looks for a
particular resource.

To avoid issues with our resource overlays, which currently use a
full resource table extracted from the APK, we set the cookie to zero
to ensure that the framework doesn't only look for resources in our
overlays.

Bug: None
Test: Manual
Change-Id: I109a0cae37e0ad79e81ecd8bbcbf9dcb7d26e9d4

Former-commit-id: c17004b7bc60441bf1d3dc14b844e42e6339dac4