Fix friend paths for Kotlin compilation
When compiling test variants, Kotlin Gradle plugin
is relying on existence of TestedComponentIdentifierImpl
in order to detect which classes with "internal" modifier
should be accessible. After http://ag/10233377 we regressed
as this was removed.

This CL creates artifact collection with tested artifacts
in order to fix this issue. This is necessary in order for
AGP to continue working with older Kotlin Gradle plugins.

Bug: 150500779
Test: KotlinTestCompilationTest
Change-Id: I2de2a2cb7b847b49d2217c844627b7fa4972d4aa

Former-commit-id: 4a74f11ab78833be78c03287f14f84c814374bce