151719769: Offer stable update versions of the Gradle plugin
Lint currently tells you if there's a newer version available
of the Android Gradle plugin.

If you're using 3.5.0, we'll tell you that you should update
to 3.6.1. However, we've also released dot dot releases to
3.5.0. It would be useful to also tell users that 3.5.2 is
available, such that they can at least update to that (which
may be simpler). That's now provided by this CL.

Test: Unit test included
Fixes: 151719769

Change-Id: Ib73370a579436b4dd888af80ca3cbe3ad52e0308

Former-commit-id: 13b7d612e2950e4aee812c119d21b6201d64917e