Decouple DataBindingExpBuildInfoTask from LayoutXmlProcessor
DataBindingExportBuildInfoTask used to generate a class that contained
the build environment info needed for data binding, but it is now no
longer the case. This task now simply creates an empty class annotated
with a data binding annotation to trigger data binding to run.

This CL simplifies that task and decouples it from the now-unnessary
LayoutXmlProcessor lambda, which will unblock upcoming work on
LayoutXmlProcessor.

Bug: Clean up
Test: Existing tests
Change-Id: I302dd689de29aecb439bb51023b9fac12379dccf

Former-commit-id: bf456632ccd2ac0ed02a748dab7bb64153defa84