Remove LayoutXmlProcessorDelegate
This is not fully needed by DataBindingExcludeDelegate (only
need the packageName) so we can skip it there.

So now LayoutXmlProcessorDelegate was only used on MergeResources
but we can instead just instantiate LayoutXmlProcessor in the
task and make sure the task has the right input.

This also make the implementation of the OriginalFileLookup
reuse the MergingLog instance instead of recreating on each
lookup. Since the instance will parse the folder during creation
this should make the process faster to just parse it once.

Bug: N/A
Test: existing tests
Change-Id: If883936b9698d19020acfb748e1531b0b29f9aac

Former-commit-id: b872280a5e9ae4cc2b3d299042f8089d0d8ed91e