removed properties elements from build outputs.
as it turns out, information written by the process
manifest task in the properties of the variant outputs
are not used by anything any longer.

this will allow separating the main manifest merging from
the multi-apk merging as the main manifest merging will not
have an output.json associated any longer.

Test: Existing
Bug: N/A
Change-Id: Ifbd70ee845a6aff10c0e480a3d711d9dfee1ae7a

Former-commit-id: 73e8158d4ec2cd66fc634cc3daa1ce2e26b446f7