Handle UInjectionHost instead of ULiteralString in Kotlin
In Kotlin 1.4, the Kotlin UAST plugin is planning to change
the behavior of Kotlin literal strings such that they are
always nested in a UPolyadicExpression. This breaks existing
code which assumes that it can directly look at expressions
(such as method call arguments) and see if they are instanceof
ULiteralExpression.

This CL does two things:
(1) Updates lint such that it correctly works in this scenario
(2) Updates the lint unit testing infrastructure such that for
    testcases which include Kotlin sources, it runs the lint
    checks twice, with and without the UInjectionHost change,
    and asserts that the results are the same.

Test: Existing unit tests
Bug: https://youtrack.jetbrains.com/issue/KT-27283
Change-Id: I8ddc460b8d69ebab736162ad26256d1c90ceffa3

Former-commit-id: 66ec07041aafb1bfac72dba809f87ff4a685f5a1