Move outputs to app prop only.
Previously on ComponentPropertiesImpl but they
only make sense for app module. All other variants
do create a single output internally but it's not
accessible via the variant API.

The outputs are still there under the hood as they
are used everywhere. This change also makes it
so that the single output of dynamic feature
module properly contains the versionCode/Name
wired from the published info of the base module.
This makes reading its version code simpler in
consuming tasks.

This change also ensures that setVersionCode/NameOveride
in the old API does not do anything for any other
type of modules than app.

Bug: N/A
Test: existing tests
Change-Id: Ib805351b87da80c189d6afa0e4e380d817169cc3

Former-commit-id: fcb004cc3f2d15cc2290cd7f86ddc7d58074692e