[iml_to_build] Support cross-module test friends
This adds support for iml lines like this:

  <component name="TestModuleProperties" production-module="<name>" />

That line specifies a module that should be treated as a
Kotlin 'friend', allowing access to internal members.

Note: Kotlin friends should generally not affect the connected
components of the module dependency graph, because Kotlin friends
will generally be declared as normal dependencies as well.

Fixes: 152503545
Test: updated //tools/base/bazel:iml_to_build_tests
Change-Id: I80524fd1c169977668a2c2bc276f9186078d2fea

Former-commit-id: fdc71cbbf4643b033d30b0443522964022fa5f89