Convert ExternalNativeJsonGenerator to kotlin.
This is the conversion with just enough other changes to keep things
working. The only real change here is that a few previously protected
members of the class are now public, because they were actually
accessed outside the class and what appears to be a compiler bug was
allowing that previously.

Test: treehugger
Bug: None
Change-Id: I5308b4e0f0038b5b9fa2fff99e25b9e56dafa939

Former-commit-id: 673b18beaacc528c752fad0480d7a60ef2e1e74e