Move Deployment Builds to /build/intermediates
When the IDE injects device information to the build and
the build produces an APK for that specific device configuration,
the APK is placed in a different location under /intermediates
to differentiate it from other build outputs, since such an APK
is not guaranteed to work on all devices.

For buildbenchmarks, this means that APKs will be in a different
location when the "fromAS" variant of each benchmark runs. The
BenchmarkProjectAssertion now passes along information about
whether or not the build is running in "fromAS mode".

Bug: b/69014026
Test: Updated tests
Change-Id: I4199a0e2b892cbf2a1f39990613789d630718580

Former-commit-id: 096e8050f4dc1a0de84e58de0727b7a1b95bb8bb