Cache loaded symbol tables
The tables end up being loaded repeatedly during the build, and this
change introduces a soft reference cache for them to save memory and
load time during the build.

This is implemented as a Gradle Build service with two concrete
implementations - one caching and one that preserves the current direct
loading behavior.

This change uses that service for library symbol tables, a following
change will migrate the application and instrumentation test
implementations.

The caching is behind an experimental boolean option which currently
defaults to false, except in one targeted integration test.

Bug: 152323103
Test: Includes new unit test, SymbolTableCacheBuildServiceTest
Test: Covered by existing AGP integration tests
Test: Integration test UnitTestingRClassLoadTest expanded to include case with the flag manually enabled.
Change-Id: I36a0700aa944c8c28c8cb28983f77c4143894c22

Former-commit-id: 7314ac21814eed38b068f759da07fe4346fd7fac