Pass user managed ADB settings in ADB#init
Currently user managed ADB server settings are parsed
from an environment variable.  We need a way to set this
from initializers of AndroidDebugBridge, so this CL changes
this to pass it in with ADB init options instead..

Usage of ANDROID_ADB_USER_MANAGED_MODE is removed.

Bug: 153094651
Test: Existing tests pass. Manually toggled the management state
to verify the port switches over correctly.

Change-Id: Ib15e733ed2cda8eb1ce21ee7abb7e8faa78e5030

Former-commit-id: 3aedc7e78b5f73164e6054272bad4ae9c4dfe0fa