Compress assets before PackageApplication task
This CL creates CompressAssetsTask, which compresses the assets files
before they're packaged in the APK, to remove some of the bottleneck at
the end of the build.

This task outputs a directory of single-entry jars (instead of a single
jar) so that the downstream packaging task doesn't have to manage its
incremental state via a KeyedFileCache.

Bug: 135275558
Test: CompressAssetsTaskTest, existing AssetPackagingTest
Change-Id: I8d94205cb5f23316f8672f802b35b3e4dc9e0e2e

Former-commit-id: f1f87281ab2cb8dc5b06b8fe12d62721d25a47f8