Use new-style path if avaiable
Sometimes we end up with an sdk package that
- doesn't have a package.xml
- is recognized by the legacy sdk support as a legacy package
- is actually a (different) non-legacy
Unfortunately this happens right now when you download and unzip
cmdline-tools: it's recognized as the old "tools" package, since the
top-level directory is "tools".

We should use Pkg.Path if it's available to set the path, rather than
relying on the legacy support.

Fixes: 150943631
Test: LegacyLocalRepoTest.testReadNewInLegacyLocation
Change-Id: I110f44e8260b832a42dc324cd35d60a875428aff

Former-commit-id: 5860c69aaba388ca908048e8e463688a9baae9a5