When building subtracting artifact collection add build dependencies
When building a final file collection in the subtracting artifact
collection make sure to add build dependencies of the main and
tested file collection.

This is necessary because artifact subtraction happens only on the
transformed artifacts. For e.g. dexing we will dex all tested
artifacts, and those will be subtracted from the test dex artifacts.

Fixes: 154545489
Test: DexingArtifactTransformMultiModuleTest
Change-Id: Ia7a72d0cf7533bf18a0952f71c1724a0acc2b9fd

Former-commit-id: df9392b3ed5dd90379a080a72b73cced831000d4