Deprecate cleanBuildCache task
In a previous CL, we deprecated the AGP's build cache and also removed
the cleanBuildCache task.

For better user experience, we need to deprecate the cleanBuildCache
task first instead of removing it in 4.1.

In 5.0, both the option to use the build cache and the cleanBuildCache
task will be removed together.

Bug: 144105090
Test: CleanBuildCacheTest
Change-Id: I919e3a7c63467ffd8c157114f5d49ac9984b9010

Former-commit-id: cfefcdb61f7cc834e3a824a52bd065c12f576d8d