Target C++14 on our C++ builds
Currently when we build C++ code we might target different C++
standards in each of our enviroments (linux/mac/win), because
we're depending on the compiler default for that environment.

This on the best case cause issues on pre/post submit that could
have been caught before hand, and on the worst case different
behaviours in each environment which might be very tricky to
hunt down.

Bug: n/a
Test: existing
Change-Id: I7028e3fc7ad632fa494d3dfaf91b8bf198c4495f

Former-commit-id: 1998d6f2b57c079bdc2484cb12d1bb6b2afdfe6c