Start using SDK build service in the build
Make sure to inject SDK build service in all tasks that need it.
This avoids any unnecessary initialization work during the configuration phase.

For some SDK components, such as `AndroidJarInput`, `@Nested` is
used to ensure correct input annotations. Other tasks, manually
inject the service and query for components. It is important to
notice that most of the SDK components should not be kept as
task properties as that would require resolving them when
serializing task graph for instant execution. To avoid this, while
still having the correct up-to-date checks all SDK components
should be returned by functions.

Bug: 152321793
Test: existing
Change-Id: If02d3a22e00d99d66be419547f7ff6a6b9ed090d

Former-commit-id: 0fe8a751a8935665952b6c2531a24c739018d850