Generate source file lists for baseline coverage
Baseline coverage refers to a coverage report without running any tests,
or equivalently, the set of all instrumentable code. The current system
instruments only code that is loaded by some test, which creates a gap
through which never loaded code is reported as uninstrumented rather
than uncovered. This can happen if either a test does not exist or
if it is not part of coverage runs for some reason (forgot, flaky, etc).

This gap can be closed by collecting baseline coverage and adding it to
all coverage reports. Since all our coverage reporting systems flow out
from the custom report system this ensures correct, consistent reporting
in all views. The baseline coverage system is opt-out to ensure nothing
is missed by mistake.

The high level flow of baseline coverage will be:
1. All source targets generate a list of coverage eligible source files
2. All these lists are merged into one
3. The results XML->LCOV conversions use the list to drop irrelevant data
4. All source targets generate LCOV of their baseline coverage
5. All these LCOVs are merged into one
6. All coverage reports include this merged LCOV
This CL covers (1) and the rest are in following CLs.

Uses macros to create two new genrule targets for source targets:
- List the source files of a target to an output file
- Filter the list of source files using the coverage master ignore list
This produces a list of all coverage eligible source files in a target.

Two macros are used to do this throughout the codebase:
- coverage_baseline: produces the two genrule targets
- coverage_java_library: wraps java_library and coverage_baseline

We use coverage_baseline directly in all our custom macros and we use
coverage_java_library to replace direct uses of java_library in BUILD
files for source targets we care about for coverage.

Macro modifications:
- iml_module: Feeds source java + kotlin arguments to coverage_baseline
if any exist.
- kotlin_library: Feeds java and kotlin inputs to coverage_baseline if
the rule is not testonly, which filters out a couple instances in the
codebase of test libraries built w/ kotlin_library for some reason
- maven_java_library: Added a new argument to flag whether to produce
baseline data (defaults to True). Only produces baseline data if sources
exist and the flag is True. The flag is necessary b/c of usage in
java_proto_library.
- java_proto_library: Flag the maven_java_library to not produce any
baseline coverage. We don't care about the coverage of generated
protobuf code.

Change-Id: I49ac3e3d3eea6c7f82c7e44290369a69d4dd05be
Test: N/A
Bug: 153913236

Former-commit-id: c2875398505eb2fe560953a34d6ee40a7258f73d