DSL: Don't parameterize AnnotationProcessorOptions
This is a potentially breaking change for Kotlin script DSL users,
as the type of the annotation processor options block when used from
Kotlin DSL is now the API type.

It's a little more subtle for builds that uses Kotlin in buildSrc.  If
the buildSrc build applies the kotlin-dsl plugin, it falls in to the
same case as the Kotlin script DSL where the type of the annotation
processor options block changes from what it was before. Without that
plugin it already changed, as the explicit addition of the
annotationProcessorOptions(action: AnnotationProcessorOptions.() ->
Unit) method, the action block binds AnnotationProcessorOptions to
this rather than it.

This is not a breaking change for compiled code, due to type erasure.

Bug: 140406102
Test: API tests updated
Change-Id: I3aef9cc936afa43113cf7d0f933ae6676fb4f766

Former-commit-id: 7e882f5e355010ce90bb4734ce658a9e94a90166