Support non-transitive R classes in app modules
Add a flag for generating non-transitive R classes in app modules. The global flag only affected libraries since the final R.txt was generated by aapt2
we'd generate the R.jars based on it. Now if the separate flag for the app module is enabled (and the global one is enabld too) we can parse the local
app resources (from small merge) to get the list of local symbols, and generate a non-transitive R.jar for the app module.
This most likely will have a negative impact on the build speed, but will greatly reduce the R class size. Because of this, and because of being a breaking
change, this needs to be opt-in.

Test: NonTransitiveAppRClassesTest
Bug: 145944388
Change-Id: I329a74975df14481efb53e5e05e51d1172de4358

Former-commit-id: c868d81a1a3860add39a9179b785000480d489ea