Merge baseline LCOV and use in all coverage reports
Merge all individual baseline LCOVs into a single baseline LCOV, which
then is used as an additional input to all coverage report definitions.

This ensures that all coverage eligible files that meet the
include/exclude rules of a coverage report definition are actually
included in the report. Previously if some file was eligible, but never
actually touched by a test then it would be omitted due to lack of data.
With this change, such omissions will be detected and included in the
report as a file with zero coverage.

Change-Id: I1a4b14ebf8f75e342786cb7917e3503d4e95207e
Test: N/A
Bug: 153913236

Former-commit-id: fa50dc42616906ce28dc667184fda538725f3830