Find and merge all baseline coverage sourcefile lists
Uses the WORKSPACE loop trick to find all baseline source file targets
by looking for their ourput and constructs the merger rule with all
the rules as the sources.

We cannot express such a dependency in Starlark normally due to a mix of
factors (no wildcards in BUILD files prevents using a genquery to find
the rules and even if we could do that we couldn't use the output file
to populate the `srcs` of a genrule anyway).
However, by using a WORKSPACE loop into `bazel-bin` to glob for files
that look like the rule outputs we can identify all the rules and
construct the merger genrule correctly.
The downside is that we need to do a targeted remove of the
rule outputs to ensure we don't pick up anything from deleted targets,
but we can mechanically enforce that in both places we invoke coverage
processing (tools/base/bazel/studio_coverage.sh and
tools/base/bazel/coverage/report.sh).

This is similar to using a script to construct a BUILD file and then
invoking a target in that BUILD file, but this method does not pollute
the workspace filetree.

Change-Id: I7132af7aea7f8b2a65b393801ca7962453cd02a9
Test: N/A
Bug: 153913236

Former-commit-id: 51157d603c2bceda3863d6464d2c7c095fcfa5c2