Extract and refactor DeviceId
CL 1/many in a larger test restructuring.

Refactors DeviceId into its own top-level class, and moves the
responsibility of storing API level and version from FakeDeviceLibrary
to DeviceId.

This ensures that a DeviceId can always be converted to its respective
API level and version, outside of the context of a FakeDevice.

This will be used in subsequent CLs to drive JUnit @Rules based on
DeviceIds.

Bug: None
Test: All tests in android.sdktools.deployer
Change-Id: I617ca63508d5e9188e868e11214a8552518e70bc

Former-commit-id: b15d8f315827f08485b826892fde198058bd79d7