Add Lint check recommending KTX extensions
The list of KTX extensions available is hard-coded from
https://d.android.com/kotlin/ktx/extensions-list#androidxbenchmark

There are a few KTX extensions which do not have a directly
corresponding Java-only artifact (e.g. androidx.work:work-testing).
For now we ignore these, because there's no obvious place to
surface a Lint suggestion for them.

The suggestion to use KTX is only offered when the project already
uses Kotlin---specifically, when it already depends on the
Kotlin stdlib. (This heuristic might fail if a project picks up the
stdlib transitively without actually containing Kotlin sources.)

Bug: 153760933
Test: unit test added
Test: manually checked that all artifacts in 'depsWithKtxExtensions'
      exist and have corresponding KTX artifacts.
Change-Id: I36eb8f9fb3d6687ce63a29c794c1751fe99b912d

Former-commit-id: 5595cb0ad5a77f1e7b0fb1d1204fbc98ea091320