Add conditional --worker_max_instances when executed locally.
The $BUILD_NUMBER environment var is expected when executed from
buildbot. If provided, allow bazel to determine the appropriate number
of persistent workers. If not provided, limit the persistent workers to
2 instances in an effort to avoid overloading a users development
machine. If bazel needs to perform a huge amount of build actions,
spawning too many actions locally can cause loss of interactivity and
lock the host machine, leading to much sadness :(

Bug: 154933248
Test: none
Change-Id: I92cc51f7eee7bb958edaa10b719a7d7651db99b7

Former-commit-id: 963e484e3112abdf2adbdea2fd7ca72d0b76e3a7