Change StackFramePointers to use Map
Using a map instead of a list saves memory in studio.
Studio previously enumerated the list of pointers and
constructed a map for quick lookup of frame parents
and frame names. Converting this to use a map means
that studio no longer needs to duplicate the structure
for O(1) lookups. This is an O(n) memory savings in
studio.

Test: N/A
Bug: 149541949
Change-Id: I9ce5534b85d4b52663ceffd805f950dc0ebb7945

Former-commit-id: f54a5ae46ec7039bcc6c20220cb4bf37481f3718