Add unhandled exception logging to agent
- Adds instrumentation to Thread#dispatchUnhandledException to detect
when an unhandled exception has been thrown by an application. Data
is written to a log file in the application's data directory.

Each instrumented process writes its own log file.

Note that this is a pass-through operation, and the normal exception
handling of the app/framework are unaffected.

A log file is also written if instrumentation fails in a startup agent.

- Adds a method to the install server to retrieve the exception logs of
a particular package. This method is called from overlay swap, after the
swap has completed.

We retrieve logs on a best-effort basis; because the agent can only
write logs to a SELinux protected space, we must rely on subsequent
swaps to retrieve the data written by prior swaps.

Moving forward, a better solution may be needed.

- Adds some basic log recovery; the logs are forwarded back to studio,
but they are only logged in idea.log. A future CL will report the log
data as metrics.

The logs only contain basic statistics about the agent at the time of
crash. Further fields may be added.

Bug: 150382039
Test: CrashLogTest.java
Change-Id: I82c8a8441ffdb73c1283d6da08427a28050d2723

Former-commit-id: aa2e2511e353fca46b9cdd92371701b2e73c8c3e