Fix profiler freeze on Android R
Added an option to kill command runner without blocking on the callback
thread.

This doesn't fix the underlying issue on statsd side where no data are
sent to perfd, but makes sure Studio doesn't freeze when stopping
StatsdSubscriber. It should also reduce the latency while waiting
for StatsdSubscriber to finish, which could take up to 1 second even
when the statsd bug is fixed due to polling frequency.

Fixes: 154958464
Test: manual
Change-Id: I890b8ce2be425637687a7c0059337997ae0abf94

Former-commit-id: 83561903d15372a88d14b20599be80ff62009e5d