Remove dexing outputs only for changed files
This CL fixes an issue with incremental dexing. More
specifically:
- we run incrementally only if file containing mapping from
  paths to file hashes (i.e. output locations) is found and
  loaded succesfully
- once the mapping from the previous run is found, we will run
  incrementally only if the list of files from the previous run
  is identical to the current list of input files.

This means that if an input changes in some way, but path sensitivity
is not considering this a change (say a file name changes for input
with @Classpath), we will remove output of this file. Previously,
we'd fail to compute the hash for the new absolute path (as it
was not reported as changed), and when determining the set
of unchanged output locations we'd crash.

Bug: 154712997
Test: DexArchiveBuilderDelegateTes, DexingArtifactTransformTest
Change-Id: I413c7fef510d4db33bbf6a842b6c4e92ca734ed5

Former-commit-id: 54f3276d926131724961aef3296b5a2b1064df41