Introduce new APIs to obtain value of BooleanOption
For each kind of Option(e.g. BooleanOption), there will be two APIs to
obtain its value. One is to return the actual value which should be used
if we want to use the value at configuration time. The other one is to
return a provider which has the actual value to be evaluated at
execution time. Note we should not invoke get() method on the returned
provider of the later API at configuration time.

This CL also introduces a OptionValue type which contains two providers
which provide the value of an Option. One provider is for obtaining gradle
property value at configuration time, the other one is for use at
execution time.

Bug: 155766476
Test: BasicInstantExecutionTest + existing
Change-Id: If0089a5a408c165d7280b1847f44f252b8a09538

Former-commit-id: 7132159b82c1841913a454b79464cd5db1f8a53e