Remove PrepareLintJar
This was used when there was a single lint configuration
for running lint inside a module and publishig it. Since we
can only publish a single jar, we needed to ensure there was
a single dependency setup.

However we have since split the publishing and the local
setup of lint jars containing rules to run. The new publishing
has a similar checks but the check on lintChecks was left as-is
making it impossible to depend on multiple jars containing
lint checks.

While we don't need a prepare task, we do need a way to
trigger compilation of the lint check so this change
introduces a `compileLintChecks` lifecycle task that
can be used to compile them. For now this is still
triggered by the generateVariantSources task but
this will be moved into its own action in the future.

Fixes: 141840950
Test: existing tests.
Change-Id: I91d27d1fb96d239c309c135587b67a10854f8489

Former-commit-id: 4163f33b0ad7420461274b6307cdfb1e72ea19bc