Add lint rule jars to the model and make lint use it.
This adds the list of rule jars to both AndroidProject,
and LmModule, as well as handling the conversion from
IdeAndroidProject to LmModule and the serialization of
LmModule.

If the value is null, then the model does not contain
the information (ie old AGP) and therefore alternative
ways must be used to find the information.

Fixes: 66166521
Test: existing tests
Change-Id: Idd48e1e0538a3f3ef5401533fb92fadeca0b7114

Former-commit-id: 0d76cc0f48bd35841015f398f4b2a422f4d77303