Replace synthetic properties with findViewById
The path name of the synthetic properties needed to be a variable
otherwise the path is considered as wrong when a template is generated
for a new flavor.
E.g. "main" in import kotlinx.android.synthetic.main.${layoutName}.*

This CL replaces them with findViewById to fix the issue, also they
are likely to be replaced by view binding

Bug: 147916973
Test: TemplateTest, Manually created affected Activities, Fragments
Change-Id: Ieec669c41f1a8e1875d79d122b113aea8ab834f5

Former-commit-id: 815089e38669bf0c7b4e1b03668cf4229080f955