DSL: Expose ABI filters as val Mutable
It was previously nullable that would make this problematic,
but consumers intended to treat empty collection the same as
null, so now it is not nullable. This change fixes various
inconsistencies where some places did accidentally.

Not a source-level breaking change for kotlin script users
as the existing setter before 4.1 was a builder-style one
so did not make the field mutable from a kotlin perspective.

Bug: 140406102
Bug: 156169950
Test: Updated API expectation files
Test: Existing AGP integration tests, e.g. NoSplitNdkVariantsTest
Change-Id: Ibdea1b01c32cb5843f3cc350aa9deac6edabb013

Former-commit-id: 2ee2029d1c3bff187d2436646902ae702e288f5f