Remove VariantBuildInformation
Now that the dependencies are out of Variant, we can make
Variant an object that's always returned in the model.

This allows the IDE to know all the sourcesets of all
the variants, even for the ones that are not sync'ed.

Also move bundle only info into its own interface to
allow Artifacts that don't have bundle to just null
it out. We may need to introduce more classes to
simplify things further.

Bug: N/A
Test: N/A
Change-Id: I1452449f01a357f8f7f34cb2869514d2aaef873f

Former-commit-id: df804f7cc305cc925cc09865edd26f7a0d66bd17