[AGP DSL] Infrastructure to generate DSL impls.
This change introduces the AGP DSL decorator,
which similar to the Gradle instantiator, generates
a subclass of an abstract class at runtime.

The generated class is then passed to Gradle's instansiator.

This change replaces the use of a kotlin delegated property for
compileSdkVersion with an automatically managed, lockable field.

(The implementation uses reflection to access the APIs to declare new
classes in existing classloaders safely, to avoid forcing AGP to be
compiled against java 11 APIs just yet.)

Bug: 140406102
Test: Added AgpDslDecoratorUnitTest to cover the existing corner cases
Test: Updated KotlinDslTest.testDslLocking for the new error message.
Test: It is used in one example in AGP, covered by existing tests.
Change-Id: I86f0a48faa0e740b5b30f37fee98fffec10d58ea

Former-commit-id: bafd946f0d47cb5e4fb9328afbe2c0e2265208de