Shutdown TPD if inactive for long periods
Add a 2nd thread that keeps checking when was the last RPC received by
the daemon and shut it down if is has been longer than the value
specified in --server_timeout flag (defaut: 1H, 3600 seconds).

This should prevent that on a Studio crash or forced shutdown, where the
services where not able terminate the process properly, we'll not leave
a dangling process in the users systems for ever.

Bug: 156027092
Test: manual
Change-Id: I374693410bb43bc95a613394c201111a5c66b86c

Former-commit-id: 71e16f1ce0727fbd4bd9097a09e8a132ffe1c27f