Don't attempt to copy runtimeFiles to themselves.
Libraries built by the module will show up in the linkLibraries list
for other targets in the module that depend on them. This wasn't
showing up in the typical case because we process linkLibraries before
a build happens and we filter out any files that don't exist in an
attempt to filter out any inputs that aren't actually libraries.

When the built libraries *do* already exist (such as when the JSON
model is regenerated without the out directory being cleaned), this
could result in the source and destination of the copy rule being
equal. When the paths were *exactly* equal (string equality, not
normalized path equality) this causes Files.copy to throw
IllegalArgumentException.

Whether or not the files were exactly equal or just equal after
normalization seems to be dependent on the environment. Within our
test environment the paths were being given relative to the build
directory rather than as absolute paths. CMake does this for source
files to keep paths short, and presumably the same is happening here.
I also fixed those cases here so the test would actually test the
right behavior.

There's another issue here not addressed by this patch, which is that
it's not really correct for us to filter out the libraries that
haven't built *yet*. I'm wary of making that change in 4.0.1 so I'm
addressing that separately.

Test: Added test
Bug: https://issuetracker.google.com/158317988
Change-Id: I32f854d406c6c793e70fc8dbc8cecc1192d80230

Former-commit-id: 4fb954a71253ae17fa5383fa0b9a03bfb8aa1237