API: Expose public android resources file
The internal artifact has the difficult semantic of
not existing if the user hasn't explicitly used the public tag.
That is, empty file means nothing is public, whereas file does not
exist means that everything is.

For API consumers we instead construct the file that lists everything,
separately from the public.txt that is included in the AAR.

Bug: 154626581
Test: Unit test: Added GenerateApiPublicTxtTaskTest
Test: Integration test: Added KotlinScriptApiTests.getPublicTxt
Change-Id: Ib616044d92364ecf2609fb69526d33987c683ed6

Former-commit-id: 98af9e606ddc949f127a36dab3b02a13becd5b82