Add synchronized wrapper around SocketChannel
Cleanup is done on the main thread. This can cause the shutdown
of the server to hang, cause a NPE or other threading issues.
To resolve this a guard was added to the SocketChannels managed
by the server. The guard allows the JdwpProxyConnection thread
to block or finalize the transition while the server is stopping.

Test: JdwpProxyServerTest.secondaryServerStartsInFallbackMode
Fixes: 157903373
Fixes: 157195752
Fixes: 155515207
Fixes: 157244508
Change-Id: I9770ab1ba1af1f7e2eebc692a555dcf9c7c65f53

Former-commit-id: 712fa5e4c9f74f927708f690b5bdc1fcb8c30b8c