[TPD] Let the OS assign us a free port
Instead of forcing a port through a flag, it's more reliable to just try
to bind to "port 0" and the OS will give us a free port.

This value will be read by the DaemonManager on the IDE, which will then
use to create the channel to the correct port.

Bug: 149379691
Test: manual
Change-Id: Ib438d1e91769446ec301a9d53045ce82e1cdccc2

Former-commit-id: 53013b8c49b55b12d734bf49cca6d5811122b7ff