Fix missing and trash  build files in Native Model
Bug: 152223150
Test: Added an integration test

This never worked after single-variant sync. No build files were
returned in the case that 'inexpensive' metadata was requested.
Instead, we should at least return the build file that we know
about.

When expensive metadata was requested, CMake server was returning
trash in the list of build files. It included stuff like .git/HEAD!?
This change simplifies that code and reduces the list to only
CMakeLists.txt.

Change-Id: I81c9602ef83ebbd589af8f2bcab9df2d6660a80c

Former-commit-id: 0f65f339f7c8f998dfb1afbb5c93fbabc29d6c74