Fix test runner arguments issue for configuration caching
Instrumentation test runner arguments can be specified from android
gradle dsl or as gradle properties. Those arguments is categorized into
two groups: standard args and custom args. This CL is to address
compatibility issue between configuration caching and specifying runner
args as gradle properties.

For standard args, we make it fully compatible with configuration
caching by introducing a comprehensive list of standard args and let
Gradle monitor them for each build.

For custom ones, it is not fully compatible with configuration caching
because we are not able to let gradle monitor arguments that don't
exist in the first run but appear in the following runs. In other words,
when users add new arguments in the "second" run, we would fail to
invalidate configuration cache and end up ignoring those new arguments.

Therefore, we encourage users to migrate away from specifying custom
test runner arguments by outputing warnings. In the meantime, to
avoid this to be a breaking change, we still allow users to pass
custom args as gradle properties and are able to capture changes
to the existing args from the "first" run.

Bug: 155766476
Test: existing + GradlePropertiesTest
Change-Id: I5dbd6319e2214e382d615dfef1688dda666e11eb

Former-commit-id: 73e53320032cbe1a2f540014aeda9e50bbe26a00