Register clean build cache task only in the current project
If configure on demand is used it is possible to start executing
tasks from root project before plugin is applied. This means that
we may end up added a new task to the root project too late.

This CL changes how this (deprecated) task is registered, and
it is added only to the current project.

Bug: 158468794
Test: CleanBuildCacheTest
Change-Id: I7bd0129c96c81a04a136f8768b47262c7151ab1b

Former-commit-id: c9b333aa07dde11dcc7d8e8a5b92c2e9400aa00e