140699627: Fix handling of generated source folders
This CL fixes this bug:
Fail when using StringRes as field of an enum used as SafeArgs argument

The root cause was that we don't handle generated source files
correctly; in particular, any Kotlin files in the generated
source files are not sent to the topdown analyzer, and as a
result, resolve() into generated sources (in the case of this bug,
a NavArgs generated Kotlin class) will fail.

(We also have a fallback in the resource reference lookup where we
assume that unresolved symbols in a compilation unit which includes
synthetic imports are probably id references. This CL will also
mark these results as being heuristic, and in the case of the
resource evaluator we'll ignore heuristic results.)

Fixes: 140699627
Test: Unit tests included
Change-Id: I734821282d433bc059f8f933ff7cb07f39db958d

Former-commit-id: 8ae31ebee5e8edce870d361b48ecbdfe91a68fdc