simplified private API for workers based tasks.
The original submitWithProfiler tried to be made available for end
users but it proved to be too complicated/constraining on top of
complicating the implementation.

Instead force AGP tasks to subclass simple interfaces when declaring
work parameters and work items. Such decorated can then be submitted
and proper work allocation will happen.

Changed only Task to use the new facility to help with code review,
will change all other tasks plus the ones that used the old
deprecated API in a subsequent CL.

Test: Adapted existing
Bug: N/A
Change-Id: If409ccca9869bb56681e8346599d41511355c79f

Former-commit-id: 5470e87978dbf6edf47dbb59e7a4f47211f45596