Fix statemachine for the terminology detector
On mismatched input it needs to jump to any state
that captures the input matched so far. For example, if
we have matched "foobar" but then get an unexpected character,
instead of just going back to the init state, we jump to
for example the "bar" state, if it exists (we look for possible
states oobar, obar, bar, ar and r).

Also took the opportunity to fix a few other things: make the
error message include the whole word surrounding the term,
if different, and fixed a problem computing error locations.

Test: Unit test updated
Bug: Not filed
Change-Id: I8de0e048678e341a81f56ab0276cd8b30e0c2d2f

Former-commit-id: 627fc04f3ada030ff798f743ba4876bc7a1d9efb