Fix the issue New Wear OS module is missing @color/dark_grey
The color is coming from the com.google.android.support:wearable
dependency, which is added through the wearModuleRecipe.kt, but
the code isn't executed when creating a new module.

So adding com.google.android.support:wearable is needed explicitly
for the blankWearActivityRecipe.

Note that there is a similar library named com.android.support:wear,
which is mapped to androidx.wear:wear for the AndroidX version.
But it looks like it provides different functionalities from
com.google.android.support:wearable.

Bug: 159103915
Test: Manually followed the steps in the bug
Change-Id: Id0eee944d20e8d888cdf77a3d0ed98576479a398

Former-commit-id: 404b8a43932b18966db6b0ac8e6673c125fcaf77