Remove usages of WearableActivity from templates.
WearableActivity is soon to be deprecated in favor for
AmbientModeSupport from AndroidX.

This commit updates the project templates for Wear OS to use the
standard activity instead. Note that this remove "Always-on" behavior
from these templates, but since we expect most applications not to
require "Always-on" support I think the sample for "Always-on"
activities is the best place to add this rather than adding by default
to all apps build using the sample and having to let the developer
remove it if (most likely) it is not needed.

Bug: 159582162
Test: tools/base/bazel/bazel test //tools/adt/idea/android-templates:all
Change-Id: I047fd10b18c3c184268b084e6d7593117d4add91

Former-commit-id: 4e33f20e09ad5e278a3cc248aaeb16868d9efa9b