Add singlejar actions to dynamic execution.
This will run the singlejar action locally, in addition to remotely.
Actions which are run dynamically are executed in a [symlink] sandbox,
so local actions should not interfere with one another.

We see singlejar actions occasionally take a long time when RBE
workers are overloaded. This halts other java and kotlin compilation
that could take place locally on persistent workers, as they depend
on the output of this action.

Bug: 158102957
Test: Checking presubmit bazel profile
Change-Id: Ic0b8c6ac78f0c76f5c01c521f222c7cf4e0ef279

Former-commit-id: e3c7caa2ca96315a13cc28d66ded2db15f94c34a