Split android-ndk into android-ndk and android.ndk.tests
This is required to configure the correct class path for UI tests.

Note, that the dependency being added to android-uitests/BUILD is the
exact opposite to the ultimate goal of this change. This change removes
removes one part of the problem only making the other part explicit.
Future changes will remove this unnecessary dependency of
android-uitests.

Bug: 158590296
Bug: 156018904
Test: n/a
Change-Id: Ia5b46a703a8d1870f05074a768f528e5af76d93c

Former-commit-id: 1ac124905c20b6d0f615c7821405ece847278613