Update Incident to be more suitable for detector usage
Until now, the Incident (previously "Warning") class was used only
inside the CLI library and was used to record all the known information
about reported incidents in order to sort and present the results.

This CL cleans things up a bit such that the information supplied to
the Incident is limited to the detector visible aspects, and the
rest is computed lazily or done outside the Incident class.

This is further prep for moving Incident up to the Lint API where
detectors will report incidents directly.

Test: Existing
Bug: None

Change-Id: I0fb87d7acd2ac1e77640158253a40d64b98c1395

Former-commit-id: b3d06f5257366247fd5cda621b6d7e2df33a8b81