Preserve order of filtered and non filtered variant operations
The filtered and the non filtered operations were executing separately,
the non-filtered actions will be all executed first and then all the
filtered actions. This means there is no way for the user to change the
order of operations that two different plugins do if each use a
different API, which is a problem for APIs that depend on registeration
order (e.g. the new ASM API).

Bug: N/A
Test: VariantOperationsTest
Change-Id: I2089bb5368a8506157533a8c3a568350055496fc

Former-commit-id: e49ebac874f3c9a7d6a3c747dcea8604b47ebe7a