Add support for multiple projects to iml_to_build
In order to build the new project (/tools/adt/idea) a new set of rules
is created with a project id prefix (unb). Once these rules become
oficial we will remove the duplication.

Also add a small tool to assist with the copying of manual attributes
between rules.

All these rules are tagged manual and unb, so they can be found and
manually run.

Test: The new project creation will have its own tests
Bug: 156497102
Change-Id: Ief5de0ae76d9215bd2714c92b00c7c36af531983

Former-commit-id: e455aad71108fbfd788df28bf4b1cc6e893d2a23