Actually disable RequiredAttributeDetector
LintClient.isGradle() is based on a global variable that is not yet
initialized when the RequiredAttributeDetector class is loaded.
Thus LintClient.isGradle() returns the wrong value, and
RequiredAttributeDetector becomes enabled by default when running
in Gradle. This results in false positives that break the build for
several users upgrading to AGP 4.0.

I filed b/160626532 to track a better fix, but in the meantime we can
disable RequiredAttributeDetector entirely. This reverts behavior back
to before Change I3f2734224.

Note: I also had to update MainTest because RequiredAttributeDetector
happens to request a second analysis phase, thereby changing the
expected CLI output slightly.

Fixes: 155088391 (24 upvotes)
Test: manual (on sample project provided in 155088391)
Change-Id: Icb0c07efd77cc6fcf4c7e4c161ffa7e31974fe20

Former-commit-id: fe1a3bf8eb41f6d7c9dcd8093a57c081b1bef29d