[JDK11] Fix bootclasspath setup in formc action
The "sun.boot.class.path" property was removed in JDK9+. Instead of
having FormCompiler query for the bootclasspath, we're now explicitly
including it in the Bazel action setup code.

Luckily the @bazel_tools/tools/jdk:platformclasspath rule extracts
the needed classes from the runtime image into a separate jar, so
we can treat it just as an extra classpath jar.

Bug: 135535322
Test: Bazel build
Change-Id: I002db2f63d4dc1f637a85eeb1b7ae3c1e1b3ac7b

Former-commit-id: a37d76e37a6a81d259d27d1ece5404c23ee91d2b