Change perfetto config to better handle verbose traces
There are a few tradeoffs to consider when capturing data using
the linux kernel. The tradeoffs are a fixed size ring buffer,
the syncronization interval, and the amount of available memory
on the device. If we attempt to create a buffer too big for the
memory available the recording will fail to start, if we create
one too small we need to increase the syncronization internval
which may introduce latency, if we don't we may drop start/end
packets. After extensive testing on a verbose unity app
These settings provide the most accurate results with the best
trade off.

Test: N/A
Bug: N/A
Change-Id: I0973d831f843382771980eea0f6d04c6e77d6029

Former-commit-id: 6db20d434849c502c2e7ec9637ab3bd27e9529e5