Improve fixtures for model comparison
Added support for checking whether all the properties are
dumped. This already caught a couple of missing items.

This also adds the ability to regenerate golden files
via the command line.

Finally, this adds a base interface to all the model
interfaces. This is mostly used at the test level for
now but it's useful to have in general.

Bug: N/A
Test: Updated BasicModelV2Test and AppModelTest
Change-Id: Iccc59eee8cda1aeaa354150f6b3c9e35f8146d6e

Former-commit-id: 07bd16a80e8154247419db6048bb3da5daf2b934