Update PackagingOptions DSL
This change is an amended version of reverted change I16f6b955, which
broke androidx because they use the top-level excludes etc from
PackagingOptions. This change does not delete those; instead, we'll
deprecate and then delete them in the future after the tasks are wired
to the new DSL.

This change adds new blocks to the PackagingOptions DSL for
dex, jniLibs, and (java) resources.

In subsequent CLs, this new DSL will be wired to new VariantProperties
objects and then to tasks.

Also in subsequent CLs, assets and (android) res blocks will be added
to the PackagingOptions DSL with the ability to disable compression for
particular assets, res, or resources.

Bug: 129428538
Bug: 149770867
Test: existing
Change-Id: I4fba4708884d0138fe83e53d7751ae5c6fffdee1

Former-commit-id: 8a1691c1d08f1d41892c89de4ad5e50a0e681c5b