Make lint checks resolution lenient when syncing
If dependencies are missing we do not want to throw an exception
as this would break the whole sync and the IDE would not get
any information at all.

This commit makes the dependency graph for lintChecks resolved
in a lenient mode so that failures are instead recorded.

The current v1 model does not have any mechanism to return the
resolution failure. v2 does not either but we can add them there.
We'll keep using bug 161774530 to track this, and this will get
fixed in a subsequent commit

Bug: 161774530
Test: missing, will be done with the commit that add the failures to the model
Change-Id: Ic8c1eda9afdd8c57057d7ca34d1562941cf3b27c

Former-commit-id: 9ee5cfe895cdace2046982daf2f4d0327843c310