Re-add generated res values directory to the model.
It was removed in 82a1103f3a [formerly de05970702] in favor of having Studio model the res
values based on the model.

However, with the new variant APIs res values can be generated at
execution time, this is no longer a viable approach. This change reverts
back to including it as a generated resource directory, as was the case
in Android Gradle Plugin 4.0.0

Fixes: 155177757
Test: Updated Android Gradle Plugin integration tests
Change-Id: Iea1fd86d8571e906ae4bf0a9e2d4fd724a05b3d8

Former-commit-id: 874bd26b58e2b5a2ed6637d3927c9618aa2318cd