Update to support view transformations
This change has two parts: changes to the agent and changes to the skia
parser.
- In the agent, we send over the absolute screen coordinates of the
corners of each view.
- In the skia parser the logic is significantly simplified. Now, the
coordinates of each requested view are sent by studio. The parser now
keeps a single SkCanvas, and reads and returns the specified pixels from
that canvas. When a view node is created the entire canvas is cleared,
which is possible since all draw commands within a view node are
contiguous.
Unfortunately we still need to support AS 4.0 and 4.1, which require the
old behavior, so the old code is preserved in the _v0 files.

Test: pending
Bug: 157150680
Change-Id: I1275277b89240abb8a913022ef461a408f065a80

Former-commit-id: 152264b4cb471e498350a1a1aea722302d7df7d8