Fix registration of data binding artifacts when using Kapt
There are 2 ways to register data binding artifacts as outputs of
JavaCompile/Kapt:
  (1) Register with JavaCompile when Kapt is not used, and register with
      Kapt when Kapt is used.
  (2) Always register with JavaCompile, and when Kapt is used, replace
      them with Kapt.

The first way is simpler but unfortunately will break the publishing of
the artifacts because publishing takes place before the registration
with Kapt (see the bug below).

Therefore, we'll have to do it the second way.

=====

Caveat: With this fix, we'll also have to remove the workaround in
DataBindingCompilerArguments for Kapt bug
https://youtrack.jetbrains.com/issue/KT-39715 (KaptGenerateStubsTask
resolves annotation processor options early), because the locations of
the artifacts are no longer predetermined. However, that bug is far less
serious than the current bug, and the workaround for it was only added
in AGP 4.2.0-alpha05 so removing the workaround is not a regression to
AGP 4.2.0-alpha04 and older.

Bug: 161814391
Test: Updated DataBindingKotlinAppTest
Change-Id: I3ed5520fc33184475a42c19d19412a87cb2d386d

Former-commit-id: cea0b5923006a9eff4b19a4530cf496dc47da092