Revamp the lint.xml support in lint
The lint.xml file is the original way to configure lint (prior to the
ability to configure it via DSL in gradle files), and it continues to
have some uses cases not supported by the DSL, such as the ability to
suppress issues based on paths and regular expressions. It's also
relevant in the Gradle DSL world; the DSL lets you point to lint.xml
files, via the lintConfig method, and this lets you share a single
lint.xml file between multiple projects.

However, there have been a bunch of bugs and requests in this area; I
searched for lint.xml in the issue tracker and found 37 bugs. Some of
them have a lot of stars as well.

This CL revamps the configuration support in a number of ways:

(1) The lint.xml configuration is now properly used in the IDE as
    well.  Prior to this, only the enable/disable settings specified
    in the Gradle DSL were applied in the IDE; now the lint.xml files
    are wired up as well. (There were also some gaps in the way the
    DSL methods for enabling/disabling by category was handled; I've
    extracted the CLI configuration implementations and moved them up
    into the lint client API such that they can be reused on the IDE
    side.

(2) Some users were depending on the fact that lint.xml doesn't work
    in the IDE to write Gradle/CI-specific configuration (for example,
    disabling NewerVersionAvailable such that the CI servers never
    look for up to date versions). My fix for (1) would make this no
    longer work. So this CL also introduces a new attribute you can
    specify on the root <lint> or individual <issue> elements to
    designate that the configuration element only applies when lint
    runs in certain contexts, like "gradle" or "studio" or cli".  You
    can also prefix with "!" to specify negation, so "!studio" means
    "when lint runs outside of Studio".

(3) Lint authors have been asking for a way to pass data to their
    checks; until now the only way to do it is via environment
    variables or system properties, which is hard to do on a per
    module or directory basis. The lint.xml file is a good place to
    store this data. This CL adds a simple per-issue key/value lookup
    to the configuration interface and persistence. Longer term
    it would be nice to let detectors declare their own options,
    such that we can create DSL methods for it, construct UI on
    the fly in the inspections UI, etc, but when we do that, this
    lint.xml mechanism to pass data to it will still be useful.

(4) Users have struggled a lot with where to store lint.xml files.
    Put it in the source root? In the module root? In the project
    root?  The safest thing to do is to explicitly point to it via the
    Gradle DSL. However, particularly when you want to share settings
    across many modules, this is repetitive and annoying.  This CL
    implements a mechanism similar to what .gitignore and
    .editorconfig files offers (and which has been requested in some
    of the below buganizer issues) :

    (a) Nesting of lint.xml files, such that you can specify lint.xml
    files in the root folder of your projects, and various overrides
    in each project directory, and various overrides in each module
    directory, and even various overrides in individual source
    folders.

    (b) A change in the way lint looks up configurations; until now
    when lint reported an error it would look up the configuration for
    the project of the source file; now instead it looks up the
    configuration for the target location of the error.

    The use case for letting you place lint.xml files in files above
    the module directory is to make it easy to share configuration
    across many modules and projects.

    There are many use-cases for letting you place lint.xml files in
    individual source folders, but it lets you keep configuration
    (such as detector options, or enabling/disabling) close to the
    source files such that they move around with those folders, and
    only affect a subset. For example, you may want to turn off
    certain lint checks in a generated or imported source folder, but
    not turn them all off.

When doing this cleanup I discovered various other bugs that I fixed
as well. (There's still some unfortunate separation in flag handling
where LintCliFlags is only available in the CI yet the concepts are
needed in many other places; I'll clean this up in a separate CL.)

Here's a list of the bugs I believe are addressed by this CL; I've
spot checked many of them, but I'll go through them now to confirm
that they're all fixed, close many of them out as duplicates, and see
if there's anything else remaining to be addressed:

Fixes: 163080422: lint.xml: synchronize with IDE, support configurations, etc

Also fixed a few related bugs:
Fixes: 131851821: Lint reports error about unused font resource in the library project
Fixes: 134987035: Use of LintOptions severity overrides breaks OTHER check validations
Fixes: 119660763: Allow configuring options separately for AGP and Studio

Change-Id: I923f8dbfb5dd8961e21979465a6295a6cfde7a5a

Former-commit-id: e48839385a9fd74f17265029dbdac3ef05c0cec6