154635330: Properly handle moved constants in API check
In some cases, methods and fields are moved up from a
subclass to a super class or interface. This is generally
runtime compatible, since the API continues to be available
from the subclass. As long as the receiver is of the subtype,
this will work on all API levels.

Lint was already handling this for methods, but not for
fields in general. This CL addresses that.

Bug: 154635330
Test: Unit test included
Change-Id: Ia2f836db4844b6d245977d26f80dc53c0bef1a91

Former-commit-id: e09ef78c049502a01fed5ccbc2ff944ab8968f44