Add --nouse_ijars to avoid building broken targets.
This is a temporary fix for b/162497186 where the root cause is still
undetermined. We do not (AFAIK) heavily rely on ijars, but may see a
performance regression in java compilation time. This change is to get
Windows builds working again until we can find and resolve the root
problem.

Bug: 162497186
Test: Treehugger presubmit
Change-Id: I23b98885f68a4ce13f623bba53b30bb40dc2a500

Former-commit-id: 6bb967b7149ff7813b4f9b944a6bb36a70660a38