In lint unit tests, make sure test files are unique
I discovered by accident that in one of the lint unit tests,
there were multiple manifests defined -- and when lint creates
the test scaffolding later occurrences will just overwrite
earlier occurrences. So I added an explicit check for this
into the unit testing infrastructure and lo and behold, there
were 23 tests failing as a result of this. Generally it was
benign but in one case the specific scenario in a false
positive check was actually reeplaced by another manifest
without the relevant pattern!

Test: This is it
Bug: None
Change-Id: I92cf358d7b32fe6bf990416bcfe1b5e7919654d3

Former-commit-id: 2435a8c7e84e8452578aa7576a255a136661a92b