JVM res compiler: understand non-html sequences
Previously when the pseudolocalizer encountered an HTML opening tag, it would run out of bounds for the string looking for the closing tag (if no closing tag was present).
With this fix in this case it will go back to the original character and will continue parsing the rest of the string.

Fixes: 162528480
Test: added new in PseudolocalizerTest
Change-Id: I362ebfa86ead3fff2762c4583c8a75519b66cfa9

Former-commit-id: cbf409d1cccb87e8b23798638a8fbf0362af070d