Move native lib extraction for profilers to separate task
This is to allow further cleanup of the deprecated transform API.
Logic for profilers transform application is moved to the variant,
to allow for easier sharing. In the next CL legacy transform API will
be cleanup up more.

Packaging tasks are now simpler as native libraries used by the
profilers are now handled when merging native libraries, avoiding
the need to inject this piece of info in those tasks.

Bug: n/a
Test: ExtractProfilerNativeDependenciesTaskTest
Change-Id: I07888e97a717456e3b90a5fd45bbacb9582d4bf9

Former-commit-id: 54d997bd982b3abaff120bd79950273f1d30421d