make assembleVariant task depends on artifact.
Instead of depending on task name that can be replaced through the
new variant API, make the assembleVariant task depends on the
actual task output which is the artifact. If the artifact is replaced,
the assembleVariant Task will naturally depend on the new artifact
producer. This allows end users to replace artifacts like APK or
Bundle still using the assembleVariant task.

Test: Existing.
Bug: 159993210
Change-Id: I983986225c0218faf1139fe02ef77746f6134133

Former-commit-id: fd3d8a445566095e601218f235e58c70919e3d7b