Add 'I' as a possible state mapping to "DEAD"
This state was introduced as part of the perfetto/TP update and also
maps to DEAD. Both 'x' and 'I' are supposed to mean a dead process/task,
but these values are reported by different kernel versions.

The new TP dependency seems to map them to 'I' when they should be, so
we also need to add it to our list of states.

Test: test should keep passing with new Perfetto/TP version
Bug: 151355530
Change-Id: If96881d1a696a00d30a3fdf31a78730ea36d70c8

Former-commit-id: fc499854332a66fb58098dbaa23803a17d24308a