switched VerifyLibraryResourcesTask to use artifacts.
The task is a verification task which does not produce anything of interest to the other task.
Thefore, a manual dependency was added on assembleRelease for library modules in order to
verify release builds.
Instead of relying on dependsOn(), make the task produce an internal artifact type and change
assembleRelease to have that artifact as a input.

Test: Existing.
Bug: N/A
Change-Id: I3cb8ea2dde225a3172d4b88838d562fd1a366a6a

Former-commit-id: 262458f2bb95467225231684a3abf79e772e1568