Support for parsing CMake file api json
There are a couple of things that make this not entirely
straight-forward:
1) The Target json schema which contains flags and source files
   also contains relevant information like type of build output
   *after* the list of source files. This means the entire
   Target schema has to be streamed rather then directly
   read into a static object. This logic can be seen starting at
   around CmakeFileApi.kt(130).
2) There appears to be no direct way to get the path to clang++
   from the model. So this implementation locates the linker
   via CMAKE_LINKER and uses that to find clang++.

Since we're forced to scan files and flags because of (1), it
makes sense to process this information as fully as we can to
prepare it for Android Studio to read it efficiently. That's
not part of this CL, but it's the reason
readCmakeFileApiReply(...) takes a callback to respond to
each source file.

Bug: 152223150
Test: Added unittests with .json drawn from several projects
Change-Id: Id319b65cf29be53959b802de9aa77337d428a842

Former-commit-id: 3913bc7495d38fa0ef35bfd6be77bdc439dd6db8