Enable cpu frequency (and idle) events capture
Changes the configuration sent to Perfetto in order to enable the
capture of ftrace events that we will expose in our CpuFreq data.

power/cpu_frequency produces a counter that register changes in the
frequency that each cpu core is operating.

power/cpu_idle produces a counter that signals when each core went idle
(no work to be done).

Bug: 162354761
Test: n/a
Change-Id: I94f39da5e1572202cfe688f14a5018f4ac433dd1

Former-commit-id: ee44f33b11f1fd4823738bc9c9162f737c6f8242