add safeguard for variant api misuses.
the new variant API is a builder pattern type of API and if users
forget to correctly chain all calls, nothing will happen which is
very misleading.

keep track of all outstanding use of the variant API until the
chains of calls is completed. At this point, generate an exception
when some calls are not completed, api council should eventually
decide if we should just warn or generate an exception.

Test: added integration tests for most failure cases.
Bug: N/A
Change-Id: Id9f6c8c917556324fd68709c695a0badccc4ad2f

Former-commit-id: 69e8a8b6b09bc8ad5e174dc6d8d7fece70a62dc9