Use relative paths in compiled resources
This change allows for the relative format resource paths to be written
to compiled resources when the ENABLE_SOURCE_SET_PATHS_MAP is enabled.
MergedResourceWriter is refactored to use a single object as
constructor parameters for ease of passing additional values such as
relative resource source set file paths.

Bug: 164090417
Test: compiled resources manually tested & integration tests pass with
BooleanOption enabled

Change-Id: Ice1ffbb87503f355ad138d7663aa832c426caab9

Former-commit-id: 1918600cd15fb5b5ee32c901288fd7cece1d785f