Migrate "tools" attribute to "exec_tools"
This prevents strange host / target CPU combinations from causing issues
when trying to build targets that use the perf_test rule.

This was occurring because the genrule tools attribute is used
internally by the perf_test rule to build a binary that transforms
a target jar by adding profiler hooks into it.

The tools attribute indicates that we only want to build that target for
the current host, but for some reason, sometime after bazel version
1.1.0 and before 3.3.0 (which we ultimately want to upgrade to), it
gets confused and tries to build dependencies for the current target
as well (in this case, Android).

After this change, perf tests build again (and pass), both on bazel
1.1.0 and 3.2.0 (which I tested this against)

See also:
https://docs.bazel.build/versions/master/be/general.html#genrule.exec_tools

Bug: 160150358
Test: N/A
Change-Id: I81150074fc9a52b2f082395e3defe6acff433cd7

Former-commit-id: 28774b69a6a6225f2de915b67bb361c6479ce780