Dex merging: Make DexMergingTask incremental
In a previous commit, we started splitting the input dex files of
DexMergingTask into buckets. The task merges each bucket in both
non-incremental and incremental builds.

With this commit, in an incremental build, DexMergingTask will
re-merge only the impacted buckets (those containing changed input dex
files).

This commit also changes the default number of buckets from 10 to 16
as we want to add more incrementality to the task.

Bug: 132615300
Test: Updated DexMergingTaskTest + IncrementalDexMergingTest
Change-Id: I33f3410e3fde08d730cb45c572e65d99197fc972

Former-commit-id: 7417f9ae354280f1a5d27eb13467a975478faf71