Lint: improve RestrictToDetector error message
...by mentioning the groupId of the current module being analyzed.
This is helpful because Gradle does not always make it obvious what
the groupId of a given module will be.

This came up while debugging Lint errors in the androidx project.
That project uses a custom Gradle plugin to set the groupId, and after
debugging we discovered that the groupId was not being set correctly.

Bug: not filed
Test: updated RestrictToDetectorTest
Change-Id: Ie6ecd9df475d03be0219fd7b78bc8fd07a9d61ea

Former-commit-id: ff6d1906b56bd3a221cb01faabfed286fc7f6e06