Delete ProjectLibrary class
Library information in a form of `Library`/`ExternalLibrary` will be
deprecated in favor of the platform project model. `ProjectLibrary` is
not used. We will merge `Library` and `ExternalLibrary` classes into
an interface providing data specific to Android resources (the part
which is used in the code base) and which can be efficiently implemented
by project systems.

Bug: 163607587
Test: n/a
Change-Id: I2d7a3f7bc03b96169875c445f7cd8d3a52190201

Former-commit-id: e2398f8279bcde22b255c2c92f3e59f6723fdaa3