[lint,Kts] pass a GradleContext to detector
The various Detector functions (e.g. beforeCheckFile()) are permitted
to assume that the context they are passed will be appropriate to the
kind of detector they are.  In the case of our Gradle detector, this
has to be a GradleContext; although we use a Java context to parse the
Kotlinscript file, make sure to pass the GradleContext to the visitors
and checkers appropriately.

This (probably) has no effect at present, but is necessary for checks
on Kotlinscript gradle files which will run after the entire file is
scanned.

Bug: 156217625
Test: existing
Change-Id: I4be1b9a3953156c35b327f7eceaca22a980ff995

Former-commit-id: f7faabd069777a963c5df997ae60707179c343a5