166113037: Lint fails with StackOverflowError with core lib desugaring enabled
Test: Existing*
Fixes: 166113037

The tests verify that the current baseline functionality is
not broken by this check. I have not found a way to simulate
the error scenario in this bug, but this fix adds some boundless
recursion prevention. Basically, lint has a mechanism where it
looks for "similar" error messages to gracefully handle drift
in error messages, and in this user's setup the baseline for
an issue contains a set of old messages that don't match the
current messages, but the pattern triggers alternatating
between two similar error messages so it searches forever (until
the stack overflows).

Change-Id: I77c5896640a34b5a6d9257b6d00b75255f60ee3d

Former-commit-id: 336dde60de3ebd6bf4c70415c21dc200c1ce06e5