Inject ILogger in AndroidLocation.
This is important to ensure warnings output during
Location computation are properly displayed
in Gradle.

Added a test to validate the warning output
by AndroidLocation is ANDROID_SDK_HOME is used.

Because this now relies on an actual logger,
this commit also implements fully FakeLogger
to simply record all things that are logged.

Test: AndroidLocationTest
Fixes: 166168605
Change-Id: If93944ceef8e1d6a390517436f7c9e7d405af91a

Former-commit-id: d7cca095738be54136388618456a1f172843e360