Change iml_to_build multi-project support
Instead of using the prefix .unb, it now uses
a property "project". There is only one
possible project active at a time and it's
configured in the project.bzl file.

This allows having the same rule name in
different configurations. This could be done
with selects, but would require a much larger
change to iml_to_build which will be short-lived

Test: existing
Bug: 156497102
Change-Id: Ica3a1e610488bb10d07d3dbb4a21c5ccb769b27a

Former-commit-id: 177f8b7938a5be8cece27b6ac8539fbfff9407bb