Wire jniLibs.useLegacyPackaging to tasks
Wire jniLibs.useLegacyPackaging to ProcessApplicationManifest such that
we'll set android:extractNativeLibs="false" when
jniLibs.useLegacyPackaging is false.

Also wire jniLibs.useLegacyPackaging to PackageAndroidArtifact to verify
that the specified value is compatible with the value of
android:extractNativeLibs in the merged manifest. In the future,
PackageAndroidArtifact will use the value of jniLibs.useLegacyPackaging
to dictate the native library compression level and throw an error if
it's incompatible with the manifest.

Bug: 69953968
Bug: 149770867
Test: ExtractNativeLibsPackagingTest, ManifestMerger2SmallTest
Change-Id: I899c859f69bf0b407df32e3c362ca9da4082975e

Former-commit-id: 67429c3f9da0e4b7baf95a51d9b125bbef222c7c