New version of AndroidLocation.
The new version (AbstractAndroidLocations) is mostly
based on the old one but does not memoize the computed
values in static field. It uses instance fields to memoize
the content.

To use, one must used an actual class that implements
the abstract base. One version, AndroidLocationSingleton,
is a kotlin singleton that provides a similar behavior
as the previous version. Instead of using static fields,
it's just a singleton instance.

From the Gradle plugin, we will not use this version,
instead we'll use an proper instance, wrapped in a build
Service (this will be done in a later commit).

AndroidLocation is now deprecated, but usages have not
been updated yet. This will come later.

The goal is to ensure that running this code in gradle will
recompute the location once on every run, in order to take into
account changes to env var or system property. Previously
because the classes of the AGP plugin are reused between runs,
the value coming from AndroidLocation could have been out of date
if changes to env vars had been made.
Even though the environment provider properly uses new Gradle APIs
to read the property and ensure they are part of the configuration
cache inputs, if the classes were already loaded, the old values
would be reused.

Also improved the error reporting in case we cannot find
a valid location, by providing the list of location
that were checked.

Bug: N/A
Test: copied AndroidLocationTest to AbstractAndroidLocationsTest and expanded
Change-Id: I834a73f4db6865b03a33bd37ee31708cc2385d2e

Former-commit-id: 8a7fb3b4a6805164bad000061bd0025533107dee