Record if config caching is enabled
This CL allows us to record whether config caching is enabled at
configuration phase in non config cached builds. For config cached
builds, it is not recorded as cofig caching enabled as configuration
phase is skipped, but we can tell if a build is a config cached
build by checking if there is "configuration" span in proto file.

This change will help us have a better understanding of how config
caching is adopted by users and how well it works in a real world.

Bug: 166254496
Test: n/a
Change-Id: Ie711d63e5cde15f5f7a36cf20bc49ce3a1105921

Former-commit-id: 0c4ee76d737060e17d44408633989bb47a16c625