Remove usage of static AndroidLocation from AGP
This change introduces an AndroidLocationsBuildService which
is global to a build, caching the resolved values, but is reloaded
after each build. This fixes the problem of the previous implementation
that used a static field on a class to cache the value. This cache
would survive across several build even if the underlying value
(env var for instance) changed between builds.

This change adapts all code used by AGP to require an instance
of AndroidLocationsProvider. This allows AGP to use the build service
while other tools (studio or cli versions of sdkmanager, etc...) can
use a more normal singleton.

Bug: N/A
Test: existing tests
Change-Id: I9ae8c1b6ce9c0724b07a637cb6c75b72c0be3b67

Former-commit-id: e975c32caf783dea7b083107ef4f486827502f99