Don't extract agent-alt on 32-bit platforms
IWI run currently tries to extract agent-alt on 32 bit platforms, which
always fails because there is no agent-alt packaged on those platforms
(and there is no need for an alt agent since only 32 bit agents are
supported).

This CL removes that and only extracts the appropriate agent on all
platforms / apps.

Bug: None
Test: Manual + Tracing
Change-Id: I5ff8b502006b3dc041db08ee31d23d140ef7ba0a

Former-commit-id: bbb959bfa65dfcb5f81c1a1a08dee1b06f6c4405