IRR: Fix installer downgrade detection
Problem: If an user downgrade to a previous version of Android
Studio, a bad version is detected but the connection is not closed.
This result in previous version of Studio waiting for a "full"
response that will never arrive.

Solution: Close connection from installerd when a wrong version
is detected.

Test: Manual
Bug: 167449438
Change-Id: I3c2d51824be975f37e32a04e989b460dc03b73fc

Former-commit-id: a97dfe1d3bde51ed7d0e94d34c638a9d75cd14a8