Save the BES file and bazel profile for coverage test runs.
The BES file is parsed by a separate service to record metrics about
tests, such as duration and test status. This information will be used
to feed queries and produce stats on how often tests flake.

The bazel profile is used by dashboards to show which actions frequently
appear on bazel's critical path.

Bug: 168054851
Test: presubmit
Change-Id: Ie830ee7301371d14ff0a711af0f61226658d384c

Former-commit-id: afac3155724b74604d77dfa13576001d893abbf8