Clean up and fix lint configuration inheritance
This CL fixes a number of problems with the configuration inheritance.

First, lint used to visit all the dependent projects as a flat list,
which means that the code to set up parent configurations did not work
correctly among sibling projects. Instead lint now visits the projects
recursively (skipping dependencies it has already seen, so the same
libraries are visited, but we get a chance to restore parent pointers
as we walk back up the dag).

Second, this CL cleans up and clarifies the exact semantics of having
multiple configurations associated with a single folder. This was the
root cause of 168261940 and required extensive changes to handle
properly and cleanly. For something like a Gradle-based project
directory, there can be up to 3 configurations associated with the
same directory (a flag configuration (e.g. for the lintOptions
configuration settings), a lint.xml configuration file if the file is
in the directory, and a lintConfig XML file pointed to by the
lintOptions. This required more careful tracking of the parenting
hierarchy, such that when we for example reparent a configuration, we
reparent the last configuration in the inheritance hierarchy
associated with the same project directory. The new defined hierarchy
for a configuration lookup is as follows: First any global override
configurations are consulted Then the nearest configuration from the
source is consulted, up to a project directory In the project
directory, the flag configuration (DSL lintOptions) is consulted In
the project directory, the lint.xml configuration is consulted In the
project directory, the lintConfig XML file pointed to be DSL
lintOptions is consulted The parent project's configuration chain is
consulted next, or if there is no parent configuration, the containing
directory Finally, any global fallbacks are consulted.

Prior to this change, several flag configuration classes inherited
from the XML based configuration, and tried to implement the above
semantics (partially). This is now cleaned up by completely separating
the configurations and instead having the parent hierarchy setup code
shared.

Another bug fixed by this CL is what happens if a configuration is
modified (which can be done by some suppress actions in the IDE, for
example for icon files); if there is no lint.xml file in the project
directory, it will now be created.

This CL also updates the lint test infrastructure to properly sandbox
the lint.xml file lookup to the test folders. Before this, I
accidentally placed a lint.xml file in /tmp/, and got many surprising
test failures until I figured this out (and added a defense against
it).

Finally, with the clear definition of how the lintConfig file is
actually a fallback (which is only consulted if there are no closer
results), this also adds a way to specify an
override-configuration. This has already been possible using a system
property, but you can now configure it with a startup flag,
--override-config. (Soon this will also be a new lint model flag; it's
not added here to conflict with in-progress work on the DSL and
lint-model on the AGP side).

Test: This CL adds a bunch of new tests (both unit tests and Gradle
      integration tests), plus updating some existing tests.
Fixes: 168261940
Change-Id: I48f61c41a45dd12340cc73c446e07fe2f44d5c7e

Former-commit-id: 696a0d13e3429b1f27a9e21c6e07ed33494b5cbd