Wire dex.useLegacyPackaging to PackageAndroidArtifact
Wire dex.useLegacyPackaging to PackageAndroidArtifact and verify that
the specified value is compatible with the value of
android:useEmbeddedDex in the merged manifest.

With this change, PackageAndroidArtifact uses the value specified by
dex.useLegacyPackaging to set the compression of dex files, *except*
when dex.useLegacyPackaging is true *and* android:useEmbeddedDex is
true, in which case we leave the dex files uncompressed and warn the
user. In the future, this warning will become an error instead.

Bug: 69953968
Bug: 149770867
Test: UseEmbeddedDexPackagingTest
Change-Id: Ia802467a0fb9371d3c8454942a5e672f13b1d16f

Former-commit-id: 630d4c4c13541e7b0f79b7789b235117eb3ac708