Case where CMake runtime file wasn't created
[This may necessarily regress Dolphin perfgate]

If CMake reports an expected runtime file and that file is not in
the expected output location then we hardlink or copy it to the
expected location. There was an issue where if the reported
runtime file was not created by the build the hardlink/copy would
cause an exception to be thrown.

Depending on where in the copy process the missing file was
encountered the perfgate regression could be small or large. I
don't expect the impact to be zero though.

Bug: NA
Test: NA
Change-Id: Id9d4b6e55c0ff4929eacc1e2b2a415cfcb015c0a

Former-commit-id: f35387daede84f7deed605598fc8555c75eeaf1b