Remove old lint cancel support
Lint had an old, cooperative cancel mechanism. It was based
on checking a flag regularly, instead of using exceptions.
This was back from the Eclipse days. It makes the code really
convoluted; the command line integrations don't use it, and
the only way lint gets interrupted now is via the process
canceled exception. This CL deletes all the cancel blocks.

Change-Id: I0d06c9fe969abdfc9e077ee9de32a0970d5bac4c

Former-commit-id: 1e2e319fd367afc1d4569ffd0324657cbcac56f9