Lint: use SoftReference to cache ApiLookup
Before, only a WeakReference was used, which meant that the API
database was usually reparsed from disk each time Lint ran on-the-fly
in the editor. This seems like an ideal scenario for using
SoftReference instead.

This was found using ide-perf and tracing the package
com.android.tools.lint.checks. ApiLookup.get was the only thing that
took up nonnegligible time (15 ms or so per call).

Bug: not filed
Test: manual / profiling
Change-Id: I3367d04244c637290c6a42109ff5aa8377baa1b7

Former-commit-id: 501616094373aab544c970bb17b684a1c260895c