Fix ddmlib proxy client when replying to multiple clients
The root cause was because JdwpClient.findPacket() mutates
the buffer. This causes the first client in the list to
get the proper reply value while other clients get a
modified version. The modified version can be an invalid
packet, this causes cascading failures in ddmlib.

The second issue was in the handling of the handshake
previously the handshake was handled on demand. This change
simplifies the workflow by making the handshake happen as
soon as the connection is open. Now the host->proxy
handle the handshake independently of the proxy->device.

Test: Updated
Fixes: 166178897
Change-Id: I05a5c3141fb1304701c53b2650f2885b6e5e2dc0

Former-commit-id: 75063141865a1e129425c3c1672b47db1c74e7c5