Add adbOptions.timeOutInMs = 30000 to connected test projects
Previously, the AGP connected tests would often hang for ~10 minutes
because there was no timeout set via adbOptions. Setting
adbOptions.timeOutInMs to 30 seconds will allow the tests to fail faster
in the cases when AGP is waiting for a response from a device via adb.

This change also adds a build.gradle file to the
.../integration-test/connected directory to allow for syntax
highlighting when working in the tools/ project in the IDE.
The added build.gradle file is based on the build.gradle file in the
.../integration-test/native directory.

Bug: 148626301
Test: ran manually with --config=remote and --runs_per_test=100
Change-Id: I1211dcea5910961a70733757c06f7a452ef4f73a

Former-commit-id: e02177dd789141c0a1630fd54e119abd5160348e