simplified VariantSelector by removing target type.
have a target type in the variant selector would make reuse between
beforeXXX() and onXXX() methods impossible since one take Builder
type of objects while the other take Variants.

removed the generic type for the VariantSelector so now it can
be reused and removed the withType() API which became useless after
migrating the APIs to the AndroidComponent extension.

Test: Exiting.
Bug: N/A
Change-Id: I8a34002f65606d9d5eb7e39aa3d9a7e7743149af

Former-commit-id: 15be9034512bbc36b9b87f0e3ae80948e09fe894