Support enabling v1-v4 signatures per variant
This change adds support for enabling/disabling v1, v2, v3, and v4
signing per variant.

This change adds a new task instead of adding to SigningConfigWriterTask
because we don't want to write the sensitive signing config info to disk
unless absolutely necessary, but we can always write the signing
config version info to disk.

This change also fixes a bug in which v3 and v4 signatures would be
disabled if any signing options were injected from the IDE.

Bug: 37137460
Test: SigningTest, SigningConfigUtilsTest, etc.
Change-Id: I7f56ff014d357c5c753b690f47578557958e25c3

Former-commit-id: 6a1e73d62bff325d5b45a911dd0c2bc7861ecd72