Add project.execute("uninstallAll") before connected tests
This change adds project.execute("uninstallAll") to the @Before block of
the connected tests in order to (1) make sure nothing is installed at
the beginning of each test and (2) check the adb connection before
taking the time to build anything.

This test also makes UninstallTask compatible with configuration caching
by removing references to project in the task action.

This change also turns configuration caching on for many connected tests
now that issue 146163513 is fixed.

Bug: 148626301
Test: existing
Change-Id: I5c2e0b4b5c0c1060e522bea0f49d30cda0a67a97

Former-commit-id: 2b101aa27e37fa99852f937c12d9749dd0927c3a