Fix dependencies on transport native code
Previously, for reasons I don't entirely understand (perhaps a bug in
bazel?), generated protobuf headers were not correctly included by rules
depending on the proto library. It seems that adding
strip_include_prefix = "" to the generated proto library rule fixes
this. This is now done, and the corresponding copt -I$(GENDIR)/...s are
removed accordingly.
Additionally, cc_libraries in transport (and other places, but only the
ones in transport are addressed in this change) didn't declare hdrs, and
so all targets depending upon them needed to add a -I copt. Now hdrs are
declared and -I copts are removed.

TODO is the same thing for profilers and deploy.

Test: there shouldn't be any impact on any build outputs: existing tests
Bug: none
Change-Id: I045df31f41fa92cf50a32a46a77d32867e83218e

Former-commit-id: b3070f72a35f8c16c45d845268901fbe878e527b