In lint tests, include secondary line content
Lint results can point to a list of related locations,
called secondary locations. These are included in error
reports, and when lint runs in the IDE and the secondary
location is in the same file (such as for duplicated resource
declarations), they're also shown in the editor.

However, in test reports, only the file, line number and
message (if any) were displayed. This meant that the
error range for the error was not shown, and sometimes
the range was not optimal. This CL fixes that; tests
can now opt into full secondary location checking, which
is turned on for all of lints tests, and a few cases where
the secondary locations were suboptimal those detectors
were changed.

(Including this was necessary because an upcoming change
will switch some location computations over to be based
on a resource repository, where locations are computed in
a completely different way, and this will help make sure
that there's no location range regressions.

Test: Existing (many updated)
Bug: None
Change-Id: I0c35c3f7f51acb1d4a4f95e92eeceaf2a096bae0

Former-commit-id: 33069905aa0837a1604d98dd18cc99dcadaa378c