First part of new lint model lint
The new integration is behind an optional boolean flag
android.experimental.useNewLintModel

If the flag is not explicitly set, the new integration is automatically
enabled when using configuration caching.

Currently, lint queries for the project structure through the tooling
model, however this approach is not compatible with either configuration
caching and resolves configurations across projects which can lead to
incorrect results and is being deprecated by Gradle.

The new integration doesn't add the global lint tast that
runs lint sequentially on each variant and then merges the result.
The global lint task is just implemented as an anchor task that
depends on the default variant for lint.

To follow:

 - CheckDependencies support via publication of models between projects.
   This will be replaced by a new integration that performs analysis
   within each project anyway, but is probably helpful as an interim
   measure.

 - The previous lint integration merged dynamic feature sources into
   the main app from lint's perspective to fix the unused resources
   detector. This is not implemented in the new integration yet.

 - lintFix support in Android projects is not yet implemented.

 - Up-to-date and cacheable lint

The lint integration tests are parameterized to run with both
integrations, except the ones related to functionality that is not
implemented.

Bug: 160392650
Test: Parameterized all lint integration tests.
Test: Added LintModelTest with golden model files.
Change-Id: I820a953ed67d91ce50e2e60073ee1981bb8e783c

Former-commit-id: 63a4817da66c67e8eae92d6cb07d4c50d8da24a9