Fix handling of annotations and assignments
This CL fixes three bugs around annotation handling for
assignments:

169610406: Strange warning for Kotlin property initialized by constructor call
80166964: IntDef annotation for variable not working anymore

The fix makes the annotation support treat assignments differently;
it will now tell detectors whether the annotated element is on
the left hand side, or on the right hand side.

It also fixes a problem where we'd end up with multiple errors for the
same problem (we had a number of occurrences of this in the codebase
so this CL also removes those errors from the baselines now that they
are no longer reported).

Fixes: 169610406
Fixes: 80166964
Test: Unit test included
Change-Id: I8a5343dac558e44919a4c90e04c6a3474dc0d4fc

Former-commit-id: 2e494ff79b114c6ec76271ab35e084599c106201