Use offset annotation to map back LL key.
The Compose compiler will move toward using annotation to map LL.

In short, the editor will identify a literal using an "offset"
which is an index of the literal within the parse tree.

The compiler will annotate the helper method with an offset annotation
for each LL's default value helper function. We are going to look
up which LL we are swapping by comparing the offset instead.

Bug: 163430998
Test: LiveLiteralUpdateTest.java

Change-Id: I373062b47396ef6836ec07ff8de9560e01a6ed6c

Former-commit-id: 5a02f7edb12c33ab3a4cfe3b5708bfbb1ff164f0