Remove agent-server from the deploy pipeline
This is a minimal refactoring CL to remove the agent-server from the
deployment pipeline and move all swap-related functionality to use the
install-server (app-server).

There is still significant code re-use between installer commands that
are not addressed by this CL, in order to keep the delta of this CL
small.

Bug: None
Test: Manual + DeployerRunnerTest
Change-Id: Ic1c99f1ef7e358acbb0ee54199048fc520dbe6db

Former-commit-id: 94042693342d0f142b0e3f6ca1a56a87ab6ba165