Add profileable_reporter to report profileable apps
profileable_reporter is a standalone program running
on Android to report a list of running profileable apps.

It writes the findings to stdout. For example, the
following output is produced in the scenario where two
profileable apps are started and then stopped. (The
output may be updated in the future to accommodate
programmatic reading.)

0 profileable processes
1 profileable processes
23514 com.example.mypmplayground
2 profileable processes
23514 com.example.mypmplayground
23553 com.example.android.displayingbitmaps
1 profileable processes
23553 com.example.android.displayingbitmaps
0 profileable processes

Because profileable_reporter invokes "/system/bin/cmd"
to check whether an app is profileable, which may write
to stderr. It's recommended to redirect stderr when
running profileable_reporter. For example,

adb shell /data/local/tmp/process_reporter 2>/dev/null

Passing 'd' as the command line argument will make the
output in debug mode, which outputs each profileable app's
start time and the stats of each query, for example,

adb shell /data/local/tmp/process_reporter d 2>/dev/null

2 profileable processes
1978 com.android.systemui start_time: 1111
13328 com.android.systemui:screenshot start_time: 17815804
    Query takes 1918 ms (850 processes, 72 apps)
    Query takes 126 ms (853 processes, 75 apps)
    Query takes 133 ms (856 processes, 78 apps)

Bug: 147112874
Test: detector_test.cc included

Change-Id: I8ae9b4f18c9695c00d85b34f01b4daf43b3fde32

Former-commit-id: 506691b9dfb17ad579fd393a25620ed69a047187