Fix Bazel rule for aapt2-proto
Problem: the prebuilt aapt2-proto jar is listed as a resource jar,
which means that it does not end up on the compile-time classpath
for downstream modules. Compilation succeeds only because the
kotlin_library rule has a bug (soon to be fixed by Change Id42366ef8).

Solution: use an intermediate java_library rule to create the
exact jar we want, then use maven_java_import to consume it.
(Is this the easiest way?)

Bug: not filed
Test: bazel build
Change-Id: Icd9699b6429511a23fe738a779479a9d589be583

Former-commit-id: 2e7c47463c67882169b5fe9782481e41a1c6d241