Don't create error message string unless error is thrown
Profiling indicates that we're spending >10ms in String.format() in
VariantScopeImpl.publishIntermediateArtifact() in
SantaTracker_jetifier_codeImplChange (and some other benchmarks) because
we're creating the string on every call, instead of only when an error
is to be thrown.

This change modifies the method to create the error message string only
when an error is to be thrown.

Bug: NA
Test: existing
Change-Id: I0f4b959bfc28c8f7911117563cfeb66f144eefa9

Former-commit-id: 97aec79d79dcbd3d1c7ada73f6e445842753324f