iml_module: refactor runtime_deps handling
We can simplify things slightly by putting runtime deps into the
JavaInfo of the prod module, rather than in testlib.

This also ensures that the prod module has an accurate runtime classpath
if it is used directly by a non-iml rule. Additionally, users of
the testlib will no longer see runtime_deps on the compile classpath.

Bug: not filed
Test: existing bazel tests
Change-Id: I5c9357fbd6ffd90462c92973f87b78cd914ad991

Former-commit-id: 46ed2ad13a0d498a595da4fa6d58c0e3c1905e49