Bring "Find Usages" to ApkAnalyzer CLI
This feature, previously available in ApkAnalzer GUI,
lets you search for references to classes, methods
or fields, in order to find all usages of said
class/method/field.

This feature can query the reference tree of a class,
method or field from command line

./apkanalyzer dex reference-tree --references-to="descriptor"

Or a list of them from a input file.

./apkanalyzer dex reference-tree --input-file="<File>"

Test: updated unit tests
Bug: 170418752
Change-Id: I95446464c88b81ebbf9b7eafff116c98bd9654a5

Former-commit-id: 44ebb8f0ead2e2e78adfdbbf1a1e80297d2b9e46