New lint integration: Model lint outputs
This means that the new lint tasks can be up-to-date as long as they do
not use checkDependencies.

The previous lint integration would have allowed lint and lintVital to
clobber the same files if the build author specifies them. Different
variant lint tasks may still clobber the same files, but this is now
done by a separate report copy task to preserve the existing semantic
while keeping lint able to be up-to-date. Lint vital and lint fix no
longer write the same reports though.

LintOptions are no longer passed in seralized form as an input,
as more control was needed.

Bug: 160392650
Bug: 117870210
Test: Gradle integration tests: expanded LintStandaloneTest & new LintUpToDateTest
Change-Id: I1c787df7e46d5c8caff716e943cad6534febe5c5

Former-commit-id: 92adafab4f263f63754b7d0a5de83b2c11a7f93d