[AGP UTP] UtpConfigFactory update for UTP Extension
Made changes to the ManagedDeviceInstrumentationTestTask to support the
GradleManagedAndroidDeviceProvider UTP extension instead of the
VirtualAndroidDeviceProvider.

This primarily resulted in changes to the UtpManagedDevice
implementation to send relevant data to match the changes in the
UtpConfigFactory.

As a small update, the unique id for the managed device is set to be the
task's fully qualified path name. Ensuring uniqueness.

Also added a basic config provider for managed devices to ensure the
default apk for the given api/abi is retrieved for testing.

Bug: 141510559
Test: Validated manually against my fixes in to the Gradle Managed
Device provider, as well as updated ConfigFactory tests

Change-Id: I95c1ca25a5d41bbfc63851a1846b495c33bdbf2b

Former-commit-id: 6a94508791fb8f0f01043330965fedf5fce84634