Consolidate logic for determining STL
Before this CL, it was defined in multiple places.

Also, the result of determining the STL should be saved in one
place so that there's no chance of drift between different parts
of this code. The STL should be determined by the full
command-line, not just the user's extra variant-level arguments.

Bug: 170485962
Test: Refactoring
Change-Id: Ic4697818b1be983a1fc64dd10b649457d860eb80

Former-commit-id: cc829dbe686eb9b77f8ba3d507791f4e961e0bb8