Add DSL to replace AndroidManifest.xml's package attribute
Add `var packageName: String?` to BaseExtension, which will supersede
any value specified via the manifest's `package` attribute.

We use "packageName" to match the existing API in Variant (as opposed to
using "namespace" which was also discussed as a possibility).

This change also stops using applicationId as a default for packageName
in VariantDslInfoImpl, which is relatively new behavior introduced in
I8dcacf905d39b990be9c95dcb38b2cd4b15cbfb9. The reason for changing that
behavior is because (1) we now have the packageName DSL for this use
case, and (2) applicationId can be different per flavor, but packageName
should be the same for all variants.

Bug: 172361895
Test: VariantDslInfoTest, ManifestMergingTest, RClassPackageTest, etc.
Change-Id: Ib57a790470ad03a27686f3d409a665580344778c

Former-commit-id: e729f4019d5de08a499cc2cb38ed6f8d28e277e4