Add IncrementalInstallSession to ddmlib
Create an abstraction that communicates with PackageManager in order to
install an APK incrementally. This abstraction invokes the binder
command "package install-incremental" to start an install of one or
multiple APKs. PackageManager requests from ddmblib blocks of data from
the APK that need to be read. ddmblib responds to PackageManager with
the blocks of data that PackageManager requested.

In order to be installed incrementally, an APK must be signed with
APK Signature Scheme v4.

This abstraction also allows the for the caller to exclude blocks from
being served to the device. The ability to exclude blocks from being
sent to the device enables testing how the device behaves when certain
APK data fails to be delivered and enables framework developers to
reproduce failures in the field caused by missing data blocks.

Bug: 160634981
Test: IncrementalInstallSessionTest
Change-Id: Ida2a08037c98e39b426f18d648673c8aefd8acd7

Former-commit-id: 73549da1632809e0bfe37f7b9d03d0600c868833