[AGP UTP] added --enable-display option to tests
Added the ability to display the emulator while tests are being
performed. This adds a new task command line option "--enable-display"
to the ManagedDeviceInstrumentationTask which dictates to the UTP
extension whether or not to display the emulator while running tests.

Bug: 141510559
Test: updated tests on UtpConfigFactory and TaskMethodModifiers... to
account for the api changes. Manually tested the boolean option to
ensure the emulator appears.

Change-Id: Ia7b83796db7b5c5d02062c1a5f4c73e64fe186cb

Former-commit-id: 2df7810ead04e37bc506226b19e51e3f333923e2