Add lint tool display name
Lint has had internal names referring to various embedding
clients (such that lint checks can do tool-specific behavior,
if applicable). However, these were really internal names and
not suitable to include in error messages (again, when
appropriate). This CL adds a new display name that clients
can override to customize the tool or product name that appears
in error messages.

This CL also adds an improved utility method for looking up
the calling detector and associated issues when an exception is
intentionally or unintentionally thrown.

Both of these will be used in an upcoming CL where there will
be error messages that refer to the embedding tool as well as
the related issues.

Test: Existing
Bug: None
Change-Id: I2943fb340c5dce47db9be1373416ab14b0760526

Former-commit-id: 38deb0adca39ef3ed1f538ae851a382a5fe5877d