Update MinifyConnectedTest to run on RBE
When reviving this test, it worked as-is with R8, but with Proguard it
failed until I added getDefaultProguardFile('proguard-android.txt') to
the testProguardFiles, which suggests there may have been a breaking
change while this test was inactive. I added a comment about this in the
test, but I didn't investigate further because we're in the process of
removing our Proguard support.

Bug: 77903319
Test: MinifyConnectedTest
Change-Id: Id91a24e11995909555bc8e7c7d78437d08cb5647

Former-commit-id: f04eed83f1c0b6913da548136433c46a1ae8a215