Improvements to lint's test project support
This CL updates lint's handling of test project configuration:

* It picks better default names (e.g. "app" and "lib" instead of
  project0, project1, ...)

* It makes sure that the picked names are unique (so if the project
  has two app modules, they would be named app1 and app2)

* It makes it easy to create multiple projects without having to
  explicitly create project descriptions and setting up dependencies.
  In a file list you can simply use relative paths, like
  ../app/AndroidManifest.xml, which would create a new project named
  "app". It also sets up dependencies by guessing from the name
  which direction the dependency should go in

* You can name dependencies by name, such that you don't have to
  initialize them in a particular order and have to create local
  variable references.

* There are methods for setting which specific variant you want to
  use from the lint model.

* Handling of duplicate incidents now is faster and has a better
  explanation error message.

Initially this code was written into the TestLintTask class, but that
file is getting crazy large or complex (so much so that Convert to
Kotlin fails), and since it can't be converted this code had to
initially be written in Java. Therefore, this class extracts existing
code to actually perform test execution and makes the TestLintTask
primarily be a configuration object for tests. The new TestLintRunner
class (internal only) now performs the test execution, including a lot
of the project setup.

There are a few other minor changes -- test tasks can now configure
the format of the output instead of always being hardcoded to
TextFormat.TEXT, some issues were fixed with relative path handling in
test output (and test root directories)

Bug: None
Test: Unit tests included
Change-Id: I2951c317c3d86c12c8bb9d3454578a1a840834a7

Former-commit-id: 057750614304f658227fc268614ee8f4178c04dd