Add lint test mode support
Lint will by default run any unit test that contains Kotlin source
files twice; initializing UAST with different behavior for the
UInjectionHost (which is going to change in an upcoming version of
UAST and we therefore want to make sure that detectors are handling it
correctly.)

This CL generalizes this concept; there is now a TestMode "enum"
(behaves like an enum but is an open class to allow third party tests
to add their own test modes) which specifies different modes of test
execution. This will be used in the future to run tests through
additional validation.  For example, with provisional support, lint
might run the test in the normal way (which will be used in the IDE)
and in provisional mode where it first analyzes the project with
minSdkVersion=1 and then merges the data with minSdkVersion=<whatever
the test declared>.  Similarly, when transitioning from optional to
mandatory resource repository support, lint might run tests that
reference the resource repository twice, with or without resource
repositories or perhaps switching between different implementations:
lint's own and the AGP one.

Test: Existing tests
Bug: None
Change-Id: I6d8375d4428c0048ad5e4aeda47c3af4eee7b074

Former-commit-id: 8b8f1da00f06ac31ea39a29a5c607e78c617b944