Lint mandatory resource repositories
This CL adds resource repositories into lint such that it's a guaranteed
service instead of being optional and only provided in the IDE (and
applicable unit tests). It also adds the ability to query framework
resources as well. It also updates various detectors which were
conditionally checking for the resource repository to access it
directly.

For now, this uses a local lint implementation to persist resource
repositories for the framework and for AAR files. In a future CL this
will be replaced with tools/adt/idea/resources-base/ code (to be
extracted from the IDE plugin). There are also a few detectors which
currently do a lot of extra work in order to work outside the IDE, such
as TranslationDetector and StringFormatDetector. These work but can be
simplified now that resource repositories are always available; this
work is left for later.

This CL also improves the position computations in several ways -- for
text nodes it now computes a full range (previously it only computed the
starting point), and resource item handles now get associated with the
actual elements and attributes.

Test: Existing & updated
Change-Id: I6cdfeb1c5bd1d74705431271f6b909aabd3f550c

Former-commit-id: 64857dabc30ce1ac5fe64c4c89439f93854a47cc