Require permissions for Manifest Class generation
Currently, the GenerateManifestClass will produce an empty
class i.e with no permissions, if there are no custom merged manifest
permissions. This chnages avoids generating the Manifest Jar entirely if
there are no permissions. This will enable to BooleanOption to be
enabled by default as this jar will only be generated when needed.

Bug: 157117028
Test: GenerateManifestClassTest
Change-Id: Ie1228c6370f24775a7bc9bf250fb9f887978405f

Former-commit-id: 25a083c72e03208a5419961bb592676979fb8e94