Stop depending on the microApkTask to register the microApkResDirectory
DependencyResourcesComputer depends on taskContainer.microApkTask being
set to set the microApkResDirectory. This means that the
GenerateApkDataTask should be configured before MergeResources task,
which fails when adding a listener to TaskContainer.whenTaskAdded as
tasks are configured when registered to be able to invoke the listener.

This change removes this dependency and sets the microApkResDirectory if
isEmbedMicroApp is set, and if the task didn't run the directory should
be empty which won't affect the current behaviour.

Fixes: 173911569
Test: Manual
Change-Id: Id73f31eb2606f07a2d70962167425d4208823d8a

Former-commit-id: 3eb91a4dc4176752c1b8d0577663219a1505cc9e