Prevent SIGSEGV when shutting down TPD by lack of activity
When TPD is shutdown by lack of activity, the server during shutdown
might invoke the callback one last time. But since both were on the same
scope of RunServer(), the callback pointer could have been disposed
already.

So we move the callback creation to main(), so we guarantee it's only
disposed after RunServer() finishes and we know the GRPC server has been
shutdown.

Bug: 158078813
Test: manual with --server_timeout
Change-Id: I098738c7b6de1e7cb0fa7faed8aedc36a8f6f906

Former-commit-id: 00d53ff9cfadea00feeb8e320d17e8f9b88c7866