Only log error message if server cannot start
Rearrange log messages so that when the TryStartServer method fails
to start the server via run-as, it simply logs an info event with the
output from run-as.

The responsibility to log an error message on server startup failure is
now in the StartInstallServer method, in order to reduce confusion and
ensure we only log at error severity when the server can't start at all,
not when the initial optimistic start fails.

Bug: None
Test: Manual
Change-Id: Ic823d07c376618f0b351bcf8ed10795af0acd616

Former-commit-id: 08341257d3aa096f214dfddef3d8c35dc3799b3f