175054126: Lint tries to create `~/.android` by default
The old AndroidLocation lookup class would create the ~/.android/ folder
if it didn't exist. That's not ideal for lint running on non-Android
code and machines.

Switch to the new location lookup API in 7.0 which does not create the
folder if it doesn't exist.

Also start flagging occurrences of ~/android/lint/*.jar files since
these will not be loaded anymore in the next release.

Relatedly, in non-Android projects, we'd still end up querying for the
Android SDK manager (as part of setting up the PSI enironment, looking
for external annotations). This causes a lot of unnecessary work, and
also creates package.xml files if missing in the SDK. (This is part of
b/175054520.)

Test: Existing
Fixes: 175054126
Change-Id: Id29d7356c2f59f9e76eedcaaf7f1bd7249fad4d4

Former-commit-id: 36983da1bac67eb9dd0e1a85cb50366797f71492