Omit native debug metadata files if they're stripped
Previously, AGP would include native debug metadata files for native
libraries, even if they didn't include any extra information (e.g., when
the native library came from an external dependency and was already
stripped in that external dependency).

With this change, the ExtractNativeDebugMetadataTask checks each output
file against the corresponding stripped native library file and deletes
the output file if it's the same size as the stripped native library
(which indicates that it's a copy of the stripped native library).

Bug: 169266007
Test: ExtractNativeDebugMetadataTaskTest, MergeNativeDebugMetadataTaskTest
Change-Id: I581e25c908dbf01e07773006cb54d9c7d76a5165

Former-commit-id: 1d7ed74086904d4c513e5cb06123e72f75e40fc7