Remove coverage post-processing dependency on deploy jars
Jacoco needs classfiles to generate a report. Until now, we used the
deploy jar of a test target to ensure we missed nothing and then
filtered the deploy jar down to only things we potentially cared about
(throwing away almost all of the deploy jar). This unfortunately
involved a lot of wasted work and could take a while for large test
targets with lots of dependencies.

Now that the baseline coverage system exists, we can directly use
the already built jars of source class files we care about and eliminate
the need for deploy jars. This change uses a merger of all the baseline
source class jars, which is more than is needed for more tests, but more
precise breakdowns are probably not possible without migrating to aspects
(too unfamiliar to be quick at the moment) or explicit lists (user
unfriendly and defeats the current automation of baseline coverage).
The merged baseline jar is still smaller than the largest deploy jars
or even filtered deploy jars so it is still a win.

Bug: 174603669
Test: N/A
Change-Id: Ia326ea019bf700b19f800cdf565ead28917938cd

Former-commit-id: 7497de6e5b790fd926653a25e1b2c9a0be914f58