Move databinding, cmakejni connected tests to RBE
Also enable jacoco connected test

Some changes had to be made to the build files for the databinding test to work.

The connected tests had to depend on publishLocal so databinding libraries would be built.
The databinding test projects had to be updated to use the new maven-publish plugin to be compatible with gradle 7

Bug: 77903319
Test: Existing

Change-Id: I2f4fc0c6ca099f51d3e2c2efc5ee236be73f64ce

Former-commit-id: 33bf8bd0c9c5f6b4bdc15781a4b2cd196320b29f