Do not leak coverage split handling
Coverage split handling is not meant to support general test_suite
usage and is intended to be invisible to users as much as possible.
Furthermore, there's no benefit to using a test_suite of top-level
targets in coverage (split handling is intended for dynamically
generated targets) and in fact adds an unnecessary merger layer that
slows things down.

Thus, this change:
- Moves the test_suite and split handling fully back inside
single_gradle_integration_test_per_source
- Gives new names to the top-level targets that had previously been
split from the autogenerated suite
- Updates the coverage report definitions to include the renamed rules
(the suite is not used in any other location)

Bug: 175035625
Test: N/A
Change-Id: I9072fa099f9112e5d75e96aacf8e631eb5ca38b0

Former-commit-id: 6b8344d2356a4fa1bd85d143c85da48b837f29a1