Better support for compiled bytecode in lint test support
Until now, lint checks could include bytecode by specifying the
binary data as a base64gzip string. However, that had several
disadvantages:

(1) Not obvious how to create these in the first place

(2) There's no requirement for the source code of the bytecode
    to be included, or be up to date, and in fact this was true for many
    of lint's existing bytecode based checks. In this CL, when switching
    to the new mechanism, I used fernflower to decompile a number of
    these test files where the original sources were lost to reconstruct
    the original source and in some other cases fixed the source code
    which didn't actually compile (e.g. removing unnecessary and broken
    imports etc.)

(3) Harder for lint to do some automated testing. Lint is in the process
    of switching to a new architecture where each module is compiled
    separately, which means that any reference to an API in a separate
    module will only be available as source. We want to make it easy to
    write tests where stubs and other calls are really binary, not just
    provided as source for convenience.

This CL defines a new test file, a "compiled" file, which can wrap test
files like Java or Kotlin. You can also specify the binary content of
all the outputs from this compilation (which in the case of Java
innerclasses means multiple class files, in the case of Kotlin package
level functions creates .kotlin_module files, etc). If you just specify
a source file, the first time the test runs it will fail and show the
binary content to be included in the test definition. This is done by
the test infrastructure compiling the sources using javac and kotlinc,
and then using ASM to visit the classfiles to figure out which ones
correspond to the current compiled file's source.

This CL goes back and updates a number of existing tests to use the new
mechanism.

It also updates lint to handle more recent bytecode versions (by
switching the class visitor from ASM5 to ASM7.)

Test: Existing and updated
Bug: None
Change-Id: I74f24d07fa7c0a438734d6108d1409c9190af78b

Former-commit-id: f27e30c458b4ee153cd2794b7eb1f127891fa35c