Run coverage genrules locally
With the restructuring of the post-processing pipeline, remote execution
is no longer *required* to reasonable execution times. Cutting the network
transfer costs via local execution brings the post-processing critial
path down to about 70 seconds. Currently, the machines invoking bazel
are about 32 cores, which is approximately the width of the
generate-XML-per-coverage_report bottleneck and thus doesn't cause issues.
As the number of coverage_report entries grows we may encounter slowdown
and need to investigate where the breakeven point is w/r/t re-remoting parts
of the workload.

Test: N/A
Bug: 174603669
Change-Id: I60c2b10f7432d4435669a7af706b8a634b1eec8a

Former-commit-id: 7ac2476345b47276afe6c6c40dfda9c7db2a479d