Make MergeNativeLibsTask a non-incremental task
This change simplifies the logic of MergeNativeLibsTask, improves its
performance, and removes its dependence on apkzlib.

Making the task non-incremental makes the task faster for clean builds
because we no longer need to save all the incremental state. The same
should be true for many incremental builds as well.

Running Dolphin_nativeCodeImplChange with this change results in a
dramatic performance improvement for the task (~5.8k ms to ~1.3k ms) and
an improvement in TOTAL_BUILD_TIME as well (~48k ms to ~45k ms)

Bug: 175337498
Test: MergeNativeLibsTaskTest, existing NativeSoPackaging(Options)Test
Change-Id: I1288a8377547b7e210aacb4009b834cd83513dd2

Former-commit-id: 54b3efd3c2fd44ee06cf63b2a303e4e49f835c6a