Reduce number of JNI global refs created by perfa
In order to safely cache jmethodID and jfieldID values perfa must
ensure that the classes they are associated with are not unloaded. The
standard way to do this is to create a NewGlobalRef for each of them.
This unfortunately can cause problems since on android (for
performance reasons) there are a limited number of JNI global-refs
available. Apps with a large number of loaded classes or which make
extensive use of JNI can run out of global-refs, potentially breaking
the app.

This changes perfa from using global-refs to instead creating and
maintaining a basic linked-list structure using Object[]s. This allows
us to hold all classes live with only a single global ref.

The linked list is made up of buckets which each contain (2 << 12) - 1
class objects when fully filled. Buckets are filled from front to back
with the first array index being the link to the next bucket (if
present). New buckets are prepended to the existing list. This list
structure is entirely opaque to perfa and no support for reading the
classes out of the list is provided.

Bug: 117829134
Test: manual
Change-Id: Ia0f9588e2fae6c6bbfe70810cd20b961feedaffe

Former-commit-id: cf837c88aadd1ca0a6a9f05eb63637197e8fe2b0