Create a new version of all repo xml schemas
This just copies the existing schema to a new version. It's not a useful
change without follow-ups to make modifications to the new schema.
The anticipated changes to be made for v2 are:
- change host-bits to host-arch with values x86, x64, and aarch64
- add "soft" flag on dependencies
- allow repeated <tag>s on system images
- allow sha256 shas
- maybe something to support mainline modules?

In this change we also add manually-run BUILD rules to generate the
java code for a schema, including for any dependant schemas.

Test: pending
Bug: 177929448
Change-Id: I0573b8f768d04cef79026f588c16773948dde489

Former-commit-id: d21ad07dac2f4ebf0f574e7a1b676300c59140b2