remove UnitTestBuilder and AndroidTestBuilder
instead of having beforeUnitTests and beforeAndroidTests APIs,
provide a simple knob on the variant to enable/disable the
unit test and android test part of the variant.

unitTestEnabled can be used to toggle unit testing, default is true
androidTestEnabled can be used to toggle android tests, default is
also true.

the beforeUnitTests/beforeAndroidTests are still present in the
AndroidComponentsExtension until androidx converts to the new
APIs.

Test: Adapted existing tests.
Bug: 177504396
Change-Id: I63b002ebd16a0dffe6282a2518dabe10b2dbd032

Former-commit-id: aa5e3b156901ba8d7b6e77dd737f36ff6055ddd7