Test GradleModelMocker via Lint models only (not using IDE models)
This is to make it possible to re-implement GradleModelMocker without
help of IDE models.

This change tests fewer configured properties than the old
implementation because not all of them are available in Lint models.

Some of variant related tests seem to expect the wrong values and they
will be fixed together with the new implementation of the
GradleModelMocker.

Bug: 37797956
Test: n/a
Change-Id: I869854992cb45017617a169c28ef1c969ddf9290

Former-commit-id: 0ec9860e53c93ebb7835287e47c9a3056700585a