Fix excessive notifications from DeviceClientMonitorTask
DeviceClientMonitorTask sometimes generated millions of notifications,
consuming most of the available memory. If an IOException occurs in the
channel.register call in processChannelsToRegister, mChannelsToRegister
is not cleared at the end of the method. When the main loop in
DeviceClientMonitorTask.run comes back around to the
processChannelsToRegister call, the offending channel will still be in
the list, and will presumably generate another exception. Since the
message is logged at ERROR level, this generates a notification each
time through the loop.

IOExceptions could also come from the select() call at the beginning of
the run() loop. While select() is usually a blocking call, if the
channels are in a bad state it's possible it could throw quickly,
leading to a tight loop generating notifications.

This change switches to lower log levels in some places, and ensures
that channels are removed from mChannelsToRegister even if the
register() call throws an exception. It also returns from run() if
select() throws an exception.

Bug: 162871030
Change-Id: Idf3633c9cfc262dfcdc9a0e2f7061394ad531a07

Former-commit-id: d266b754b7ab814891f0674a98045cb23d290c9c