Be stricter about threading access for root views
This fixes a race issue where sometimes you get no root views at all, if
you try to query them too quickly when the inspector first starts up.

This commit additionally introduces the "computeOnMainThread" method to
allow it to block and return a value, allowing for a fair bit of
existing code to be called in a more ergonomic manner.

Bug: N/A
Test: N/A
Change-Id: I587387a3855f3f952a4d0c8770bc60637e36597d

Former-commit-id: d01749346a24cdaaf559a54ce8f39cbc91fd18d7