Reintroduce targetSdk in libraries
It was omitted as it is essentially ignored for main,
it is written in the manifest, but consumers don't do anything
with that information.

In the future it would be helpful to validate, perhaps in a lint check,
that when consuming libraries their target SDK, if defined, is equal to
or higher than the consuming project's target SDK to verify that those
libraries have been tested on the newer platform.

For now, just re-introduce it in the API to not regress things for
build authors.

Bug: 177236412
Test: API surface validated by StableApiTest
Change-Id: If4677235ae20bc2a0cbf40e8f693fb0e8ef95861

Former-commit-id: 930c0cac82c434388375d2b00a0f52a36bd133f2