Change default from Release to RelWithDebInfo
Starting in NDK r23 and CMake 3.20, the NDK toolchain will
defer to CMake's Android toolchain. One effect of this will be
that the -g flag is no longer passed when CMAKE_RELEASE_TYPE is
set to Release. This would affect debugging in existing projects
and create upgrade confusion.

This CL changes the default to RelWithDebInfo in cases where AGP
would have automatically chosen Release. If the user explicitly
specifies CMAKE_RELEASE_TYPE=Release then we won't override that
decision.

Bug: 37128614
Test: Existing
Change-Id: I2973b21a0fa608d90f73c8d7de264fe2b0daac1c

Former-commit-id: 7defa2bbc55962d948fdcc16942e7aaab76607b9