Allow repeated tags on system images
Right now we have a single "tag" that has to indicate multiple
things--e.g. that an image is play store compatible, or that it's a wear
or auto image, etc. So when we have e.g. wear images that are and aren't
play store compatible, we have to have a special "wear without play
store" tag, and in studio we have to know that the normal wear tag
indicates play store compatibility, etc.
With this change we allow multiple tags, so we can simplify the logic
in studio and be more flexible with different image types going forward.

Test: pending
Bug: none
Change-Id: If3cfeca63e585d8bacae131e0e1667991192d304

Former-commit-id: 9ffe2647f8dca3b368f79063079f3813d2e81f56