Move fake-android under transport/
... and update all references to the new location.

It turns out, as I get ready to write tests for the app-inspection
agent, that I'll need to modify fake-android a bit (Looper stuff), but
this has already started breaking the transport agent tests, so I've
backed off. This seems to me to be a sign that each module should have
its own fake-android module, so I will be forking this code in a
follwup change.

Bug: N/A
Test: N/A
Change-Id: I1163133562ec1ed396540bf24694630f39b92987

Former-commit-id: 8eb60b2d38bbd08fdf4f08718dc40342b8fd39ce