Moves bundled dependencies to plugin def
This change moves the bundled library dependencies
from being defined on each module to the plugin definition.

Additionally this change introduces the same check for libraries
than the one we have for modules: If a module needs a library
dependency, then plugin layout is used to verify that the dependency
is satisfied in production.

Test: Added check at build time
Bug: 179064916
Change-Id: I1982d16e3b01a21f5508ddfbe92f7ca042fb7949

Former-commit-id: f9a95e9796cd12e74bc04d239ab9ea3c4d7786a7