Migrate some fake-android classes to Java
As fake-android is trying to mimic the Android Java framework as much as
possible, we need to avoid exposing statics that are called from our
Kotlin inspector, as this expects to find a Companion object.

That is, with this:

```
// example.kt
class Example {
  companion object {
    @JvmStatic
    fun doStatic() { ... }
  }
}
```

when it is called like this:

```
// main.kt
fun doExampleStatic() {
    Example.doStatic();
}
```

this generates bytecode like:

```
Example.Companion.doStatic()
```

even though the JvmStatic version is there. This is fine in most
codebases, but that bytecode explodes when running on an Android device,
because in the framework version, "Example.Companion" does not exist.

Bug: N/A
Test: Covered by existing
Change-Id: If8b49dd780f59d77e1fec1353865c827dcae7ca1

Former-commit-id: ab2ed178d6982250b5b6bd2764bd988cb4d387ac