Removes Guava library
This change removes the Guava library from
the tools/adt/idea project. Guava is provided
by the IntelliJ platform instead.

Currently we use a custom Guava as a dependency
but then we don't bundle it and use what is
provided by the platform. In fact both Guava libraries
are in the class path when running tests.

Bug: 179064916
Test: Removing a library, existing tests should pass with the other one
Change-Id: I4590bcf921aad25836c6c3eacc1680c84bee9ab0

Former-commit-id: fdf9c5e6368370e5b3c7b7765223b5c8bdf989ba