Run AndroidXDependencyCheck in more configurations
AndroidXDependencyCheck checks whether the project uses AndroidX
dependencies but does not have the `android.useAndroidX` property
enabled.

Previously, we performed this check only on the compile and runtime
classpath, but this may not be enough.

We now perform this check whenever *any* configuration is resolved.

Also clean up a few related places to make the upcoming changes easier.

Bug: 130286699
Bug: 167700246
Test: Updated AndroidXJetifierMatrixTest
Change-Id: Id536d6d6e4cd2c01ad82a02b9b830202874e4ecf

Former-commit-id: bb09dc6d780c72d0b99b065334e9f554b3362705