Handle large packet sizes in jdwp proxy
This change adds a flag for setting max jdwp packet size.
This change also changes the flags to be an env variable instead
of a java property.
Finally when a packet is greater than max packet size an exception
is thrown and the client is dropped mirroring the behavior of
ClientImpl

Test: manually, and updated
Bug: tbd
Change-Id: I19c3a77d7f7bd615caa62b35d1db6631d900e9fd

Former-commit-id: 227a52ad42d5ec9b9017a92041eb928660798e5f