[AGP UTP] fix error message when tests fail.
A non-zero return value is used by UTP when tests fail. As such it's
important to not throw an exception in this case so the Test results
file can be generated by the instrumentation test tasks. Otherwise it
is very unclear why the task failed.

We should still rethrow if an unhandled exception occurred however.

Bug: 141510559
Test: ran managedDevice and connected tests with a failing
instrumentation test on a project. This is how I located the issue.

Change-Id: I54e76066e53b8e72d0d21c41975ca39a110e2c8d

Former-commit-id: f4f96ff68585d0777c6957be1ec4df9bc693c9f0