Include all AAR metadata issues in the error message
Previously, the error message would only include the first AAR metadata
issue, but that is not ideal because a developer should be able to
consider all of the issues when deciding how to resolve the issues.

For this change, I considered only including the highest minCompileSdk
error message in the case when there are mutiple messages with different
minCompileSdk values, but the problem with that is that a user seeing a
single minCompileSdk error message might decide to resolve the issue by
removing the dependency, and it would be unpleasant to be surprised by
another minCompileSdk error message after thinking the issue had been
resolved, so I decided to include all of the minCompileSdk error
messages instead.

Bug: 150132272
Test: CheckAarMetadataTaskTest, CheckAarMetadataTaskTest
Change-Id: I2be04dd3a93d24c7cbfb6351481f9268a34280ae

Former-commit-id: ae955ac4d342818e6d2ce237402b299aa214dcb4