Allow detectors to vary severity per incident
Historically, detectors did not have the ability to dynamically decide
on a severity for an incident; the default severity is specified on the
Issue, not when you call report() for a specific occurrence. The reason
for that is that severities are left up to the user -- severities can be
changed via lintOptions, via flags like -Werror (turn all warnings into
errors), and so on.

This CL makes it possible to specify a default severity on a per
occurrence basis. This severity takes priority over the default severity
specified on the issue, but will be ignored if there is any user
configuration (such as a lint.xml file setting the severity) for this
issue.

There were two lint detectors that had slipped through the cracks and
had defined their own "error" and "warning" issues. This CL replaces
these issues with a single one for each of them, and to support that
properly, adds the capability of issues to declare "aliases", such that
when lint comes across one of the older names (in baseline files,
lint.xml, etc, it will correctly map the the canonical id instead.
Similarly, issue registries now have the ability to declare issues which
have been deleted. Prior to this change that was possible but only for
the builtin issue registry where it was hardcoded.

Fixes: 180017974
Fixes: 180018551
Test: Unit tests included
Change-Id: I1b04409e5e964c9bdf686d015a30e912d5494286

Former-commit-id: 9312a29cf6fcdb53f9728445ac2966cfde323409