[Lint Partial Analysis] Add partialResultsDir to LintModelWriterTask
This change modifies LintModelWriterTask to include the partial results
directory location in the lint model. The partial results location will
be included in the model even if we're not doing partial analysis. I
considered an alternative approach in which the partial results location
would be added to the model only if we're doing partial analysis, but
always including it is much cleaner, and in the future we'll always be
doing partial analysis when this task runs.

LintModelWriterTask.partialResultsDir is a File instead of a
DirectoryProperty, and I used the ArtifactsImpl::getOutputPath API
instead of the usual ArtifactsImpl::get API. The reason for this is
because the task should depend only on the location of the partial
results directory, not its contents, and we don't need this task to
depend on the AndroidLintAnalysisTask that produces the partial results
directory.

This change also registers the lint analysis task in the standalone
plugin.

Bug: 178810169
Test: existing, will add integration tests after adding new report task
Change-Id: I535494f544a10b67e72dff33d534fa92a2a6d467

Former-commit-id: 21b5d12f4c29a68d60c0d88dc627415896a60559