180078655: Default project has "does not specify a vendor{...}" warnings
Recent versions of lint requires issue registries to provide a vendor,
such that warnings can be attributed to the right place.

However, there are a lot of existing AndroidX libraries in the field
which do not specify a Vendor; including some that we reference in
default projects.

Lint was already guessing a vendor for some group id prefixes, so in
this case, don't log a warning. We should still add vendor registration
to the androidx libraries such that they can properly set up their
contact information, bug URLs, etc.

Test: Existing (updated) unit test
Bug: 180078655
Change-Id: Ia92b5351a95e3d23e22e4b987f9fe258ab15f140

Former-commit-id: 0089270496b7890950f683e80c6715c1915b2126