Update model dumper to handle default build tools version
In order to not have to update all the snapshots everytime
we update the default build tools revision, make the snapshot
recognize the default value and output it as such insted of
the actual value.

This also tweaks the dump of enums to not have the string quotes
in them.

Bug: N/A
Test: updated existing snapshots
Change-Id: I95607b60c3f5bae582966ce168a8d7c3426dac40

Former-commit-id: 070d4bb4962173d2b4a459c7454c09293cf9d1f7