Hook up bytecode transformation code to network inspector
This change hooks up the HttpURLConnection transformation code so the
inspector can start sending events to Studio. OkHttp not finished in
this change.

The flow of how HttpUrlConnection is tracked:
1. Http(s)URLConnectionWrapper wraps user's object when they call
URL.connect().
2. Wrapper delegates the tracking logic to TrackedHttpURLConnection. It
provides the fully substituted API of HttpURLConnection and
HttpsURLConnection.
3. It uses HttpTracker (implemented as ConnectionTracker), and
StreamTrackers to record and report request/response headers, payloads.
4. Reporter (new concept that didn't exist in profilers) facilitates the
sending of events to Studio.

This change maintains the structure of the code that was in profilers,
with the exception of the new reporter classes.

Test: To be written in future changes.
Bug: 178232811
Change-Id: I4d0a6aa9a20b776ba0338c62af5b836f04e12a55

Former-commit-id: 6ca4107fc4621997b3d9e5e7a610c0a7d09fbdd7