[Lint Partial Analysis] Fix checkDependencies in dynamic feature model
Previously, the LintModelWriterTask would always set checkDependencies
to true, but we want checkDependencies to be false when generating the
dynamic feature model when checkDependencies is false in the app module.
In the future, we may want to revisit this and sometimes set
checkDependencies to true when running LintModelWriterTask from dynamic
features (e.g., when checkDependencies is true in the app module), but
we're setting it to false for now in dynamic features because
SantaTracker_withPartialLint_lint fails otherwise.

This change also moves the tests from DynamicAppLintTest to
LintDynamicFeatureTest and adds some more test coverage to demonstrate
the current shortcomings of running lint with dynamic features with
partial analysis enabled.

Bug: 178810169
Bug: 180672373
Test: LintDynamicFeatureTest, SantaTracker_withPartialLint_lint
Change-Id: Iecf865d8aa48fc2510f609a136d0e2034391c8ef

Former-commit-id: 48c802064bc253019c076e9501b202b3dda6d846