Add a test to ensure features are not in SUPPORTED stage
When a feature is in the SUPPORTED stage, users can enable or disable it
without seeing any warnings as AGP will support both cases.

However, the SUPPORTED stage doesn't specify a clear timeline and
features may stay in this stage for too long, thus increasing
maintenance cost to AGP and users.

Instead, we should use other stages such as `SoftlyEnforced` or
`Deprecated`, which have a clear timeline.

Bug: 144917849
Test: Adding a test in BooleanOptionTest
Change-Id: I893f7372bf45ad34faf73955487f103af7432676

Former-commit-id: 792fbf69a90cb82b0d2aacffb82776af16192c0f