Move miniconfig file to build/ folder
Moves the location of android_gradle_build_mini.json to underneath
the build/ folder so that it can be cleaned up by clean. As part of
it, I move the definition of this file's location under CxxAbiModel
(which is the standard). This, in turn, required a smallish refactor
to plumb CxxAbiModel to the locations it was needed.

Bug: 168805638
Test: Existing
Change-Id: I1d2063777c8ca385495d88b7adabd30e5de2f83c

Former-commit-id: 4688a5bdfb312d5d46f3165e746349a2b28f2ebf