Remove remote:test only jvmopt configuration
testCustomDockerImage is removed as it relies on an environment variable
only present in the remote docker image. This would never pass in a local
test execution. We want to avoid tests that only pass on remote execution.

This test relied on a --jvmopt flag configuration which invalidates
all test actions when switching between configurations. We want to
minimize unnecessary configuration changes to improve caching. Having a
separate configuration for remote tests means cached test results
cannot be shared by non-remote configurations.

Bug: 168721038
Test: NA
Change-Id: Ie2eb78c73dbfc4272d6510227ec6ca2cf357aced

Former-commit-id: 153a2fd3c4b174b448cc988166f3d694a447715f