Add x86_64 build variant for transport daemon
32-bit Android emulator system images are not supported
in Android 12 (API S). At the same time, 32 bit ABIs
("x86" or "armeabi-v7a") are dropped from the 64-bit
emulator system images. As a result, on an API S AVD
running on x86_64 (e.g., Intel based MacBook), "arm64-v8a"
would be the only compatible ABI for transport daemon
before this change. It appears to be running but is slow,
and it may freeze randomly especially when it invokes
system calls.

Adding a x86_64 build for transport daemon is the
most performant and cleanest solution.

Fixes: 178797055
Fixes: 179464588
Fixes: 181506492
Test: existing
Change-Id: I02516657e4e11b2f16c459b658a7b06ad936bc9d

Former-commit-id: 6b422a043d134aa639ca1f6493745541fc691a1c