remove VariantScope::getProguardFiles
some tasks were still using directory connections to wire task
output to tasks inputs. This will be deprecated in Gradle 7.0
where task dependencies must be explicitly declared.

The difficulty here was related to defaultProguardFiles that
can be added to the list as a file reference in the DSL.

To retain the behavior, the list will be reconciled at execution
time where reference to the expected locations of the extracted
default proguard files will be swaped for their final values
coming from the InternalArtifactType.DEFAULT_PROGUARD_FILES.

The consumerProguardFiles and testProguardFiles will be handled
in subsequent CLs.

Test: Existing.
Bug: N/A
Change-Id: I8b78334cc808d940b4dc27899c5cc8b671c25014

Former-commit-id: 9c70134890773a93f588502c08c57fcaf58cc9f0