Add RepoLinker
RepoLinker is similar to RepoBuilder in that it takes a list of
artifacts and resolves the Maven repository paths for each artifact.
Instead of creating a zip file, it creates symlinks instead, speeding up
the process.

This library will be used as a replacement for RepoBuilder and for a
rule that replaces the maven_repo rule.

Bug: 148081564
Test: RepoLinkerTest
Change-Id: I71fac129230e0a54bc45f75dac79cdbad4c041df

Former-commit-id: a877a51b275863c88c7aa97cb22abe644be8de91