Fix use of Files.copy(Path, Path) in MergedAssetWriter
Files.copy(Path, Path) is problematic because, on Windows, it forwards the readonly bit (see b/182063560).

This CL introduces a replacement method, FileUtils.copyFile(...), that removes the Windows readonly bit if it was forwarded. It also fixes the immediate use of of File.copy(Path, Path) referenced in the bug.

There are many other cases that will be covered in followup CLs. The final CL (ag/13877946) adds a lint check to prevent new cases of Files.copy(Path,Path).

Bug: 182063560
Test: Added unittest for FileUtils.copyFile()
Change-Id: I3c1c23238d3baf0733e04c72d237e6db0300b641

Former-commit-id: 948b6eadfbeefaa86ba21355bb31785defd0cfa4