Include rendered version of the API guide
It's in HTML so should update reasonably well after future updates. This
is similar to how Filament includes both the markdown and the rendered
version in their sources.  (The rendered version has the advantage that
we can link directly to it from gitiles etc which normally doesn't allow
scripting.)

Also includes some simple docs tools, which in addition to the
rasterizer contains some tools to clean up trailing whitespace,
fix quotes, etc.

Test: N/A
Bug: N/A
Change-Id: Ie131733fb5352b067bc6c80ca09bc79356b6bebc

Former-commit-id: 2ad77db3689d0b587887b99dd7c926effb228d69