Don't check dependencies when running lint from dynamic features
Previously, when running lint from a dynamic feature, AGP would respect
lintOptions.checkDependencies if it was set in the dynamic feature's
DSL. With this change, AGP no longer respects the value set in the DSL
and insteads sets it to false in any of the dynamic feature's lint
tasks.

The catalyst for this change was
Ib990333ebdf54d4e2ea9df2f4deecadaad76d370, which improved lint's
handling of dynamic features, but broke running lint from a dynamic
feature with checkDependencies true. This change adds back a modified
version of the test deleted in that change.

In the long run, the plan is to remove lint report tasks from dynamic
features and also analyze all dynamic feature dependencies when lint
runs from an app with checkDependencies true.

Bug: 178810169
Test: LintDynamicFeatureTest
Change-Id: I3d6c84c972799cb10376d244096b2ae2ece99cb4

Former-commit-id: 25cb34a15909c6a2845954759151d1968c8e4600