Warn when testApplicationId is used as namespace in test modules
For test modules, if there is no namespace from the DSL or package
attribute in the manifest, we use the testApplicationId as the
namespace. This change adds a warning when we do this because we plan to
discontinue this behavior in AGP 8.0.

This change also modifies the error message for all module types when
there is no package attribute found in the manifest; the message now
includes instructions for specifying the namespace via the DSL.

Bug: 170945282
Test: NamespaceAndApplicationIdTest
Change-Id: Ie8e66d7978f468a9afa56e800b7dc0c73e4c06b1

Former-commit-id: 3242f9e99b745b05613777e43513229168609894