Fix bug where CMake project was incorrectly invalidated
In this bug, the file metadata_generation_command.txt was deleted by clean task.
This caused configure task to delete the module/.cxx folder as 'stale'.
This test checks for a build side effect: the presence of .ninja_deps file.

This bug was found during C++ build cache investigation. The bug # below tracks
that work.

Bug: 182809209
Test: Added integration test
Change-Id: I17f4790e188e3597004cd7d379bb64f5b7cc2001

Former-commit-id: d4ef24852d3501c5bcf0e0b926bfa761919939d2