[tools/base] Remove The Deprecated LogCatTimestamp Field of LogCatHeader
The origin of the duplicity is that there are 2 ways of parsing a timestamp.

One way using a "MM-DD HH:MM:SS.mmm" format and another using an Epoch integer. When parsing from an Epoch, we store an Instant but when parsing from date-time, we store a LogCatTimestamp instead.

We can get rid of the LogCatTimestamp in the header if we simply parse the date-time into an Instant. The date-time string is missing a year so we have to "make one up".

Using a single representation of the timestamp also means we don't have to have 2 ways to format it into a string. We current have 1 way to format from an Instant and another to format from a LogCatTimestamp.

Bug: 185479433
Test: Existing tests provide coverage
Change-Id: Ifdbbaeb01212d2c865ff847fa82ffd2ac98cdf4c

Former-commit-id: dab3969562fef37a6d87ea50c021fec6e644f089