Fill v2 dependency model with proper graph
Because there isn't a single API in gradle to give us
what we need, this combines the old code that creates
ResolvedArtifacts (Which provides the type of module
and path to artifacts) with the ResolutionResult API
that allows us to traverse the resolved graph.
This lets us creates a proper graph for the model.

However, there's currently an issue in Gradle 7.0
where ResolvedArtifactResult coming through
ArtifactCollection do not have any Capabilities
listed, so there's a need for a workaround.
This is achieved by looking at the artifact name
since we mostly need this for test fixtures and the
test fixtures capability is appended to the artifact
name
Bug reference is https://github.com/gradle/gradle/issues/16919

This change also refactors the processing of the dependencies
as we don't need to share that code with the v1 model, making
things simpler.

Bug: N/A
Test: Updated the snapshot code and the existing tests' snapshots.
Change-Id: I075b667d7fb605eacf235b87780007847f68d095

Former-commit-id: 963302a67d050898df50dc7976d2a77c5deeae86