Revert "Stop running targets tagged as `very_flaky` in coverage"
This reverts commit 5aea9b426fee49e59aec3ddf180415b5ceeda03a [formerly b7aac2995d30e1a39d85d2b7cce872e98b49e3c2].

Reason for revert: This change is a no-op, which is good because if it had its intended effect it then coverage post-processing would have exploded. Since the added build tag has no semantic effect it should be removed to avoid confusion by later readers.

The coverage bot invokes a test_suite and --test_tag_filters will only check if the test_suite has the appropriate tags rather than whether tests in the test_suite do.

In the future, please do not change the coverage bot's behavior without my +2 unless it's an emergency and I'm unavailable.

Change-Id: I61206b85f2647db9765af67a583402d7be70328d
Bug: 153972155

Former-commit-id: 3b7238a88e699e82d3b13cd81fa93099078316c1