Replace "host" cfg with "exec"
The host config is a deprecated config which does not take into account
that builds may run on a different remote machine. See
https://docs.bazel.build/versions/master/skylark/rules.html#configurations

Other rules should be replaced as well, but for right now this is only
focusing on protoc. This is because we are seeing some edge
cases where remote builds fail on protoc actions for some users, but
this has not happened in CI builds (which always uses remote execution).

Bug: 182172121
Test: Presubmits
Change-Id: I77d7f473ea1354dc081217be2efe7e7fe2e00583

Former-commit-id: 24adf6583843871103df416f2b638677afc46016