Log installation exceptions as warnings
Installation failure typically is not a product bug as there can be
many reasons behind the failure. In this case, user dismissing the
install dialog reports installation failure. The installer handles this
correctly, by reporting that installation failed.

Logging this as warining is more appropriate - it will not be filtered
from the log, but is not flagged as a bug (i.e. error).

Fixes: 146228187
Test: N/A
Change-Id: I1bc2268e98f6d95ca0f3057acd09ec8fa442071c

Former-commit-id: 75617c5aed7e06195ed559d9ef5e2c4b8612abd2