Warn if an intent-filter data tag has more than 1 attribute
Because all data tag attributes are combined with all others, it can
be confusing if any data tag declares more than one attribute, since
that would imply they're required in combination, not individually.

This means that a declaration of

<data scheme="http" host="example.com"/>
<data scheme="https" host="example.org"/>
<data scheme="customScheme" host="companySpecificHost"/>

will also match "https://example.com" (maybe intentional) and also
"customScheme://example.com" and "http://companySpecificHost", which
are definitely not intentional.

Bug: 176909931

Test: com.android.tools.lint.checks.ManifestDetectorTest

Change-Id: I5c7af01f75348927f6dce4893ea8f661902e66b8

Former-commit-id: c714c6443e15639b9e7983c0cefa97fb7af35a15