Implement C/C++ structured log support
This implements a protobuf-based structured log format. It's
intended for granular logging and supports a string table to
persist string values as int32. Structured logging is turned
off by default and can be enabled by the end user by creating
a folder at:

  $PROJECT/.cxx/structured-log

The first structured log message is just the old LoggingMessage
redone in protobuf format.
A future CL will implement specific log messages for build
cache to make it possible to debug issues in the field and to
write integration tests that check whether caching happened
or not (and why).

Bug: 182809209
Test: Added new unittest and adjusted existing tests
Change-Id: If77679c9e3f0622e627b221709ad660911b65b32

Former-commit-id: de8b451ce40d43336339002bc46b4c2504d847b0