Add support for checking view roots in a thread
This change introduces an inner RootsDetector class which both
encapsulates all the root detection logic from before and adds a way to
start / stop a thread that additionally will periodically trigger the
check on its own. This will allow us to capture cases where a window is
removed from the system but we aren't notified about it.

Note: We remove a handful of "empty event queue" checks from tests, as
they might actually not be empty depending on the timing of the check
roots thread.  However, events will continue to be received in the
same order.

Fixes: 184973796
Test: Added + covered by existing
Change-Id: I50c4548c2f8876fe72911afc561a60a15a99df9f

Former-commit-id: 48b6c77613d26040523e78d22d3180f2ec1efa85