Set lnzipper strategy to local
The execution requirements of the lnzipper expect a filesystem
representation similar to that of the workspace. Other
execution strategies like sandbox or remote use symlinks when building
the action's execution root directory. To have lnzipper correctly build
a ZIP with the proper symlinks, we must use local execution to avoid
archiving sandbox symlinks.

Bug: 185519599
Test: NA
Change-Id: I121ae406079a099ad546489424068c75f1066e9b

Former-commit-id: 9d2347cf85dd54921725df33a1a77745a770d03e