Prevent disposed callbacks from throwing exceptions on GC
While it is considered an error to never reply to or cancel a command,
as this indicates a programming logic error, a case that shouldn't
count is when commands were interrupted by an inspector shutting
down.

The only way we could detect that a command was never replied to was by
throwing an exception at GC time. However, what was happening was you
could start an inspector, stop it, then restart it, repeat a couple
times, and eventually a command callback would be GC'ed that had been
interrupted by a previous run. This would throw an exception, killing
the current inspection session.

We now handle disposal explicitly.

Fixes: 186459719
Test: This is not trivial to test. Manually ran a case in layout
  inspector that used to repro fairly consistently but is not
  anymore.
Change-Id: I1c6faad79659a7dbafc8a63e0a5c416016a3d6fe

Former-commit-id: cb3f02415081b0275a47d26d5651913e509db6b1