IRR: Fine-tuned AdbInstallerChannel timeout
Timeouts are provided for whole operations but at the socket level, we
can detected a device failure with fine-tuned write timeout.

This CL introduces the concept of "per-write" timeout which applies on
individual write operations. If not data is transferred within this
this time period, the operation is considered to have failed.

Test: None
Bug: 183285087
Change-Id: If8e358030af1baa9930f63ddc2dc34893e3f6169

Former-commit-id: 29dfee4374caced543d88f867b92197624303193