Allow non-unique package names in manifest merger
Test manifest merger will create a temporary manifest file and is added
to merging as a library manifest if a test manifest already exists. In this
case if the original test manfiest has a package name = "xxx.test" and the
temporary manifest has a package name = "xxx", there will be a warning message
indicating that two manifests both have a package name "xxx.test". This warning
message is undesired.

Bug: 186901700
Test: ProcessTestManifestTest
Change-Id: I18e101e2321c534523baa6417d3c7bfc9570bd42

Former-commit-id: fc60dba74e3de8e36463642ad32a040e3613b2f7