Implement reader\writer for Ninja build dependency file
For build cache support. This CL implements a codec for reading
and writing .ninja_deps file. This is the file where Ninja
stores, for each .o file, the list of dependencies. The file
format is binary with strings interned, but the effective format
is a list of lines like this:
.../WiiUtils.cpp.o|1617220563|WiiUtils.cpp,.../string,etc.
.../IniFile.cpp.o|1617220563|IniFile.cpp,.../jni.h,etc.
The three segments delimited by | are:
1) The path to the object file relative to the .ninja_deps file
2) The timestamp (which is custom ninja and platform specific) of
   the .o file.
3) A comma delimited list of dependencies, including the source
   .cpp file. Paths are relative to .ninja_deps file or absolute.
Since this information is per-source-file, it's important that
at least the reading part is very efficient.

Bug: 182809209
Test: Added unittests including some with real-world snapshots
Change-Id: Ifb7ab64d848954ec93160cf0ad3c74cb23d4d2f2

Former-commit-id: 7fe61ef512a3adad7a99f2e56132760e2bd9b016