Multiple vairant publishing for library projects
With this change, we are adding multipleVariant publishing support for
library projects. The multipleVariant publishing support for application
projects is not added for now because currently we don't create all
component for application projects.

With multiple variant publishing dsl, users are able to specify which
variants to be published and we add attributes(buildType, flavors) if
necessary to disambiguate those published variants. If not component
name is provided, we publish variants to the "default" component.

In terms of testFixtures, we maintain the existing behavior if users
don't opt in the new publishing dsl. When using the new publishing dsl
feature, we publish the test fixtures with its main variant by default.
Whether its main variant is published is determined by the new
publishing dsl. Users are able to disable test fixture publishing by
using gradle AdhocComponentWithVariants api.

Bug: 150132264
Test: Test: existing + LibraryPublishingTest + TestFixturesTest
Change-Id: Ie1addbd7333d8b31a341de7b121c2930735f6315

Former-commit-id: a1f5cc28053b48019af45969159099dabf3531e9