Data binding: Sort files to ensure deterministic order
java.io.File.listFiles() does not guarantee consistent order of the
returned files. The same is true for
org.apache.commons.io.FileUtils.listFiles() which data binding uses.

This commit makes sure the files being used are sorted to ensure
deterministic order, which is important for deterministic and faster
builds.

Bug: 187340555
Test: Existing tests (new test will be added in data binding)
Change-Id: I10ae032812cce93a862fbe24bee01b918eef4d6d

Former-commit-id: 0e6207fc37b87adda3e044606646d64e9a462f7d