Skip data binding base library in AndroidX check
We have a check to ensure that when android.useAndroidX=true, there
should be no legacy support libraries in the dependency graph.

The exception is com.android.databinding:baseLibrary, which
androidx.databinding:databinding-compiler depends on (baseLibrary
doesn't need to be replaced with AndroidX).

Therefore, we now ignore com.android.databinding:baseLibrary in the
check.

Bug: 187448822
Test: Manually verified that
    bazel test //tools/data-binding:runtimeLibrariesAndroidX doesn't
    produce a warning (it was showing a warning before)
Change-Id: I8ebb28c77afabaf478cadbe10ad9aa69c1559555

Former-commit-id: 7c58248f99312f443a28d1220e6322d342640d12