Add build target to compile_commands.json.bin
There's a lot of files but most of them are
updating out or Dolphin baseline file API reply.

When restoring from build cache we will need to
know the build target for each source-to-object
translation so that we restore only the targets
the user has requested.

This CL adds build target to compile_commands.json.bin.
It is a required field and there is exactly one for
every source-to-object translation.

Finding the target name is done differently for
ndk-build and CMake. For ndk-build we know the eventual
library name and this becomes the target name. For
CMake, we have to inspect the output path that CMake
creates. CMake file API doesn't provide this
information elsewhere.

Bug: 182809209
Test: Added unittests for the new logic
Change-Id: I41fbd0e2e9174de6a350ae2d8955da8fb2aab28b

Former-commit-id: 53ed05485fb35694ac60b360c3c1df46eb83e9ee