Update agent to support new version of coroutine lib
In newer versions of the coroutine library, DebugProbesKt is moved
from the debug module to the core module. So we don't need to rely
on DebugProbesKt.bin to do the bytecode replacement.

This change updates the agent to check if DebugProbesKt from
kotlinx-coroutine-core exists. If yes, it instruments DebugProbesKt
from the std library to call functions from it.
If not it falls back on the old behaviour of the agent,
to be backward compatible with older versions of the coroutines
lib.

Test: Will do separately
Bug: 182023904
Change-Id: Icb4fe86a6ac1a9f2d52c9e5f185f4696d9713c24

Former-commit-id: 19f1c1b00b9cbe6106f7eeb058b8565ab14a531b