Fix concurrency issue with Icebox and Orchestrator
Icebox queries process ID using the process name, which does not work
very well with Orchestrator because Orchestrator relaunches the process
multiple times and causes Icebox getting a stale process ID
occasionally.

Previousl we worked around it by adding a delay, but it make the
integration test flaky. Here we have a more appropriate by querying the
logcat to see when it starts to wait for debuggers.

Bug: 184536229
Test: FailureRetentionConnectedTest#connectedAndroidTestWithOrchestratorAndFailures

Change-Id: If641299d0bad1cb90aca7d3a78bdbec5cdb98d19

Former-commit-id: f3f8366b40c0301cd5fd57c96f15808e0069b080