Fix SubtractingArtifactCollection for android test
With mixed java-kotlin projects there are 2 published
artifacts under the same variant. This causes ArtifactCollection
to return inconsisten results (https://github.com/gradle/gradle/issues/17213).

Because of this, Kotlin classes were missing from the test APK,
as we were using the resolved artifacts to get the list of files.

This change computes the list of removed files, and any
artifact file that is not in the removed list is kept.

Bug: 187353303
Test: AndroidTestDependsOnKotlinProjectTest
Change-Id: I80adfe86202d8466e712956d7a31590a040200ee

Former-commit-id: 447e6cb62d1520074572fb26c27f3d40e3ee78dd