[AGP DSL] Clean up new lint block
The move from lintOptions to lint means we can clean some things up to
be more consistent for kotlin use.

This change separates Lint from LintOptions and simplifies how things
are exposed, especially in kotlin script.

* All collections are directly exposed as Set<Strings> to allow consistent use from
kotlin script

* Severity overrides and the enable/disable sets are unified, using a
custom set implementation that removes the issue id from other sets when
it is added to a set.

* Locking is fully supported.

Also, switching to the new type for the standalone plugin would be a
breaking change without a rename, so introduce a new 'androidLint'
extension with the new type. Eventualy the old type will be deprecated
and removed.

|          | Android Plugins      | Standalone lint plugin |
|----------|----------------------|------------------------|
| Existing | android.lintOptions  | lintOptions            |
|      New | android.lint         | lint                   |

Bug: 140406102
Test: New unit tests
Change-Id: I00afbf95d8b2f67bd08f800fc022d978f15208e7

Former-commit-id: 80da7430ce5c4942b774ca16c892a5f31134b40a