C/C++ build cache protobuf definitions
These are the definitions for protocol buffers used
by C/C++ build caching. Each one has a corresponding
'Encoded' version that is used for writing that
information to structured log.

DependenciesKey -- contains the information needed to cache
#include dependencies for a particular .o file.

ObjectFileKey -- is a superset of DependenciesKey and
additionally contains information about the (implied)
#include files.

Bug: 182809209
Test: NA
Change-Id: Ie263c3166f3ea8900609ce7649c08de1e79e22cc

Former-commit-id: bb8395a7442b35d33975eb3da60be71adab0bad4