Refactor dep graph builder code for testing
Extract the actual implementation of the global library
service into its own class, out of the implementation of
Gradle's BuildService so that it can be used in tests.
Also introduces some interfaces for it.

Extracted an interface for ArtifactCollectionsInputs as well
so that tests can use fakes.

Tests will be in the next commit.

Bug: N/A
Test: existing tests only.
Change-Id: I412ec1acbb089813862bec54b5cd4b8cf064ec6c

Former-commit-id: ff27db56740db476387e47aaddfeff7d0d8365ea