Settings plugin to set default values for AGP DSL
The settings plugin (com.android.settings) is applied in the
settings.gradle file. It exposes a basic DSL with compileSdk
and minSdk. If set these values are applied to all projects
in the build automatically (as long as they apply an AGP plugin).

The DSL is somewhat similar to the normal project DSL but there
are small difference. First it must be common to all projects so
there's no customization per plugin type. It also does not allow
creation of flavors or build types so minSdk is directly under
android {}.

In order to test this, the fixtures are changed to not include
a default compileSdk or minSdk version. Therefore quite a few
tests are updated to manually set these values where applicable.

Also had to change the fixtures around settings.gradle generation
in order to setup plugin management repository. This whole
area could use some work...

Bug: N/A
Test: new model tests to validate the final value is correct.
Change-Id: Icef6d3c958949e67012afa04f7d670f148856605

Former-commit-id: 4bca5a333a5213db231974ee6c22f24db254674a