android.disableAutomaticComponentCreation not to be experimental
In order to encourage users to use the new publishing dsl, we ask
users to use android.disableAutomaticComponentCreation=true to "hide"
the warning about the upcoming behavior change. Therefore, when the flag
is set to true, users should not see a different warning about using the
flag.

With this CL, we only raise deprecation warning message about this flag
when users are explicitly using the unrecommended value(false).

Bug: 189367188
Test: existing
Change-Id: I6e7850f87b76ba7cd6a7cd6ac5a70bc8def3a9d4

Former-commit-id: c40cf5b950ec60a48aca68326b4eb046637c6c33