Add lint test modes for compiled source files
Lint's testing DSL lets you provide a testfile for bytecode with the
corresponding source code. The primary purpose for this is to preserve
the source code that the bytecode was originally compiled from such that
tests can be reconstructed or updated as necessary. In addition, there's
support for easily creating the bytecode markup.

However, the testing infrastructure also takes advantage of this in the
partial analysis test mode; for any compiled files that are found in a
library, the corresponding source code is deleted (to simulate what
happens in AGP, where source code is not made available outside of the
current module). This mechanism is specific to libraries.

However, there are cases where a detector attempts to work with both
bytecode and source code, interchangeably. This CL adds a couple of new
test modes -- source only and bytecode only -- where tests that contain
compiled() test files are run extra times, with bytecode and source code
deleted respectively, and then the output is compared.

Test: Unit tests included
Bug: None
Change-Id: Id0913134ea05efe193a814b2e169952f5ce79e48

Former-commit-id: d6c5b0831eb148cc492347959ae9260a058bf26e