Check that all imports in lint unit tests can resolve
This CL adds a check to lint's unit test runner which visits most of the
import statements and makes sure that they resolve to classes, methods
or fields; this helps catch the common mistake where you reference a
library but forget to add a stub, which a normal compiler would prevent
but which can lead to tests not working as expected (or worse, detectors
being written to match the broken behavior observed in the test, using
simple instead of fully qualified names, etc.)

For detectors that specify a set of method names or reference names they
want to visit, it will also flag any references found in the test files
that match these names but cannot be resolved.

Turns out a number of the built-in tests also had some unresolved
symbols, so this CL updates the tests to pass this new check.

Test: This is it (and many updated tests)
Bug: None
Change-Id: I111025a82950b5fac2981289af82c083190aa889

Former-commit-id: a2d0a4ca4d8939ab1751f18a617ff7a80b6e0431