Fix synchronization issue during snapshot capture
If there are many redraws happening and you tried to capture a snapshot,
we could serialize the skp, start processing in captureExecutor
(removing the snapshot request and thus no longer requiring
serialization of SKPs), and then have the
StreamingPictureCallbackHandler fire again before we actually wrote the
serialized SKP to the output stream, causing it to be dropped and
resulting in an empty SKP coming back in the snapshot.

Changed so that we don't stop capturing SKPs until one is actually
written out.

Test: none yet
Bug: 189149509
Change-Id: If1e2c5c5678d566479e801b8d2795d6f5c0762c3

Former-commit-id: b527871174cf1fb60b28a3028dcef2b5e7d25c8d