Don't fail build when AndroidLintCopyReportTask inputs missing
Previously, the AndroidLintCopyReportTask would fail if it tried to copy
lint report files that didn't exist. There are at least 2 ways this
could happen: (1) if the copy task is invoked directly or (2) if the
lint task is skipped or fails before writing its reports.

In the long run, AGP should remove the ability to write lint reports
from different variants to the same file, which will make the
AndroidLintCopyReportTask unnecessary, but for now, it's better to
warn when a lint report don't exist instead of failing the copy task.

Fixes: 189877657
Test: AndroidLintCopyReportTaskTest
Change-Id: I6ad69a92baba480ea2ae34363dd351999f1938db

Former-commit-id: e47fb94d2425d9468d036cbfe5c47b8cf3048e59