Add issue documentation generator
This tool generates markdown pages with documentation for all the lint
issues (either built-in, or in a provided list of lint jar files).

These are currently hosted here:
https://googlesamples.github.io/android-custom-lint-rules/checks/index.md.html

This CL also fixes a few issues revealed by this work: the constant
evaluator wasn't combining characters and strings correctly, and was
made to understand the .trimIndent and .trimMargin "built-ins", as well
as String.format (when in allow-unknown-segments mode).

Finally, various issue explanations and unit tests (which are pulled
into docs as samples) were edited for clarity, formatting or
correctness. Some other minor cleanup.

Test: Unit test included
Bug: None
Change-Id: I52fe21e4944f2ca3c1ea59a54ddab5d649308757

Former-commit-id: 0a8e0a85add969077a0a099ec2990585894bb378