Fix NRE in CMake server when buildDirectory is missing
I don't know the cause of a missing build directory; we
haven't seen that before. I elected to make it a warning and
ignore case rather than trying to recover by using the relative
path because, downstream, we're going to try to copy the file
to the output folder and it won't work. We have the best
information here to give a good report.

Bug: 188860472
Test: Unitested
Change-Id: I27f4b69819ef744c93e8a7a05b974a1bee8d79f4

Former-commit-id: 7dc93f34e60ee6a59cdf76de66e612e10bbd941b