Remove rendered docs
We don't need to keep pre-rendered versions of the Markdeep docs in the
source tree. Originally, the idea was to make it easy to host these
pages directly, but cs.android.com doesn't let us render HTML. We
separately host the pages on
  https://googlesamples.github.io/android-custom-lint-rules/index.html
now, which involves a separate publishing step (where we also generate
docs for all the individual pages), so leave the HTML rendering for that
publishing steps. This also avoids having to update HTML pages every
time we make edits to things like the changes file (which has a high
likelihood of creating merge conflicts, especially during cherrypicks.)

Also fixes a few issues with the publishing scripts.

Test: N/A
Bug: None
Change-Id: Ib149abdf76546bea4ed656434de444212c493f0a

Former-commit-id: fd119045a91d63372b61919c62e594391d8295ae