Include alwayslink archives as input
It seems like link.exe doesn't like it if you pass in only a
/WHOLEARCHIVE argument without any normal inputs, which will currently
happen if you try to build a cc_binary from a single alwayslink=1
cc_library with no non-alwayslink dependencies.
Including the /WHOLEARCHIVE library as a regular input as well should be
ok, I think? And make link happy.

Test: build with I808a4653 should work
Bug: 129401912
Change-Id: Id28c9eb15a88ca1405852d125869a509efb0d760

Former-commit-id: 0ed736d9bae7392940fbfb281a16318b8a292f59