Lint override: Don't use maven versions
For historical reasons, the maven artifacts for lint use version
numbers, which are 23 higher than the Android Gradle Plugin, so lint
7.0.0 is published at com.android.tools.lint:lint-gradle:30.0.0

If a build author specifies 7.0.0, this is now automatically mapped to
lint maven artifact with version 30.0.0. For simplicity, this change
always adds 23 and hides the fact that the artifact has a different
versioning scheme.

Also validate that the version specified has at least the same major
version as the default, with a warning if lint is behind AGP to avoid
creating expectations of unlimited compatibility with older versions of
lint.

Bug: N/A
Test: Overrides unit tested in AndroidLintInputsTest
Change-Id: I2fb6c3af56f6326868b6bc1a2b126dcedd62763f

Former-commit-id: 95de7eebfdb73d1dbb738bd4e804ced7a973a711