Delete calls to UastEnvironment.disposeApplicationEnvironment()
Previously, ExtractAnnotationsDriver and LintIssueDocGenerator made
explicit calls to UastEnvironment.disposeApplicationEnvironment().
It is pointless to call UastEnvironment.disposeApplicationEnvironment()
if the process is going to exit anyway. I.e.,
UastEnvironment.disposeApplicationEnvironment() is only relevant for
daemon processes. And for daemon processes the disposal has to be
managed elsewhere anyway, since there could be multiple jobs using the
app env in parallel.

Fixes: 190615886
Test: androidx-studio-integration
Change-Id: Icae994dd2a9fbe972bdc1178f5b2fe9e72eba4a3

Former-commit-id: 108670b68f32485afeb372544ad22efea934641c