In the NewApi check, immediately skip source elements
The API detector mainly enforces Android SDK APIs like calls and field
references. These will always be found in android.jar, and the resolved
element will be of type PsiCompiledElement. This means we can quickly
bypass all the other checks by just checking for this first (well, right
after we do a couple of other checks that *do* operate on source
elements, such as @RequiresApi enforcement.).  This is useful when lint
runs in the platform source code.

Test: Existing
Bug: None

Change-Id: I6321a628d28606d27db99ba11a2126819f69f84e

Former-commit-id: 8ceb9a8aa873a5181bceaacbd57c9394012d3058