Suggest updating to the latest version of lint
When using the support for unbundled lint the version to be used is in
gradle.properties. This CL updates the Gradle dependency check to also
flag versions in this file.

It also adds some documentation for using unbundled lint, fixes a bug
related to the Incident(context, issue) construction method, and finally
makes the .properties lint check suppressible.

Test: Included unit tests Bug: None
Bug: None
Change-Id: I3c21a70e3a5a4e447aa4c2f74739559aa3ac0431

Former-commit-id: 8ecdc7d7b399ab677e19b2fcadfb0c5621c7b706