Replace RepoBuilder with RepoZipper and maven_repo_zip
RepoZipper is a replacement for RepoBuilder. It takes the output of a
maven_repo manifest and converts it into a zip representation of the
Maven repo.

This change also adds a maven_repo_zip rule that generates the manifest
for and calls RepoZipper given a maven_repo instance.

With this change, instead of using `use_zip = True` to generate a Maven
repo zip, maven_repo always produces a manifest file and the new
maven_repo_zip rule converts it to a zip.

Bug: 190732162
Test: RepoZipperTest and manual creation of maven_repo_zip
Change-Id: I168cecda5bf5cec9e26e6c5268c27ac23fb3c144

Former-commit-id: fb16d2ca32b4f38c29e77ce2510d51de8fb1cef1