190089699: Don't flag missing api/minApi for androidx
There are a couple of older issue registries in androidx which do not
specify the api and minApi properties of the issue registry.  Lint will
flag this, but for AndroidX we've manually verified that this works and
a lot of existing projects use these libraries so silence the check for
these older libraries.

(Note: Bug 190089699 also requests that the obsolete issue registry
warning should only be issued once, not once per project. That's not
addressed by this CL.)

Test: Existing
Bug: 190089699
Change-Id: I8dec4e73d7b45dc817bf29c984b12565316f2eab

Former-commit-id: ced34f84846563b954722b5801fff4a69814eddc