Add new lint check to track down file encoding problems
This lint check will look for cases where streams of bytes
are converted to or from strings without specifying an
explicit encoding, which means the default character set
will be used, which is often a bug.

This is not an issue on Android, where the default charset
is always UTF_8, but that's not the case everywhere else.

This check is off by default. This CL also updates the lint
testing infrastructure to (a) make it easier to test
non-Android-platform checks, like this one is, (b) to
flag invalid regular expression backreferences in quickfixes
and (c) fixes static import handling in the quickfix
verifier.

Bug: 191265156
Test: Unit tests included
Change-Id: Ic5cb80e1ed6ad703d220ad6e320ef8998961206c

Former-commit-id: f8bcbb7b9377d44e4fd6472e0524fe6a707f8749