Fix bug in overlay diff
To determine which files need to be removed from the overlay on an
overlay update, we subtract the set of files in the newly built APKs
from the set of files currently in the overlay. Any remaining files
should be removed.

Our subtraction was implemented using only file name; that is, two files
with the same parent APK and name were treated as equivalent, but this
is incorrect. The fix ensures that only files with the same parent APK,
name, *and* checksum are treated as the same for the purpose of set
subtraction.

It also ensures that we don't delete files that are also being added in
the same update; we currently delete before we write, so this isn't an
issue, but this prevents future bugs and saves some work.

Fixes: 182882359
Test: OverlayDifferTest
Change-Id: I39031b6dd7736d82cc251ef8419532ff5f1b76b6

Former-commit-id: 12eed72b0ef502a51e4800ccbecd239bc1ddb97e