Fix @CheckResult if-statement handling
This CL fixes a regression in the CheckResultDetector around if
expressions (and ternary expressions in Java which is also handled as an
if AST node in UAST), and covers these with tests.

It also fixes the problem that wrapping an expression in a parenthesis
would treat that as a "usage" of the result.

Fixes: 189970773
Fixes: 191788196
Test: Unit tests included
Change-Id: I4051908549d9150f5aadb6a9b1ec2b431f882508

Former-commit-id: 17fe9bf5869ca0d4509838a5ea72ae9fc818e04c