Refactor handling for namespaces.
- Move all computation logic into VariantDslInfo so that
  it's centralized.
- Introduce VariantDslInfo.namespaceForR for now since we need it,
  and its computation is a bit special too.
- Removed VariantDslInfo.testNamespace has it was used in a weird way for
  test variants to get their values from the parent. Instead feed each variant
  their DSL namespace directly.

Also moved VariantDslInfo(Builder) to use the new native DSL interfaces
since they have the namespace and we need to access them. This impacts
quite a few classes that had to be tweaked but it's somewhat minor and needs
to happen at some point anyway.
This removed a weird case where VariantDslInfoBuilder was initialized with
an instance of BaseExtension and then one of its method was passed an instance
of CommonExtension!

Test: existing tests
Bug: none, refactoring before actual bug fixes.
Change-Id: I7f4256ed5980e86d2868e1105fb2824889c7f7d9

Former-commit-id: a5e3ad621cbd99a563b6bf8cda469da951c4e1a4