Change the Android Test namespace for a corner case.
AndroidTest namespace for the R class uses applicationId
instead of the package from the manifest. This is wrong but
we cannot change it without breaking some projects.

The newly introduced (test)namespace DSL will force this
value to be used over the appId.

However if the project does not use the DSL and have flavors
that customize the appId (directly of via suffixes) then the
different AndroidTest variants are not sharing the same
package name for the R class which is problematic.

Now that the model contains the R class package name, we
really need this to be consistent. This change breaks the
existing behavior *if* there are flavors *and* their
appId aren't consistent. In that case we use the normal
namespace value (which is either the DSL value or the
manifest package).

Bug: 176931684
Test: updated existing tests.
Change-Id: If93004801dfdc83cfa3702beb7536f7884ea1c24

Former-commit-id: 0c3b59f53ce578cd18a232b79225651404f63968