Lint: Process libraries before the main project
When lint runs in partial analysis mode, it will first process the
libraries, and then the consuming modules. However, when lint was
running in global mode, it was processing the main project before the
libraries. This normally doesn't matter, but it makes some checks harder
to write, such as an upcoming fix to VectorDrawableCompatDetector which
needs to record whether vector icons in libraries are rasterized before
usages are encountered in dependent modules.

Test: Existing
Bug: None
Change-Id: I76d3b86e3dd5408aac9317b96dba69f9e9f890ed

Former-commit-id: 62febc5db26f4117fbbdd453022b6cdb08a8b57f