Add Android Build ID to --build_metadata in CI scripts
This makes the Android Build ID more easily visible in the resultstore
invocation details page. All build_metadata values are listed as
properties on the bottom of the page. Even though we pass the ID in
other places, it is hard to find the Build ID inside the full command
line string.

Bug: 192042331
Test: NA
Change-Id: I5871b8555eeb77b3d956a0f1509205ade4672095

Former-commit-id: 1fa98f99e1a6aeb3a8295dec03307db35db634f0