Move lint test utility from unit test to testing library
Lint's unit testing infrastructure had some code to make it easy to
obtain a UAST representation from one or more source files. However,
this code was sitting inside lint's own unit tests instead of in the
testing library, where it can be used by the testing library itself
(such as in the new parenthesis-testing mode, which was the push to get
this extracted) as well as by third party tests. This CL moves the code
up to the testing library.

Test: Existing
Bug: None
Change-Id: If840d8c06ad61ef18a2e8e7d1157fb8b01ceaaf8

Former-commit-id: b0ce1142cff3040ad08991e6b8a4c3827bf7f9e0