Add parenthesis test mode
This CL adds a new test mode, "parenthesize mode", where it looks at the
ASTs for all Kotlin and Java files, and figures out where it can(*)
insert unnecessary parentheses, and then it does that and runs the test
with these modified source files.

This helps shake out bugs in detectors where the detectors are
incorrectly doing things like

    if (node.uastParent instanceof UCallExpression)

since parentheses are present in the AST, and you therefore need to be
prepared to skip through these.

There were actually quite a few problems in this area (a bug report
mentioning that when adding optional parentheses a lint warning went
away was the trigger for not just this test mode but the whole source
code test mode transformation changess), so in addition to the new test
mode, this CL also systematically fixes all the problems with the
built-in checks related to parenthesis handling.

*: Obviously you can add unnecessary parentheses infinitely. This mode
   will only add parentheses in places where they seem somewhat likely.
   For example, it's unlikely somebody would ever write
   `(this).toString()`, so we allow detectors to reasonably just check
   `call.receiver is UThisExpression`, and so on.

Test: Existing (this re-runs all existing tests with modifications) plus
      new tests for the test mode transformations
Bug: None
Change-Id: I0cf593791a5648be5a92b25b5dd0cf2f1a93447a

Former-commit-id: 913121594a59a713ecf50c5d663687477fa6d869