Change default value for targetSdkVersion in the model
It was set to -1 if not set which is not good value. This
change renames the model property to be targetSdkVersionOverride
to make it clearer what it is: only the value coming from the DSL
and not the value coming from the manifest. If null, the IDE
could read the manifest and attempt to figure out the value.
This is temporary as AGP 8.0 will rejects values coming from
the manifest and require the value to be coming from the DSL
so that it can be known at configuration time without parsing
the manifest(s).

This change also improves the support for min/target sdk
value in the Variant API. The different stages of the variant
API (VariantDslInfo, VariantBuilder, Variant) all used different
values and data structure and there was a lot of objects created
for no reason.
This changes creates a MutableAndroidVersion object that's created
once in VariantDslInfo and used throughout the stages.

Fixes: 192434852
Test: existing tests.
Change-Id: I4d6e3c4d6e97dac10d8bc8cfa38d29aa0884ad2c

Former-commit-id: 2ace9f0647684c1009b9e20d5a9b57a729728388