Add fully qualified imports test mode
This CL adds a new test mode which replaces all(*) symbol references
with fully qualified names instead. This helps find bugs where detectors
are incorrectly checking for example the call identifier instead of the
resolved method.

This revealed bug in infrastructure too, like TypeEvaluator for Kotlin
fully qualified references in constructor calls, which is also fixed in
this CL.

Test: Existing (this re-runs all existing tests with modifications) plus
      new tests for the test mode transformations
Bug: None
Change-Id: I507269e0318cb343bdb0ffd05dbda8c8dc7df655

Former-commit-id: e783c44168082de1976704ee5090098d8582ed7f