Split minSdk and targetDeployApi on dex-related configurations
minSdk is the one passed to the .minApiLevel() of R8/D8 command builder
and should be the same as configured in the build.gradle (or the value
that ends up in the manifest).

The targetDeployApi is instead used to only control which entries
we need to add to our classpath, for example if the minSdk or the
targetDeployApi is 24(N) or higher we can make the process lighter.

Bug: 157701376
Test: existing + changes in DeploymentApiOverrideTest
Change-Id: I454a2423b8a57e0096158a12ad40e7a3acb57c6e

Former-commit-id: deafcb5e0a23556d07252b271ffc07b65caff4e0