Rework minSdkVersionWithTargetDeviceApi into targetDeployApi
Simplifies the computation of targetDeployApi and make it clear its
intended usage.

One noteworthy effect is that now the injected property from the IDE is
also available on non-debuggable builds (in regarding to dexing, it was
already the case for signing), so it works the same when deploying
either a debuggable or a non-debuggable (e.g. Release) artifact.

Bug: 157701376
Test: existing + changes on DeploymentApiOverrideTest
Change-Id: Ia123e319e294f22f956b91bac78728a68b33111a

Former-commit-id: e63620980b3678078fa211570a0fe715f64949d0