Source transforming test modes
This CL adds some lint test infrastructure to support additional test
modes which transform the source code in some compatible way and makes
sure that the tests continue to pass.

This CL only contains the basic support for this; subsequent CLs add
additional test modes, such as a test mode to insert unnecessary
parentheses, a test mode to replace imported symbols with fully
qualified names, etc. The idea here is to shake out bugs where detectors
are making assumptions about the AST that are not correct.

In order to be able to run through a lot of extra scenarios without
making the test suite much slower, the test infrastructure also has the
ability to combine multiple test modes into one, such that it can
simultaneously test reordering Kotlin method call arguments and
converting methods to expression bodies, and only if there is a failure
will it run each one individually to pinpoint the problem.  With that
facility in place, adding 5 new test modes (described in the
documentation) only added 13% to the test suite run time on my machine
(4:21 to 4:55).

Bug: None
Test: This is test infrastructure
Change-Id: Ic1010993ede610dd95e94b3b70352691890bd0a3

Former-commit-id: 2c59ea11180935fac69f7bb31fd44d7faa88bc59