Add typealias and import alias test modes
This CL adds two related test modes: typealias mode and import alias
mode.

The import alias mode will go through all the import statements in a
Kotlin file and replace each with an import alias, and then it will
update all the symbols in the compilation unit to the alias instead.

The typealias mode will go through all the types in the compilation unit
and on a per package basis, create type aliases for each, and then
replaces all type references to the alias instead.

In both cases, this will catch cases where detectors are improperly
looking at the identifier name at the source instead of the resolved
type name.

This CL also fixes a couple of detector bugs which were found by these
test modes.

Test: Existing (this re-runs all existing tests with modifications) plus
      new tests for the test mode transformations
Bug: None
Change-Id: I3c4a9c7a3c39cfa2ee24b3cdabbb5fb57f6fa9a5

Former-commit-id: c30c0432d5a29f0a161ed3ab7a9039fad98615cd