Create a bridge between the two maven rule sets
This allows the old rules to import the jars
created by the new rules, so we do not need rule
duplication. Once all artifacts have been
migrated we can delete the old rules.

As this makes the artifact built by the new rule
be the only one (including its pom), this adds
a version consistency check as well.

Test: existing + new one for the version
Bug: 190268212
Fixes: 193919937
Change-Id: I0b337e9818256e632d9b673b2e183f29af4facdc

Former-commit-id: d2c01b757336cfbe1ec42028e267110452c26609