Moves gmaven tests back to .zips
This is a fast test that validates the structure of the generated gmaven,
and is the only one that uses gmaven built in bazel.
This move is in preparation of switching gmaven to be built with gradle+bazel
and the output of that is a zip.

This reverts part of commit bfe62efcdd9e23a162e471d6592945b1e28f4a1a [formerly 27d48c25d151da49ef53ea2c031bee4bb8d5e460].

Test: this is it
Bug: 190268212
Change-Id: I32cecb7cccb461643165a0e2359a6890fc168e9c

Former-commit-id: f289d75ad0e9836fa3d7ce2654697bcd61e3beea