Stop checking uniqueness of projectName + group
AGP iterates all subproject of a build to ensure the uniqueness of
projectName + group, but this behavior is not compatible with project
isolation. Consider Gradle plan to raise a warning for duplicate
project names, we could remove this check. See
https://github.com/gradle/gradle/issues/17784

The downside of removing this check is that users will get a circular
dependency error when they have two module having the same
"group:name:version" instead of our warning message. But this is a known
issue on gradle side and an expected behavior, see more related issue at
https://github.com/gradle/gradle/issues/847,
https://github.com/gradle/gradle/issues/16986

Bug: 190518042
Test: existing
Change-Id: I0e257463d0987868ef9d566ebcb94bc1b92c513b

Former-commit-id: eba0d063de0cbb787232574e7c8f9563061f20e6