[AGP DSL] Stop breaking groovy blocks
This consolidates all of the fixes for the issues caused by our use of
kotlin delegation for CommonExtension etc.

The AGP DSL decorator automatically generates an Action taking method
that triggers Gradle's instantiator to generate a groovy closure so
users get 'this' style rather than 'it'.

However, for CommonExtension we use Kotlin delegation from
the legacy BaseExtension hierarchy of classes, so that
generated method isn't on the extension as accessible by the user.

Adding it to InternalCommonExtension, makes the Kotlin compiler generate
a delegate method, and causes Gradle to generate the appropriate method.

This change moves all the implementations to use delegation through the
Internal<PluginType>Extension classes, and adds a test to verify that
all methods that are needed are present there, and implemented in
<PluginType>ExtensionImpl. Those manual implementations will be replaced
by the DSL decorator eventually.

Bug: 194369303
Test: Added new GroovyBlockInExtensionTest
Change-Id: Id3fe79f5742338e47ecd3be2dcaec64b42884eb5

Former-commit-id: 55312fd2e4fe25242ce23bd4fcd4240a74769b7b