Use the same minSdkVersion for dexing and deciding
... if desugaring is needed. In AGP 7.1.0-alpha04 we
introduced a bug where:
- the api version used for dexing is minSdkVersion e.g 21
- because we deploy to a device 24+ (e.g 30), desugaring
is not requested

This means that D8 will strip interface methods and make
them abstract, but it will not perform full rewriting,
which leads to AbstractMethodError.

This CL fixes that by introducing a single minSdkVersionForDexing
value that is used in the dexing pipeline. When deploying to
device 24+, we snap this value to max(24, minSdkVersion), otherwise
the value is minSdkVersion.

Also, computation whihc dexing mode to use is moved to
ConsumableCreationConfigImpl and native multidex is used
if we are deploying to a device 21+ and multidex is enabled.

Bug: 194289155
Test: D8DesugaringTest, DeploymentApiOverrideTest
Change-Id: Ifcdc39995ff4f05c4bdfd4622127e15dd0f33d4d

Former-commit-id: 229177ceaf8cc2daddf62473f5ec1e5b932f66c7