added redirect file capability to listing file loading.
AGP will soon add a level of indirection to the listing files as they
are sync'ed in the model but we want to change their location based
on whether the build is invoked from the IDE or not.

On the loading side of things, the method that the IDE is using to
load these listing files can transparently support this level of
indirection. Therefore no model change nor IDE changes should be
required to support this level of indirection.

Test: Added unit test.
Bug: 69014026
Change-Id: Ib492414e9a8c3f9fa79cad51e0cbde1a0918d608

Former-commit-id: 263489f24f29a55bb48b0ca52fe45d7620edecfd