Allow module-to-module C/C++ references via Prefab
This change allows module-to-module references to be published
using the Prefab protocol. This allows libraries and headers
to be automatically discovered and consumed by receiving modules.
Additionally, publishing and consuming modules can mix and match
ndk-build and CMake including multiple CMake versions.

Design: go/prefab-m2m

Under the covers, this is accomplished by splitting publishing
into two phases: Configuration and Build:
- Configuration: Everything in the Prefab module is constructed
  except for building .a/.so. This is published as a Gradle
  artifact named PREFAB_PACKAGE_CONFIGURATION.
- Build: Same as Configuration but also build .a/.so. This is
  published as an artifact named PREFAB_CONFIGURATION.
On the consuming side, the CMake (or ndk-build) configuration
phase references the publisher configuration phase outputs so
that nothing is built during the configuration phase.

The requirements for this to work:
- Consuming module must be CMake (for now).
- Consuming module must have android.buildFeatures.prefab=true.
- Publishing module can be ndk-build or CMake.
- Publishing module must have
  android.buildFeatures.prefabPublishing=true.
- Publishing module must have an android.prefab.[LIB] clause
  with one or both of libraryName and headers set.

For ndk-build to work for consuming modules we need an update to
ndk-build and Prefab CLI. That is future work.

Bug: 191778470
Test: Added integration tests

Change-Id: I116a83d4ae6f4d5113813a51d5cdd6ca656d5a60

Former-commit-id: c1c7c6549c20b5c204978321a51c53e44706f180