Fix the need to put NOTICE next to BUILD file
The sdklib/sdk.bzl has an aspect that collects NOTICE files using
the .notice output of the maven_java_import rule. Since we weren't
injecting notice file location in our bridge, it was defaulting
to look for the NOTICE file next to the BUILD file. Since that file
does not exist, we were getting errors like:

Action tools/base/build-system/tools.manifest-merger_bzl.NOTICE failed:
   missing input file '//tools/base/build-system:NOTICE'

In this CL, we propagate the NOTICE file location from the bridge
to the maven_java_import rule.

Bug: 190268212
Test: bazel build //tools/base/sdklib:commandlinetools
Change-Id: I7af48490d3092bf0703d076f7d95b009df471a18

Former-commit-id: 070c79773ac3cb3fd1559bbcfd657210a6b299f9