Add Task to produce redirect file.
Add task to produce the redirect file for all IDE listing files.
These task will be used for all public ArtifactType that also
have a listing file like APK and Bundle. Private types like
APK_FROM_BUNDLE will use a different solution in a subsequent CL.

The redirect file is still not passed through the model to the IDE,
this will also be in a subsequent CL.

Test: Added unit and integration tests.
Bug: 69014026
Change-Id: Ib28c354c8fbe296658b86c7af8e5b2e0b20456c2

Former-commit-id: 8f5a24182d0490840a912f6e05899990867b893f