Remove unecessary needsMainDexListForBundle overrides
ComponentCreationConfig.kt already sets this as false by default, so we
can remove the overrides from *Impl.kt that were also setting it as
false.

We need to also remove the override from VariantCreationConfig.kt, so
its implementations are not *required* to override it (but they still
can if they want, like VariantImpl still does).

Bug: n/a
Test: existing
Change-Id: I8a606c9c2b7a5475443391137bd64b03487e9b8b

Former-commit-id: 4c85c501fb7ea13fef5e9ed2c3aec552c37d5d6f