[AGP DSL] Decorate variant dimensions
This change decorates
 * DefaultConfig
 * BuildType
 * ProductFlavor
 * SigningConfig
with the AGP dsl decorator.

This allows locking, managed properties etc, and includes locking the
NamedDomainObjectContainers. Flavor dimension is also a special case, as
that is modified by AGP after the DSL is locked, that is encoded so the
main property can still be locked.

Does not include migrating blocks within the variant dimensions to also
be decorated, that will follow.

Bug: 193452960
Test: Added new cases to KotlinDslTest
Test: Updated undecorated API expectation file
Change-Id: I9cfa3823cd0e5eb28c2e78da3501589e1b71a3df

Former-commit-id: 894fee68ff1d7043bec0299a797f6f889046849d