195004772: Fix package lookup for project-xml projects
This CL fixes two problems:

First, when analyzing project xml projects where no source roots are
provided, lint has to try to figure out the source roots on its own by
looking at the package statements inside Java or Kotlin files and
matching segments by the surrounding path. This code was just looking
for the first "package" reference in the file, which breaks if there is
a preceeding comment mentioning "package". This is bug 195004772. To
fix this, it adds a simple state machine which tokenizes the source
code such that it's skipping comments.

Second, lint already has a couple of other similar state machines; one
for the syntax highlighter and once for the test infrastructure. Both
of these were not correctly handling Kotlin nested block comments, so
this CL fixes that.

Test: Existing + new unit tests
Fixes: 195004772
Change-Id: Ie4942faa42f6bce4e85809762be8fae4e49d728a

Former-commit-id: 2637fb75f772a00dfd0bfeed7dccf234d175c44b