Annotate generatedSourceFolders as an input to the lint tasks
Previously, it was annotated as @Internal, which led to incorrect
behavior for some corner cases.

This change is a modified version of
Ia78e0c004bc2c40aa8bb915910a1a33f4283f966, which caused a regression
because marking the AndroidTest generated sources as inputs caused the
AndroidTests' Java compile and resource pipeline tasks to run, doubling
the total build time for some build benchmarks. This patch instead
removes the AndroidTest generated sources from the lint model entirely
because analyzing them is not worth the drastic performance regression.

Bug: 183998131
Test: LintGeneratedSourcesTest
Change-Id: Ie316151519959ecf929b17fdeb29049475365d86

Former-commit-id: 7c4b3f77c0242de459626c3960cfda634bb8c12f