Lint improvements for overloaded functions
This CL fixes two things related to overloaded functions.

First and foremost, it makes it easier to uniformly support calls. With
overloaded functions, many method calls will not show up as call
expressions in the AST -- instead they can be UBinaryExpression,
UPrefixExpression, UPostfixExpression, or even UArrayAccessExpression.
This meant that code which wants to perform some checking on calls
would have to not only handle UCallExpression, but be prepared to map
this support to all the other element types as well.

This CL makes this a lot easier; it can now create a new UElement which
wraps these other types of call expressions as an actual
UCallExpression. That means a lot of code (such as the API detector)
can now uniformly just handle UCallExpression, without having to just
pass around objects of type UElement and perform a number of accident
prone instanceof checks here and there.

Lint's built-in support for method callbacks is now also using this
mechanism, so if you for example register an interest in the method
name "compareTo", you'll be invoked for "x < y" if x defines an
overloaded comparison method, and so on.

Second, this CL improves the resolveOperator() function for array
access expressions by leveraging some of Kotlin's own type resolution
machinery (and adds a regression test covering one specific scenario
where the heuristic fallback was incorrect).

(This CL also gets rid of some obsolete UastContext usages and cleans up
some formatting from other CLs.)

Test: Unit tests included and updated
Bug: Not filed
Change-Id: I296200453c1acc7017b5116fb68558d556480960

Former-commit-id: 6a9de1a646b7d9aa6884579a6d08f72de77cea71