Share lint classloader even when buildSrc changes
As the underlying UAST machinery always ends up leaking the classloader,
this logic uses a MBean (via JvmWideVariable) to cache lint
classloaders, so that the class can be reused even when AGP is loaded in
a new classloader because of a buildSrc change.

The current caching was already problematic for -dev versions, but now
caching more keyed on the classpath URIs will likely cause unexpected
reusing when developing AGP and lint. To address this, the key is now
the explicit version, plus the hash of the files for -dev versions.

Isolate the lint invocation from class loaders leak by resetting
the thread context class loader before and after invocations.

Bug: 194911959
Test: Expanded LintClassLoaderTest
Change-Id: I804a2e3d9c6a27bc910c5c9cad6ea0d125f8c033

Former-commit-id: 7fbb0be57bfad40ce160de5ab09707132b15f63e