Create a third version of all repo xml schemas
This just copies the existing schema to a new version. It's not a useful
change without follow-ups to make modifications to the new schema.
The anticipated changes to be made for v3 are:
- change api-level to be a string
- add "extension-level" as attribute to api-level

In this change we add BUILD rules to generate the java code for a
schema, but do not generate the code.

Test: In upcoming cl, when the code is generated
Bug: 191364256
Change-Id: Ia1f38a91d577f1022845aae55532d6b557a771f0

Former-commit-id: e29705058e2f503ff6757fcaed7b0f9f67f2d3b0