Add if-to-when conversion test mode
This test mode converts if statements in Kotlin into when statements, to
ensure that detectors which are looking at UIfExpressions also work
correctly when the if statements are expressed using when statements
(which show up in the AST as USwitchExpressions).

It then fixes a handful of detectors whic where not handling this
properly.

This CL also updates the test infrastructure to run quickfixes through
all the test modes as well, and fixes a bug in the parenthesis test
mode (which should have been skipping case values)

Test: Existing (this re-runs all existing tests with modifications) plus
      new tests for the test mode transformations
Bug: None
Change-Id: Ie198c4d97bfcafec10eb20951906f511ae65fae9

Former-commit-id: 24236caf9d2df03075457084d4b26b0962cb80cb