Fix v2 model for better composite build support.
In order to properly have a unique identifier on the
composite builds, this change also adds a build ID
map to the model. The build ID map is (name, rootdir)
which allows us to reconcille the API in the model
action which only has rootDir with the information
coming from the dependency model which only has name.

This also include new model tests for
composite builds, using a new  infrastructure
to create such-projects

Bug: N/A
Test: new tests
Change-Id: If79b7626b02e08db3013f4b5228d1d540b361da0

Former-commit-id: e7c000517eb155ec09b53fcb0c14e5f621d0cf35