In incompatible-jar file error messages, including referencing class
If lint attempts to load a custom lint.jar file which references an
invalid API, lint will skip that lint.jar file with an error message
which lists the missing/incorrect API. This CL updates that error
message to also include the class in the jar file which referenced this
symbol.

Second, it adds back the getUastContext method. It has been deprecated
for a while, and is calling a method slated for removal from
intellij-core this year, but it happens to still be used by some
androidx libraries (the androidx fragment library was updated as soon as
it was removed from lint, but the current stable version in the field is
still using it so leave it alone a little longer. However, the method is
now marked hidden, such that it will be included in the binaries but no
longer deleted. We'll use this for future removals when practical.

Test: Existing unit tests (updated)
Bug: None
Change-Id: I1525a32a3e3df72abbf2eb15d56f580757375780

Former-commit-id: 89d48746e6498bd7cd39851cb24aae4c4dfdda3b