Use --google_default_credentials
Instead of using checked-in credentials, use Application Default
Credentials. The current credential allows writing to the remote action
cache, and we want to restrict this so that only CI builds can write to
this cache.

Obtaining default credentials on Windows relies on the %APPDATA%
enviornment variable, so we add it to the wrapper script.

Bug: 196437640
Test: NA
Change-Id: I9b61653db4107c3ea3bc64bc4750959312f07e0a

Former-commit-id: c8bce89e7be7b59c63d2ba210d5e914b1051e798