Data binding: Pass api-versions.xml from AGP
Before this change the data binding compiler was finding the
api-versions by hardcoding the path in the SDK in platform-tools.

The api-versions file has been moved to platforms, so it is versioned
now. (There was some overlap, so this is avaliable in platforms since
platform API level 26, and was removed in the platform-tools several
years later in 31.x)

Data binding will fall back to a fixed api versions file shipped as a
java resource in the unusual case of compiling against an older version
than 26.

This fixes the linked issue where data binding currently fails if
platform tools is updated, and removes any implicit dependency on the
SDK state.

Bug: 196847367
Test: Included unit tests
Change-Id: I5782f91a0e73dbcc4f7976ebfa788789bdd933a9

Former-commit-id: c3f0081c667ead9cee4ed4d59c5bc9f966431b8a