[AGP UTP] Remove boot completed check in emulator
So, when loading from the snapshot the emulator does not output "boot
completed", so this check is redundant at best, and at worst slows down
the time to run tests on the expected test case. The adb connect check
should be sufficient for most cases.

As a backup, if cold boot occurs, there is now an adb boot check after
the adb connection is established, where we call "adb -s <serial> shell
getprop sys.boot_completed" to check for boot, with retries.

Bug: 197146281
Test: ran local tests, manually verified. Adjusted tests to support boot
check

Change-Id: If6f3a04d00e242240ca79d89c350abe7a59f34b0

Former-commit-id: cbfb55d5a04acbae15c6d03562d2200ceef3e2b8