Add support for compileSdkExtension
This reworks a bit how the different compile* properties
and methods are working together to make things more manageable.
This goes from using a single storage (compileSdkVersion) to using
storage for each scenario (api, preview, addon). This simplifies
the code quite a bit but makes things less efficient (more backing
properties).
This also allows us to be a bit more strict on the new DSL items
which is a good thing overal.

Bug: N/A
Test: Updated the DslTest with the new scenarios
Change-Id: I3f1d894c422d4e337c0e571123cea41a6d689d27

Former-commit-id: e2de8d797ad800af2dd5fe8fad05640f7e0bc529