Handle URL errors on Windows for local repos
On Windows, when we resolve version range, we use a local
repository like this:
  "file://C:\users\emrekultursay\_bazel_emrekultursay\el2qqrlg\external\maven\repo"

Which causes the following exception to be thrown:
   IllegalArgumentException: URI has an auhority component

Sample error showing how this causes version range resolution to fail:
https://source.cloud.google.com/results/invocations/53fd2f08-5075-4a44-aca2-e3dd53064f7b/log

This CL manually removes the file:// prefix. It also skips repos that
don't start with file:// as they are not interesting to us.

Bug: 190268212
Test: Add "io.perfmark:perfmark-api:0.17.0" to @maven and build anything
Change-Id: I1490c01c0f751fae7ad1e2ff9856a707a2e618bb

Former-commit-id: 6c22130eb6da9ad7ca44f294c06bbe43ea266000