Add src jar support
... and fix exports.

Now kotlin_library generates source jars like java_library does,
and they are taking into account when used in a maven_artifact.

Also fixes an issue with exported MavenInfos. There was no exports
attribute in MavenInfo before, so we merged all the providers.
However this had a problem that merged more than it should,
including source jars. This is fixed now.

This also adds the NOTICE file to the source jar, as it's
required and covered by tests already.

Test: existing
Bug: 190268212
Change-Id: I80ea22e48790fc25bde42f03c371f72bc80ef614

Former-commit-id: e731bc3acfad73f511531d3eea700f223d4400b7