Halve the job count of coverage post-processing
We suspect the flaky post-processing failure described in b/197834990
may be due to remote-cache actions not being properly counted by the
default jobs limit and resulting in the buildbot running Bazel to go
OOM during post-processing as it executes HOST_CPUS*2 actions. Halving
the limit (default is HOST_CPUS so we use HOST_CPUS*0.5) should bring
it back down to just HOST_CPUS actions in practice. If this hypothesis
is true, then this should resolve the flaky post-processing OOMs.

Bug: 197834990
Test: N/A
Change-Id: If9ace5c35c618af19534efd4ca8e24135e0917fb

Former-commit-id: 08815bbb27a39d0fcd5edd8b4c7c27bf094d1922