Try setting AgentInstallationType before accessing AgentPremain
AgentInstallationType will be available in future versions of
the coroutine lib (probably 1.6.0), see this PR
https://github.com/Kotlin/kotlinx.coroutines/pull/2912

This change updates to agent to try to set
AgentInstallatioType first. If the class doesn't
exist (because the dev is using and older version
of the coroutine lib), we fallback to the old way
of setting the field through AgentPremain.

Test: NA
Bug: 182023904
Change-Id: Ia21430eb19f015b7845998d9d0c689ea82f3380c

Former-commit-id: e6b50849f14b1128d68bab1c828a297a4ce252ca