Remove use of Gradle internals
This removes the use of Gradle internal interfaces for C/C++ build cache.
I'm searching for a replacement design that doesn't require internals.
This removes the least code needed to avoid internals. I'm leaving the
remaining build cache code intact until it is clear what can be reused
in the new design.

Bug: 182809209
Test: Removed dead tests
Change-Id: I3e85e6b897cba9e036e20f1248dcd26ea654f5bb

Former-commit-id: d4817d9c56df99733e6a4eda0c3fa39f76cad67c