Add foldable information
This change adds the ability for the view inspector to report device
fold angle, fold orientation, and posture.
Angle is reported by a Sensor, but is apparently inconsistently
implemented across devices.
Fold orientation and device posture are provided by the
androidx.window:window library. For several reasons it's difficult to
include that library directly in the inspector, so we require it to be
included as a dependency in the app in order for that code to be used by
the inspector. If the device is foldable and the app is interested in
responding to fold state it will need to include that library anyway,
and in the future we will show a notification in the inspector when
inspecting a foldable device without the library present.

Note that since the app inspection framework jarjars inspectors and
rewrites kotlin.* and kotlinx.coroutines.* packages, and the
androidx.window:window library APIs use coroutines, the inspector in
this case must use code that's packaged in a separate kotlin_library and
then invoked via reflection.

Test: Added some tests
Bug: 180228959
Change-Id: Icac353cde612c67a20e0c24b10973fb81cdfc083

Former-commit-id: 89a57eb646b8753ed38e0a31c9f4c042725cc579