Overhaul lint's annotation analysis API
Lint's annotation support grew organically from the needs of additional
checks added, and it has some known limitations. One of these is that
the callback method takes a lot of parameters, including 4 lists of
annotations: the list of annotations to analyze, and for reference, the
other member annotations, the class annotations, and the package
annotations. Providing these annotations is important such that lint
can handle "scopes" (e.g. if you call a method which is not annotated
but is in an annotated class or package, you want that outer annotation
to apply).

Besides being messy, this hardcoded list approach does not work,
because we need to handle an arbitrary number of outer classes as well,
along with potentially additional annotation targets.

This CL makes a number of changes:

* Creates a new callback with fewer parameters (along with some new
  data structures which will let us provide additional context as
  necessary in the future without breaking compatibility). The new
  abstraction makes it easy to see where an annotation originates from.

* Adds a fallback mechanism such that older lint checks using the old
  callback continues to work. (Initially, none of the built-in lint
  checks have been updated to the new API, so all the tests pass via
  the compatibility bridge. In an upcoming CL the built-in checks will
  switch to the new one, which will then allow fixing some bugs.)

* Adds support for Kotlin's file-level annotations (@file:MyAnnotation)

* Automatically handles filtering out repeated outer annotations, and
  adds support methods to handle related annotations. This helps clean
  up various other detectors as seen in a followup CL which switches
  the existing detectors over to the new API. They are not migrated in
  the current API to ensure that the compatibility bridge works -- all
  the existing tests pass using the old API as well.

* Updates the lint check which looks for lint implementation mistakes
  to ensure that you don't call super.visitAnnotationUsage from the new
  API since that will unnecessarily run wasted compatibility code

* And perhaps most importantly, adds documentation to the API Guide
  for how to write annotation-based detectors!

Subsequent CLs will use this new API to fix various bugs.

Test: Existing tests
Bug: None
Change-Id: Id5482b02e5e743f44b5aa91ac0cb79e379a7cc28

Former-commit-id: 8190da2fc7ac0b3785ed2353baee243bdf7a4012