Handle null comparisons in the Single Access Method detector
The SAM check tries to figure out if a method which takes
a single access method does instance comparisons or stores
the instances into a field. However, this was also including
comparisons against null which is generally fine.

(Encountered when turning on lint checks on lint's own
unit test codebase.)

Bug: Not filed
Test: Unit test included
Change-Id: I3751361f526a7bde85e90fb6b0e41c8286544c85

Former-commit-id: 03d96bb8b2aea13d23b6865f843a318de4a1532f