Upgrade Kotlinc used for building to 1.5
Minor note: the Kotlinc version used for building in Bazel is no
longer tied to the Kotlinc version used in the IJ platform build.
This reduces coupling (giving us more flexibility to adopt new
compiler versions without meddling with the platform build) and allows
us to remove some patches in the platform. The only constraint is that
our Bazel Kotlinc version should generally be newer than (or within
one major version from) the Kotlinc version used in the platform.
This constraint is enforced automatically because Kotlinc will refuse
to compile if it finds bytecode on the classpath from an incompatible
compiler.

Bug: 200560565
Test: bazel build
Change-Id: I9f1ef57094543b6b5c88b76e62f37e8bf94a76d3

Former-commit-id: ad2892cb791b12a1291339652e7604905b404b35