Android Lint: Add lint-tests to the Studio project
To allow editing lint all in one idea project, and unblock the removal
of lint from the AGP gradle project.

Includes fixes:
  1. Adjust KotlinClasspath test fixture to handle manifest jar
     indirection
  2. Adjusted some Truth calls as the version of truth in idea is
     slightly older

Remaining failures when run from the IDE:
  1. MemoryLeakTest
      - Tried adding -Djdk.attach.allowAttachSelf=true to the runconfig
        as the test suggests but it still fails
  2. LintDetectorDetectorTest
      - Something related to TestMode.PARENTHESIZED

The IML Bazel test targets are left as manual, as they are only used for
development and duplicating all the lint tests would consume a very
large amount of test executor time.

Note that the studio-sdk dependency of these modules contains the PSI
and UAST classes that are also present in studio-sdk with a different
version. This conflict is difficult to avoid as almost everything in
tools/adt/idea depends on studio-sdk (even common) and so it will always
be brought in as a transitive dependency anyway, but the jars in
prebuilts/tools/common/lint-psi also bring some extra methods that lint
uses. Therefore all of the added modules depend on the lint-psi jars
before studio-sdk to ensure the right jars are loaded when running tests
from the IDE. Because of these different dependencies, the existing
lint-tests:tests target is kept, and the bazel iml_module test targets
this change introduces are marked as manual.

It makes things a lot less noisy if kotlin-script-runtime is excluded
from the test runtime classpath in the test run/debug configuration as
that currently has a different version than the rest of kotlin

Everything else seems to work.

Bug: 143864616
Test: Manual, iml_to_build verifies the tests compile.
Change-Id: I2f6e726227c6bf7e2662879c72135d9a1dc38f30

Former-commit-id: 30772be8df656ea23ed32bdd08713ddd94a00dc3