Unroll deprecated lifecycle-extensions dependency
This CL updates data binding runtime dependencies to not
depend on deprecated lifecycle-extensions.

It was rather a wrapper to get consistent dependencies for
other lifecycle artifacts but since that package is deprecated,
it is useless.

I thought about updating dependencies to more recent versions
but that might affect apps. Instead, I just unrolled to remove
the deprecated dependency.

Fixes: 197553072
Test: existing tests
Change-Id: Id37023558327f6bf4b165dd50e51d13cf20d47f8

Former-commit-id: b15357eca9c54f11668e2b6e9b7654c586f317e9