Stop using SkiaQWorkaround when possible
Previously, we always used the ViewDebug replacement SkiaQWorkaround to
listen for changes on device. This used private APIs that we couldn't
access on non-debug builds, and was only meant to be a hack around a bug
in the initial version of Q.
After that, the ability to turn off the expensive serialization of SKPs
was added in SkiaQWorkaround, and the ViewDebug method needed because
inaccessible due to a change in what APIs are accessible.
In this change we use the normal ViewDebug call in T+, where the method
has been made available again, and we
ViewObserver.registerFrameCommitCallback to listen for screen updates
when we don't need SKPs.
This allows us to show live updates on non-debuggable processes, though
we will only be able to get SKPs for those processes on T+.

Test: updated existing tests to cover the new and old cases.
Bug: none?
Change-Id: Id4baa924bb98436ee4f5a130e4a21038f65eda3f

Former-commit-id: e2fc91eb66e240f76e922f171107f821b94fada4