Build AGP docs with Dackka
For now the directory layout isn't right, but the layout used for AGP
requires changes to Dackka. It makes sense to land this now to avoid
more source-level regressions with docs generation, as the old Dokka
fork is not compatible with newer Kotlin Gradle plugins.

The Gradle docs build is wrapped in a bazel target
//tools/base/build-system:android_gradle_plugin_dackka_docs
so the docs generation runs in presubmit on linux.

Bug: 202837927
Test: Included
Change-Id: I9b32514ced0b088ba44248372da945e58302f9ee

Former-commit-id: 545aa44d7b8419e81f0e0a556a77de0c8b222978