change path sensibility to absolute.
RELATIVE path sensibility for files is the same as NAME_ONLY which
is problematic for the ListingFileRedirectTask since the input
file name never changes and the content might be the same between
optimized and release versions.

Since the task is not cacheable, change the RELATIVE path sensibility
to ABSOLUTE to ensure task is rerun when the redirected file path
is changing from optimized to release.

A subsequent CL should go over all of our RELATIVE + Regular File
usages and change the path sensibility to a different value as using
RELATIVE is very confusing.

Test: Existing
Bug: 202894113
Change-Id: I30004d6cab94982bc956d182daecf5791d33b540

Former-commit-id: f77a35427325ad325af8e35a0c933acd3eb40995