196406778: Fix false positive for class registration check
When running lint on a release variant, there are scenarios (and one
such scenario is replicated in LintMissingClassTest) where the compile
class path passed to lint seems to be missing key classes. This leads
inheritance lookup to fail, so it incorrectly reports an activity as not
extending android.app.Activity.

This CL adds a regression test for this, and adds a workaround to lint.

Fixes: 196406778
Test: LintMissingClassTest
Change-Id: I4b59fdc111bf2a9678de0b6590c4a85ad3d56605

Former-commit-id: ee79c752729727da7cd070bfe32be395db5e349c