Add custom C\C++ build type enum
This adds the enum value 'CUSTOM' to the set of build types alongside CMAKE and NDK_BUILD.
Otherwise, it doesn't change any behavior. This is in support of opening AGDE-driven build.gradle in Android Studio.

For overall design, see: go/agde-as
For design of Custom C++ in AGP, see: go/agp-custom-cxx

This CL also removes ifCMake and ifNdkBuild functions and converts their use to use 'when(buildSystem)' instead.
This is to make it easier to catch places where logic for 'CUSTOM' is missing.

Bug: 192006965
Test: No behavior change

Change-Id: Iea9519dc008f57da86fd758073568b0ac3181e22

Former-commit-id: a1abff2ab371e4ae81dac4e922804a261054764f