202193843: Improve range detector error messages
This CL makes a number of improvements to lint's range check.

First and most importantly, it improves the error messages to make it
clearer what the problem is. Until now, if lint knows that a particular
value is for a known constant, it will display that constant in the
error message, such as -1 in "X must be >= 5 (was -1)". However, when
lint is comparing ranges, such as parameter must be >= 0 but you're
passing a value known to be >= -1, here lint would only display the
required range. It now includes the actual range as well, and in fact,
tries to point out the discrepancy between the two, so in this example
it will highlight that the value can be -1 (which is outside the
allowed range of >= 0).

Second, it improves the range handling such that if there is a
surrounding if statement which makes specific range comparisons, such
as x != -1 or x > 3 && x < 10, then those are merged with the other
known constraints before making the overlap computation.

Fixes: 202193843
Test: Unit tests included
Change-Id: Ide85165828bda6c6c674f4cf85f8bd9e8d11fe74

Former-commit-id: d6211b9a699344cd25a23fcdfa311dcda727e958