Revert "Use testFixtures variant name for testFixtures resolved artifacts"
This reverts commit 15216ed1b8e38efde7b0286969aae7d760fb9e76 [formerly 5a3a1f6186eadd96d92f4582bed4b447b0b8a33e].

Reason for revert: variantAttr is no longer used in v2 model, however
in v1, we still need to select the correct variant in case of
testFixtures module dependency. And since testFixtures is not
supported in v1 models, it's better to set the variantAttr to the
main variant instead of the testFixtures variant so the IDE could
pick up the main variant directly.

Change-Id: I0e2613b76ec2f6076201c8256d3c109923476322

Former-commit-id: e31caf3508f3f4b0af508131059c291f90683b88