Properly handle relocated artifacts in model V2
Gradle's module format, as well as Maven's POM format
have features to relocate artifact. When this happens,
the dependency graph contains a node with no artifacts
but with a dependency on the new artifact.

The code in FullDependencyGraphBuilder did handle cases
where the artifact would be null, but only treated a single
case (where the dependency is on the project itself in case
of a test component). This did not handle the case explained
above.

This CL now checks for externaVariant on the variant to detect
this case, and if so will create an empty node with no artifact.
The Library model interface already had artifact has nullable, so
no change is needed there. However, it felt safer to add a new
LibraryType to indicate this particular scenario.

Bug: 204165878
Test: added new integ test and unit test.
Change-Id: I53e29a3d4407b78f8d156e5b5424ccbcfe39454a

Former-commit-id: 71520f58f95f58f32056857bbeb89315c72efa7d