[AGP UTP] "host" abi changes
This change moots the abi part of the ManagedVirtualDevice DSL. This is
replaced by the requires64Bit and changes how system images are changed
entirely.

Instead of allowing the abi to be specified, we determine the abi by the
host system. That is:

arm machine -> arm image.
x86_64 -> x86 or x86_64 depending on the dsl and availability of the
image.

By default, for x86_64 systems, we will choose the 32 bit image over the 64
bit image. This can be changed by setting the requires64Bit boolean in
the virtual managed device dsl. However, for newer api levels (30+ for
aosp, and 31+ for google) only a x86_64 image exists, and will be chosen
regardless of the value of require64Bit.

Bug: 200825554
Test: Added CpuArch Test for determining the cpu's architecture
Change-Id: Ida6ee303a372f460f65d96bebcc9b033e511844f

Former-commit-id: e4e038f1d4b9ea0401d1545b1930ff93f91fc43f