Don't break lint when category is used
Since
https://github.com/gradle/gradle/pull/11997/commits/f112bd4ca2f8c03f2751b36904407f110e47e5dc
gradle is applying the org.gradle.category attribute to compile and
runtime classpath configurations in the java plugins
(org.gradle.category=library). Before this they were only used for
publication and not consumption.

This change didn't immediately cause issues as the standalone lint
plugin does not currently support checkDependencies=true, so we didn't
notice immediately.

However, when kotlin followed and started applying
org.gradle.category=library to AGPs consumption configurations,
lint could no longer find the published lint models and partial
results, leading to warnings like the following, dispite the lint plugin
having been applied:

Warning: Lint will treat :annotation:annotation-sampled as an external
dependency and not analyze it.
  * Recommended Action: Apply the 'com.android.lint' plugin to java
    library project :annotation:annotation-sampled. to enable lint to
    analyze those sources.

This change removes AGP from using category for the lint publications,
it was not needed at the time and is now actively harmful.

The project for LintStandaloneCustomRuleTest now manually adds the
attribute, which causes issues not to be reported, failing the test
without the fix.

Bug: 204552946
Test: Expanded LintStandaloneCustomRuleTest
Change-Id: I9ce3e4a187b4560fc0ff7e13f4bc3bb51a424c15

Former-commit-id: 9cd5b0b1d1e500b176568d7beb3bed6a23c4e8ba