Add "-test-fixtures" suffix to the aar name published for lint
In ResolvedArtifactCollectingVisitor, the artifacts are identified by
ResolvableArtifact.getId(), so in case of
ComponentFileArtifactIdentifier the artifact is identified by the
componentId and fileName which could be the same for artifacts in the
same module (the same ProjectComponentIdentifier) and producing
the same fileName (but the path is different).

The problem exists only for artifacts with different capabilities
as a node in the dependency graph has to have all the artifacts
attributes matching, but they could have different capabilites, and so
it's important to use the capability when identifying the artifact.

This change adds "-test-fixtures" suffix to the testFixtures aar name
published for lint and so avoiding this problem until it's fixed on
Gradle's side.

Bug: 139438142
Bug: 201391004
Test: Incoming
Change-Id: I5d522672b22fc034f681b43da6450b9ec3318c76

Former-commit-id: dab031fe5c5d842af8bcf763b2e0248677750189