Request diagnostic information for b/194020297
This CL detects that b/194020297 has occurred and requests the information we would need to further diagnose the issue.
It does this by adding a new logger message type, BUG, prints out the information we need and the bug number to attach it to.

Here's an example of the message:

[bug 194020297] /project/CMakeLists.txt : Handshake with CMake Server failed, please consider upgrading to CMake 3.18.1 or higher.
  The raw CMake Server log was:
  [elided]
  Please refer to bug https://b.corp.google.com/issues/194020297 for more information.
  If possible, please also attach a zipped copy of /project/.cxx to the bug to assist in diagnosing the issue.
  The current Android Gradle Plugin Version is 7.2.0-dev.

The BUG message type is general purpose and could be used in similar cases in the future. Note that the message reported to the user includes AGP version number because it may be different from the original bug number's reported AGP version.

Bug: 194020297
Test: I don't know how to repro b/194020297 but I did manually test that the diagnostic message above will be generated.
Change-Id: I0d67b282126464b1b18210313f7b397b370496bb

Former-commit-id: b29129898f96a4038d77360119cf16a51ada5ae1