Add a boolean value in ComponentInfo to mark a testFixtures component
Add a boolean value in ComponentInfo to indicate if a library or a
module is a testFixtures component. This is needed to eliminate the need
for the IDE to check the capabilities for a testFixtures capability and
move the checking logic to AGP where we know more information about the
component.

Bug: 139438142
Bug: 202838863
Test: model snapshot tests
Change-Id: I3d09e457a27951121c32ad7a85bad41b7a75c906

Former-commit-id: e351c360a75223d97eec4480f0f18ff334251f04