Revert "Add a sourceset field to LintModelModuleLibrary"
This reverts commit 71fd91a8dc8147ce38fe53161270e797f4a53dac [formerly e17d5172783b294ca2be7ca0962c226e273003f3].

Reason for revert: Moving from using a sourceset field to rely on the
artifact address to distinguish between different sourceset artifacts.
There is no need currently for lint to know whether a library or a
module is a testFixtures component of a library or a module. By
moving to using the artifact address lint will deal with them as
separate artifacts.

Change-Id: Id2b5cdee0d13c36d15d53b85b9674329c6c6403b

Former-commit-id: 5d1e9772a703656ab7ca178f1c1fddcf2149cd1f