205017694: Fix range detector false positives
This CL improves the RangeDetector to handle a few more scenarios for
range handling. In particular, it better handles else-clauses, it fixes
an off-by-1 error for > and < comparisons, and if an actual value range
is inferred (from surrounding if statements), it will only be confident
if the range is has both upper and lower constraints.

Test: Unit tests included
Fixes: 205017694
Fixes: 205016549
Change-Id: I700e241b1f65909ff042e77d1492b4896a36559d

Former-commit-id: 7382dcce03ea7e12f601c5211ffd27a459c47f52