202559682: UnknownNullness check false positives on kotlin properties
If the platform type is nested, it will now include the type in the
error message; this makes it easier to understand what's going on.

For example, this code:

   val ANSWER_THROWS = Answer { ... }

will now generate this warning:

   Warning: Should explicitly declare type here since implicit type
   does not specify nullness (Answer<(Int or Int?)>) [UnknownNullness]

which makes it clear that the problem here is the type parameter to the
Answer interface.

Fixes: 202559682
Test: Unit test included
Change-Id: Ic1b063eadc53c488ee1693ceee3949b79b1da4b6

Former-commit-id: 0f0ffb376177b2d7522783edcfd0bf471187febc