Ensure correct handling of empty dirs for Java resources
This is to ensure empty dirs are packaged as Java
resources. Also, this fixes an issue where SkipWhenEmpty
ignores empty directories, while up-to-date
checks do not.

Until we update to Gradle 7.4, ProcessJavaResTask will
be skipped if its inputs are only empty directories, which
is not intended. However, this CL adds a test to encode that
behavior, and it should fail once we update to 7.4.

Bug: 204425803, 205285261
Test: MergeJavaResourceTaskTest, BundleLibraryJavaResTest
Change-Id: Ia37a501a60844d2d30525e1c1c400523283c7932

Former-commit-id: 9f56c6304fe2a34a0f6bd6c850087709612cb960