Include library classes in the JacocoTaskExtention
The previous method of Jacoco offline instrumenting library classes i.e. by a
task replacing source classes with instrumented classes. This meant that
library classes needed to be excluded from the input classes to the
JacocoPlugin, as the code was already instrumented. This case will be
supported until legacy transforms are removed as they are not supported
in the current artifact transform based instrumentation.

This change avoids excluding library classes when the JacocoTransform
can be executed on the project.

Bug: 195860510
Test: JacocoTest, JacocoWithUnitTestReportTest
Change-Id: I61afcb37b8304b15679ef9fee01ac51ba8d58df2

Former-commit-id: d0b844e3485045fec7e07bf045b2695af3f03b48