Start cleaning ProjectInfo.
Start exposing more project information on ProjectInfo
in order to be able to fully remove getProject() from it.

This also removes a bunch of usages of ProjectInfo.getProject()
but not all of them.

Also start transition from APIs that return File in favor
of more lazy APIs that return Provider<Directory|RegularFile>

Bug: 202449978
Test: updated existing tests
Change-Id: If12138824111afcb40939098b2bce5ebe5a26c08

Former-commit-id: 6cd3c7b43717520c1d208f15e20107879b8a7f64