Don't depend on relative path separator for identifying relative res paths
Avoids using the relative resource separator ':/' for identifying
strings as relative resource paths. Instead the boolean options will be
used to determine if a path is a relative resource path. The exception
is in AAPT2ErrorParser where usage of relative paths are used there is
potential for a string to be an arbitrary string or relative path.

Bug: 141301405, 155218379
Test: MessageRewriteTest, integration tests
Change-Id: I35918ed3294b9087a7cbbd9b3d7b41fba997f98b

Former-commit-id: 8e8b3a15e82c41a1e59ffbd5256548628bec3e0d