Fix CDATA handling in XML files
First, this CL fixes
145767092: ImpliedQuantity should search CDATA_SECTION_NODEs

It also adds a new special test mode to catch these types of
bugs, and fixes a handful of other places where detectors were
not correctly handling CDATA sections.

This is the first source-modifying test mode that is not
modifying Java/Kotlin files, so it required some tweaks to the
test mode machinery.

Fixes: 145767092
Test: Unit test included
Change-Id: I33c0ffd4067f25af02a1e31697d4aa3703b1506d

New test mode and fixes

Change-Id: Ifbdf90f1140f6d1e60d0b85d5faa5b869cce3cb2

Former-commit-id: e130f91bc4edda14e6742f322cc12a4908d2ef60