Fix issue doc generator and add by-year index
First, this CL fixes the doc generator which had stopped working when
the jar file verifier was added.

Second, it adds a new index to the generated docs which segments the
checks by year added (inferred from the maximum copyright year found in
each file) -- and updates a couple of lint checks which were missing the
copyright notice.

Test: LintIssueDocGeneratorTest (updated)
Bug: None
Change-Id: I99ae93c3a67935e363eb86dc7de3ecebd9eb4eaa

Former-commit-id: 0fe2ce22d9c9e7c09ae7fef4cc435670e4cd0bbb