Include version in lint cache directory
The lint cache is neither an input nor an output to the lint tasks
so it needs some manual handling to avoid lint trying to load cache
items written by a different version of lint.

The cache path now includes a version marker file, sharing the same key
function as was used for lint classloaders. This includes special casing
of hashing the lint classpath for -dev and SNAPSHOT versions to ensure
that different versions of lint do not attempt to share cache
directories, even for -dev version users.

If there is a mismatch the cache is deleted before invoking lint.

Fixes: 204437054
Fixes: 204437195
Test: Manually verified. Existing tests, Added LintToolTest
Change-Id: I9421e3cd72f4a62a5048ad6d18ddd0121396d43c

Former-commit-id: 759c750bc26dfdf131cdb97f1b5b68b6f7bd6e76