Enforce order of resolving configurations in Jetifier task
Because of b/204421586, we need to resolve AGP's configurations
first. This is required as AGP tries to sync versions between
configurations in ConstraintHandler. If an external configuration
extends both main and test configurations, those two may be
resolved w/o ConstraintHandler running. Following that, when
we resolve test configuration on its own, ConstraintHandler will
run and it may try to add dependencies to it, which is not legal
and Gradle will fail the build.

Bug: 204421586
Test: CheckJetifierTaskTest
Change-Id: I16e4da4e2d91881a88b42dcdfd6a22ca5ce6d8d0

Former-commit-id: d2c4cdb8fc148929c43932dab3f4c550f4faab33