API to create ListProperty for internal use.
there was an API to create ListProperty instances which would
not be protected against late writing and unsafe reads but it
was decided we should replace with simpler APIs that make it clear
those APIs should be only used when instantiating internal or
private Providers/Properties to the AGP. Such instances can
never leak out in any public interface.

A future enhancement could be to have lint check to enforce
proper usages of the VariantPropertiesAPIServices APIs.

Test: Existing
Bug: N/A
Change-Id: I87cbd86fd34a3b02deff17ba53ffb23d121ae24a

Former-commit-id: ef9271c5dc9eaab1b0e63323c2ab84b95cd349ae