RIP GlobalScope.
This commit removes GlobalScope in favor of
more focused interfaces available at different
stages of AGP's lifecycle.

GlobalTaskCreationConfig: this is available
to tasks via TaskCreationConfig.globalConfig
and is meant to provide access to things that are
not variant specific. This does not give direct access
to the extension but rather provides access to parts
of the extension that are safe. Things that are not
meant to be consumed directly from the extension (either
because they are combined via some process, or are
configurable per-variant) are explicitely not exposed
(these are exposed via TaskCreationConfig which is variant
specific).

GlobalVariantBuilderConfig: this is similar to the global
creation config interface but is meant for internal usage
of the VariantBuilder classes. This exposes only what these
should use. Again these must be non-variant specific and
also should not impact the task graph.

BootclasspathConfig: This object provides access to the
bootclasspath information. It should be part of
GlobalTaskCreationConfig but because the old DSL has
getters that need it, this is, for now, a separate object.
Once the old DSL goes away in 8.0, it should be possible
to merge it.

TaskManagerConfig: Provides access to some information
for the task managers.

In addition to the above, this CL also makes progress
toward using the new DSL interfaces rather than the old
one. This is not the case everywhere yet, but DSL elements
exposed via the interfaces above are the newer type which
triggered some changes here and there. In general, this
should make removing the old DSL easier.

Bug: N/A
Test: existing tests
Change-Id: Id6ebcd304ada6f64e1a51ab1c6c32eb90515ad65

Former-commit-id: 5778dae21439b6353fbf10a73d1393afcafafc93