206982645: UMethod#isConstructor returns false on actual constructor
This bug is really a UAST bug; this CL simply boils the original
scenario down to its essence and tracks it as a UAST bug, now
filed as https://youtrack.jetbrains.com/issue/KTIJ-20200.

The workaround for now is to use sourcePsi is KtConstructor instead.
(It turns out that this method isn't used much in the built-in
checks and none of them need the workaround.)

Bug: 206982645
Test: This is it
Change-Id: I6e220f963bb919c86b23a4a2e71041aa3247003a

Former-commit-id: 0c5b4d90fcfcf0c8981961d6e90f61562448a949