Allow constraints to be used during the merge phase
Context.report(Incident, Constraint) was intended to be used
during the analysis phase such that lint can later auto-filter
incidents without consulting the detector again. Calling this
method during the merge phase was a no-op; incidents would
get stored in a list that is no longer consulted.

However, there is no good reason we can't also support checking
constraints during the merge phase, so this CL adds that.

Note however that calling the other deferred computation method,
Context.report(Incident, LintMap) will *not* work, so this CL
also adds a check for this and throws an error with a suitable
error message.

Test: Used by WearableConfigurationActionDetectorTest
Bug: None
Change-Id: I0fff34a3ce582051dabeead4182e1b867e09e0a8

Former-commit-id: dfbbd2f70233a9072e862ff59e9162e38aa8cb75