added a facility to declare artifacts dependencies.
Some artifacts need to be produced when other artifacts are produced,
but by a different task. We used to rely on anchor tasks that the IDE
calls to make sure all artifacts are produced.

However, in certain cases, like instrumented testing, the IDE is
not calling an anchor task but the test task directly. This created
situations when the redirect file is missing.

Instead of relying on anchor tasks, a new facility can be used to
declare that an InternalArtifactType is finalizing other artifact
types. That means that when those other artifact types are produced,
the task producing the InternalArtifactType will be use to finalize
the producer.

In order to implement, it became necessary to start bookkeeping the
task producing the artifacts in order to use finalizedBy.

removed all the un-necessary anchor task dependencies on the
redirect files.

Test: unit test and existing.
Bug: 208320988
Change-Id: Idfbdef6321d165f4c4cdd7c74c05f53822358c73

Former-commit-id: f99d30681e93f49c20f8bd35387c09cca2ad2b1a