Remove desugaredMethods from model v2
This is not yet used by the IDE. Only lint will
use it and the current API of passing a (rather large)
List<String> is not what we want. We want to move to
passing a file that contains the content. Therefore
we remove this unused property from the model until
we update it to something better.

Bug: N/A
Test: Updated tests
Change-Id: I0ad968542092ade8999adcaea52999b6987d3d6d

Former-commit-id: 55f17564aa83c19b517229ae7e648149b55aacf9