Add experimentalProperties to ExternalNativeBuild
This adds per module experimentalProperties to android.externalNativeBuild block.
This is in support of custom C and C++ build in Android Gradle Plugin.

Experimental properties will allow the DSL to be added last, after the feature is implemented and ready.

I noticed one of the comments was incorrect (referred to the wrong build system). So I corrected them as well.

Bug: 192006965
Test: NA

Change-Id: I82ed11ca82d30cdfafdf544f7d475e1b668ee743

Former-commit-id: e739ee8b01b5435342a400013559d988f034a6a7