Move bundled_deps, manifest_lines, jar_name to maven_library
For historical reasons the kotlin rules were able to produce a single
output jar with bundled dependencies. This responsibility is moved
to the maven_library rule to better seperate implementation concerns.

Another attribute, neverlink_deps, is added to the _maven_library rule.
This attribute is necessary to capture and expose the transitive java
deps used to build the output jar. This is different from the
bundled_deps since those are merged into the _maven_library output.

To test this change and verify the maven_library output is unchanged, I
built maven_library targets with and without bundled_deps, such as:

`bazel build //tools/data-binding:tools.compiler
//tools/data-binding:tools.compilerCommon`

Then compared the jar files using `zipcmp` to verify the contents match.

The resulting output jars for `tools.compiler.jar` and
`tools.compilerCommon.jar` are identical in contents, however the order
of the zip entries may be different. The same is true for `*.src.jar`
output.

Bug: 200690965
Test: zipcmp on outputs, all presubmits pass
Change-Id: Iab0058174aef411273075903b4c498ceed3c1137

Former-commit-id: c02b49676ea0079f7c695571146ef0cfa8299629