Allow AdaptNinjaToCxxBuild to recognize libc++_shared.so
When there are SO files that are not outputs of the current build.ninja but are inputs to something that is packaged, then those SO files should be treated as runtime files so that they can also be packaged. The exception is the system libraries in NDK (like libEGL.so). These system libraries are not intended to be packaged because they are just stubs to APIs that are already available on the devices

Bug: 192006965
Test: Unittests added
Change-Id: I54c4910102cc3f121e16c843f48074e97fd1b6a8

Former-commit-id: 930e380a64ea7625b60dd1ecfb8a1a0df8bf3010