Logic for naming AGP tasks based on ninja script arguments
This code examines the arguments passed to the ninja-generating script and attempts to derive a reasonable name for the corresponding AGP tasks (build* and configure*).
It does this by first assuming the arguments look like MSBuild arguments. If that fails then it assumes the arguments look like CMake arguments. If that fails, then the resulting task names will be less readable (like buildNinja-5 rather than buildNinjaDebug[x86]). It's a heuristic and we may need to extend the heuristic in the future.

We do something similar for ndk-build and CMake. The large difference is that we explicitly know those are the tools being used. With Ninja-generating script the script author will decide the names of parameters and their meaning.

Bug: 192006965
Test: Added unittests
Change-Id: I18e4c5848f5e45f66da6db2e40d1509a00c63b69

Former-commit-id: 2c2c4f0b18703749a4dab1f9eeb727928992d807