Implement Ninja Metadata Builder
This enables the use of a Ninja-generating script in AGP using 'experimentalFlags' DSL.

After this, the feature is complete except for the Gradle DSL which I plan to add after the feature has baked a while.

The feature is tested end-to-end in AGP with additions to the existing CmakeBasicProjectTest. The additions are:
- 'Ninja' which runs all of the CMake tests except that it is done through a Ninja script that is told where to emit build.ninja
- 'NinjaRedirect' which is the same as above but the script decides where to write build.ninja by writing it to build.ninja.txt (which AGP then reads to locate build.ninja)

Bug: 192006965
Test: Added integration tests

Change-Id: I50ec1c2cef74016e5d286816aefe37a432bbeaad

Former-commit-id: 64cf475fe6752ec187e03ca83e5a3ce539241d37