Implement DSL for building Manifest files in tests
Tests for the Manfest Merger are over-reliant on string
literals for test setup. This test data is difficult to
write and difficult to maintain, as the XML content it
describes is not subject to any syntax highlighting,
automatic formatting, or IDE refactoring.

This CL implements a suite of classes, applying Kotlin's
"Type-safe Builder" idiom, that allow for the construction
of Manifest XML content using pure Kotlin.

Bug: n/a
Test: ManifestBuilderDslTest
Change-Id: I926db045322117d1a92e0369d06c01e21cdba316

Former-commit-id: 7cf29a07476c42c9087d046a9286d66dea4fc711