Add Component.getCompileClasspath() API
This change adds an API for users to get the compile classpath for a
Component.

To avoid performance issues, AGP throws an exception if a user tries to
resolve the returned FileCollection before execution time when the
BooleanOption DISALLOW_DEPENDENCY_RESOLUTION_AT_CONFIGURATION is true.

This change also updates DesugarLibUtils to avoid resolving the
coreLibraryDesugaring configuration too early.

Bug: 158777988
Test: GetCompileClasspathTest, GetClasspathTest
Change-Id: I3b2dead363322538157b16db14332f7e7ecf4b07

Former-commit-id: c0d8d9916d019c023eb8d685c3717eac1530c584