Add android:localeConfig attribute to ManifestModel
The localeConfig attribute is used to specify an XML resource
from which the system can gather information about which
locales are supported by the application.

This attribute can be specified in the top-level (application)
manifest, or in an overlay (i.e. buildType or flavor) manifest,
but is ignored completely when originating from a library
manifest.

Bug: 204023976
Test: LocaleConfigTagTest.kt
Change-Id: I522d5cc382f3f143fa0917393f1d3aa6518a2808

Former-commit-id: 375079b54d20075211690401ae5a421bb1d2cfe9