Fix incremental inputs in MergeResources
Android resources are unusual in that they both (1) depend on ordering
and (2) can have duplicate files across paths.

This means that there is no inbuilt Gradle normalisation strategy is
correct for relative resource paths. Instead, split the resources up in
to a @Nested input on DependencyResourceComputer, which preserves
ordering of sourcesets (as validated by ResPackagingTest), and correctly
handles duplicates (see expanded MergeResourcesTest)

Bug: 209574833
Test: MergeResourcesTest, benchmark: Base100_Java8_addDrawable
Change-Id: Icfdb74a6565236f642de0f92209a942869b4e426

Former-commit-id: fa31914e26ea7a336ef583120a42575852aad2d8