Remove redundant test SDK requirements, and require SDKs to be present
Lint tests can request a specific compileSdkVersion. When this is done,
(a) the lint detectors will observe that specific version when they ask
for the compileSdkVersion (and some lint checks gate their behavior
based on this version), and (b) the testing infrastructure will also try
to find the corresponding platform in the SDK install, and if found, use
it. However, if it's not found, it will silently just use the default
platform instead. There's a test flag to switch this to an error
instead, but it was off by default.

This CL changes the flag for the built-in tests to true, and also
removes most of the tests that were asking for a version because it was
not necessary for the test's purposes.

In a few cases the compileSdkVersion *is* important, and those tests
have been updated to turn off the enforcement.

Test: This CL only touches tests
Bug: None
Change-Id: I59710fdd6060606ac481667e13821576ce608c86

Former-commit-id: 6c9dc891058518b77c13234e54f1323746848947