API check should not report return expressions
First, lint generally translates any platform annotations into the
corresponding androidx annotations when looking up annotations, but
this was missing in the place where the API detector looks up
annotations to compute the effective API level (e.g. outer @TargetApi
or @RequiresApi annotations).

This actually exposed the fact that lint is processing return
expressions as part of API check. This ended up not mattering, since it
will always find the same method annotation as a surrounding API level,
but it's wasted work. This CL therefore also cleans up the exact
relevant set of annotation usage types that are relevant for API check.

Bug: 214255490
Test: Unit tests included
Change-Id: I54a6d4c5df0a8ee13ffe17230363b8c4b2cf0ab0

Former-commit-id: d770f268954283913e00dfd9d083d1511dd04690