Generate a set of material3 themes when creating a new Activity
The set of themes for the BasicActivity (Material3) are created
when creating a new project.

But when creating a new Activity under a flavor, the corresponding
themes are not created.

This CL creates the set of themes every time a new Activity is created
(when a new project is created, the themes in the Activity recipe are
ignored)

Bug: 209907794
Test: Manually followed the steps at the bug
Change-Id: I1731b62fb77432e71a75fb517773dac13606dba4

Former-commit-id: 40461000df8a963ccf4f3aeecd066ce12546377a