Add metalava
For use in API tracking for the Android Gradle Plugin

Metalava is built and needs to run against a specific version of lint,
so this change brings in a prebuilt of that version of lint. When
metalava is updated, those older versions can be removed from prebuilts
in the same change.

Bug: 217551499
Test: Use to follow
Change-Id: Iacaba38326540330617dd9afaf1f59a48d502062

Former-commit-id: 68e3ddb7cfb241d2eb6d649553b92a9b42dbf08e