Don't require all caps for AndroidVersion codenames
Previously, lint was failing when run from AGP when targetSdkPreview or
minSdkPreview was set to "Tiramisu" because AndroidVersion required
codenames to be in an "ALL_CAPS" format.

This change instead uses a regex similar to the regex used in SdkUtils
to validate compile SDK codenames.

In a subsequent change, this validation code will be unified, but it's
better to make this change easy to cherry-pick.

Bug: 216736348
Test: AndroidVersionTest
Change-Id: I1a5963685e016cdd44a2d37bca7bf1a497b25d11

Former-commit-id: 52c1d018baef44c49669be21f2b7f9bca53626fb