Retry start fetch if there aren't any roots
Before the inspector connects we set system properties, which causes the
current activity to restart, and then we try to start fetching layout
inspector data. Sometimes the activity restart happens at the same time
as the start fetch command is being processed, leading to there being no
root views on which to register view invalidation handlers, and also
causing an exception if folding support is active.
Now we retry a few times if there aren't any root views.

Test: Added a test
Bug: none
Change-Id: I2e586542e1fd5140c62036acdac4540872b3a6d5

Former-commit-id: 19191584bbbff9d6b5cd513728d919acbd0928b4