ddmlib: Don't expose `ClientImpl` as public API
There should be no need to expose `ClientImpl` and `ProfileableClientImpl`
as return values of public method in `DeviceImpl`.

A change made a couple of years ago changed the return value
from `Client` to `ClientImpl`. This change is essentially reverting
that.

Reverting this change now is going to help with implementing
the ddmlib compatibility layer of adblib, so we can switch the
current (ddmlib based) `Client` interface implementation to a
new (adblib based) one.

Bug: 188067646
Test: Existing
Change-Id: I068b54ede63375fe82c3fd7f99c5cba01b006ce5

Former-commit-id: 411567000a665f3e9ee2542d18943d0f94e411ad