[AGP UTP] fixed lock error on missing folder
Fixed an IllegalArgumentException when a synchronized file was attempted
to be created in the avd folder, but couldn't because the avd folder did
not exist.

Found this Bug using the integration test in presubmit, but verified by
deleting the .android/gradle/avd folder and attempting to create an GMD.

Bug: 206798666
Test: Verified the test fixes the bug locally. Will rebase the
integration test on this change.

Change-Id: I5d9b3456dbfda3321269f46116ded0072657de19

Former-commit-id: c6442f0bd30c094b8f8e5076b43ce501e085d9be