Test tools/adt/idea/... and publish to AnTS
This canary target will now run a subset of studio tests. This is to
start testing the new AnTS Watcher service and sponge integration. The
former only processes test results, so we need to execute tests.

This should not put significant load on our existing RBE workers since
we allow test results to be cached, and there is only 1 BYOB builder
available to run studio-linux_canary builds.

Bug: 158230441
Test: results are available in sponge/ants
Change-Id: I7c6aa1fcd254f4a5e308f4706de4626178d4982e

Former-commit-id: 2684de21897ba2b416b62c48c82e3beabb0904b8