Check forceCompileSdkPreivew in CheckAarMetadataTask
When an Android library module is compiled against a preview SDK (e.g.,
Tiramisu), then AGP will add that information to the AAR metadata file
(e.g., adding the line `forceCompileSdkPreview=Tiramisu`).

With this change, AGP enforces the rule that any module
consuming such a library must be compiled against an SDK with the exact
same preview codename.

The reason to enforce such a rule is because if an app uses a library
compiled against a preview SDK and is uploaded to the play store, the
app might crash on devices running the finalized API (if there are
changes between preview and final), but devs are prevented from
uploading an app to the play store if compiling against a preview SDK.

Bug: 218900545
Test: CheckAarMetadataTaskTest (unit and integration tests)
Change-Id: I1c5eb9248fea137adb4c722625ce15c1c41f4e0e

Former-commit-id: 665a35b5807844ea625d7977cd52469f76e7ec7d