Disable tests that fail under r23b
These tests fail when upgrading to r23b. The tests passed by
coincidence under r21 because the version of objcopy in r21 didn't
extract just the symbols from the SO. In r23b, the symbols are
correctly extracted and the result is that the size of the stripped
SO is not the same as the size of the symbols that were stripped.

Bug: 169266007
Test: Are tests
Change-Id: I900fcde028dfd860dca79fdb5d4bc35caf29fd59

Former-commit-id: 108869018df8329241ecf2d61777e783351cddfc