Restrict deprecated studio_repo to current users
studio_repo is a maven repository representing what used to be packaged
inside Studio, but now doesn't correspond to anything we publish,
and is obsolete.

Tests that depend on this should migrate to using the Android Gradle
plugin as built with Gradle (wrapped in Bazel), i.e.

    "//tools/base/build-system:android_gradle_plugin.zip",
    "//tools/base/build-system:android_gradle_plugin_runtime_dependencies",

For an example of a migration, see
https://cs.android.com/android-studio/platform/tools/adt/idea/+/599c84251b2a6e5acefbe94015bebb278d4b5db8

Bug: 143864616
Test: Updating bazel visibility
Change-Id: I045b4b1ac1d281a5fad9d8c1ee0b6b9058bf1eae

Former-commit-id: 5803b18ff33e7dac7e06dbd17967c6d9e211a77f