220161119: IconMissingDensityFolder uses absolute path in lint-baseline.xml
Lint has a flag to turn on absolute paths in reports. This flag was set
in a central place which meant that it was used everywhere when a lint
check asks for a path to be formatted. In the case of a couple of icon
checks, that meant that setting the flag would cause error messages
themselves (referring to files or folders) to contain absolute paths,
which then showed up in baseline files, which then make the projects
non-relocatable.

The original intent of that flag was really only for it to affect the
paths shown on the left hand side in report files, e.g.

/full/path/to/res:Warning: Missing density variation folders in
                  `res`: drawable-hdpi, drawable-xhdpi

This CL updates this such that the absolute path flag is set when
generating paths for reports, not from detectors.

Since we've changed some error messages, it also updates the baseline
code to account for this such that it can continue to baseline match
files generated prior to the message change.

(Also updates formatting to replace javadoc @link tags with proper
kdoc links.)

Fixes: 220161119
Test: Unit tests included
Change-Id: Iaec00635c9b07e1c6a982d67173c4723a189aaf6

Former-commit-id: cfab120648d667e8310615029fcb3ad28416575d