first implementation of the fused libraries plugin.
So far, it only does a casual class merging and use normal jar task
to bundle things up. This will be refined in subsequent CL.

Plugin itself sets up various configurations needed to properly
support consuming libraries that are merged and continue having
the other libraries as dependencies.

Minimal extension is registered that allow setting the desired
namespace for the new fused library (not used yet by tasks).

publication is suppored although ran into issues with
configuration caching (to be resolved in subsequent CL if it
is not a Gradle issue).

Test: Added some integration tests and unit tests.
Bug: 62121508
Change-Id: I3b43c133feceaaf2ea3a0553fbb19ca6eb7312df

Former-commit-id: f3473c240082733842cb2970bc0abab5bf2ea7f9