Add task to update lint baseline
This change adds the updateLintBaseline task, which updates the lint
baseline file.

Previously, if a user wanted to update their lint baseline files, they
needed to delete all of the existing lint baseline files in their
project and then run lint. With this new task, they can accomplish the
same thing without needing to manually delete all fo the old baseline
files.

The baseline file is an ouptut of the updateLintBaseline task and an
input for the lint report tasks, but we don't want the
updateLintBaseline task to run every time a user runs lint, so the lint
report tasks do not depend on the updateLintBaseline task. Instead, we
use mustRunAfter to ensure proper task ordering.

Bug: 188666845
Test: UpdateLintBaselineTest, UpdateLintBaselineStandaloneTest
Change-Id: Ibab20e2125c3f3d8e02e631cdf6a76fa38293eb5

Former-commit-id: e638ae1c670eb68e48c5212f2e526b57ed71439b