Make sure we skip kotlin.Unit from @CheckResult check
UAST normally maps Kotlin's type Unit to void, but there are
some exceptions. This CL includes such an example, and updates
the CheckResultDetector to treat Unit the same as void.

The included example is contrived and is more likely to be
encountered with the @CheckReturnValue annotation which can
be applied to whole packages.

This *may* fix issue 214582872 (which I cannot reproduce in
isolation).

This CL also updates the TypesTest AST pretty printer to include
types for variables and methods to help track any further
unintended changes in this area.

Bug: 214582872
Test: Unit test included
Change-Id: I0c4a8a91c29124a158a961860698788a51ecd822

Former-commit-id: 5e03bccb80faf61c486be0dfd5b919c9f430be90