fix path relativity for dex transform input.
The code is using FileChange.normalizedPath API to caculate and
compare dex file names from .class file names. However, because
the input property was annotated with name_only relative path
sensibility, Gradle would only provide the file name as the
normalized path.

This becomes a problem when code was trying to find associated
dex files for deleted .class files as the normalized path would not
contain the package name as part of its path. The result was that
the associated dex file was not found and would therefore not be
deleted, making its way into the final APK.

Changed the path sensitiviy to relative so normalizedPath provides
the full class name including the package.

Bug: 223365583
Test: added integration test
Change-Id: If04c848a174fb2a981f402a855c5132433f6a0fd

Former-commit-id: 97c89ea7553593760265df89355610eaf25e86c2