Fix mechanism to exclude useless native debug metadata files
The previous mechanism didn't work with NDK r23. Instead of comparing
the extracted metadata with the stripped native lib, we compare the
input file with the stripped native lib and don't try to extract native
debug metadata if they're the same size (indicating the source native
library was already stripped).

Bug: 169266007
Test: ExtractNativeDebugMetadataTaskTest
Change-Id: I003451fae7e9fc1b1648cc60d7ee2086809ec800

Former-commit-id: c33830aba175cf12d3889c1bfe1bd7a80f70be3e