Support manifest placeholders in deepLink URIs
With this change, users can use manifest placeholders in their deepLink
URI scheme, host, and/or path.

For example, a user could have a navigation file with the following
element:

    <deepLink app:uri="${scheme}://${host}/${path}" />

Previously, AGP supported using the specific ${applicationId}
placeholder in the host and/or path. This change adds more general
support for any manifest placeholder.

If the manifest placeholder is used in a navigation file for a library,
AGP will substitute the library's value if it's specified in the
library's build.gradle file. If not, AGP will use the value specified in
the app's build.gradle file. This is in line with existing manifest
placeholder substitution behavior when used directly in the
AndroidManifest.xml files.

Bug: 36994900
Test: NavigationPlaceholderTest, DeepLinkTest
Change-Id: I1fc57c2cd40658452ec30efe4eb5c78803702e11

Former-commit-id: a2dc5cc27b36d0aecf9e75448218f2f0afdadefa