Convert MergeJavaResourceTask to NewIncrementalTask
The new API could also help simplify the java resources merger, as more
information is avaliable about changes from the new API that would
enabled its implementation to share more with the packaging task
implementation.

For now this is a small change just to get AGP off the old API which is
blocking its removal in Gradle.

No behavior change.

Bug: 218478028
Test: Existing AGP integration tests (e.g. MergeJavaResourceTaskTest)
Change-Id: If0b273cdc94fa7a6e4167cba93b5ded99188e2f5

Former-commit-id: 5b993b482b9948dc34abe5c9b25c977d6a18498f