224869924: CommitTransaction Lint False Positive
With a correct class path the code referenced in that bug resolves fine,
but my theory is that there must be some sort of resolve problem related
to the class path which causes this to go wrong (and placing an
unresolvable call into the chain results in the reported error) so this
CL adds some proofing where the data flow analyzer remembers whether it
encounters unexpected resolve problems and if so refrains from reporting
problems -- unless it sees that there aren't *any* possible cleanup
calls in the method (without taking instance tracking into
consideration).

Added this to the handful of detectors which uses the data flow
analyzer.

This is probably the same root cause for 224618492 as well (added test
confirming that scenario works.)

Test: Existing tests for the existing uses of the data flow analyzers,
      and a new unit test in CleanupDetectorTest specifically using an
      unresolved symbol (and a missing cleanup call to make sure we're
      really dropping the report due to the missing resolve call.)
Bug: 224869924
Bug: 224618492
Change-Id: I4ce246a6e1fdb4d73520814d8ae468bdbbf22222

Former-commit-id: e8d356b5b21be9c6a8f595306e6d69c2ee325bb8