Split VariantDslInfo into multiple interfaces per component
The split corresponds to the creation config interfaces, each creation
config implementation should use the corresponding dsl info which should
be enough for the creation config to get the needed dsl values.

Currently, the implementation still implements all interfaces. Once we
use the new interfaces we can split the implementation as well.

Bug: 226912999
Test: N/A
Change-Id: I07f6a572f6c684d7ed632368e4a5271693d0a2ba

Former-commit-id: 265e5afc3849cf40e86a347ddd8b940ff313b43f