[GMD] fix lateinit var issue in snapshotHandler
Fix an issue where the emulator process closes unexpectedly before we
could detect the serial on adb, resulting in attempting to read a
unitialized var for logging.

Improved logging around when this event occurs and prevent the variable
from being read in this case.

Bug: 227461645
Test: Manual, tested in a case where I remove the underlying system
image of a valid GMD which will cause this issue (See bug: 227431640)

Change-Id: I94dc53a3128e094c4b38d9e5c89136a6222b01bd

Former-commit-id: 1e668edf15e9dc4ee9f84d6bfa63b55f967e7eff