Move away from configuration warn mode
AGP integration tests should move away from using warn mode because if a
build is able to finish with this mode in the first run, this build
could break in the second config cached run. Therefore, we need to aim
for making all the tests running with the ON mode so that our build
would never break due to config caching.

Bug: n/a
Test: existing
Change-Id: I7073b9c918494d5f3083cdb08e902e1032e1e4f2

Former-commit-id: 7dbddff68a90b5a8dc397880602d0e4f62248127