Print warning about .bazelrc
A future change will update the repo manifest and introduce a
toplevel.bazelrc file which includes common.bazelrc. This will eliminate
the need for the bazel wrapper to explicitly point to the bazelrc file.

To avoid custom .bazelrc files being lost, print a warning and provide
a fallback for user custom settings (user.bazelrc).

Bug: 187294442
Test: n/a
Change-Id: I85d6b1063d1c09cc1c0d16100135946efbbb88de

Former-commit-id: 4227ad744c7efdf2d3c3f158046a6d5480fb1dfb