Create a bazel target for NDK 21
... so it can be used in DolphinV1 sync perf tests.

DolphinV1 perf test uses AGP 7.1.0, but this version of AGP uses NDK 21
by default. When AGP was bumped to use NDK 23 in ag/16872261 DolphinV1
started having sync issues since NDK 21 was no longer available.

This change adds a new target and makes it visible to sync-perf-tests
so DolphinV1 can depend on it to have NDK 21 available.

Bug: 228208989
Test: New target used by sync-perf-tests
Change-Id: I0db366be5218535a9f6cbd84b019455950b93b1e

Former-commit-id: 93dcb7a43d5c7caabe5550db60d3d14a07da3e95