189354516: AppLinksAutoVerifyDetector does not have a timeout limit, can hang
This CL updates the AppLinksAutoVerifyDetector to use lint's wrapper
around opening URLs -- not just so that it will correctly use the
user's configured IDE proxy settings when running in the IDE, but also
such that it will obey the new --offline flag (see issue 228976904).

While doing this I also set it to have a timeout, as requested in
bug 189354516.

I also removed the hacky way the unit tests for this lint checks had
been written and replaced them with using lint's test DSL support for
testing URLs, which had to be updated to make it straightforward to
test error codes.

Test: Unit tests included
Fixes: 189354516
Change-Id: Id2768d55fa8707908b81139b994192c25e9e1049

Former-commit-id: 2ef9791c031b6bba3dd65fe0bf5a51eeba07dd7d