Add bytecode validation to LE runtime
When class updates are received, LE will ensure that certain conditions
are met before adding the classes/bytecode to the LE context. If
conditions are not met, an error object is returned detailing the issue.

Current checks:
 - Cannot add/remove methods (covers renaming, changing signatures,
   changing return types)
 - Cannot add classes

If any class in the update fails validation, none of the classes are
added to the context. If the classes have yet to be primed, failing
validation will also prevent priming.

Bug: 228632150
Test: BytecodeValidatorTest
Change-Id: Ib9976c94f626aa9fa7dddbd25df45fa5f610c938

Former-commit-id: b2fbb616af7872cdececa50905672e05a5b2ef0f