Delete AGP check of Studio version
The version of Sync that ships with Chipmunk (and above) has checks in
its Gradle worker process for AGP/Studio version mismatches, including
the check for the AGP version being too new for Studio.  Exceptions
thrown by these checks are captured as BuildIssues and formatted with
quick fixes and links to documentation where appropriate.

Therefore, versions of AGP 7.3 and above, corresponding to Studio
releases of Dolphin and subsequent, do not need to perform their own
version checks; the check in Studio will be sufficient to detect
version skew of at least one release version.

Removing this check in AGP 8.0 potentially leads to confusing messages
for users of Android Studio Bumblebee (and below) attempting to use
projects configured to use AGP 8.0 (and above), following the release
of Flamingo and AGP 8.0; this is probably a relatively unlikely
scenario.

Bug: 227452887
Test: n/a
Change-Id: I06881a62651beed0e836aa569cc60ae8a47c3069

Former-commit-id: c23fb705a9ef8c6cbbe55af1b98fbb0baaf2d980