Adds AndroidX test mode, and fixes androidx handling in checks
A couple of older lint detectors were only checking the old support
library class names (for example, to check whether fragment transactions
are committed). This CL fixes that, and to help audit that there aren't
other similar problems (not a plot twist: there were), it also adds an
"AndroidX test mode" which takes test cases written with support library
packages, maps them to androidx and then repeats the test.  This
revealed some other checks that also needs to be fixed.

(Note that this is an internal-only test mode, not one added for all
external lint unit tests too.)

Test: Unit tests included, both for the new test mode and various
      existing tests had to be updated.
Bug: Not filed, though issue 224869924 is what prompted this analysis
     when I was wondering a correctly identified violation in google3
     (using old support library names) didn't fail when I wrote an
     "equivalent" regression test case in the androidx namespace.
Change-Id: I142ff1fa7de238f51fd457f37c462dd8c9eb870b

Former-commit-id: a7d88d2dbeef5bc6c3dfc1dc7c5e091f13d9a8f8