229713185: API level error for using TiramisuPrivacySandbox
When lint deserializes the lint model handed to it from AGP, it's only
being told about API levels with codenames using just the codenames;
from there, it has to figure out the corresponding API level. It wasn't
doing that, so it was using an API level of "0", which mean the API
check did not work with minSdk=preview.

Fixes: 229713185
Test: Unit test included
Change-Id: I4488b57f23b7287a5eb0bfa05124528b7136f51d

Former-commit-id: c4381d0cd90418d94a8eab80929faee2d84853d8