Handle dependencies without file, but with transitive deps
It is possible to have a dependency without file, but with
transitive dependencies. This change handles dependencies that
match that criteria (such as kotlin-test), and creates a new
DependencyType to represent these nodes in the dependency graph.

Also, snapshot test is updated to depend on annotation lib, to
make it simpler.

Bug: 230648123
Test: FullDependencyGraphBuilderTest, DependencyWithoutFileWithDependenciesTest
Change-Id: Ic68c7e86e885bb2ca0f92124d9275db9c8200046

Former-commit-id: c943569eb755549d48690169c81ce28eec6a3e40