Delete obsolete visualizing utils
This was a complaint about tiny texts on an old version of A.S., and
the user had to patch hard-coded text size of `8.0f` in
`CallHierarchyRenderer.class` (discussed on StackOverflow). But the
usage chain `CallHierarchyRenderer` -> `TraceViewCanvas` -> `TraceView`
is not anywhere in production code today. This change removes these, as
a witness that this problem is obsolete.

Bug: 134447818
Test: n/a
Change-Id: I2aa38947d73a9c551a6ab9c9a2f2830d5bd764e6

Former-commit-id: 54ceac5f879ac074ab2dc9bc16e4bf40b7b45b9a