Support global synthetics generation from d8
This change is to support the d8 feature of generating global synthetics,
which can be split into two parts: generating global synthetics along with
dex archives in dexIndexed or dexPerClass mode; merging global synthetics
into dex files with or without dex archives.

Incremental dexing in task and transform is supported when global synthetics
exist. Dynamic feature module is also taken care of in this change.

Bug: 227211690
Test: GlobalSyntheticsTest, IncrementalGlobalSyntheticsTest,
GlobalSyntheticsFeatureTest

Change-Id: I2ef1562b9887814ee13ea601efc2fa5eb3058392

Former-commit-id: 9a8594abf97b65d2ac4555773a1fc4d9df197d71