Inverse the projects to settings relationship
Previously the Settings plugin would push the values
into the Android projects. This requires the gradle-api
artifacts to be a dependency of the settings plugin artifact,
which can lead to classloading issues.

This also prevents the settings plugin to set values that are
not publicly exposed by the gradle-api artifact. This means
that the settings plugin could never expose propertis that
are *only* set in the settings plugin as it would rely on
the public API of the app/lib/... plugins to send the values
there.

This reversal permits us to have settings-only properties that
are copied into private APIs of AGP.

Right now there does not seem to be a good way to reach into
the Settings extension from project though, and so we use
an extra properties on the project to get access to it. This
should be improved somehow...

This CL also adds support for compileSdkExtension

Bug: N/A
Test: same integ test, but expanded the settings extension tests.
Change-Id: Ie398b4903c9967a4e45a41d022b0c07a9aab7cfd

Former-commit-id: 4372e4ac3261c74462cc81005cfac34d362c54d7