Allows building without vendor repos
By doing the following things:

Introducing a new gmaven_without_vendor target that
only contains the maven artifacts that can be built without
our vendor repos (Note that the Android SDK we currently
use is in a vendor repo)

Introducing a new --config=without_vendor that controls
the execution of vendor.bzl to set up blaze and ndk targets.

Note that WORKSPACE conditional execution is not easy to do.
In order to do that a new repo is created differently based
on different arguments and that repo is in charge of setting
up vendor calls.

Test: Modifies aosp_studio_linux.sh to use this
Bug: 220328663
Change-Id: Ib78d26397b2ba4be31299adf34c1c8e75e60cd6d

Former-commit-id: eec96a867a13811d1b218824f6b261e650b26492