Remove ProjectInfo.getProject
The remaining usages were related to the old variant API
that required access to container creation. In order to enable
this, this changeset replaces ProjectServices and VariantServices
in these classes with DslServices which has container creation.

One particular usage also required getting access to DirectoryProperty
so this change also adds it to DslServices. This is not actually
something we want there, but it less problematic than keeping
Project access to all Services classes so it's fine. The addition
is already deprecated and will be removed in 9.0 when the old
API is fully removed. Filed a bug to track this (b/232471680).

Fixes: 202449978
Test: existing tests
Change-Id: I78c8a1193359b1e745800c92233ec7cd0c14d702

Former-commit-id: fc9d3b684b7db59f5ecf8f19485093ced24da9e9