Fix regression for using asm transforms with old transforms
The regression is caused by registering the asm transform task after the
external transforms which made both run on the compilation output and
have different outputs and only the output of the asm transform is
picked up from the transform pipeline to be dexed. This change swaps the
order of registeration to have the old transforms get the output of the
asm transform task.

This change also fixes an issue for old transforms when coverage is
enabled, as the filtering was done on the scopes instead of the types
which ended up always returning an empty file collection and made
project classes not included in dexing.

Bug: 232438924
Test: Added intergration tests for both cases
Change-Id: Ic0c71e94bc7d81726084159f983266e648ec13f6

Former-commit-id: ea1c4cf15ac27d318ea384f3402cafbb1bf620a6