adblib: add trackDeviceInfo extension
The `AdbLibSession.trackDeviceInfo` extension allows tracking changes
to the state of a device connected to ADB, as long as the device is connected.

This change also fixes a bug in `get-serialno` in FakeAdbServer to correctly
handle the case where a device serial number is invalid.

This change also includes a `SuspendingLazy` utility class that is
similar to `Lazy`, but allows using a `suspend` function as the
initializer.

Bug: 188067646
Test: Included
Change-Id: I3291a1ce7c4ab6d64d50052ee9f5af9a2ce83b9e

Former-commit-id: aae2cce7381fa52f1684eb7bd824b679d9565774