LE: Detect closed channel and discard them
The AdbInstallerChannelManager can hold on close channel in its cache.
This is not a problem since the AdbInstaller is built to detect failure
to communicate with the installer binary but it triggers a push of the
binary in DEAMON mode.

This CL detect when a channel is closed, dicard of it which will trigger
and optimisit start.

Bug: NA
Test: AdbInstallerTest.java
Change-Id: Ife0c09614f6fa720fdc84146744143562b1d2b21

Former-commit-id: b2f3ca6d0260e1548314b54e483278de692a6d07