Fix snapshot problem
In the implementation of the CaptureSnapshotCommand we called a
updateAllViewChangeNotifications function which is supposed to
change the callback to SKP from BITMAP.

The code would setup the SKP callback and then call close on the
old handler. The implementation of the close in ViewDebug would
then clear the callback in the View renderer thus essentially
stopping all further events including the events that was
supposed to be captured for the snapshot.

After this change: close the old handle first, and also keep
the callback if it already has the correct type.

Fixes: 234407838
Test: Added unit test
Change-Id: I6e6d9a54a1920619abd362db6b5da5d55aa37333

Former-commit-id: 84f968be2726fb879a1f981d370d0ac7858eaa2d