Add testFixtures prefix for some testFixtures published artifacts
Gradle has an issue when different artifacts with the same name are
produced from the same variant with different capabilities, if these
artifacts are transformed by an artifact transform, they will be seen as
one and only artifact gets transformed (https://github.com/gradle/gradle/issues/18458).

This change adds a testFixtures prefix to artifacts that gets published
and are potentially transformed by an artifact transform as a
workaround.

Bug: 234629577
Test: Incoming
Change-Id: I81ddf8a6ee952fa198c011a0b4f43729f61378cd

Former-commit-id: 3e87e8eb2bdae4592add02ffe1ceae18267c9c4e