Privacy sandbox sdk: Generate APKs for local dev
Fix the incorrect publication of the runtime classpath in the fused
library base plugin to not publish dependencies between projects too, as
well as adding support for publishing the ASB between local projects (as
a placeholder for the asar).

Adds a task to trigger the transformation of the privacy sandbox sdk in
to APKs and extract them to an intermediate directory.

Bug: 235469089
Test: Covered by PrivacySandboxSdkTest
Change-Id: Ica169f3898d91c12e945926f044a1c85ae06dd7d

Former-commit-id: 84adfbbdc107ebcf98db6371ca96a21d7480e527