Improve handshake
If a device is connected to Studio while locked, the current handshake
implementation does not work.
`dumpsys activity processes | grep top-activity` would be empty.
Therefore the handshake would terminate saying that the device does
not support foreground process detection, which might be a false
negative.

To work around this limitation, this change enhances the handshake
protocol in this way:
If top-activity is not found, check if there are awake
(dumpsys activity activities | grep isSleeping=false) or sleeping
(dumpsys activity activities | grep isSleeping=true) activities.

If there are sleeping activities, but no awake activities,
it means that top-activity might be absent because the
device is locked. In this case the handshake terminates
with an UNKNOWN supported state.

If there are awake activities, it means that for some reason
the top-activity is not returned. In this case the handshake
terminates with a NOT_SUPPORTED state.

If there are no awake and no sleeping activities, it means
that dumpsys is not working as we expect. In this case the
handshake terminates with a NOT_SUPPORTED state.

Test: foreground_process_tracker_test.cc, ForegroundProcessTrackerTest.java
Bug: 232067042
Change-Id: I51ffbefeca54039da8064d49212ea44285e860f1

Former-commit-id: 97db61e2e37a2659f9caf8ab85cf12409da0f9cb