Remove lint check for old build tools
This check is not accurate as latest version for studio and for
AGP can be different in case of different versions.
It can only lead to confusion but not as much valuable (normally
user should not use this at all), hard to implement the right way
(would need to maintain the mapping from agp version to build tools
version), and user gets the proper sync warning anyway.

Bug: 183650753
Test: updated
Change-Id: I114a120b705021719226ecfc841786146997d259

Former-commit-id: 80e71a705769a77ac6b75788caa0aea8156a391c