Add IgnoredTests
When adding the @Ignore annotation, it should be linked to a bug so action can eventually be taken i.e re-enable the test, remove the test etc. This change adds IgnoredTests to keep track of currently disabled tests in a way that mean groups of related disabled tests can be easily found by using intellij's find usages feature.

Bug: 236811549
Test: A later change will use the const strign for disabling tests
Change-Id: I8de5d63b44485f5fff5904b2a2c1c00d0b064947

Former-commit-id: 11c62dbe45e9415e5b46ca0df1c5a61d21fe8b1c