237819574: Cache platform lookups
Lint was caching the SDK platform handling on a per lint client
basis, but when it's running on a tree of many individual lint
targets, this meant it would recompute the platform for each one
(which involves parsing all the XML files in the SDK foldet etc).

This CL caches these. They don't contain references to outside
data structures so this isn't going to turn into a leak.

Bug: 237819574
Test: New unit test added (plus existing tests covering functionality)
Change-Id: I2372910fdac1eb3619997809514f172f361dc422

Former-commit-id: d0011c07a3b0eea554c79f54508f01bca763bda1