adblib: Java bridge utility class
Implementation of a `JavaBridge` singleton class to make
it easier to invoke any adblib suspending function from
Java.

* `runBlocking` allows running a suspending function while
blocking the calling thread. It is illegal to call this from a UI
thread.

* `invokeAsync` allows running a suspending function in
a worker thread, returning a `Deferred<T>` instance
to access the result of the function.

* Also change `AdbSession.scope` to be a `SupervisorJob` by default
so that cancellation of child jobs does not cancel the scope.

* Also add a `isEventDispatchThread` property to `AdbSessionHost` so
that adblib[.tools] can check and ensure long running operations are not
executed on the UI/Main/Edt thread.

Bug: 188067646
Test: Included
Change-Id: I59ba9168d278be19d106eed27ea9a3c554171fb8

Former-commit-id: a59fe348a401f82e7d42366e718f9962f037e0dd