Handle input file deletions when configuration cache is enabled
Previously, some tasks would fail if one of their input files was
deleted when configuration cache was enabled.

This change fixes the issue by always setting the input file as the file
itself (instead of setting it as null when the file doesn't exist). The
check for whether the file exists is instead moved to the task action.

Bug: 237450413
Test: ProcessManifestTest
Change-Id: I608453de0f4f71d1ff057a2d4f05de4140e31bfe

Former-commit-id: baaa55163de77093b1e5566b99d73606ea41012c