Replace :kotlin-stdlib-jdk8 with :org.jetbrains.annotations
:kotlin-stdlib-jdk8 includes multiple extra compile time jars that were
unused. Only the :org.jetbrains.annotations dependency is required.

I was getting errors from kotlin-stdlib-jdk8 when trying to get compile
jars working with maven_library rules. Use of exports should be avoided
as it causes all users of a dependency to also depend on the exports.

deployer was implicitly relying on the export from ddmlib, so the
dependency on :org.jetbrains.annotations is made explicit.

Bug: n/a
Test: n/a
Change-Id: Idfb401b206bf9a363a99e2b2e38ab3168f5ebbb3

Former-commit-id: d51a7ffb13583b0462ae1832201ef27dd8e465b9