DDMLIB: Make Processor non-blocking friendly
The current implementation of Processor use busy loop as soon as a
socket contains data. This result in high CPU load when only a portion
of the expected message is in the socket buffer.

This CL gives an internal buffer to each processor where bytes are
stored until a full message is received which make them selector
friendly.

Fixes: 231790953
Test: ProcessorTest.java

Change-Id: I51c42254119606acb52f5542f3003c23de44e7bf

Former-commit-id: b0dd73abb87d9904087df7cba24e940234c5b6cd