Fix processing bug in NodeUtils.updateNamespace
When a node has not been processed by processSingleNodeNamespace, it is
removed from the attributes of the node. However, this can cause an NPE
since the next node that should be processed will now have the same
index value as the node attribute that was just removed. Therefore, in
this case we must ensure the index does not change in the next iteration
and reduce the total number of attributes.

Bug: 237788762
Test: NodeUtilsTest
Change-Id: I84af41862d619ca9e774db9ad8fed76098655231

Former-commit-id: 1f3ab0df40ea87fbc77818a82bf95fbfcacf09d8