Allow late addition of sources
A common idiom using the older applicationVariants/libraryVariants API
is to register generated sources through the sourceSets container, e.g.

    android.applicationVariants.configureEach { variant ->
        generated =//Output of a variant-specific task that generates sources
        android.sourceSets[variantName].jniLibs.srcDir(generated)
    }

In the process of adding the sources to the new variant API, AGP started
reading the sources before creating the components, meaning such
registrations were ignored. This wasn't a major issue for java and
android resources, as there are targeted APIs for registering generated
sources on the variant API that users were using, but for other types
this is a breaking change that will affect users.

Delaying the reading can also cause strange effects when using the older
variant API, so this change is a compromise, the sources are still read
from the DSL as part of component creation, and the component
SourceDirectories is passed to the DSL object, allowing it to forward
any subsequent additions to the new variant API.

Technically this means that setting the sources

    android.sourceSets[variantName].jniLibs.srcDirs = generated

will not remove the previous sources when triggered from the variant
API, but for the real use cases that seems acceptable.

Bug: 237783279
Test: Strengthened SourceSetsMixedApiUseTest
Change-Id: Ib733a28a8195a3080048cea9e6b65707847d3e1e

Former-commit-id: 21679fcbbafd4af740fdc9427fe548cff0f15918