move the finalizeDsl blocks execution.
Some checks were performed before the finalizeDsl blocks ran which mean
that a user that potentially would want to set the extension objects
entirely through the finalizeDsl block would experience validation
failure preventing him to do so.

Moved the exection of the finalizeDsl blocks first in the
createAndroidTask() to make sure the Dsl as expressed by the use is
complete and final before validating the extension objects.

Fixes: 215407138
Test: added a regression test
Change-Id: I3677947797c238cd1bb2f12d5ff4c99f35490921

Former-commit-id: 35f32a291d73c8c2e009456e0142a22dc4ed3300