Close inputs and output in IncrementalFileMerger::merge when exceptions
Previously, the inputs and output would not be closed in
IncrementalFileMerger::merge when an exception was thrown before they
were manually closed.

This change is a modified version of
Iaa211da7fadac1af60b48406e7aee3fcd006a76e which caused a performance
regression because the inputs were being opened multiple times.

Bug: 239148417
Test: manual, monitor flaky FusedLibraryMergeArtifactTaskTest
Change-Id: I9e795bafc0bb69e527983565ec761b83a8639c66

Former-commit-id: 050d8e24de1ad0ae771c1330f4e16cc856420b5c