Better diagnostic for exception during model configure
Adds a structured log proto that records the outcome of C/C++
model building for Android Studio sync.

The underlying issue was related to b/221231432.

Repro steps to gather this type of log:
1) Set up androidx repo
2) cd frameworks/support
3) mkdir -p .cxx/structured-log
4) ./studiow native

I also improved the end-user experience by adding a message that
indicates the sync failure along with a callstack into AGP code
for any exception that occurs during 'configure'.

Bug: 221231432
Test: NA
Change-Id: I5b01342de7a7c562a2d9b05551b81af89ba5f7c8

Former-commit-id: 01bdbb0885eb55ad6df78d862236e38c4430dff0