Rework VersionChecks and fix many bugs
The VersionChecks utility methods are used to figure out if anything is
known about the SDK_INT API level at a given target element. The code
was pretty convoluted, because it was computing the constraint *based
on where it was going to be used*. For example, when computing what the
constraint should be for the expression "SDK_INT > 23", the answer would
depend on the context -- if you're coming from the else clause of an if
statement for example, it would return "SDK_INT <= 23". It also tracked
whether you were looking for the lower bound or the upper bound and
varied the result based on that.

This made the code really hard to reason about, and there were many
bugs. I initially set out to improve handling of Kotlin when statements,
and realized I had to make a much bigger change.

This CL rewrites the VersionChecks class to handle this in a much
more straightforward way: it no longer takes into account where
the constraint will be used -- it just returns what the constraint
is (e.g. "SDK_INT > 23" in the above example), and then we instead
handle manipulating the constraint (such as inverting it) when we're
processing constraints as we traverse elements like if-statements and
when-statements.

As I was fixing this, I took the opportunity to make many other
improvements:

- Improve the handling of Kotlin's when statements. It now computes both
  the constraint to be checked within each case (which is based not just
  on any SDK_INT conditions on the current case, but also subtracting
  out any constraints from previous cases), as well as the constraint to
  be checked *after* the when statement; it does this by checking
  whether each case potentially falls through and then computing the
  worst-case constraint.

- It now handles Kotlin range statements (e.g. "SDK_INT in 19..21").

- It can now handle more complicated constraints than just "x < SDK_INT
  < y", which it was limited to before. It now models all the API levels
  in a bit vector, such that it can handle for example x != N, or
  combinations of spans of API levels such as (when (SDK_INT) 14, 18,
  20..22 ->).

- Instead of just returning whether an API constraint has been met, it
  can now pass back the API constraint. The API detector uses this to
  make error messages more useful. These error messages tell you what
  the current min is, which used to just refer to the module's
  minSdkVersion. Instead, it will now tell you what the known minimum is
  at the local location, showing you that it took SDK_INT checks into
  account.

- Fixed many bugs:
  247146231: VersionChecks doesn't handle Kotlin range checks
  242479753: False positives in VersionChecks when logical
    operators and constants are combined
  and many more as I scrutinized existing tests and added many more;
  see the updated diffs in VersionChecksTest and ApiDetectorTest.

- It also updates the PrivateApiDetectorTest with some new golden files.
  Most tests pass, but two of them required updates. I don't fully
  understand exactly what this detector is doing with API levels, but I
  have more confidence in what the version checks code is doing now so I
  suspect these test results were wrong (like many of the other golden
  files I've updated in VersionChecksTest which I do know were wrong
  before).

- (Fixed a bug in the body-removal test mode which I encountered for one
  of these tests; the scenario is described in the associated unit
  test.)

Future potential work here:
- Switch the version check exit finder over to also using (and
  returning) constraints, and consider merging the two version checks
  into a single one.
- Stop using null to represents no known constraint and switch over to
  using VersionChecks.ALL_API_LEVELS instead (and get rid of the null
  checks.)

Test: Many new unit tests included, and older ones updated
Fixes: 247146231
Fixes: 242479753
Fixes: 244824912
Change-Id: I965a450ed96a38fe6a5e142cbc225dad94a54645

Former-commit-id: 88ee4d52b99f72187b8893456b9e23eb55f74bf6