[SDK Index] Log cache exception for empty  message
We log the message that comes with the exception when a caching error is
detected, but if that message is empty then we do not have more
information about this error.

This CL uses Exception.toString in those case so we can have some
information to debug caching errors for the SDK Index.

Bug: 241252867
Test: N/A
Change-Id: I2e0f8d04d271906d4bcb00f55e08a870cf28a5bd

Former-commit-id: 232d445973faaea21fb9625413d6afb7e7b749d1