Allow lint.jars to include libraries in the com.android namespace
Lint's bytecode verifier determines whether something is
part of the lint API surface by checking a series of package
prefixes. This could cause problems for some lint custom
jars that happen to use the same internal prefixes, such as
the framework build jars. These were previously handled as
hardcoded checks in the bytecode verifier.

This CL changes this such that it also checks the class names
included in the jar itself, and if a referenced API is found
in the same jar, we don't flag it as a Lint API.

Test: Existing (tests updated and some removed; we no longer
  disallow including lint checks in these namespaces)
Bug: Not filed
Change-Id: I3b6e4931a98472d63266abbcdc55f709a3131214

Former-commit-id: c1f35dfa2477e7004b600b7471b68e25184e0376