Add bazel target for PrivacySandboxSdkConnectedTest
This changes the emulator script to generate keys before the initial run (required for Play Store images, which are user, not userdebug) and also disables some emulator features to make it work in a sandbox.

Also increases the timeout for emulator connection since the newer images take longer to boot.

Bug: 240533883
Test: Test changes
Change-Id: I1c3f546dbe106d16dc7edfad85f73e05cc41dca4

Former-commit-id: 7dbabf4a64732c5d795f63859fa3e0584926d9a1