Remove an incorrect restriction when computing test application id
Remove a restriction on the tested variant to be an apk producing
variant so that the test application id can be inherited from it. This
caused the test components of library variants to not have correct
application id when the test application id is not set, which causes the
namespaceForR to be incorrect when both the namespace and the
testNamespace are not set through the extension.

Fixes: 242832889
Test: VariantDslInfoTest2
Change-Id: I665df4e85abc0cb1a5e860a4f99c0b1307962fa0

Former-commit-id: c1ad01fca3ca47f1a9daff828577e37f21046a8d