BazelMultiplexWorker: simplify/fix error handling
Previously we would try to shut down the worker process if an
exception occurred, just in case the error somehow tainted future
jobs running in the worker. However, the UX when this occurs is not
great: Bazel only presents a truncated snippet of the worker log
to the user.

So, now we simply catch all exceptions and include their full
stacktraces in the WorkResponse that is returned to Bazel.

In the rare case that an exception *does* taint the worker, e.g. an
OutOfMemoryError is thrown, then users can use `bazel shutdown` to
manually restart the worker process. Or, if this happens on CI, then
it is not a problem because we use --worker_quit_after_build.

Bug: not filed; noticed while debugging an internal Kotlinc error
Test: manually tested several failure scenarios
Change-Id: I49dad0df9d2dbe926a99a031ef5ca5299fe71bd0

Former-commit-id: 84683405a935d9b9f54908b8ecf429dcd2cf23e4