Use common storage for Java related options for all variant APIS
The new variant API is using ListProperty or MapProperty for Java
related options for annoation processors. The old variant API uses
List and Map instead.

Having separate storage is an issue as it creates a split brain
situation with old and new variant API resulting in settings present
in the old or new storages only.

Use a common storage and bridge the old API usages to the new
storage.

Test: Added regression test, unit tests.
Bug: 223367107
Change-Id: Iaf43c10252d99b9e731d706c994e3778aeb560ea

Former-commit-id: cc08edcdfa7ecfe2941766d195aa5c459e499f3a