Add src/doc/sample Jars to the Library model
Currently we request src/doc/sample Jars in a seperate model and perform
caching of this information on the IDE side. The current implementation
has several limiting details.
  1. The cache is cleared on the IDE side
  2. We need to map the library to the correct src/doc/smaple Jars
  3. This will not work once the key to do the mapping in 2. becomes
     more complicated.

Gradle has introduced new APIs in order for this caching to be done on
the Gradle side, this will allow us to avoid having to perform the
mapping in the IDE.

This change adds the Jar files as extra fields to the v2 ide models,
more specifically the Library interface. These will be filled in by
using the new Gradle API and read directly by the IDE. For versions
of AGP where the new model is filled in the IDE will stop building
the AdditionalClassifierArtifactsModel and caching the results.

Bug: 197636221
Test: empty api update, nothing to test
Change-Id: Ife8b1d9ebf6d4bc04639bd8e2f86014688a35815

Former-commit-id: 67fc280c4dae72b8fee66b319da827f7171f5bf3