Accept iterable source directories in DefaultAndroidSourceDirectorySet
Previously, setting source directories in
DefaultAndroidSourceDirectorySet would eventually call project.files on
the list of passed objects, which gradle was able to flatten and parse
individually. As we're move to pass the individual files to the variant
sources API, we want to continue supporting the current usecase of
passing iterable objects.

The problem caused androidx integration branch to fail after
introducing java resources in the sources API as androidx used to do
resources.srcDirs(setOf(...)) which worked previously.

Bug: N/A
Test: Manual
Change-Id: I59457577fe935e30c44077b2bea021b135f75891

Former-commit-id: 7c8314d97e401c1c90f6b50a967d1e98470e68fb