adblib: implement channel based shell collector
For shell commands that return arbitrary binary data, having a raw
collector that emits `ByteBuffer` of arbitrary sizes is not convenient.

This change adds a `InputChannelShellCollector` that wraps the output
of a shell command as as `AdbInputChannel` for `stdout` and `stderr`.
The caller can then use these `AdbInputChannel` instances to
consume (as suspending function) the command output as long as the
command is active.

Given `shellCommand` returns a `Flow`, the API is a little awkward to use
with `execute`, so we also introduce a new `executeSingle` operation on
`shellCommand`, which allows processing the output of a `ShellOutput`
as a single value, keeping that value alive and active as long as the
underlying shell command is active.

This change also introduces a general purpose AdbPipedInputChannel
interface and implementation, which allows writing arbitrary binary
data to it (similar to `java.io.PipedInputStream`).

Bug: 188067646
Test: Included
Change-Id: I49d156bb3ff707b07af159887abbe05698d10a2d

Former-commit-id: b5fa1c2b68d284ebd400aaeae7e3d4b58fb9450a