AGP API: Check incubating DSL annotations
The new IncubatingApiTest validates that all methods defined in
incubating DSL classes are explicitly also marked as incubating.

There are currently 139 exceptions, which are listed in the test
explicitly to be cleaned up in a following change.

The test also validates that there are no non-incubating methods that
refer to incubating types.

Bug: 242312390
Test: Included
Change-Id: I76f36e24874d059762f1b118d6e647ea924f0025

Former-commit-id: 65d18720488af0084c6c4709c9854f928051fb5d