Stop using old transform APIs for external libs scope.
removed all usages of the transform APIs for the external
libraries scope for CLASSES.

A new internal scope EXTERNAL_LIBS has been added to store
the artifacts that used to be under the old scope in the old
transform pipeline. It is unclear is such a scope will be
necessary once the old transform API is removed but time will
tell.

Test: existing
Bug: 244224699
Change-Id: I3dce410016e2e8072ba3eba8c12959645bf70082

Former-commit-id: edb356d641f9bcb1f92070131929267a2b261d9f