Return the DeviceHandle from claim()
Currently, the DeviceProvisioner is not explicitly told which
DeviceHandle claims a ConnectedDevice; the DeviceHandle is added to the
plugin's devices, where it can be observed indirectly.

To support devices that transition from offline (but connected to adb)
to online, we need to know when a ConnectedDevice that was previously
claimed by an OfflineDeviceHandle has been released, so that we can
re-offer it to other provisioners.

However, since flows are asynchronous, this indirectness makes it
impossible to distinguish the case where a ConnectedDevice has been
claimed by a DeviceHandle but not yet propagated to the flow consumed by
the DeviceProvisioner from the case where the ConnectedDevice has been
released already. This causes the DeviceProvisioner to offer the
ConnectedDevice again, or fail to re-offer the ConnectedDevice.

Instead, we need to provide the DeviceHandle that claimed it; we can
reliably determine from the DeviceHandle when it no longer holds the
ConnectedDevice and re-offer it at that point to the plugins.

Bug: 241591222
Test: existing
Change-Id: Id924074e65132da8d6deb0db8bb3d794046552c3

Former-commit-id: 6f3b637c7dfdc992bb9cd8d13e35ac1f1e8a00e0