Fix privacy sandbox connected test assertion.
Former method was not working because the waitForIdle call was mistakenly assumed to be synchronous where it isn't. Since the remote view is not updated by the app itself, there really isn't a signal to wait on provided by the platform/testing infra, so I switched to a retried assertion with timeout instead.

Also included:
1. Change uninstall cleanup to not ignore errors and do it only when required. This was swallowing an actual error in the uninstall command.
2. App now propagates any failures to main looper, so that when it throws it's detected by the instrumentation test.

Test: PrivacySandboxSdkConnectedTest
Bug: 237278996
Change-Id: Ic40b52b38a2c7ec5f805edc8b3f8c2aed2859909

Former-commit-id: 82b27cbe35c4b0d92281ac0ae344ba3f66ffefca