Define a forward-looking constant
To help us to remove support for old AGPs without unneeded
inconvenience to users, we need to be able to notify users in advance
of that removal that something will change.  Introduce a new constant
specifying the first version of AGP that will be supported in a
future (presumptively the next) version of Studio, in order that the
IDE can provide suitable deprecation notifications and encouragements
to upgrade.

Bug: 180703127
Test: in accompanying CL
Change-Id: Ic1fdf8b7ea13f5b7ae119999d888f92a0d7ac7da

Former-commit-id: 87074670b509e842274fa1449107629470053c50