Generalize CPU Trace Status Protos
CpuTraceStatusData is now renamed to be
a more general term of "TraceStatusData".
Also, the related protos for cpu trace status
events have been moved out to a "trace.proto"
which is a file of tracing protos meant to be
shared.

This is part of a larger refactor to have
unified trace status events for both cpu
and memory profilers. Upcoming changes will
merge the memory profilers status events/protos
into these new, unified protos/events.

Bug: none
Test: existing
Change-Id: I2bad547a68e17deaa71d79a95796f1e15335b0a6

Former-commit-id: 6dd4e8a6c206048a125431a775459fba9e1d5418