Add Bluetooth LE Lint Check
b/244343735 requested a lint check for the new API for selecting devices to be used for communication purposes. Now developers must call `setCommunicationDevice(AudioDeviceInfo device)`, and later call `clearCommunicationDevice()`. If `setCommunicationDevice` is called, we check to make sure that `clearCommunicationDevice` is called somewhere else (it is not feasible to check the order of these calls, and they may not always be in the same file), and if not issue a warning. Because we need to look everywhere, this issue is not available on-the-fly in the IDE.

Bug: 244343735
Test: CommunicationDeviceDetectorTest
Change-Id: Iaf3b28624ab15279935eb171adafdeb67ce66842

Former-commit-id: 297c88203d0231f6b8b480f4096c1d3c4ed27c69