Fix false negative of type def checking in sources
When users do wildcard import for annotation classes, checking the
full qualified name including the package wouldn't work and
would lead to not processing those source files.

This change fixes it by only checking the class name. As a quick fix,
it is fine to have false positive for now where there are some class
name same as "IntDef" because it would only take a little more time
which is much better than not processing those files.

Bug: 228751486
Test: ExtractAnnotationsTest
Change-Id: I5af3f35f49f52cd66bbf4bf09ed7aa9ed030abf6

Former-commit-id: bd51304590610435c4276ec2ca34d77780f647d5