Add Suppressible Test Mode
This CL adds a new test mode to lint's testing infrastructure which
helps ensure that all lint checks correctly support suppressing issues
via suppress annotations and attributes. There's more information in the
updated test-modes document.

This revealed around 50 tests (out of 3122) which failed. This CL fixes
these as well. Generally, the problems were in the lint checks
themselves (passing in the wrong scope element for example), but there
were a couple of general problems that had to be fixed:

* Suppress annotations on properties suffer from the same problem as
  other annotations with default use sites in Kotlin: the suppress
  annotations only showed up on the private backing field, not the
  getters and setters, so warnings issued on these were not suppressed.

* In XML files, warnings outside the document element were not
  suppressible. This CL adds the same treatment that we have for import
  and package statements in Java files: it will look at the document
  element for these.

* In UAST catch clauses, the uast parent of a type reference is the
  catch clause, not the variable declaration, so suppress annotations on
  the variables were not consulted; this is now special cased.

A few other tweaks here and there.

Test: Many, many tests.
Fixes: 226909879
Change-Id: I6ed5a674bf753bebc28d15bae671c3508130307c

Former-commit-id: ced1745b82aa1259a705b2d569b154fbb89d02f0