Define AgpVersion class
Many (though far from all) uses of GradleVersion are in fact
specialized to computations on versions of the Android Gradle Plugin,
for which we have substantially more constraints: we always have
numeric major, minor and micro versions, and any preview information
is appended after those.

Define a specialized class, for now mutually compatible in its
operations with GradleVersion, to represent versions specifically of
AGP.  In the near future the two classes will be decoupled and no
longer mutually-compatible, once all specifically-AGP versions use the
new class.

Bug: 242691473
Test: in accompanying and next CL
Change-Id: I8ef7105fb45824d92b2a8eca4dab730a9f714062

Former-commit-id: 2d9067a6f77a10aecca9f7d3428dda2c6a615563