Add a task to collect build analyzer issues
Add a task that is responsible for collecting configuration time issues
that should be sent to the build analyzer via the
BuildAttributionService. This makes the support for dynamic warnings
possible, where we can check some values at configuration time and based
on it send warnings.
This change also introduces the concepts of partial results that are
generated for each BuildAttributionService, which means one per
classloader. Each service should read all partial results files
generated by previously closed services and based on the combined
results the final attribution data file would be generated.

Bug: 147093479
Test: incoming
Change-Id: Icc830df8922b2590243001f05d2dad1b05d2399e

Former-commit-id: daacac2ba1248ae4cf71dccc1d7bc77bf2e14c1f