Prevent offline devices from appearing online
The ConnectedDevice of an OfflineDeviceHandle updates asynchronously
from the device provisioner, so when the offline/unauthorized device
becomes online, it's possible for a client to observe this and attempt
to use the OfflineDeviceHandle as an online device. (We want the client
to instead use the proper DeviceHandle that will be momentarily created
by one of the plugins.)

To prevent this, proxy the ConnectedDevice and filter out the ONLINE
state from it.

Bug: 241591222
Test: updated
Change-Id: Id8f7db5a1091f1edaed4a0d336dccaeedbf14d40

Former-commit-id: ca6e67dbce76eb4289eb8711aa00ae29880b8a3c