Remove the UsesMinSdkAttributes lint check
This is one of the original lint checks from 2011, and makes sure that
you specify a minSdkVersion as well as a targetSdkVersion in the
manifest, which users sometimes forgot.

With Gradle, this is no longer the right place to set it; you set these
values in build.gradle via the DSL and then Gradle will inject it at
build time, so we stopped reporting these from Gradle a long time ago.
They often have false positives in other build-system integrations as
well, and do not really provide any real world value anymore, so this CL
just deletes it. (The issue is also added to the deletedIssues property
of the IssueRegistry such that lint will gracefully accept references to
this issue id in lint.xml files, baselines, command line flags, etc.

This lint check was used in a bunch of unit tests (as a simple issue to
trigger), so this CL also replaces these test usages with the
MultipleUsesSdk issue instead.

Test: Existing updated
Bug: Not filed
Change-Id: I7fb8d505322f5f4798bf7cc903bb549d7e619cce

Former-commit-id: 5d4c1b30945f286f69b55ed1c4ba2af33b5dec38