Add power counter tracks to trace processor dameon
The query and respective processing code for power trace data has
been added. In order to test this processing code, a test data file
of 'power.trace' has been added. This file was attained by taking a
perfetto trace with the option Power > Battery drain & power rails
enabled.

The processing method of PopulatePowerCounterTracks was placed under
the counters request handlers as the data table names for the power
data are 'counter' and 'counter_track'.

Access to the request handler has also been added into the trace
processor service, so that it can be accessed on studio-side via
grpc.

Bug: 254116554
Test: added
Change-Id: I85696df0e2ecc3c5b3bfaadcc57ea52a34dd62be

Former-commit-id: 8e968562339940f5266b44f962404ee513159664