adblib: read ahead input channel
Implementation of an `AdbInputChannel` that reads data from another
`AdbInputChannel` in a coroutine that runs concurrently with `read`
operations, hence "reading ahead" data from the input. This is the
similar to a `java.io.BufferedInputStream` except read ahead is done
in a concurrent coroutine (and marking is not supported).

This change also update `AdbPipedInputChannel` to support forwarding
errors (i.e. `Throwable`) from the source stream to the piped stream.

Bug: 188067646
Test: Existing + Included
Change-Id: I38c324e0f1abdb9dde52ecb0571bf870eece04b2

Former-commit-id: 71bef1ddf41178d11a41dcdbba7aa3bc9301e187