Fix support for round screens on Device creation
When creating new devices with round screen, the build properties were
not correctly setup at creation time. The properties would be correctly
created before serializing the data to disk so, when Android Studio
reloaded, the device would have the correct round screen.
This moves the setting of the properties to the correct place so the
Device does not need to be serialzed before hand.

Test: DeviceManagerTest
Fixes: 253422409
Change-Id: I0e9b0908c757b88df3daf263f761e5af3e5d461d

Former-commit-id: 6357843974048b1ca9f50afa9f6dd87ed6672b65