Defer CMake version checking to build time
Calling "cmake --version" is needed to determine the strategy for
using CMake to obtain build metadata. Before this change, CMake
version was determined at configuration time. Since Gradle can't
know that a particular cmake.exe hasn't changed it had to call
"cmake --version" during each configure call just to figure out
if the configuration was up-to-date.

This change moves cmake (and ninja) discover to build time and
out of configuration to avoid this cost.

Bug: 233947045
Test: Adjusted existing

Change-Id: Iccadd2196d79c8d8655154e21f4fe9cfad4df9e7

Former-commit-id: 734e311684c5e963a004472e2f7d4b3467675af9