Fix flaky test
The treeWithEmptyRoots test was checking for 2 ROOT_EVENTS received
at the end of the test. Expected was:
- one from the startFetch command while there is 1 existing root
- one from the screen capture command after all roots were removed

The problem was that the RootsDetector is running on a separate
thread. And the execution from the startFetch command sometimes
didn't execute before the roots were removed. One of the ROOT_EVENTS
were not sent as a result.

The fix is to check for the 1st ROOT_EVENT before removing the
root views.

Fixes: 223378768
Test: ViewLayoutInspectorTest.treeWithEmptyRoots
Change-Id: I330feb1d2ab5e3e5ca97e4eede6e0cec0b90f69e

Former-commit-id: 9b457756acee1413eee6085dccceac95a59b6d0e