Ensure ReverseDaemon sends connections in order
When ReverseDaemon starts a new connection, it kicks off a new thread
that first sends an init message and then starts streaming the
connection data.

The mirroring agent starts two connections one right after the other,
the first for the video stream and the second for the command stream.

If the ReverseDaemon happend to run the thread of the second connection
before that of the first, the connections could be initiated in the
wrong order and thus be swapped on the studio side. This would result in
the studio command reader reading "vp8" (v=118), which is the initial
message sent by the video stream.

This fix makes ReverseDaemon send the initialization messages
synchronously, so they are guaranteed to be sent in order.

Test: none yet
Fixes: 250669299
Change-Id: I1a8941136b143b3365e29f43c67d2749abfaa73b

Former-commit-id: 70456eb7ada08b9b893f565ccf6470c32690948e