Return last seen foreground process
If foreground process detection started polling, then receives
a start tracking command again, instead of ignoring it, it should
send an event with the last seen foreground process.

This is to cover the scenario when one project starts the polling,
then another projects open and connects to the device. But during
the whole time the foreground activity on the device hasn't changed.

Test: ForegroundProcessTrackerTest
Bug: NA
Change-Id: I5981c1d6282daeb12cb4ec4f7dd6d4f4754cc92b

Former-commit-id: c2faaefd2bb0011bd668890bba27bfe3f7305644