Return last seen foreground process
If foreground process detection started polling, then receives
a start tracking command again, instead of ignoring it, it should
send an event with the last seen foreground process.

This is to cover the scenario when one project starts the polling,
then another projects open and connects to the device. But during
the whole time the foreground activity on the device hasn't changed.

Test: ForegroundProcessTrackerTest
Bug: NA
Change-Id: I42a3703d4e3fe449ab49d13cf3960f7c92842605

Former-commit-id: 199d14984810fba98766f6866f67235594629e10