Reduce lint's API database size from 8.5 MB to 0.94 MB
This CL optimizes lint's API database to be significantly smaller, and
also faster to process (not just because we're reading and accessing a
smaller table, but also because we omit a lot of data we can infer.)

There are two separate optimizations which allow this:

First, it's very common for class members (methods and fields) to have
the same metadata as their containing class (the same introduced-in API
level, the same deprecation-in level, etc). Therefore, we can just
*omit* these entries from the database. When we're looking up the method
or field metadata for a given containing class, if we don't find it, we
just return the class metadata instead. This works because the API
database should only be used with valid APIs (in the API checker for
example we're processing type-resolved method calls) and it doesn't need
to be able to distinguish between a valid method and an unknown method.

This reduces the API database (which is read into memory whenever the
API detector is enabled) from 8.5MB to 3.3MB.

The second optimization is to observe that we generally don't need to
know the actual method or field names in the database*; we only need
to know if they're present. That means we can store something smaller.

I tried a number of things:
- Storing only the unique prefixes within each class (e.g. if you have
  the three methods {toString, toXml, reverse), the unit prefixes would
  be toS, toX and r and as soon as we've matched this far we know we
  have a hit.
- Similarly, we can even do camel caps matching, so here we could use
  tS, tX and r.
- I also tried only storing the argument list for members where the
  method is overloaded.

These attempts all yielded some nice reductions -- 20-40%. But by far
the best one is to just switch to hash codes! Now, we compute the hash
code for signature, and store that instead. This makes the whole
database significantly smaller -- down from 3.3Mb to 0.94MB!

(There is the possibility of a hashcode conflict. So we preserve both
code paths; in the event of a hashcode conflict in the class, we use
the old member entries. But this doesn't actually occur for the current
75,000 members in the database.

We also sort the member table by hashcodes, such that we can continue
to use binary search to quickly narrow in on the right one.

*: There were two places where we actually needed the real member names.
ApiLookup had two methods for looking up removed APIs: getRemovedFields
and getRemovedMethods. It turns out getRemovedMethods is completely
unused. And getRemovedFields is only used for one very narrow scenario
(for a single class). So I've special cased this such that this still
works, but only for that class, and I've deprecated these methods.

Test: Existing
Bug: None
Change-Id: Id1646f11e2563b1dd94260d16673c91a61acf469

Former-commit-id: b52626606c5ad4667bf3024fa101a4461fd4ec68