Implement Version prefixInfimum factory
It turns out that we need this to support relatively simple semantic
questions, such as "is this version's leading part numeric?" which is
equivalent to "is this version greater than the prefix infimum of
`0`?".  This question is asked in various places, particularly when
parsing something which may or may not be a version (e.g. it might be
a string interpolation incautiously passed to be parsed as a version).

Uglify the comparison and equality code in Version substantially in
order to handle prefix infima.  Make it less ugly than it could be by
using the trick of computing an effective parts list one bigger than
the parts list of its counterpart, to ensure that explicit sequences
of DEV are correctly handled.

(Multiple-dispatch would make this substantially clearer.  Alas.)

Bug: 242691473
Test: added
Change-Id: If2498f9d32314fb9fc089b50ab91c440563da9bc

Former-commit-id: 1629a3293f87e9b095724a3298edd3de4da980c4