Fix a possible array bounds exception
This is currently unlikely to be triggered from real-world use, because
parseCoordinateString() and parseRevisionNumber() defend against
producing an empty revision list, instead producing a list with a
single 0 component.  However, there is a public constructor which
allows constructing a GradleCoordinate with an empty revision list, so
defend against that possibility in acceptsGreaterRevisions().

Bug: none filed
Test: added
Change-Id: I840e9b9f567d753d13fceeab0e6fe96df0471d62

Former-commit-id: 614e9b74c9c4f878e7cc1e7a086af7fd16728f05