Generalize lint jar verified package access
This CL updates the jar file verifier to check API access in a more
general way; instead of the hardcoded known case of KotlinUClass.ktClass
being protected, it now checks every single API that is called to ensure
that the API is either public, or that it is protected *and* that the
caller is in a subclass of the accessed API.

In most cases, it can trivially check check that an API is protected
because the call is a direct subclass of the accessed API. In some cases
it has to look harder; for this, it uses reflection instead, creating a
class loader for the current lint.jar and then walking up the class
hierarchy to see if it finds the target class.

This sounds expensive, but in practice, it's fine. For the
860 lint.jar files in maven.google.com, bytecode verification took 10s
before this change, and 10s after (around 80ms slower), and the indirect
path is only taken for a handful of libraries.

Test: Unit tests updated
Bug: Not filed
Change-Id: I0c85306b94e5324f1bfd33e972dcb46f001dfc8d

Former-commit-id: 3135583c1ec461d7fd37e5b71b7dae6211955606