Use passed in buffer size for system trace
Both perfetto and atrace were using constants
to define the buffer size, ignoring the passed in
buffer size value. In a past change, the perfetto config
passed in (which includes the buffer size) started to
be used, but atrace still used the hardcoded value. This
change makes the atrace manager use the passed in buffer
size from studio-side (as part of TraceConfiguration).
Because of this change, we no longer need the constants
for perfetto and atrace buffer size.

Also, because the value is read from studio-side, we can
no longer run the test to make sure the atrace/perfetto
config buffer size does not change. These type of tests
will be implemented on studio-side as the hardcoding of
the buffer size will be done there.

Bug: none
Test: removed
Change-Id: I70945421e43be3438cc36b29432d95a75f3e7acc

Former-commit-id: 12e3acb537b9f300ae080d0be8142493cf44c3e3