reintroduced lateAdditions for old variant API support
This reverts commit 8d7a398f06a1d12ed6a31d45b79a09f3550c3693 [formerly e367a0b6b53b0ece7f9c193f1ffc06798dd0b0fc].
which itself was a revert of adding lateAdditions capabilities
to the sourcesets.

The lateAdditions are still useful but non functional without
the other changes introduced in this change.

The directories will be added into an existing DirectoryEntries
(identified from its sourceset name) so it is added with the right
priority.

Using a FileCollection to retrieve the list of directories was not
functional as task dependency was not properly carried. Instead,
use the Provider.zip method to flatter a List<List<>> into a
single list and provide the source directories as a
Provider<List<Directory>> which carries correctly task's
dependencies.

Finally, some Tasks that need to have access to the final list
of source directories at configuration time (which mean before
the old variant API has completed) can now register an action
to the CreationConfig. That action will be ran after the
variant API has completed and can be used to obtain the final
list of sources during configuration time.

Test: Added regression integration test.
Bug: N/A

Change-Id: Ia081409358123fd5a251fcd0801b8c5b7d2376c2

Former-commit-id: 6730ea6a9e0c15c1f85515d07b1878549b116a4b