Add android.lint.api-database-binary-path
In certain setups, we can pre-generate the lint Android SDK API database
and point every build job to it.  Normally, lint generates the database
if needed and reads/writes this database from/to a cache directory
(which is configurable).  This CL adds the system property
`android.lint.api-database-binary-path` such that lint will only read
the binary database file from the given path.  If the file is not found,
or uses the wrong binary format version, lint will fail.

It also adds a private command to generate the database for a given SDK
platform. For example:

lint --XgenerateApiLookup <api-versions.xml> <output-database-file.bin>

An example of <api-versions.xml> is:

$ANDROID_HOME/platforms/android-31/data/api-versions.xml

Test: Unit test included
Bug: 262585154
Change-Id: Iab38d9c8cac31786454449429bf9ac6a56c9528d

Former-commit-id: fbb5eebed69e90393ae123c17f9ad7238d2b5d86