Make SDK version checks handle local variables
The SDK_INT check looks for surrounding and preceding
SDK_INT checks, and it allows some indirection here, where
the SDK_INT check is abstracted into a utility method or
stashed in a field.

This CL improves it such that it also handles scenarios where
the SDK_INT check is stored in a local variable (for example
before being referenced multiple times in a switch statement).

(Also updated the check which decides if a field is constant
from looking for the static modifier to instead looking for
the final modifier.)

Test: Unit test included
Fixes: 262376528
Change-Id: I10aff9832be2040113ceaa9268d6adea6f3d5984

Former-commit-id: b8f2523014144d928839006362649ca8e28c1ec6