Fix API check bugs 263526184, 262851206, 263526227
This CL fixes several bugs related to the API level checks:

263526184: Lint only checks safe casts for directly implemented
	   interfaces, not inherited ones

When lint constructs the API lookup database, it inlines all the
inherited methods and fields from super classes and implemented
interfaces and computes the right API level for each (which is
influenced by where the member came from; if for example a class added
an interface in API level 21, then the methods inherited via that
interface will also require API level 21 even if they were introduced
earlier in that interface).

This inlining of inherited members was however not done for the set of
interfaces implemented by a class. Lint uses this list to make sure
casts (both explicit and implicit) are valid. However, since inherited
interfaces were missing, these weren't checked. As an example, if a
class started implementing AutoCloseable in API level 31, attempting to
cast this class to AutoCloseable on older API levels would trigger a
lint error, but not attempting to cast it to Closeable (which is an
interface inherited via AutoCloseable).

This lint check fixes this. It now inlines all the inherited interfaces
in the database. This had almost no effect on the size of the database
(it grew from 0.943 MB to 0.944 MB), and it doesn't change the schema of
the database so I didn't need to bump its revision number.

262851206: TypedArray#close (API 31) not desugared but AS does not
	   display warning when used in try-with-resources

Lint had support for try-with-resources in Kotlin (using the `use`
extension function), but this needed to be special cased for Java
try-with-resources statements. This CL makes lint go and check each
resource variable's type to make sure it provides a close() method that
it can call. (It isn't necessary for the variable to actually implement
AutoCloseable as it does with the use extension function; I tested this.
The compile dex code doesn't perform a cast, it just dispatches to
close()).

263526227: Lint doesn't check valid casts for call receivers

If you have a call with a receiver, it's possible that even though the
resolved call has an API requirement that isn't met, the receiver
provides an alternative. For example, you may be calling
activity.isDestroyed(), which requires API level 17. However, activity
may implement an interface which provides `isDestroyed` at a lower API
level. Lint goes and looks for this. The problem was that if it found
that method, it would conclude that the API call is safe. But this made
the assumption that a cast to the interface is safe. In this bug, the
problem is that this code finds the `close` method via an interface on
the receiver, AutoCloseable, but that interfaces was added in a later
API level! This CL goes and adds an explicit check to make sure that the
interface providing the potential method is a safe cast, and if not, it
ignores this potential solution.

Finally, there was a test in ApiLookup which was attempting to
comprehensively test that all elements in api-versions.xml can be found
using the binary database and that the since values match up. This test
had been turned off because it was flaky. This CL fixes that (by making
sure it uses the same XML file as the binary database), and extends it
to check removed members as well.

Test: Unit tests included
Fixes: 263526184
Fixes: 262851206
Fixes: 263526227
Change-Id: Iad1841f7de81c143121436e4f32cddcb660f7f63

Former-commit-id: 125598ae3efac36da61631ad2e87a010ee064df5