fixes to recipes following upgrade to Gradle 8.0
After upgrading to 8.0, a task started failing due to missing
input file which was a bug in the recipe : we forgot to save
back the metadata file after transforming APKs.

Another started failing due to the lack of source files to compile,
the task was not successfully executing but issuing a NO_SOURCES
status that threw the test off.

Test: these are tests
Bug: N/A
Change-Id: Ic965eac79b71aa7d04e029ea14abdda4700dbdb0

Former-commit-id: ad4a48fda4c12a5d3be55a3aa47322c7a1aa6efe