[Refactor] Avoid repeated full manifest reparsing during manifest merger
During manifest merger, manifest DOM is modified repeatedly at each step
of merger. When manifest file is loaded, each node(Element, Attributes
etc) of manifest DOM are wrapped into a corresponding XML
object(Document wrapped in XMLDocument, Element wrapped in XMLElement,
Attribute wrapped into XMLAttribute). These wrapper objects contains
additional information that help with merging and conflict resolution.
Because these wrapper object exposes the mutable DOM object, we modify
the DOM object directly making the wrapper objects out of sync with the
corresponding DOM object. In our current code, we arbitrarily reparse
the manifest when we suspect that the DOM is modified to generate the
whole tree of wrapper objects. This poses performance
problem(reparsing aggresively), fragility of code(problems due to state
mutation all across the codebase) and maintainibility of this code. This
change centralizes all the mutations of DOM through the wrapper objects.
Now we can manage mutation and update the wrapper objects accordingly
removing the need for whole reparse. For ex, if an attribute value is
changed in DOM, we just update the wrapper XMLElement and XMLAttribute
without reparsing. The mutable DOM object is still exposed and this
change does not prevent us writing code in future that modify internal
DOM. In future, we want the method that expose internal DOM protected
but there is still many places where this DOM is read. In this CL, we
have addressed all DOM mutations and most of the complicated reads. Rest
of the DOM object reads can be addressed in future work.

Bug: NA
Test: All Existing Tests

Change-Id: Ief39ba53ace11976b4046f656dfa066fbad6396f

Former-commit-id: f838f60f2fdabed5b1cd52aa33f8e957af9bd296