Fix liblog.so copied to APK on Windows
The issue was that CMake seems to not guarantee consistency in slashes even within the same reply.
Normally, libs that come from the NDK are detected as special case and not copied as regular dependencies.
In this case 'sysroot' had forward slashes but liblog.so had back slashes so the lexical check failed.

Bug: 257976107
Test: Added unittest

Change-Id: I560620baa344e3ecd29b7bda4141cd9e0a55f5f1

Former-commit-id: 06ea583b983f14e09d7407f383c811b486a16d24