Introduce global (cross-classloader) BuildService in AGP
Build services in AGP so far are classloader-scoped to avoid
dealing with incompatible types across classloaders.

Since global (cross-classloader) build services are more beneficial
in many use cases, we will provide them in this change.

The key idea is to use an interface/proxy with reflection underneath to
access the target object from a different classloader.

We'll introduce the first global BuildService called
`RunOnceBuildService`, which allows running an action only once per
build, and use it to fix bug 265296706 as an example use case.

In the long run, we can migrate more classloader-scoped
build services to global build services as needed.

Test: Updated GradleVersionCheckTest (integ. test)
      Updated VersionCheckPluginTest (unit test)
      Added BuildServicesTest (unit test)
Fixes: 265296706
Change-Id: Ia604c4c58b1a8a13205604f5085b3a5df1b4deac

Former-commit-id: cefd161b590f8a0533d78b53b34466d648870fd3