Add delta view query to tpd
This change adds an additional query to the
trace processor daemon to show the delta value
between power rail values from time t to t + 1.
This query is unioned with the raw counter values
of the battery tracks as we still want to display
battery counter in a zero-based view.

Bug: 254116554
Test: added
Change-Id: Id3dfb5f2344c2243ccd51e6a3a2a3ee6e1aa292d

Former-commit-id: 2da465811c5d54b03747adc1dd0991e87eba80c5