Provide the workspace location from an env var
tests need to have the location of the Android Studio codebase
provided in order to find prebuilts or other dependencies. We can
always find the location when the test is located within that
codebase. However, for experimental projects that can located
somewhere else, we should be able to provide the location through
an environmental variable.

Test: Existing
Bug: N/A
Change-Id: I56461760229d65c7af29cfa0753991c314e0f1a7

Former-commit-id: b135276a5bd0a68c627383736f2c6d95463cd431