Deduplicate dependencies in deps.txt file
For some runtimeClasspath dependencies (aar and jars) Gradle seems to resolve the dependency twice via different variants. The effect of this is that we end up writing the same dependency twice in the deps.txt file of that dynamic module, which leads to an error when running :app:checkDebugLibraries
Even though the dependency is resolved twice, both resolutions point to the same file. This change does not address the double resolution of deps (the issue may be with Gradle or with how AGP handles AARs in particular) but it dedupes the dependencies before writing them to deps.txt

Bug: 247843123,246529493
Test: na
Change-Id: I0462611ab27df7db4a10586d2914855662ef05c3

Former-commit-id: 793b73b4188350c17f4558594bc272b1cb8aa3e0