adblib: Increase timeout for process discovery
In stress test environments, 2 seconds is sometimes not enough
for the AndroidVM to send the "WAIT" ddms packet, so `adblib`
sometimes miss the fact a process is waiting for a debugger.

This change increases the timeout to 15 seconds, to match the
time Android Studio waits for a process to show as
"Waiting for debugger" after deploying and launching an
application for debugging.

Bug: 188067646
Bug: 265975214
Test: Existing
Change-Id: I92994f9a35ab3f36701dd57dbfa4d70317e56f98

Former-commit-id: 2751319e2d9a89b7c9520d4411541eb52e70cbb2