266116266: No Lint warning about kotlin.text.MatchNamedGroupCollection#get(String)
This CL updates our special handling in the API detector to include a
few methods from the kotlin standard library which have implicit API
requirements because they turn around and call an Android SDK method
with minSdkVersion requirements. The way this is handled in the output
is to list the problematic Android SDK method, and to (in the error
message) indicate that it's called via the given Kotlin stdlib method.

This also required updating the baseline message matching for NewApi
(such that it gracefully handles the new suffix and matches them up
with previous calls). It also fixes the method name identifier lookup
for array access nodes to handle the UI injection mode where it has
to search a little harder for the left bracket; this was exposed by
the test modes for the new unit test.

Fixes: 266116266
Test: Unit test included
Change-Id: I52dcf857744addc6e700c794d9fbe2fb441895a6

Former-commit-id: 2553f154c9148529f7bdad459553ca33fffd38d2