Revamp Android version name computation
The main goal of this change is to bring extension level
support to all the method computing Android name, in order to
remove some confusion where the extension was not displayed.

Unify all methods computing Android name versions inside
AndroidVersionUtils.

There are now 2 methods:
- one focused on API, possibly extended with release/code
  names
- one focused on release name, possibly extended by code name
  and API level

Remove obsolete name computations and updated all code to us
the new methods.

The format of the name also changed a bit, so updated tests
to match the new formats.

Bug: 266121725
Test: Updated existing tests, added new unit tests for new methods
Change-Id: Ibe521b73e1d1f448eab0c9b0bc0315b3ba6dac5b

Former-commit-id: 0d16b6bf232fac66f94b67502d8910dbb451357f