Disable incremental dexing in dynamic feature
To get the dex artifacts of dynamic feature module project dependencies,
we could transform classes into dex in two flows: 1. jar -> dex
2. class-directory -> dex. Due to https://github.com/gradle/gradle/issues/17213,
some artifacts with same componentID and fileName would be deduplicated,
which would lead to missing artifacts if we transform in
class-directory -> dex flow. Therefore, to avoid hitting this issue,
we choose jar -> dex flow for this case.

Bug: 246326007
Test: DynamicFeatureDexingTest
Change-Id: Ia029820623ada6a24b4d74024c962925b0387b53

Former-commit-id: 246a2d52c80a7ff9c921ed9195645b40d0865b01