Use BigInteger to parse numeric parts of Versions
There is no formal upper limit to the numeric part in a Gradle
version, so we have to be as permissive as possible.  (The current
implementation in Gradle represents numeric version parts as Long,
treating numeric versions outside that range as non-numeric; strictly
speaking this implementation deviates from the one in Gradle, though
arguably this one matches the documented behaviour more closely).

The use of major/minor/micro version accessors in Studio implicitly
assumes that they are nullable-Integer-valued (as that corresponds to
the previous definition in GradleVersion).  Return null from these
accessors if the numeric part is outside the Int range.

Bug: 267033059
Test: added
Change-Id: Ib79a6d2bf96580772a976e550c4ae6e92c5cd067

Former-commit-id: c423f2d3d8e5558d82b3b35e675c6eb4b80949c8