Fix startup method tracing stop being ignored
There was an issue where the a startup ART method tracing
was not able to be stopped. This was caused by the
profileable discovery stop command prematurely ending
the startup trace, leaving the startup trace waiting forever
at the filesystem notifier poll() method to return/signal
that the trace file exists.

This fix checks if the app is currently having a capture
ongoing, and if so, prevents the profileable discovery
commands from executing. This is okay as, if the app
is being profiled already, we know it is profileable
already, rendering the check unnecessary.

Fixes: 266854081
Test: updated + added
Change-Id: Ibf84f7371338db41757a2399974a59a43b1003e9

Former-commit-id: 1f9af15e0e8eac20b8ea2eed7b6ff1b0132bab83