adblib: handle DDMS protocol for API >= 28
Starting API 28, a bug (b/267961031) was introduced in the
Android debugger related to sending back "empty" replies
from DDMS commands.

Essentially, before API 28, anytime a DDMS command would
be done executing without an explicit failure and without
an explicit FAIL response, the Android Debugger would send
an empty JDWP reply packet, i.e. a JDWP packet with the
same ID as the DDMS command packet, a length of 11 bytes,
no payload, the "isReply" flag set and an error code of zero.

The behavior was accidentally broken starting API 28, i.e.
replies containing a payload or a `FAIL` chunk are sent by
the debugger, but empty JDWP replies are not sent anymore.

The absence of empty JDWP reply means that the client,
i.e. adblib, does not always know when a DDMS command
is done processing in the Debugger. Some commands return
empty replies or a FAIL chunk in case of error.

The workaround implemented in this change is to wait
a maximum of 2 seconds for a `isReply` packet or a
FAIL chunk after sending a DDMS command. This
timeout is only used for API >= 28.

In the future, let's say the behavior is fixed in API 34,
the workaround behavior still works as long as DDMS
commands return their empty reply/FAIL chunk within
the given 2 seconds timeout.

Furthermore, if the behavior is eventually fixed, `adblib`
can be updated to use the same behavior for API < 28
or API >= 34 (for example).

Bug: 241411829
Bug: 188067646
Test: Follow up CL (see b/266699981)
Change-Id: I72e2af5d829c89d923c79c1e5c825813b3a82d89

Former-commit-id: 5d42d7c0c78619c30b19411c6d254cca38bc4b3d