support for dex layout optimization from baseline profiles
R8 added a feature to optimize startup performance by putting
all necessary classes in the first dex file.

Provide a flag to turn on the feature and feed R8 with the
startup profile.

There can only be one startup profile and it is located in one
of the application baselineProfiles folder under the name
startup-prof.txt. There is no merging, first one is picked so
src/release/baselineProfiles is before src/main/baselineProfiles
for instance.

The resulting APK should have at least 2 dex files (even if
everything would fit in one), with the classes referenced in the
startup profile packaged in the main dex.

Test: Added integration tests.
Bug: 255793809
Change-Id: Id4ad54d23ec8564b3702dd49854c1f2b45c3eee7

Former-commit-id: bac0fbc386f1702218ee566b3a08940aebd38d76