[AS-K2] Add AB target script for K2 build.
Add a `studio_linux_k2.sh` script that will run the same tests as
`studio_linux.sh`, but with the Kotlin K2 plugin enabled. This will be
added to a separate postsubmit target.

Explicitly-configured K2 targets are skipped in this script, because
they are copies of other targets that will already be run with the
globally-set K2 flag.

Bug: 268111702
Test: tools/base/bazel/studio_linux_k2.sh
Change-Id: Icb4f9d8b7fafcfbf73ea9dbf763ca316038caf1a

Former-commit-id: f8525d969437caf89932150dfa542aed3758c2ab