Fix race condition in device state latch
Race condition: If the device state changes right before the waiter
lock in waitForOnline() is locked, the state change unlocks all the
waiters. Waiter lock is then locked which results in a lock that
should not have happened.
Fix: Replace DeviceStateLatch with a stateflow that acts as a latch
Value for this state stateflow is updated when the state changes.

Bug: N/A
Test: Manual
Change-Id: I8b2e24b6691779f3834b6f016a9fd17550f54bfd

Former-commit-id: 8dfc339f6dce0e7b3afd2dbae6b76340e74ecf04