Add build setting for studio release flag
Using --define is a legacy method for declaring custom flags. One
benefit of user-defined build settings is they can be be overridden
instead of accumulating by default like --define flags.

For example, the following can be used to disable the flag even though
it is enabled by the bazelrc release config:

  bazel build --config=release --no//tools/base/bazel:studio-release
  //...

https://bazel.build/extending/config#user-defined-build-settings

Bug: n/a
Test: n/a
Change-Id: Iad1e4d81f068744463b893f27a318329eedd1a75

Former-commit-id: a9024fe72de909f10eed58f3fda7f24812bdf506