Fix merging of child nodes with same type and key in manifest merger
During manifest merger, when we merge two XML elements, children of the
low priority node that does not exist in high priority node are simply added to
high priority node one by one. We introduce a bug when we started
keeping our in memory nodes always up to date with underlying DOM node.
Once we add a child node, if the next child node has an identical type
and key with the previously added node, we would detect this as
a merge conflict and warn user to add the merge rule marker. Fix for
this problem is to identify all the matching child nodes before we
start adding children to the higher priority node.

Bug: 270201653
Test: ManifestMerger2Test
Change-Id: Ib44c1c06e8a4eeb0cd3b461e49046574b622bdb7

Former-commit-id: d88bdcf235792ed24b403e941f074e4564975017