Zipflinger: Improve inflate performance
Problem: Inflating is almost guaranteed to result in a bigger byte
array than its source. We create the NoCopyByteArrayOutputStream with an
initial buf size equals to the src size which is likely to result in an
extra allocation.

Solution: We know the resulting inflated size ahead of time so we create
the NoCopyByteArrayOutputStream using that value.

Bug: NA
Test: NA
Change-Id: I9ecbc000796285af3e57db39b41a72176eddaeaf

Former-commit-id: 0de7cf8708185d7a796b74562361130c1da2aa43