[toml] Set up key offsets for tables
Most of the use of key offsets is I think for key / value pairs, where
the (literal) key precedes the value.  For non-inline tables, the key
is defined in a header which is treated as part of the value (the
value beginning at the open bracket introducing the header).

Initialize the keyStart / keyEnd fields when the table has a
key (which should be always, except for parse errors).  When
generating a location for a table (or in general, but this
particularly addresses tables) use the minimum of the keyStart and
start offsets, provided keyStart is initialized.

Bug: 271454831
Test: added
Change-Id: I04a94cbec93456cca324720ec1ef820dce42ad1e

Former-commit-id: 105e0745c035bae132913e09a6f76d1f6f49bf29