Support deeplink URI hosts starting with *
Previously, deeplink URI hosts could begin with `.*`, but not a plain
`*`. For example, <deepLink app:uri=".*.example.com/path" /> was
supported, but <deepLink app:uri="*.example.com/path" /> was not
supported. The latter would silently result in no host being specified
in the corresponding intent-filter in the merged manifest.

With this change, both syntaxes are supported equally.

Fixes: 268128036
Test: DeepLinkTest
Change-Id: Iff015d7e2cf2f247e3c4e5f6cb9955713ad9d5d2

Former-commit-id: b82d95f5d4c8564d01c7fdc31010cd1d9e13318b