[FTL GMD] Make StaticTestData Serializable
Make the StaticTestData class serializable, so that
ManagedDeviceTestTask can be run in parallel.

AndroidVersion, and the apkFinding and bundleFinding functors are
causing StaticTestData to be unable to be serialized as a parameter to a
gradle worker. This fix focuses on trying to make those components
Serializable so that the StaticTestData class is truly static.

Bug: 249817356
Test: Updated Existing, and the TestLabPlugin
Change-Id: I99a41763d4ba335f2b09a680045e7ee3b42f4fcf

Former-commit-id: b8d70cf972cf9cd0607d09537d9a2d9ee75b1457