Use new VersionRange class in GradleCoordinate
Make getVersionRange() return a VersionRange, not a
GradleVersionRange.  The benefit of this is that we can manifestly
handle the cases of revisions that GradleCoordinate supports: a
specific revision, or a prefix match.  They are interpreted
differently:

- a specific revision implies "this version or anything greater", and
  we compute a version range based on our understanding of the
  compatibility between different versions of the specific artifact in
  question;

- a prefix match (if the coordinate acceptsGreaterRevisions()) already
  has an upper bound, and we can neither impose additional
  restrictions on the implied range nor relax it.

On the other hand, disable a single part of a test that implicitly
asserts that GradleCoordinate can parse a maven version range.  Before
this CL it could, and you could get a versionRange back out of it, but
none of the other GradleCoordinate methods would return anything
sensible.

For the convenience of the one user of GradleCoordinate version
ranges, install a couple of convenience fields in VersionRange (to be
removed in the next CL).

Bug: 242691473
Test: added
Change-Id: I0c1431598bb4dcfbe47bfa170c84f0942a86cd23

Former-commit-id: 3cfa153daaf620da9039df734a2162f112db58b1