New Component class
This class represents a specific version of a Module, for use in
describing specific releases.  It is not suitable for expressing
generic Dependencies, which have potentially both more specificity (on
the kind of artifact required) and less (a version range or a Rich
Version, rather than just one specific version).

The question that entities of this class may be able to answer is:
"can this entity potentially satisfy a given expressed dependency?"

Bug: 242691473
Test: added
Change-Id: I0d7c111bd67b9ca556e1377a8d071b5940f9da6f

Former-commit-id: 4d1edffe35981a307f60bf57cce83277ccff2f49