adblib: use common factory for JdwpProcessImpl
`JdwpProcessTracker and `AppProcessTracker` now use a device
cache entry to store `JdwpProcessImpl` instances, so that
both tracker can be used concurrently (for a given device) without
competing for a `JdwpSession` to a given process on the given device.

A reference counting mechanism is used to ensure both
trackers can shared `JdwpProcessImpl` instances.

Bug: 188067646
Test: Included
Change-Id: Ia5bb227b9ed9fcafbd0dcbb2dfc8f167ad57574a

Former-commit-id: 00a9c0b6e999adb1b7f9ee7f98ee75ee81816cb3