New RichVersion class
This models the parsing and representation of Gradle's Rich Version
concept.  We support `contains` as an operation to compute whether a
version is strictly within the range expressed by the RichVersion, and
`accepts` to compute whether a version would be considered acceptable:
in each case handling the set of excluded versions, while allowing the
difference in semantics between a strict declaration and a required
one.

We also provide convenience functions for detecting and extracting an
explicit singleton version, if any, from a RichVersion; this is useful
for various IDE manipulations of Gradle build files, where singleton
versions are treated somewhat specially because of their natural
string-based representation.

Bug: 242691473
Test: added
Change-Id: I2bc5481e0638678920bb25c0c501eb1ddb2ce616

Former-commit-id: 88b33f3fba897de1fd372cea1eed22cafab32b76