Expose DDM STAG status in adblib
Make use of app "stage" which indicates where the app is in its boot
process. Stage can be received as part of a HELO reply and also
through a STAG packet is emitted spontaneously by the Android VM
on each stage transition.

Update adb property collection logic to use the app boot stage
when it is available instead of relying on a WAIT package.

For more info see go/ddm-stag

Bug: 274973991
Test: unit tests
Change-Id: I5427b98f1cf1d7437905fa67e1481367882c453d

Former-commit-id: d53975f43f1a582d4be22c4d6a96651953fb1299