Always include kotlin plugins in tools srcs
This prevents different persistent workers from being destroyed and
re-created when tool srcs differ. However, both kotlin_library and
iml_module rules need to have the _compose_plugin and _jvm_abi_gen
attributes. It would be nice to decouple this in the future.

Testing this locally with --worker_verbose --worker_max_instances=1 to
verify the warning message is no longer present when switching between
different kotlin actions. The kotlinc action for the following target,
//tools/adt/idea/adt-ui-compose:intellij.android.adt.ui.compose
went from 8.1 seconds to 1.7 seconds when the worker was re-used.

Bug: 275414761
Test: multiple local bazel invocations
Change-Id: Ie426094a28275e1e2adaa228a023549bafdcacc9

Former-commit-id: b7e3e9a02eee2034eb2d3e61936c91156ffbdeb3