Refactor BUILD File
Replace the `iml_module` with handcrafted rules. This makes it possible
to separate the various components into different targets.

Most importantly, the test classes can be built and bundled into a
single jar file. This will allow us later on the compile it in different
ways. For example, with a different Kotlin plugin, or with Dex etc.

Bug: NA
Test: NA
Change-Id: Iff7fb3303ab66a8a9a088db3480821fdc588dcd2

Former-commit-id: 9f7fb6fbaeaf6ef3f9aa34cf573b4f54145d9f42