change newly added API to include dimension.
Last week, a new API in VariantSelector was added to only use the
flavor name to select variants. This worked fine because today we
do not allow to reuse the same flavor name in different dimension
since they are part of the same NamedDomainObjectContainer.

However in the future, this may change so we should not make such
assumptions. Added the dimension parameter so Groovy users are not
forced into using a Pair<> which is clumsy.

Test: unit test.
Bug: N/A
Change-Id: I17be7fe66556aafceb804a25097a3d334a52b250

Former-commit-id: b93ed4b0d48a65c462c1652a3cd6dc243eeb09fe