Fix Jacoco agent detection for coverage runs of Gradle integration tests
This check is used to attach the Jacoco agent to the integration test's
spawned JVM. It was probably broken when the coverage system switched to
using a vendored Jacoco, but wasn't noticed.

Added a note explaining what the check is looking for so hopefully it's
not missed again, or is at least easier to notice later.

Kudos to hummer@ for noticing this!

Fixes: 279980534
Test: N/A
Change-Id: Ie7e352d69cb9afc7ed0eabb6bbb5149e0baa99e0

Former-commit-id: 1c32f7fa34cec13350b3ff70bef25e57c81078bb