Restore origin comparison for implicit property annotation
When propagating annotation from property to accessors implicitly
(e.g., @property use-site or default use-site, instead of @get/@set),
the owner of the annotation (either LC element or source PSI of
KtProperty) and the annotated element of interest (either accessor
in terms of UElement or its java PSI) will never match, but we don't
want to determine the annotation as inherited.

PsiManager#areElementsEquivalent, which seems general enough, eventually
calls KtLightMemberImpl#isEquivalentTo, which in turn compares `this`
origin and `that`'s "member" origin, resulting in that property source
and its accessors are equivalent. Generally speaking, this is undefined
behavior, but technically incorrect, and removing all of previous origin
comparison made the isInherited logic rely on that wrong ULC logic.

In SLC (K2 UAST), `isEquivalentTo` for accessor clearly filters out this
comparison. This commit reintroduces member/origin comparison logic only
for implicit propagation of annotation on property.

Bug: 278573413
Bug: 271032477
Test: RestrictToDetectorTest w/o and w/ K2 UAST
Change-Id: I914ba1142005624cef49abd74fb890c923f3155e

Former-commit-id: 5760345e62d5c3a7d7adcf012f067c93f8882fc1