Add ColdBootAction and remove ActivationParams
For uniformity, define cold boot as its own action rather than having a
parameter to ActivationAction. The existing ActivationAction is
incomplete, as it is lacking a way to indicate whether a particular
activation mode is supported. Cold Boot would also have a different
presentation, since its name is different. Defining it as a separate
(optional) action accomplishes this, and is consistent with the way we
generally indicate availability of actions.

Bug: 276974127
Test: added
Change-Id: I89afe511efe237329b7cfde5840f2a10870b9040

Former-commit-id: 3e5c258f628ec54698c90836678f2ca5958f4a56