Capture the origin info for downloaded artifacts
This change adds the basic infrastructure for recording origin
information for downloaded Maven artifacts.

With this change, when the repository generator tool downloads a
Maven artifact, it also writes an `origin.json` file in the same
artifact folder. This file contains the repository information
for each file in the artifact (e.g., pom, jar).

Bug: 280336914
Test: Updated LocalMavenRepositoryTest
Change-Id: I93a24825b4857aeeffacb42bede4e441673098e1

Former-commit-id: 399fd51c2f59911f65f5e2551bb395f1bf55a70f