Introduce ExtraFileUploadTask to testlab
Added the DeviceExtraFilesUploadTask class to upload and validate the
shared files specified by the ExtraDeviceFiles DSL parameter.

This task pulls double duty, 1. to have a single upload location for the
extra files, so they don't need to be re-validated every test. and 2. to
have cacheable validation on the content of all extra files, whether
they are local files or google storage references.

Test: Added new Unit Tests
Bug: 276517167

Change-Id: Ia4443b48754698f79e1ff2917116e0cb3409b336

Former-commit-id: df624e96314304bdc50de15823ec50bd53707183