Add core library desugaring info to aar metadata
CoreLibraryDesugaringEnabled and desugarJdkLib artifact name are added
to the aar metadata. The build would fail if consumer project doesn't
enable core library desugaring when one of its dependencies requires to
do so. However, we don't check desugarJdkLib artifact variant(three
variants) and version in CheckAarMetadataTask. Instead, we choose the
correct variant and version of the desugarJdkLib for library desugaring
based on aar metadata in library desugaring related tasks which is
handled in a following CL.

Bug: 203113147
Test: AarMetadataTaskTest, CheckAarMetadataTaskTest, etc
Change-Id: I9203bd7919b7b55cd3b46ea5465ecc93f55771e3

Former-commit-id: 87769208d55a09803aa6265673c6a23cf12a9956