[lint] Filter commons-io spuriously-high version from Gradle cache
Correct implementations of repository queries should never return this
spuriously-high version, as there should be metadata indicating that
it is excluded or otherwise lower than other versions.  However, if a
user has incautiously accepted a suggestion to "upgrade" to this
commons-io version in the past, information about it might still be
present in their Gradle cache, which does not have sufficient
easy-accessible metadata or provenance information.  We therefore
filter out this version specifically from the query to the Gradle
cache.

Test: manual, existing
Bug: 218605730
Change-Id: I9e76cc3c9f6512db13002078fa49dfee8ae4627b

Former-commit-id: 4b8d82d41a58e201cdf5eea81653c7074adc71ad