Always add the task provider in variant api operations output paths
Not having the task provider in the output path of the `toAppend`
operation meant that all of the directory will be an input to the next
operation. Meanwhile, `toTransform` operation will also output to the
same directory that is used as an input, which creates a case where the
output of a task is an input to the same task on the next run, and also
incorrect clean runs since some files will be processed as an input
after they were just been created.

This caused an issue with plugins that use the transform APIs (like
Hilt) and plugins that use the toAppend variant API.

Fixes: 285170632
Test: AsmTransformApiInstrumentationTest.testInteractionWithVariantApi
Change-Id: Iee3a81b66d2057f2d3c821e3e6a02e252a5b9600

Former-commit-id: 3772afc298c2381fc2f15dd15054304e24c07bb2