Remove check for projects using legacy support libraries
Previously, if a project still used legacy support libraries but the
user set `android.useAndroidX=true`, we would ask them to migrate to
AndroidX.

This check was useful when only a small number of users were aware of
the transition to AndroidX.

However, according to our stats, currently 97.98% of users have
already migrated to AndroidX, so this check is no longer necessary.

Removing this check now to improve build time.

Test: Removed the corresponding tests in AndroidXJetifierMatrixTest
Bug: 278602339
Change-Id: I65bcd1744fe876fe52fa2f0559b6e2cf7a988611

Former-commit-id: 0b4df94afa088f20344b793193a78e5822296fc0