[Toml,lint] name libraries in quickfixes more cautiously
If generating a name incorporating version information, filter out any
potential separator characters, and ensure that use of the filtered
version has an alphabetic prefix to allow it to be used as an
identifier in build Dsl languages.

Also perform this sanitization on the generated numeric id suffix in
case of collisions with all other possible names.

Test: added, existing tests updated
Fixes: 275605166
Change-Id: I526701acd44564de84dc01f0527ad7167aa5b815

Former-commit-id: db3006871924d3f0dd15d24cbe497cc3e245b382