adblib: Simplify JDWP receivers implementation
The contract for receivers of `SharedJdwpSession` does not quite
match what the `SharedFlow` API offers, because

* Receiver collectors are serialized, i.e. collectors are guaranteed
to run non-concurrently. [SharedFlow] are asyncronous and concurrent.

* Receiver collectors end "normally" when the underyling
`JdwpSession` ends, i.e. when `EOFException` is thrown.
[SharedFlow] never complete.

* Receivers are cancelled when the `SharedJdwpSession` is closed, or
when the session scope is cancelled. [SharedFlow] don't directly
support cancellation and `close`.

The current implementation of `SharedJdwpSession` implements these
requirements using a somewhat complex approach of `StateFlow`,
`Channels`, `CoroutineScope`, custom `Flow` and so on. This
leads to somewhat complex code.

With this change, the implementation has been rewritten on top of
a new general purpose `SerializedSharedFlow` helper class,
simplifying the overal implementation while maintaining the
contract around `Receivers`.

`SerializedSharedFlow` is similar to `SharedFlow` with
a couple of additional features

* The `emit` operation terminates synchronously *after* all
collectors are done processing the emitted value. For context,
`SharedFlow.emit` is asynchronous and does not guarantee
that all collectors are done collecting a value when the `emit`
operation completes.

* A `SerializedSharedFlow` is `AutoCloseable`, with the `close`
method cancelling all active collectors.

Bug: 188067646
Test: Included + Existing
Change-Id: I0838201fb1deb0dc2838d01728489d19bf31a501

Former-commit-id: a9a96513c3dbcfdcdb6b3a24820878a7dde08b69