[lint] Mostly remove GradleCoordinate from GradleDetector
Its primary replacement in this context is Dependency, representing a
declaration.  In some cases we perpetuate the "pun" of a particular
Component to a Dependency, in particular when handling Dependencies
with a textual declaration involving a non-compile-time-constant
string (e.g. variable interpolation), where Lint requests the resolved
component from the project models and then treats that as if it were
the user's depdendency declaration.

Even after accounting for this pun, this implementation is not correct
because the mapping of identifiers to Dependency entities is not
one-to-one: there are multiple ways of writing certain RichVersion
declaration identifiers, for example "[,]" and "+" both mean "any
version", but the entities themselves do not preserve the original
notation.  This means that attempts to perform string replacement of
the version identifier for a new one may fail.

Bug: 257726238
Bug: 259279612
Bug: 242691473
Bug: 279886738
Test: existing tests adjusted
Change-Id: Ie21b0add2075875080d582b1f167ef836d0e42c9

Former-commit-id: 33fa726b70e5ad732e620a001c65197aca99b85d