Always send AppContext and Configuration
This optimization is source of bugs and provides minimal performance
gains.

I tested the average size of an event from Layout Inspector on the
Sunflower app, using the embedded Layout Inspector. First with the
optimization ON, then OFF.

Average event size with optimization ON:  2114 bytes
Average event size with optimization OFF: 2229 bytes

Averages are calculated on ~300 messages in each configuration.

Test: NA
Bug: 287265444
Change-Id: Ib03d656b7dc404f8490e028f3f89ae1b167f63fe

Former-commit-id: 5ccf71cb71a5d2de925c85a27e8ec0f40b03d7aa