adblib: Make `JdwpPacketReceiver.flow` safer to use
`JdwpPacketReceiver.flow()` is somewhat unsafe to use,
as many standard flow operations (such as `buffer` or `take`)
lead to the inability to access the payload of the JDWP
packets.

With this change, the `flow()` method automatically calls
`JdwpPacketView.clone()` before emitting JDWP packets
to the flow, meaning packets can be safely used, as the cost
of having to copy their payload in memory.

Performance sensitive code paths should keep using the
`receive()` method instead.

This change also cleans up unit tests related to the `flow()`
and `receive()` methods.

Bug: 188067646
Test: Included + Existing
Change-Id: Idd4687b947ae2a699fa1cf60f1c040d696398cd1

Former-commit-id: ee2c88197611f16d62218be7e496a311b83e7ced