adblib: fix flakiness in `JdwpPacketReceiver.activation`
In the "activation" coroutine, we use a busy loop waiting for
the receiver to start. The only suspending code in that
loop is the `emit` call to the underlying `SharedFlow`.

That `emit` call is not always suspending, specifically if
there are no active subscribers (see documentation of
`MutableSharedFlow.emit`). To work around this, we
added a `ensureActive` call that will check for cancellation,
which is the only case the busy loop may never terminate.

Tested with:

   bazel/bazel test --runs_per_test=1500 --test_filter=SharedJdwpSessionTest //tools/base/adblib-tools:adblib.tools.tests

Bug: 188067646
Test: Existing
Change-Id: Ia42c8dfa8c8be4aacc55a76a308fd94da95c8187

Former-commit-id: db8db645e393e0c8bf974b26b3554e38b07b1015