[lint] Do not recommend snapshots found in Gradle cache
A snapshot version tends to be used for temporary artifacts, and (at
least in common use) need not refer to the same artifact at all
times.  We already do not suggest snapshots from Google's maven; given
that we don't have provenance information from entities in the Gradle
cache, we should not suggest snapshots found there either.

Bug: 282127516
Test: existing test updated
Change-Id: I8d112a197aa44bd66a004df01b598d1e816de88b

Former-commit-id: 02f6c07de68a844d5a26ec45ef811d1dc33cc094