adblib: Ensure correct behavior on cancellation
Accessing the payload of all implementation of
`JdwpPacketView` should properly support cancellation,
meaning a `withPayload` block executed after a cancelled
`withPayload` block should get access to the whole
payload, not partially read data.

To ensure this, the `releasePayload` is now a regular function
(as opposed to the `suspend` function) so that `releasePayload`
is called when any exception is thrown from a `withPayload`
block.

Bug: 188067646
Test: Included
Change-Id: I33677e19a843c598bebe6a4833804af0a712ee1e

Former-commit-id: 772845a715f4c0e288b6d580b1abb4b5a0998229