Check all possible delegated call to PsiMember.containingClass
When UElement is used as a receiver for containingClass access, without
specific override, it comes from underlying javaPsi of PsiMember type.

However, due to overly complicated interface hierarchy, there are many
ways to reach to PsiMember, e.g., from UMethod, IJ go-to definition
suggests three candidates:
* UMethod <: PsiMethod <: PsiMember
* UMethod <: PsiMethod <: PsiTypeParameterListOwner <: PsiJvmMember
* UMethod <: PsiMethod <: JvmMethod <: ... <: JvmMember

Even though PsiJvmMember and JvmMember are eventually subtypes of
PsiMember, they have their own override (why?!).

Interestingly, PsiJvmMember is a subtype of PsiMember _and_ JvmMember
while JvmMember is a subtype of PsiMember:
* PsiJvmMember <: PsiMember
* PsiJvmMember <: JvmMember
* JvmMember <: PsiMember

Therefore, PsiJvmMember is a least upper bound (in a subtype hierarchy).
The resolution in K2 UAST chose PsiJvmMember, whereas that in K1 UAST
chose PsiMember.

Now, even funnier part is, when LintDetectorDetector aims to find the
usage of PsiMember.containingClass (and warn it w/ the error message
which points to that specific declaration), it uses PsiJvmMember as an
expected container (why?!), which made:
* K1 resolution didn't hit PsiJvmMember.containingClass while
  it's really coming from PsiMember.containingClass (hence false -)
* K2 resolution hit PsiJvmMember.containingClass
  which seems like a false + at a glance, but actually correct

This, though, is somewhat unspecified behavior. The best solution for
Lint is to handle all possible delegated calls so that it can catch both
of K1 (PsiMember.containingClass) and K2 (PsiJvmMember.containingClass)
as a violation of LintImplUseUast.

To that end, this commit:
* changes the detector to use PsiMember (the top-most one) as an
  expected container so that the search goes up to that point;
* changes the detector to look at member in sub class, not exact match;
* and updates newly found violations with access via javaPsi

Test: bazel test \
  $(bazel query 'kind("lint_test rule", //tools/...)')
Bug: 276495493
Bug: 288737678
Bug: 288737874
Bug: 289168726
Change-Id: Ic2048a029e7c0e3f67c250aa4387a16d57870f21

Former-commit-id: 48a740890db0b301607fa4f8d899c391c45b0939