process configuration cache reports and flag new errors.
When running integration tests with configuration cache turned on,
Gradle will produce configuration cache reports that contains
all the violations.

Process those reports and compare the violations to the list of
existing ones to ensure we do not add new ones.
When fixing current violations, one should remove them from the
list so we don't regress in the future.

Test: existing.
Bug: N/A
Change-Id: I70bf08bb2243e0d5f233c184146390ba53591838

Former-commit-id: 408fb5e60ae844a2c098fbe59319a27eb6a2dbe6