New kmp android plugin can consume artifacts of kmp + library plugin
Set the KotlinPlatformType attribute of the incoming configurations to
be `androidJvm` in the new kotlin multiplatform plugin. This allows us
to be able to consume published multiplatform libraries with the android
library plugin.
And since the kotlin plugin has a compatibility rule that equates
`androidJvm` to `jvm` and a disambigution rules that always prefers the
same type, we are able to consume the artifacts of the new multiplatform
plugin.

Bug: 289214845
Test: added tests
Change-Id: I7c814444fae4b0eba1f74230e9ec2223b6851347

Former-commit-id: 02c1e1fe60605206b511c131b88d9f8838b78f52