adblib: Make `JdwpPacketView` thread safe
When a `JdwpPacketView` instance is based on a read-only
in-memory `ByteBuffer`, there is no reason to duplicate the
packet content when initially received from the `JdwpSession`
through the `ShareJdwpSession`.

To allow re-using the same `JdwpPacketView` instance across
layers, this change introduces a `ThreadSafetySupport` interface
that can be queried at run-time to see if an object is thread safe
and immutable so that it can be safely shared across threads and
coroutines.

Bug: 188067646
Test: Included
Change-Id: I50b50b430aa546bc5984b3fffa638b4a9e9aeff3

Former-commit-id: 827969cefb38171a35a08297892dd01d2a80ea98