Added lifecycle tasks to variant API.
The difficulty is that onVariants() block are running before the
Tasks are created so we cannot give access to anchor tasks in the
Variant API directly. Instead, we provide the ability to register
a Task to become a dependent of a lifecycle task once it gets created.

Lifecycle tasks are now provided by the LifecycleTasks interface which
only contains preBuild so far, more later once the API passed the
API council.

Test: Unit test and integration test (gradle-recipe later).
Bug: 280783683
Change-Id: I33ddcc4a6ad1302b6e67123a14eb17933d91877e

Former-commit-id: a866aa546d32077797335c39c0c2fd730c6fe98c