adblib: Improve process close to avoid delay
With change I99fd340e3caa7d51f11794e5c6c26d09acc786a1,
a 100 milliseconds delay was introduced to ensure the EOF
from JDWP connection can be received before the corresponding
JdwpProcess instance is forcibly closed.

With this change, the workaround is improved by making sure
the delay is not fixed, but behaves more like a timeout waiting
for EOF to be received. This is achieved by allowing `JdwpProcess`
implementations track whether a JDWP session is active (using a
ref. counting mechanism) and thus allowing the
 `JdwpProcessTracker` to wait until there are no more active
sessions before forcibly calling `close` on the `JdwpProcess`.

Bug: 291272975
Test: Existing
Change-Id: I14a9e1849a2e21d119d5bff56bae33242c84cb0f

Former-commit-id: b95f7735a6617b0641e642581360a23b2aacc5be