Collect main display size instead of window size
Screen size in the AppContext was a misleading name, since it
actually contained the window size.

This change removes screen size in favor of main display size,
which is the size of the main display.

Embedded Layout Inspector needs to know the screen size in order
to successfully render view bounds.

Test: NA
Bug: 293145625
Change-Id: I62f818ae3d652e2dbe17fa7e0a67957211abc7b2

Former-commit-id: a60a48b46cb39f6d99ca4a24cba61d59c4c80d30