Package RuntimeEnabledSdkTable asset
Play requires assets/RuntimeEnabledSdkTable.xml to be present in the base apk of a non-sandbox enabled device that is using a Play SDK privacy sandbox library.

There will be additional work required to only include this asset for deployment to devices that do not support privacy sandbox.
However, this is a small XML file that is only packaged for privacy sandbox projects deployed locally, so this is not high priority.

Bug: 295172892
Test: PrivacySandboxSdkTest
Change-Id: I9893f5e6ffa8c53d51e313c6a3699c7982af2e4f

Former-commit-id: c36bef8f5353c1f97abde896b9d492271acdcdb6