Fix Manifest file path determination during test config generation
When abi split is enabled, application merged manifest artifact is
generated as ONE_OF_MANY output type. But after cl ag/19853967, we
assumed that there will be single merged manifest with output type
SINGLE. Because of this, when we run tests with
testOptions.includeAndroidResource as true, test config generation
fails. This change simply fetch the first element of merged manifest
output which works regardless of the abi splits enabled or disabled.

Bug: 293547829
Test: GenerateTestConfigTest
Change-Id: I2faa9b0ab98a1b150a04661fcb4be98efc2ebdfc

Former-commit-id: fec9320702408247265cffba24510b1faef7453d