Separate PreviewScreenshotTestTask into two separate tasks for image diffing and updating golden images
Currently the screenshot CLI tools handles both rendering screenshots and image diffing using a record goldens argument. So these tasks are set up the same except for whether or not to provide that flag. These tasks and the CLI tool will be updated in following changes to separate the logic for image diffing and updating golden images.

Bug: 296100736
Test: ScreenshotTest
Change-Id: I871d4011a5b27237b6ee0337d874e772b660855d

Former-commit-id: c43a07d32735356877d378005d40ed7403858cf3