Handle Lint suppression on companion objects and properties
There were a few problems in how LintDriver handled suppression
annotations:

1. There was a special check for annotations on Kotlin properties
because of UAST/PSI nuances, but it was checked in the wrong
place. It's been moved so that it is done at every step of the
AST traversal.

2. We need a new special case for companion objects, because UAST
makes them invisible when traversing the AST upwards (the UAST parent
of a member of a companion object is the class in UAST, not the
companion object). So now we switch over to the PSI checking logic
when the UAST traversal encounters a child of a companion object.

3. There was a redundant code path for handling UAnnotated elements.
This path traversed the UAST tree upwards by repeatedly selecting
the next parent of type UAnnotated. The other code path simply goes
to the next UAST parent regardless of type. There are special cases
in both code paths, but since the type of the first element determines
which path will be taken, only a subset of these special cases apply
each time. We can instead use only the regular UElement code path,
which goes up the tree in the normal fashion without skipping any
elements, and include the special cases from both of the original code
paths here. We can then switch to only looking for UAnnotated's once
we encounter one.

Bug: 293334438
Bug: 296288411
Test: testSuppressCompanionObjectAnnotation, testSuppressPropertyAnnotation
Change-Id: Ie22ffaf45e0442d7a065921acfce1a614c5198e9

Former-commit-id: 84edc46acfaa19bb7ff2811d0fa1770c2e4256cf