Exclude packaged local file dependencies from dynamic feature artifacts
Not excluding local files artifacts that are packaged in the main apk
from dynamic features made classes and java resources be packaged twice
in both the app and the dynamic features where those dependencies
exist.
This issue doesn't appear except when minification is enabled as we do
an explicit check in the R8 task for entries that are packaged in both
the app and dynamic features.

Bug: 295205663
Test: DynamicAppPackageDependenciesTest
Change-Id: I2e019bb812c730e43f00a20dbb83983f7f03a9da

Former-commit-id: e67a1aee167d99b042005ca111d82c364ad2ea58