Add a kotlin extension method to access the kmp android target
When applying the kmp android plugin, the gradle `ExtensionAware`
extension function will not be generated until a successful sync, which
means that users need to know the dsl and specify the namespace and the
compileSdk version before auto complete can happen.
Adding a kotlin extension function makes it possible to have a typed
closure without syncing.

Bug: N/A
Test: N/A
Change-Id: Ide7be67ff962eef39eb9267f41b9a66da327bc62

Former-commit-id: 5f4bd3d45e65c7741a6a7820b4f1d55a7908acf6