Avoid leaking DeviceProvisioner via ConnectedDevice
Don't launch in the device scope, which may outlive the device
provisioner scope. (Even though the job was launched from within the
device provisioner scope, launching with an explicit receiver means that
the scopes are unrelated.) Instead, launch in the device provisioner
scope, but cancel the job if the device scope is canceled first.

(Whether the device scope may outlive the device provisioner scope
depends on how adblib is set up; see Id1cf5b5163)

Bug: 275572248
Test: in change Iba64012d2b
Change-Id: Ic032919527180a67f159908f3d3f2879aad696e8

Former-commit-id: b083ac845d8413937a95e9fc70fe691b0ca3bf31