Add explicit-implicit support to PendingIntentMutableImplicitDetector
This change adds support for cases when intents are changed from
explicit to implicit and vice versa via setComponent() and setPackage().
For an intent to be implicit, both component and package have to be
null, so the detector now tracks intents' state. Previously, all
analysing methods returned a single boolean value representing if an
intent is implicit or not. Now, they all return an object of IntentType
that stores intent's component and package existence.

Bug: 286042003
Test: PendingIntentMutableFlagDetectorTest
Change-Id: Ic0b157d0dcae766da86a4060c3e887f32d6945c9

Former-commit-id: 289b0e88ffa57d26739b627fe1f6703d548b6ac1