Use ide build target to determine main split
When building multiple Apks, build tasks that work with variant outputs
pick the first variant output that match the output type as the main
split. For ex, when processing an app
manifest(ProcessApplicationManifestTask), we pick the first manifest
from the collection of abi specific manifests ignoring the target device
that the user has selected. This change modifies the logic of picking the
main split by considering the target device selected in ide and choose
the variant that matches the best with this selection.

Bug: 296250245
Test: ProcessApplicationManifestWithSplitsTest, VariantPathHelperTest
Change-Id: Ic392607ba84b05c688826c7a850f90d57cbc88ed

Former-commit-id: f9947cc217538531875cd0400a77b3a2ab910cf8