Fix TypeDefDetectorTest.testIntDef
This test was relying on an old API that changed in Android 34 (in
particular, one of the parameters no longer has a typedef annotation).
So now it uses a current one that works for what it is trying to test.

Also, an old "regression test" from 2015 that has been silently ignored
for a long time has been removed. The test was added by ag/836120 to fix
b/37075080 and the code it tests no longer exists.

Bug: 291269705
Test: is a test
Change-Id: Ia21f703c24a220e72d759a4061962e9514a4daf8

Former-commit-id: 36db0af88d4f0c5be3d949dc023b88ff29625ddb