Suggest suppressing inheritors of abstract detectors throwing errors
If an abstract detector's visit method is invoked by the dispatch
visitor and then throws an error before returning control to code
in the implementation detector, the implementation detector won't
appear in the stacktrace, and the error message unhelpfully suggests
disabling the abstract detector and doesn't list any issues.

A natural solution is to look for inheritors of the abstract detector
and suggest suppressing all of their issues.

Bug: 263289356
Test: testNoImplementationDetectorInStacktrace
Change-Id: Ib18cb0c8bbf29a6e82a58abf2eac0e75f617202c

Former-commit-id: 85404fc3b686d9f5f747b4dbc87cff9a5394cde9