Annotation quickfix should skip comments
The quickfix for inserting annotations when running in the IDE is using
IntelliJ's machinery, but from the command line there's a custom
implementation. This CL updates it to skip past any comments (since
frequently the associated element's offset includes the comment, but
users want the annotation inserted between the command and the beginning
of the modifier list), and it also will omit the package declaration if
the annotation has already been implemented -- and finally it adds tests
for this.

Bug: Not filed
Test: Unit tests included
Change-Id: Iced4521f6b0f1611844ce8c3090f22e0237054f7

Former-commit-id: dfe2577ea1274c006a1fb693323c93f1b0e3c0e7