Make plugins responsible for closing handle scope
Currently, the DeviceProvisioner runs a coroutine that monitors the list
of device handles, and whenever one is removed, it cancels its scope.
This is fundamentally unsound, because StateFlows conflate consecutive
updates, and a device could (in theory) be added and removed very
quickly on a plugin's list of handles, such that the DeviceProvisioner
never sees it, and thus the scope would never be cancelled.

Although it's unclear how often this occurs in real usage, it is the
(partial) cause of some test flakiness, so this change moves the
responsibility of cleanup to the place where the scope is created.

Fixes: 297361128
Test: runs_per_test=2000
Change-Id: I5c9bbba41c0919827e761c8bdd6d423747470148

Former-commit-id: 3241b2d8bdedd8ef8ea2ed508878af6eb55a6bbc