Disable failOnWarning when running Compose tests
Compose tests are currently using KGP 1.7.20, which is not fully
compatible with Gradle 8.4 due to
https://youtrack.jetbrains.com/issue/KT-53823.

When run with Gradle 8.4, these tests will fail if failOnWarning is
enabled (the default mode).

To unblock the Gradle 8.4 update, we will disable failOnWarning for
these tests.

Once we update Compose tests to use KGP 1.8.0+, we should re-enable
failOnWarning (tracked at bug 298678053).

Bug: 298008231
Test: ComposeHelloWorldTest,
      KotlinMultiplatformComposeTest,
      ComposeHelloWorldConnectedTest,
      ./gradlew :base:build-system:integration-test:application:automaticTest --tests=*[composeHelloWorld]
Change-Id: Iac7cd7e486b2a5e9df288baf32f6ed6d58f6671e

Former-commit-id: 011b4d8aeaa8b4a8455b549a09d06eb19d81a99c