Create ValueSource for checking if analytics are enabled
This allows for checking whether analytics are enabled in a
configuration cache compatible way.

Previously, checking if analytics were enabled was not compatible with
configuration caching because (1) AnalyticsSettings.initialize() checked
for the existence of the analytics.settings file and (2) AGP checked for
the existence of the same file. This change (1) moves the call to
AnalyticsSettings.initialize() to the obtain() method of the new
AnalyticsEnabledValueSource and (2) removes AGP's check for the
existence of the file.

Fixes: 278767328
Test: ConfigurationCacheReportChecker
Change-Id: I6d7e3c757e6cf6ea854982de1879a8b1ae24673c

Former-commit-id: abebf9d204f8e768ed7b1c35b6673f11882862fe