FakeAdbServer: add `wait-for` command handler
Add support for the `wait-for-<transport>-<state>` host command.

Implementing this new command handlers requires refactoring
how `DeviceState` are handled when invoking host command
handlers.

Basically, `wait-for` is special in the sense this command accepts
a device selector (e.g. "serial xxx") for a device that may not
be currently connected. This implies checking the device selector
for a given host command points to a currently connected device
cannot be done eagerly for all `HostCommandHandler`
implementation. The solution implemented in this change is to
introduce a new `DeviceStateSelector` interface that allows
to perform this verification "on-demand" so it can be delayed
until actually needed by any `HostCommandHandler`.

This change also contains other minor refactoring:

* `HostCommandHandler` now receive a `CoroutineScope`
tied to the lifetime of their `Socket`, so that they can run
asynchronous tasks if needed.

* FakeAdbServer now uses `ListeneableFutures` internally
for it main server thread executor, so it is easier to use
from coroutines (i.e. `await()`) and allows for proper
cancellation.

* Clean up error handling in `ListDevicesCommandHandler`

* Add support for the `tport` device selectors, which is similar
to `transport` but requires FakeAdbServer to write the device
transport ID to the socket before executing the requested
`HostCommandHandler`.

Bug: n/a
Test: Existing (`wait-for` will be used in a follow change)
Change-Id: I47926aef43433f02aba9f202459e4f36e283dfdb

Former-commit-id: 3493e4b9f7b00f3a6a748afc28d7051d09d450f7