Model local.properties as config cache input
When the content of the local.properties file changes, config cache
should be invalidated. The content of an empty local.properties file
should be considered the same as an non-existing local.properties file.

Note we still cannot remove the entry in the "report checker" due to issues
from kgp side. However, we can verify the fix works by removing the
entry and run a test without kgp like CMakeBuildSettingsTest.

Bug: 293321312
Test: CMakeBuildSettingsTest, etc
Change-Id: I1e6177bfd8b56e8bf52f131d667f2d7400cd7699

Former-commit-id: e86703bdf6a2a936cc829b2067fd39e7317f303c