Add a reactive toListenTo API to variant API.
It can be beneficial to have a facility where tasks are automatically
invoked once the final version of an artifact has been built. This can
be used for instance to check that the merged manifest file is correct
without having to explictely call the verification task.

While implementing, I discovered that some artifact can be built without
a Task being involved. This is problematic in the context of this
feature since we cannot rely on Gradle's finalizeBy facility.
Added dummy tasks to produce such artifacts so the Artifacts implementation
can rely on the fact there is always a Task backing up the production of
an artifact. Other solutions will be discussed before merging this.

Test: Added unit tests, and integration tests (recipe to follow).
Bug: 232323922
Change-Id: I8740dbbc8b2572b3e0e5df41a9761defbab70a7c

Former-commit-id: 8bc5d584d1de836119def4067868bf840335ba1b