added a testedApks API on the separate test module variants.
When using the androidTest variant in an application module, you
can request the testedApks but it was not possible to do this since
the tested APK is coming from a different project.
Added a `testedApks` API that provides access to the APKs coming
from the module defined by the `targetProjectPath` DSL entry.

Bug: 270201189
Test: Added integration test.
Change-Id: I00e35702e9894368e93cdeaa1a97024b6137ed8a

Former-commit-id: 8bf465ec3e865ba9bb82a16836ebe94166766587