Unify CLI baseline writing
There are two ways that Lint baselines can be written:

Lint CLI has access to the XmlReporter which can can write the
lint report to a baseline file.

If the LintBaseline itself has writeOnClose=true, then calling
baseline.close() will write to the output baseline file, if specified,
using the info already collected in the LintBaseline.

These two codepaths are similar but do not have feature parity.
For example the XmlReporter will add path variables like
$GRADLE_USER_HOME into applicable location paths, and adds extra
metadata like the contents of the location line and the column number
of the incident.

The intention might have been to use XmlReporter from Lint CLI and
LintBaseline.close() in the IDE, but actually Lint CLI will use
LintBaseline.close() when it is revising an existing baseline, and
XmlReporter when creating a new one. It is not just the newly added
reports that get changed - the entire baseline is rewritten. So the
"better" version of the baseline only exists the first time it is
written, and is replaced with a worse version on all subsequent runs.

This CL makes it so that Lint CLI always uses XmlReporter.
It can be argued that this is a performance degradation because we
don't take advantage of the LintBaseline which has already been
accumulated, but having better baselines and consolidating this
confusing split is a good tradeoff.

Bug: 301290432
Test: LintBaselineTest
Change-Id: Ib1848efacce8278625a4958ca050e6ed3ebd5747

Former-commit-id: 7c57bde67a3587bdcc69fbfdab4b4e3dc9b52bc1