Handle path variables and Gradle cache paths in baselines
If a path in the Gradle cache makes it into a baseline (it shouldn't)
we should handle it by comparing the part of the path inside the
cache

This change is related to ag/24806637 - LintBaseline.write() didn't
handle path variables in location paths the way that
XmlReporter does when it writes to baselines, so that has been fixed.
Right now many regular lint tests end up writing baselines through that
method, but in light of ag/24806637 we need an explicit test case.

Bug: 238892319
Test: testPathsOutsideProject, testHandlePathVariablesInLocationPaths
Change-Id: I5cd30c6d84e396f1d008c9323b32141ee11a5262

Former-commit-id: b24456c8456522ee3e3c1714e7d7420ee21e3866