Fix main split determination when device support multiple abis
When a device support multiple abis, all supported abis are passed as
project option in form of comma separated string joined in order or
preference. We now consider the target device when choosing the main
split and this change improves this logic to also consider preferred abi
when multiple abis are supported. See also: b/24541408

Bug: 296250245
Test: ProcessApplicationManifestWithSplitsTest
Change-Id: Ia1125045972cf499a48b45c38cc0ddcedbe147d7

Former-commit-id: a302967bba1418efc9adebd90e6b9b9ca0bbbe48