Refactor `NotificationTrampolineDetector`
We were passing around boolean `isBroadcastReceiver` and repeatedly
dispatching on it for various things (e.g. what to check, what to
display, etc.), with the implicit understanding that "not broadcast
receiver" means "service".

This change replaces the boolean with a more informative enum that
documents what the notification trampolines are, and reduces duplicate
code.

Bug: n/a
Test: existing
Change-Id: I847abad8b135cf965860cc27034bb0cb64d08a7a

Former-commit-id: 7a54cd420f6724babb9a05a4e5e39546f347c595