Fix annotation fix location
Follow-up to ag/24829320 - we should still use the given element
as an anchor for the quickfix even if it isn't the type of element
that was causing problems before that change. Callers should be able
to specify the PsiElement and have the location taken care of without
their help.

Right now we test annotation quickfixes with individual detector
tests, but this change doesn't affect any of our detectors (the change
to ParcelDetector here is just to fully migrate it to the new annotate()
API). So we may want to add dedicated tests for annotation quickfixes in a follow-up CL. If the AndroidX detector in b/301598518 switches to the
new API then its test can count as a regression test. This change has
been tested using that detector.

Bug: 249043377
Test: manual, and existing tests for annotation quickfixes
Change-Id: I5026bada5e07eec1cd8b18d1455fad02bb8f56da

Former-commit-id: e7d285ca80d37087706f409e339662fad5a0bea7