Lint check for framework FlaggedApi lint checks
This CL adds a new lint check which ensures that all code using APIs
marked with @FlaggedApi(feature-flag) is always inside explicit "is this
flag enabled?" check.

Note that this new lint check is not part of the standard Lint
distribution. However, because it's intended to be used in a number of
integrations which are not using AAR libraries -- such as Android
platform builds, google3, and ASfP, it's part of the lint artifact but
explicitly pulled in using a command line flag.

(It also fixes the handling of $ANDROID_LINT_JARS for global lint checks
-- which was relying on the same mechanism the new include-aosp-checks
mechanism does.)

Test: Unit tests included
Bug: 303434307
Change-Id: I891a04e7cf5019d09465a1ecc2d618136f00ff43

Former-commit-id: 28ced012782b60ddd5965dec5f8726ca324dbfd5