Add end session call to stop task command handlers
Because tasks and sessions have equivalent lifecycles under
the Task-Based UX, when the task is considered to be finished,
the correspnding session should be terminated. Thus, this change
adds the behavior that under the Task-Based UX, if the underlying
recording for a task has been stopped, a call to end the current
session is made as well.

Bug: 277797528
Test: updated + added
Change-Id: Iead8cef58db12eba79ba864c1d8bb4fe33ff0861

Former-commit-id: f5210161a0a593f4ce4a9c0f0ba202ef612042cc