Tolerate removed intermediate CMakeLists.txt
In this bug, CMakeLists.txt refers to another subfolder/CMakeLists.txt when the CMake project was generated. Then the user modifies CMakeLists.txt to no longer refer to subfolder/CMakeLists.txt and deletes subfolder/CMakeLists.txt.
Before this fix, this would result in an error about missing subfolder/CMakeLists.txt the next time the CMake project is generated.
After this fix, this kind of transition won't result in an error.

Bug: 262077903
Test: Added integration test to cover the case.
Change-Id: Ia34b0cfc09f4873c6bf76ed427541846a2005f00

Former-commit-id: 99ac5e60ca79c5a6f577e1e827adbc9a4325ce24