wired Sources.jniLibs to unit test task.
the `variant.unitTest.sources.jniLibs` was completely ignored and
therefore the unit tests could not load a JNI library that is
placed in those directories or added via a Task.

The unit test Task has a systemProperty() method that can be used
to set the `java.library.path` system property for the spawned JVM
running tests. Use the variant jniLibs values to set that property
at the Task execution time.

Bug: 306420288
Test: Added integration test.
Change-Id: I7d0414f74348e12b553f6610102302bc1642c318

Former-commit-id: 195fdd99249e4e1e9a42590a319a6b9459a140c1