Update NewApi message for casts to make clear when they're implicit
Lint's API check will also check "implicit" casts -- for example,
passing a Cursor to a method which takes an argument of type Closeable,
or assigning a cursor to variable of type Closeable.

The error message was talking about "casts" which may be confusing.
This CL updates the error message in these cases to instead call it an
*implicit* cast.

Test: Unit tests included
Bug: None
Change-Id: Ib80b0a4fd8ca757838a9c209b0e803e369bb6278

Former-commit-id: 44fb599f1329679fae0b55a23b66416874100ff0