Improve Nested resource detected warning
Validate happened in packageDebugResources
and mergeDebugResources but we wanted to see
only one message. So validation was moved to
GenerateResourcesTask task as it used to be
anchor task that both package and merge depends on.

Algorithm is fairly naive in favor of simplicity
Trie model (with folder names) can be created to
make it faster for big number of resources.

Test: NestedResourceFoldersTest
Fixes: 285041677
Change-Id: I60f7a7d10e97720abd0ee75f54d5136eaab482cc

Former-commit-id: 135f9f7d5716f5826fb7af8a7967eea466e6ed81