Pass minSdkVersionForDexing to the IDE through output-metadata.json file
In order for LiveEdit to work correctly, it needs to know the minimum
API level that `.dex` files in the APK can support.

To provide this info to the IDE, AGP will now store
minSdkVersionForDexing in the output-metadata.json file next to the APK
in the build directory.

Note that minSdkVerionForDexing may be different from the minSdkVersion
specified in the DSL/manifest (as AGP aims to improve build performance
when the IDE is deploying to a device where the API level is known).

Test: Updated ProcessMultiApkApplicationManifestTest
      + GenericBuiltArtifactsTypeAdapterTest
      + a few others
Bug: 284201412
Change-Id: I57c309651fd7dec9bdcdbc0f099d4c75eadce638

Former-commit-id: 69cda6890852309a96b961218be014f8c3a772e9