Revert "Only deploy SDK runtime SDK apks on Android U+"
Revert submission 25017675-android-t-sandbox-compat

Reason for revert: When deployed by studio, the manifest is generated before the device spec is generated so the user-sdk-library is always omitted so SDKs aren't loaded in 34. So to unbreak native sdk runtime support we'll need to include the tag with the downside of needing to install standalone sdk apks on 33 devices.

Reverted changes: /q/submissionid:25017675-android-t-sandbox-compat

Change-Id: I4fbbeeee35e191cbf965ed4b7b97f0eff363f510

Former-commit-id: 6f1a1769fdeda4307db7df1f623fca82f85d34d4