Make physical devices go away after disconnection
Since we have DeviceId now, rather than attempting to preserve the
object identity across connections, allow the disconnected devices to go
away. Upon reconnection, we have a new DeviceHandle with the same ID as
the previous connection.

Since devicesBySerial is updated in multiple contexts now, introduce a
mutex to synchronize changes to it.

Bug: 293436517
Test: Updated
Change-Id: I8ccefea5e246e8c9d3fc4a6528f3eeca2a977b5c

Former-commit-id: 3b0072532b441b64c9240b00d260dda1133d31a5