Pass the current targetSdkVersion from Lint to Studio
When these checks run inside Studio, use the LintFix to pass the
detected current targetSdkVersion to the IDE. This enables Studio to
determine whether the "open assistant" quickfix shold be available
depending on whether the current targetSdkVersion is supported by the
assistant.

Outside Studio, the previous fix behavior is preserved.

Bug: 297422963
Test: N/A
Change-Id: I9b3d99a6ee4933ab59cd3331cbfd093a77a093e4

Former-commit-id: 194a5b1036d4d8e98c8558137055eaaa2299267e