Override minSdk of SDKs as 33/T
Sets new BuildSdkApksCommand.setMinSdkVersion to allow SDKs to be be installed to 33/T devices.
This is a temporary workaround to allow for Privacy Sandbox deployment to 33/T devices that require
standalone SDKs and backwards compat split SDKs. This change should be reverted once a better alternative
has been found, such as deploying a split with <uses-sdk-library> that is only deployed to 34/U devices

Bug: 307903093
Test: Manual
Change-Id: I71a54632f1eea9e4973244977b151c3410620654

Former-commit-id: 4a91e790e32879109c4bd07aa6c97afe21f6b684