Simplify Headers
In the existing implementation, we have a structured headers object
(Map<String, List,String>>) but we serialize it into a string only to
deserialize it on the Studio side.

Instead, we can just pass the structured data in a proto. We can't use a
Map, but we can pass a list of entries instead.

We also change the way we handle the response code. Rather than
extracting it from the headers, we pass it explicitly.

Bug: NA
Test: NA
Change-Id: Ie6d4aec7ed90d5e45a6ee1c43f6ea2d3e36cc8c4

Former-commit-id: b1bfed50278649fc0fda566dcd3bc33a086506a3