Overhaul the quickfix machinery
This CL updates the lint fix machinery significantly.

Lint quickfixes are typically *described* by detectors, and then
performed by the lint client. There were two completely separate
implementations of this:

- The IDE, were originally all quickfixes were performed. The quickfix
  machinery in the IDE can leverage important facilities like import
  management and code formatting of modified regions.

- The command line, via the LintFixPerformer. This is used both for the
  `lintFix` Gradle task, as well as the `LintFixVerifier` which is used
  in lint's unit tests.

Unfortunately, there was a significant discrepancy between these, which
meant that your unit tests may pass yet in the IDE the fix doesn't work,
or vice versa. I ran into exactly this scenario when I recently added
more complex quickfixes for a new detector (not yet submitted), and
while all the unit tests were passing, when I decided to try it in the
IDE I discovered they didn't work at all. This CL contains a unit test
for that exact scenario: ReplaceStringQuickFixTest#testCompositeString.

The root problem is that in the IDE, when you have a composite fix (e.g.
multiple nearby edits), we were just running the quickfixes in
succession. But here, the first quickfix would trigger code cleanup,
which would invalidate the PSI elements, and when we get to the next fix
in roughly the same part of the file, the edits fail.

We need to compute all the edits up front, then sort them and then apply
them in the right order, and *then* perform a single pass of import
cleanup, formatting and selecting text at the end. This was all already
done by the CLI fix performer.

As I started fixing this, it snowballed into a larger set of fixes as I
addressed various other related changes. This unfortunately makes the CL
pretty big, but attempting to untangle this into separate CLs would be
difficult.

Here are the high level changes:

- Use the same fix basic fix engine from both side. To do this, I moved
  LintFixPerformer up into lint-api, in the process moving the
  CLI-specific changes (such as inserting selection markers for unit
  tests etc) into a subclass LintCliFixPerformer, and a new
  implementation on the IDE side, LintIdeFixPerformer. The big
  difference between these two is that on the IDE side, we operate on
  PSI files and documents.

  This fixes the root issue that kicked this off -- we can safely handle
  multi-edit operations in the IDE.

- On the IDE side, we also hook into the IDE's facilities such as import
  management and reference shortening. There was a basic implementation
  of this in the CLI fix performer, only performed from unit tests. This
  CL improves this quite a bit, such that imports are inserted in order
  (rather than always prepended right below the package statement), and
  some basic reference shortening. This makes unit tests more accurately
  reflect what will happen in the IDE. There is also a mechanism to turn
  this behavior on, since this has been requested from people performing
  large-scale refactorings etc. Similarly, I improved the XML attribute
  handling in the fix performer to insert attributes in the
  Android-canonical order instead of always just prepending to the
  element attribute list. And the annotation-fix has been fixed to
  correctly handle the replace=true attribute (this was only implemented
  in the IDE side fix.)

- The string-replacement quickfix (the most powerful and most commonly
  used quickfix descriptor) now supports having a fix be applied
  repeatedly (e.g. "/g" in sed, or "*" in wildcards), as well as
  whether at least one match is required (e.g. "?" from wildcards).
  This is done with the new `repeatedly` and `optional` attributes on
  the replace fix descriptor. (This was motivated by rewriting some of
  the IDE quickfixes -- for example, the quickfix to rename an XML tag
  relies on also matching the closing tag, which may or may not be
  present.)

- Revised selection handling. Selection had unclear semantics and was
  handled inconsistently between attributes and text replacements; it
  was also set per file instead of per edit which made it unclear what
  should happen in a multi-edit situations.

- The quickfix support for creating new text files (for example used by
  the backup descriptor lint check) now reuses the same machinery as
  editing files, such that it supports the same selection, import
  management and formatting machinery.

- There is now a flag, `--X-include-imports`, to opt lint into using the
  import cleanup machinery from the command line when running large
  scale refactorings, as requested in b/263058327.

On the IDE side:
- Properly handling multiple edits as a single unit
- Improved handling of formatting the edited regions
- Various bug fixes to the XML handling code
- Many quickfixes were switched from using a custom fix implementation
  over to using lint fix descriptor on the lint side

Test: Many new tests, many updated tests.
Fixes: 263058327 (and many bugs not filed)
Change-Id: Ic7e86bd6f7ce6e901177caa84b934f1fff7bccaf

Former-commit-id: a6d6a777b7f523b922d38cacf61fe88c0c2e4777