Avoid registering ExtractSdkShimTransform for all variants
Registering ExtractSdkShimTransform can be expensive as it creates multiple detatched configurations. It is
currently possible to set experimental properties on each variant so the transform is registered many times.
This isn't necessary, so this change will now throw an exception when setting different variant values for
the same property, so it can just be registered once.

Bug: N/a
Test: Manual, leads to build performance improvement in local benchmark
Change-Id: I004be51e00fa24364b91664f5bb6655536d412cb

Former-commit-id: 76ef85ad611056a7801b05f8e7be9a805126f8a1