Make maxApi non-nullable in baseline profile installation
After change in profgen to always have a maxApi value, the processing of
this value should be updated so that this value is not nullable anymore,
and also having a profile for only one api (i.e. minApi == maxApi) is
supported. Also, move the baseline profile related integration tests to
a dedicated package for better organization.

Bug: 310971470
Test: existing
Change-Id: I74e121859b4405e1e5d0af96471dae89fb0c5155

Former-commit-id: 04f1efc95f84ca2de8f238fbc0033cef5b05b1dc