[SDK Index] Suggest version that has no issues
... from SDK Index.

We currently add a lint warning saying that there is a newer version of
a library if we find one available, but that particular version can have
SDK Index issues and thus prevent users from publishing their app.

This CL causes the upgrade suggestions to pick the newest version that
does not have errors or warnings in the SDK Index.

Bug: 301295995
Test: GradleDetectorTest, GooglePlaySdkIndexTest
Change-Id: I9f5fffdbb0b2eaf2367ae8e4dbc4d4a45eba0355

Former-commit-id: 3c8dbf3b58c780790e8a56b733de95a9380502f4