[SDK Index] Only show upgrade message if no issues
... related to SDK Index are present in the current version.

We currently show NewerVersionAvailable issues if we find that a neer
version is available and that new version does not have SDK Index
issues, regardless of the current version used, this causes similar two
similar messages to be shown: one that says that the current version has
issues, poiting to the SDK Index (where users can look for a version
without issues) and other saying that a new version is available (with a
quickfix to change it). It is more similar when the SDK Index issue is
because of an outdated version.

In order to reduce the number of shown issues and to not imply that we
suggest to upgrade due to the SDK Index issue, with this change we will
not show the NewerVersionAvailable issue if there are SDK Index issues
present in the current version (we already pick a newer version that has
no issues).

Bug: 301316600
Test: GradleDetectorTest
Change-Id: I510fc5be5b838e87a4fb75514737042319244c5b

Former-commit-id: bf7d1b7c19c72259ed0853ac9f5cbe242105ea40