Fix lint test infrastructure diff computation
If you see the following diff
    @@ -3 +3
    + import androidx.test.filters.SdkSuppress
    - @RequiresApi(29)
    + @SdkSuppress(minSdkVersion=29)

you would conclude that the old source contained an import statement
right above the RequiesApi annotation, right? However, there's a bug
in the diff algorithm where there could be a real line there. If we
switch to window=1 (e.g. "diff -U 1" on the command line) we'd see

    @@ -5 +5
      import android.widget.GridLayout;
    + import androidx.test.filters.SdkSuppress;
      // Comment
    - @RequiresApi(29) // ERROR: don't use in tests, use @SdkSuppress instead
    + @SdkSuppress(minSdkVersion=29) // ERROR: don't use in tests, use @SdkSuppress instead
      public class UnitTestJava {

So the diff with 0 lines of context is misleading. This CL fixes that,
and does it using a compatibility mode -- so if a unit test fails,
it runs it again with the old algorithm and only if that one fails
does it fail the overall test.

Bug: Not filed
Test: Unit tests updated
Change-Id: Ia5b4a067bb53dfe204f710ec033c9548ab8918f2

Former-commit-id: 68276e28a874eaa4f2d9737052d3bf68a051e47a