Change suppressMinSdkVersionError to refer to NDK's minimum platform
Before, the value was to be set to the user's choice of android.minSdk.
After, the value is to be set to the current NDK's minimum platform.

If the value is correctly set, then the error about android.minSdk being too low is not sent and the user's chosen android.minSdk is used instead.

This CL also creates an experimental property that can be used to set suppressMinSdkVersionError from build.gradle. This is to match parity with user's chosen NDK version which is also set per-module rather than per-project.

The Gradle property is also still allowed but is no longer a comma-delimited array. If both the property and the DSL are set, then the latter is used and the former is ignored. An info message will be sent to the log when this happens.

Since experimentalProperties is a Map<String, Any>, I also added an explicit warning when the value cannot be converted to an integer. This should be enough to signal to the user that they can't use codenames like "Tiramisu" here.

Bug: 310718265
Test: Added unittests for the new states
Change-Id: Iaca64afc9a6c9df0084b38b7d2296bdf6beec18c

Former-commit-id: da449327b4f0565ea2c20479314ee26ce3483fde