[lint,.kts] Look through BinaryExpressions to find parent block
The Gradle construct

    plugins {
      id("foo") apply false version "3.14159"
    }

is represented in Kotlin Psi (and UAST) as a left-skewed tree of
binary expression nodes, meaning that the actual call id("foo") is at
the bottom of arbitrarily many BinaryExpression nodes.  Walk up the
BinaryExpression nodes before attempting to find a suitable parent
block name.

This reveals that we were effectively doing the same scans twice in
the UastGradleVisitor, only shielding ourselves from having duplicate
reports by failing to find the correct block name in some cases.  With
the improved finding of parent blocks, we can always handle cases like
this in the method call, so remove the special-case handling in
handleBinaryExpression().

Bug: 279885639
Bug: 297608723
Test: existing test slightly strengthened
Change-Id: I43771f2de200c19c5164abb92fca3efd5631e232

Former-commit-id: 43f9b235e5191e9c298a2acd9f193668dfdb86ed