[GMD] Ensure snapshot is deleted when invalid
In rare cases, snapshot generation succeeds in
creating a snapshot that can pass the "-check-snapshot-loadable" check
but causes a segfault when the emulator is next booted.

Although we detect these snapshots in the SetupTask, the snapshot can be
left behind in the GMD avd folder after the task completes. This can
cause an issue with subsequent calls to the setup task passing,
resulting in obscure errors in the Test Task b/c of the invalid device
state

Test: verified locally (by reducing retries to force the invalid
snapshot to exist)
Bug: 315209934

Change-Id: I5ee89aef29175588aced624e7a82142da91d4f4e

Former-commit-id: d5006b4cff95aecf60b7e692eb8cf9bd24e726fa