Ignore maven_library in coverage baselines
We don't care about maven libraries for the coverage baseline since the
point of the baseline is to grab all of our source code.

This is accomplished by adding an arg to `kotlin_library` about whether
to generate the coverage baseline. It defaults to True to maintain the
default inclusion behavior. `maven_library`'s use of `kotlin_library`
overrides it to False so all `maven_library`s will be excluded by
default.

There was an error b/c of a duplicated classfile pulled in by a
`maven_library`. This change fixes the coverage post-processing again.

Bug: N/A
Test: verified locally
Change-Id: Ied3f6247340de7cff9ebfafac8f5975660260b75

Former-commit-id: 441902d9db31c71192057ae9c2e8f52bdec37e2a