[GMD] prevent fallback to cold boot during validation
Check the snapshot succeeds `emulator -check-snapshot-loadable`
before attempting to validate the emulator snapshot by booting
from it in the Managed Device Setup Task.

Emulator will fall back to cold boot when the snapshot does not
exist. Instead this should fail the validation and retry the
snapshot generation (if retries are available). However with the
fallback snapshot creation is considered successful and fails
a later check.

Bug: 315483338
Test: Verified manually that snapshot generation is retried instead
of failing the task.

Change-Id: I62b1cdd771518984076ef46255b2889f112fc82a

Former-commit-id: 729e4fd5bbe53321501702df88bdb22c19212cef