Fix false positives related to notification manager
This CL fixes two bugs related to the POST_NOTIFICATION bugs
* 271249297: Lint check wants POST_NOTIFICATIONS permission for
  unused class in dependency
* 315440581: Code should not explicitly check to see if permission
  is available when calling areNotificationsEnabled()

The first bug makes sure that if we check the bytecode and find a
bytecode reference which shows that the app is posting notifications,
then we check whether this is coming from ExoPlayer. If it is, we check
the merged manifest, and we only complain if you're also registering a
download service -- which is an optional part of exoplayer responsible
for the notification.

The second bug makes the notification permission check, *and* the
regular permission check, aware of the special areNotificationsEnabled()
methods available on NotificationManager and NotificationManagerCompat,
and use these to suppress warnings about missing POST_NOTIFICATION.

While there I also discovered the code was using the wrong classpath for
AndroidX' notification manager (so it had not correctly flagged
notification usages of AndroidX).

Finally, it fixes a bug where in some cases the location range shown for
a missing permission would include the surrounding call.

Test: Many unit tests included
Fixes: 271249297
Fixes: 315440581
Change-Id: Ibf78248dc5dbca63e6e35a8a499561581ec4a485

Former-commit-id: 2136da4e6efe1fda193303296c5c3bb20526b266