Suppress Unchecked Cast warning
Casting from Provider<Any> to Provider<Directory> is not safe in Kotlin
as there is no way to check the generic parameters at runtime in general
case. Since we are reasonably sure of the type and believe the cast is
safe. So adding a suppress warning to these lines and using as? for
additional layer of validation

Bug: 315835138
Test: Existing tests
Change-Id: Idc8c364aafd04d6df8c8af7543e37ad45ba984e8

Former-commit-id: 7f77d878838465f6d1af5dabb5e7863a9af3bd90