316596193: Allow class references from @RequiresApi
If your code does this:
   if (event instanceof BackEvent)
on an older device, the app will crash with a NoClassDefFoundException.
Therefore, lint enforces that classes referred to in type expressions
(such as instanceof, casts, etc) adhere to the API requirements.

However, users can also annotate their own classes and libraries with
@RequiresApi. This is generally treated identically to the runtime API
requirements. But for class references this is not ideal. Annotations
can only come from classes already present locally, and in these cases,
performing an instanceof check *is* safe.

**Unless**... the class extends (possibly indirectly) a runtime class
with a higher minimumSdkVersion requirement.

This CL updates the API check to consider this scenario. When we check
a class reference from an annotation, allow these unless the referenced
type has a super class with a higher runtime requirement.

Fixes: 316596193
Test: Unit test included
Change-Id: Ifce66c3189301ac68c820430902d378797b8f54e

Former-commit-id: 3668db39b7d9f640aecab61c91f46ee48545f726