Update the property names of KeepRules to be consistent with BP
During API meeting for introducing BaselineProfile to the Optimization
DSL block, we created the names "ignoreFrom" and
"ignoreFromAllExternalDependencies", which are similar in function to
the properties in KeepRules. We decided these names are better than the
current properties of KeepRules, so those properties should also be
updated so that they are consistent.

Bug: 279648472
Test: existing
Change-Id: I378a1f8f08ae19f267ac24c41d2028112efba8f0

Former-commit-id: f7b8c3917cf370e5078f85b73e98eaf04bcd6301