Make DISCONNECTED the final state of ConnectedDevice
This aligns the behavior of adblib with IDevice, which sets a device's
state to DISCONNECTED when the device is no longer present. OFFLINE is
a state that can be reported by adb, whereas DISCONNECTED is a sentinel
value indicating that the device is no longer reported by adb.

This is also helpful for readers of ConnectedDevice's deviceInfoFlow.
Currently, the only way to detect that a ConnectedDevice has become
disconnected is by checking if its scope has been cancelled; this
provides an in-band signal that the flow collection can be aborted.

Bug: n/a
Test: Updated
Change-Id: Iec3d7c0fc852880edc05bd99af2a15a8627be48d

Former-commit-id: 0c2b84e790aeff74d11d058129ee86faa1c5781a