Use BrowserUtil for link navigation in ImageEditorPanel
BrowserUtil.browse() seems to gracefully handle errors coming from Java
when browsing the URL fails. Furthermore, it's used in many places in
tools/base, so if an OS (e.g. ChromeOS) doesn't support it, at least we
are consistent with other instances, and will leverage any fixes to the
root cause.

Bug: 317337602
Test: N/A
Change-Id: I421bfaba8e294a2dfdda6b8534223a07a74021e8

Former-commit-id: 0a25fcf102488e9f0a900e8f547d9575bc6c87b1