AST-based control flow graph
This CL updates the control flow graph, used by the WakelockDetector,
from being bytecode based (which means it doesn't run in the IDE,
doesn't have as accurate information to source offsets etc), over to
being AST based. It also updates the WakelockDetector to operate on
source code instead of bytecode.

Test: Many new tests (ControlFlowGraphTest, along with
      WakelockDetectorTest)
Bug: None

Change-Id: I4347d7e59bbddf928b5295db11e2f2a877bb8f5c

Former-commit-id: e476e81a80488de8bb2015554c40be453f3e6782