Add isStartupTask to session command and data protos
The isStartupTask is a boolean flag passed in the BeginSession
command and is used/extracted from the resulting SessionStarted
data retrieved on successful session start. This allows for the
detection of whether the session/task is done on startup of a
process or not. If it is a startup task, then we will only invoke
a subset of the profiler task handler behavior.

Bug: 277798531
Test: none
Change-Id: Icde8822a0506ae82282fac27e172f4909d31b6c2

Former-commit-id: 83a54fc5dddd8903ad8371866fac6beff6b36943