Allow path variables in XmlParser.getLocation()
Previously, there could be a LintError when XmlParser.getLocation() was
called for a source file of a library module dependency. This is because
library dependency source file paths contain path variables during
downstream module lint analysis.

With this change, XmlParser.getLocation() just returns
Location.create(file) if the file's path starts with a path variable.

Fixes: 303215439
Test: LintStringFormatDetectorTest
Change-Id: I7e0dd2f059df7ef7b79de2dabec895ac9d0752c7

Former-commit-id: 10a81216cb4291701505be0ac432b94ea3262326