Implement DeviceState.isReady in AVD plugin
Rework the implementation of LocalEmulatorDeviceHandle's stateFlow:
The current version allows the state to be updated in various contexts,
and uses `update` to make updates atomic.

However, as more state updates are added, it becomes difficult to reason
about the various concurrent operations that could be going on and how
they could interact.

This changes the LocalEmulatorDeviceHandle to use an actor-based design:
a single flow-generating coroutine maintains the state of the handle as
local variables, sequentially processes messages sent to it, and
publishes a subset of the state in the flow.

Long-running operations like activation are requested by the client via
a message, launched by the actor, and the completion is signaled by the
actor on a Deferred. Since the actor has the full state available, this
makes it more practical to support things like cancelling activation or
monitoring the state of the emulator process in the future.

Bug: 317411917
Test: LocalEmulatorProvisionerPluginTest
Change-Id: I09578e7f37495962653aa521b37fda9fc7b6b061

Former-commit-id: d8c7e25608fac73abb7c44ebabfef2d4dd03970d