Publish minified runtime classes from lib
With this change, inter-project publishing for libraries is more
consistent with publishing libraries as AARs. In both cases, the
libraries program classes are minified if applicable and the local file
dependencies classes are never minified.

To help users adopt this new behavior, a boolean flag is introduced and
will be enforced in 9.0.

Bug: 263197720
Test: MinifyLibTest, ProguardAarPackagingTest
Change-Id: I5b773131437720291fe02a2897d4171596e33acd

Former-commit-id: ff354eed2b1d4faa5dd4a839baebd6230c92a812