Return null when storage object does not exist.
Need to ensure that the download fun works properly by returning null
when the storage object does not exist.

This can occur when the logcat object is specified but not written to by
firebase. The function was supposed to return null when the storage
object does not exist but instead was throwing an HttpResponseException

Bug: 319510292
Test: Added

Change-Id: Iefe6db65e5b8eae3b8958d7f19223d6f13cdd249

Former-commit-id: 9614f35c35103d5ce9968b776f58b63fb18d77ae