Set isReady after getting properties
The previous change separated setting bootStatus from reading device
properties, meaning that we get two separate state updates: one setting
is isReady=true, the other updating to the newly-read properties. This
is not necessarily wrong, but it causes test flakiness, and it is
simpler for clients to know when they receive the ready status that the
properties are up-to-date.

Bug: 320514147
Test: LocalEmulatorProvisionerTest, 1000 runs
Change-Id: Id48fb4038c3f98cb6bce02212e349bbfb67b6bda

Former-commit-id: efc6315259ba928cd3bab752e53c024e974886c1