Add support for jar dependency to pom artifact
Example:
 - We want to depend on nl.jqno.equalsverifier:equalsverifier:3.15.6
 which is a jar dependency.

 - However, the pom file declares packaging=pom.

Before this change, we used to overwrite the 'jar' dependency with the
'pom' packaging we read from the pom file. However, that's not ideal as
it makes us unable to handle the equalsverifier case.

This change reduces the jar->pom overwrite case to be exception, and
only applies it for known cases (we only have one, and that's
groovy-all).

Bug: 320775555
Test: n/a
Change-Id: I3ad326848c8c3112a81c958a465e99e17ab99eb0

Former-commit-id: ff9f511b5a8f32207ddb380de4350409fd19484b