Use Unix-style relative paths for dexing transform desugaring graph
To make the desugaring graph used by dexing transforms relocatable,
we previously used relative paths of the files (as opposed to
absolute paths).

However, that was not enough. We also need to make sure the relative
paths are in Unix style (using "/" as separators) so that they can
be used across filesystems.

Bug: 266599585
Test: Added DexingTransformTest.testDesugarGraph
      + Existing CacheableDexingTransformTest
Change-Id: Ie79e9337dba5bea2ed392bb8cdb16444eb1c0f04

Former-commit-id: 8473e5fa2f36600947d29a49b6cce17ee46dbc45