Allow sync to finish when namespace not specified
Previously, when namespace was not specified in the DSL, there would be
a RuntimeException thrown during sync.

With this change, VariantDslInfoImpl instead reports a sync error and
uses a placeholder namespace value so that sync can finish.

This change also adds TYPE_EXCEPTION to SyncIssue because without it,
SyncIssueUsageReporterTest was failing. SyncIssue and IdeSyncIssue were
otherwise in sync, so adding TYPE_EXCEPTION to SyncIssue seemed like the
right thing to do.

Fixes: 321392253
Test: MissingNamespaceTest
Change-Id: Idcc9eba6449b98c9786d4ed5a2312eb0d7b1b62c

Former-commit-id: 6831417966caea3b21510eb866919e2ce2d829b8