Add packaging.jniLibs.testOnly
This change adds DSL and Variant API to force specified .so files
to be excluded from the main APK or AAR and be packaged in the
test APK(s) instead.

The testOnly DSL is added to the global packaging.jniLibs DSL
alongside the existing excludes, pickFirsts, and keepDebugSymbols.

For the Variant API, new interfaces were added because it didn't
make sense to add testOnly to the existing JniLibsPackaging
interface for test components or test modules.

In order to support the existing nested structure of the packaging
Variant API, this change adds TestedComponentPackaging and
TestedApkPackaging interfaces, along with corresponding
JniLibsTestedComponentPackaging and JniLibsTestedApkPackaging
interfaces.

Bug: 236913987
Test: JniLibsTestOnlyTest, CmakeBasicProjectTest
Change-Id: I54373614d71555fcf4431497ebf0c69bdde595f8

Former-commit-id: c1a00e31e8fa207319fd75f382940a995b6e3786