Lock down access to indirect resources in lint
This fixes
  321010878: Handle library layouts in MissingInflatedId checks
as well as similar problems that can happen in some other lint checks.

In Iguana, lint can no longer look at indirect module resource paths
while in analysis mode. (This is because AGP recently changed the way
the resource repositories are initialized; for caching purposes, the
paths now contain path variables, and these can't be resolved during the
analysis phase.)

This CL fixes the existing lint checks to handle this gracefully --
generally by only using project local resource repositories during the
analysis phase. Usually the resources are in the same module as the
resource or code reference, so this solves it efficiently. However, it's
possible for the resource to be in a module dependency. For some checks,
when the resource isn't found locally, this is handled via partial
analysis instead; store partial data and then look it up during the
reporting phase. In some other checks, this is deemed a niche use case
and therefore we just gracefully ignore these.

There is a third case (such as the StringFormatDetector and
PxUsageDetector) which access resources from other modules, but only to
report locations (and/or including the qualifier and filename in the
error message). We continue to allow these to include module resources,
and to properly support these locations we're now tracking locations in
the resource repository persistently.

There were a couple of lint checks which were using the resource
repository to make sure that a resource URL is valid -
NetworkSecurityConfigDetector, ManifestDetector. This is no longer
necessary; aapt enforces this now -- e.g.
  AndroidManifest.xml:5:5-24:19: AAPT: error: resource
  xml/data_extraction_rules2 (aka
  com.example.myapplication:xml/data_extraction_rules2) not found.

In order to catch future bugs and third party checks affected by the
same problem, this CL also updates the lint test infrastructure to lock
down indirect resource path access. Attempting to look up ResourceItem
source paths from any other module than the current one will trigger an
error during the analysis phase, similar to (and reusing) the same
mechanism already there for accessing the main project or merged
manifest.

This CL also adds multi-module tests for all the checks that were using
module+dependencies resource repositories. In order to make sure no
detectors were accidentally overlooked, it also adds a new test mode
(TestMode.MODULE_RESOURCES) which takes projects and turns them into
multi-module projects where resources are split into a separate module
from the code (and value resources separate from file resources), and
ensures that the above lint errors aren't reported. (This revealed one
more detector that had to be fixed, PxUsageDetector.)

Example:

    ../lib/res/layout/casts.xml: Error: The lint detector
        com.android.tools.lint.checks.ViewTypeDetector
    called ResourceItem.getSource() during module analysis.

    This does not work correctly when running in test.
    You can only call this on resources in the current module, not
    library resources.

    In particular, there may be false positives or false negatives
    because the lint check may be using the minSdkVersion or manifest
    information from the library instead of any consuming app module.

    Contact the vendor of the lint issue to get it fixed/updated (if
    known, listed below), and in the meantime you can try to work around
    this by disabling the following issues:

        "FindViewByIdCast","WrongViewCast"

    ...

    Call stack: LintResourceRepository$Companion$removeFileAccess$
      withoutSource$1.invalid(LintResourceRepository.kt:366)
      ←LintResourceRepository$Companion$removeFileAccess$withoutSource$i
       1.getSource(LintResourceRepository.kt:372)
      ←ViewTypeDetector.getViewTags(ViewTypeDetector.kt:389)
      ←ViewTypeDetector.visitMethodCall(ViewTypeDetector.kt:268)←...

Bug: 321010878
Test: Many new tests included
Change-Id: I5b9c09a99260b4f605b817734844d142e241b91d

Former-commit-id: efe489085c9e2ba858ac3d7f9d6d1862e7e90df4