Allow detectors to write error messages with indentation and continuations
Lint's string handling generally automatically trimIndents everything --
issue explanations, test results, etc. This wasn't the case for
error messages, but I've seen a detector attempt to do it, so we might
as well support it.

(When trimming error strings this also revealed that one built-in
warning (lint baseline issues fixed) actually had a trailing newline;
this updates the tests for that.

Test: Unit test included
Bug: Not filed
Change-Id: I5780bd2829fda5fa1e65c8a3e8d9b3aa21b91875

Former-commit-id: 5f7f27c938e2807f1965dbb0293db26057d7f3e5