Call `tryTrackResponse` From `contentEncoding`
This method needs to be called from all locations that complete a request.
We already do this for other places like `getHeaderField`, `contentLength` etc.

Not doing this will cause user code to throw if `getContentEncoding()` is called before accessing the input stream.

This cl also adds code in the test app that triggers this bug.

Fixes: 289769163
Test: Added
Change-Id: I7110c23bf4c2d00f8148741721f992ffcec49024

Former-commit-id: 267a124bd3df7faefb9c31933517f720004382c7