Ensure UTP configurations are eagerly created
Configurations that are created during dependency
resolution may cause issues like
https://issuetracker.google.com/324631534.

In this particular case, while resolving the
androidTest runtime classpath, KGP may trigger
confiugration of UTP-test tasks. These tasks create
new configurations and add them to the configurations
container. This results in ConcurrentModificationException
in Gradle, and failed sync.

This change ensures all configurations are added
to container as soon as the task is registered.

Bug: 324631534
Test: synced androidx-main successfully 12 times,
      previously it was failing 1/3 times
Change-Id: Id16c8e5163f086510f761e4529eeb2a2f86a0297

Former-commit-id: 35a1c5c34b26d7557886e6daaf3217c2a188a618