Remove Fake Android From Prod Deps
This causes a crash in the agent (b/327727908).

But not having the dep in prod makes a test fail (b/328052863)

Turns out there is some strange behavior where if `ApplicationInfo` is a Kotlin class, the public field `uid` is for some reason seen as `private` when run from Bazel.

Changing the class to Java fixes it.

Fixes: 327727908
Fixes: 328052863
Test: NA
Change-Id: Ia317370900aab36deaa789ec87677ebe822dd342

Former-commit-id: eed5aee1ff8e27a667a5ef274e8ca92ebd237027