Add test for UastGradleVisitor
This test is a mirror of GroovyGradleVisitorTest and
LintIdeGradleVisitorTest: checking that the expected checkMethodCall
and checkDslPropertyAssignment calls happen in the course of visiting
the file.  It is a distorting mirror, though, which itself reflects
the differences between Groovy and Kotlin, both syntactically and
semantically.  The test cases were adjusted by hand, with several
being deleted as essentially testing features that do not exist in
Kotlinscript Dsl.

Adding the test cases reveals some meaningful differences between the
Uast visitor and the Groovy ones.  Address those differences by
modifying relevant parts of the UastGradleVisitor.

Bug: 279885639
Test: added
Change-Id: I48b9eaa5fde6f1637e27ad91f5c51581f1bc8167

Former-commit-id: 8aea229743f15fc116271484301b9423410f4eed