Refine handling of google repository declaration check
We want to hide suggestions to update AGP if the project might be so
old that it does not have the correct repository declarations.  This
is a niche case, but if it still exists the project will have
buildscript repositories declarations.  More modern projects do not
have any buildscript repositories declarations, with repositories
being declared in settings under pluginManagement instead; if we do
not find any repository declarations at all in a Gradle build file, do
not hide the AGP version warnings.

Move one of the maybeReportAgpVersionIssue to afterCheckFile.  In
practice that's still too early because the agp version check will
happen in a deferred runnable, but it's still clearer to have it there
than in the middle of the repositories block update.

Bug: 144442233
Test: existing test simplified
Change-Id: I1aaf27991004f4b41ebb0799ae582bb327dd092e

Former-commit-id: d9fcf45a934ef28f18e5332e052df4303d685ed6