Fix internal declarations access in test compilations
Pass main compilation classes.jar to test compilation task's
friend paths flag. This is used by the kotlin compiler to give access to internal declarations between different modules.
Currently, friend paths is populated by KGP and it defaults to
main compilations classes dir which does not match with the classpath flag (which is controlled by AGP and uses the jar)

Bug: 328227851
Test: added regression tests
Change-Id: I24b673a4a554057d6c9b103ee188306bac62c0ef

Former-commit-id: 2534b26de883dc1e496e2dd9d8cc74b082c9d70b