267449090: Lint information for library desugaring is missing fields
This CL updates lint's support for core library desugaring to also
handle field desugaring, now that R8 handles those (which is
particularly important for NIO methods).

It also updates the testing support for core library desugaring (so for
example the gradle test files can turn it on via DSL).

It also updates lint's fallback desugaring database (which is not used
when lint is run with AGP where we get the exact desugaring files from
the build system.)

Finally, it also fixes a bug (327670482) where inherited methods were
not correctly handled (e.g. Collection.removeIf is in the desugared
method list, but code references to ArrayList.removeIf were not matched
with the inherited method.)

Fixes: 267449090
Fixes: 327670482
Test: Unit tests included and updated
Change-Id: Ia2bbdb580239ceb39719e0e837f3fdbc8e790e94

Former-commit-id: 7cd900c1d6ace2584979fae35e3abda3cda47343