Upgrade skiko-awt from 0.7.85.4 to 0.7.85.5
...by adjusting the artifact versions in the bazel/maven/artifacts.bzl
and then running the bazel/maven/maven_fetch.sh script.

This new version contains a hot fix for a fatal bug that would crash
studio when any skiko dependent UI (so compose UI) was invoked. This
fix defaults to software rendering when the hardware rendering fails.
This issue was only present on linux, however the linux specific
and other OS specific artifacts had no change in the native code,
so no new version of them was made (they are kept 0.7.85.4).

Bug: 272822954
Fixes: 326131113
Test: manual
Change-Id: I3732d69dad3189dc807b9ed77ab73b420a50a1a7

Former-commit-id: a0a926165d76a7d40df51f3b9038dfa61ba97a3c