Add Force Open Option
If true, the agent will handle unopened databases differently.

Without this flag, unopened DB's are reported as offline and AS is not able to perform any SQL on them.

With this change, the agent will force-open a connection to the DB.

When performing any SQL on the DB, the agent will use the forced connection unless the database has since been opened by the use in ReadWrite mode.

Test: Added
Bug: 173312205
Bug: 230166342
Change-Id: I596f647cc40b2217944b7046137f56f4b7afa28e

Former-commit-id: 2db2ec0171aba79b56fe11de5f7ac0f0fe4fb806