[lint] Handle .gradle.something files like .kts files
At least to the extent of having those files included in the
gradleKtsContexts UastSources element.  This change doesn't do
anything beyond that: in particular, the UastParser will not generate
a UFile for these files without a Uast language plugin implementation
for declarative Gradle files ("Something files") being available to
the lint client.

Bug: 330888875
Test: existing
Change-Id: I880991b6efa0c336c0bc9b9260485e6018ff932e

Former-commit-id: 0941afc048af3ff619c767352840ba881d99067d