Fix DSL decorator
Signature parameter is meant to provide information about
the generic types assosicated with a field. An empty string does not convey
any meaningful information and it might even be interpreted as an invalid signature by
a reflection system leading to GenericSignatureFormatError

Passing null, implies that the field is a plain, non-generic type.
Indicates to the reflection system that there is no generics to process.

Bug: na
Test: using javap
Change-Id: Ib4358df126cbd7b6225050c37def1961df4ac5a8

Former-commit-id: 6384a03067864eae4c0a3d19d9b21120d4e34672