Support test only module with Privacy Sandbox
To support test only module for application that use privacy sandbox
sdk, we need to first determine if the app under test uses privacy
sandbox. In this change, we introduce a flag in TestExtension that
indicates that the app we are testing uses privacy sandbox. Secondly, we
need to expose Privacy Sandbox related Apks necessary to run the
instrumentation test as Android Artifacts for test module to consume.
To fix the connected tests, we are using 1.9.2 version of kotlin because
the androidx api-generator is unable to read kotlin class metadata for
higher version.

Bug: 331635507
Test: PrivacySandboxSdkConnectedTest
Change-Id: I42c45af093e1b2bebdf6d3c543191151fe98df51

Former-commit-id: e1302723a9bfc130d6418cf901ffc05b59831c70