Fix `IDevice.getClientName` for DeviceImpl + adblib
To match ddmlib implementation `IDevice.getClientName`
should return a package name . It appears that this bug didn't cause
any actual issues because `IDevice.getClientName` is used in
code that is handling older android devices for which the
package name is derived from a process name. For example this
is used in legacy profiler java/kotlin allocations tracking.

Test: existing
Bug: 330938911
Change-Id: I290179b4852e09b83e7fd7a2c6845556ac4a2851

Former-commit-id: a538187551940a941e8383723e720df393f405c4