Make Force-Open Imply Keep-Open
Using a non-native connection has its issues so we want to avoid using it if there's a
native connection available.

We still keep the forced connection for now. We will get rid of it in another cl. Note
that we should keep the forced connection around if the only available connection is
readonly.

Bug: 173312205
Test: Added
Change-Id: Ic1b15a2eef1a64ae33a1e0df0e3aac451b896020

Former-commit-id: 8f9e66a335b8473962c48d508072e8f74f8568d0