Make res folder added through old API visible
res folder that were added through the old variant API used
a dedicated storage on the old BaseVariantData. This created
a situation where the folders were not visible to users of the
new variant API. Eliminated the storage and register the folders
to the same storage used by the new API.

Bug: 331946533
Test: Added regression test.

lease enter the commit message for your changes. Lines starting

Change-Id: I6df032572029957f6a1330c87452decabfd1c15e

Former-commit-id: 4596fb8c56e57b11d793b3c8a39a4d797f8f90fd