added code coverage to device tests in variant API
added APIs to set the code coverage in the DeviceTestBuilder and
a read only version in the DeviceTest interface.

Internal structures are now initialized from the DeviceTest
version.

There is potential future code optimization that can be added,
by refactoring or removing some CreateConfig interfaces but this will
be done in a subsequent CL.

Test: Existing, new unit tests
Bug: 317492905
Change-Id: I1d9fd93b5f7af687a69a4c19175f655fef587638

Former-commit-id: 0ac56d50b7695cdaffa3fa10108536bdc4d5246b