adblib: fix bug with orphan JDWP process
The `JdwpProcessFactory` would sometimes create
`JdwpProcessImpl` entries for JDWP processes that
have already terminated on the the device, leading to
failure to open JDWP session every 2 seconds for
each "orphan" process (until the device is disconnected)

This started to happen when the concept of "delegate
adb session" was introduced to allow sharing a single
JDWP connection to a given process between multiple
`AdbSession` instances.

This change addresses the issue by making sure
each `JdwpProcessFactory` keeps track of the list
of active JDWP process IDs, and uses that list to
cancel `JdwpProcess` instances that should not
be active anymore.

Note: This is commit is the same as
commit 64c767e4e69778ac37c78c47b6d4e2e627e4d5b8 [formerly 1e15805cfdf2111d27499374b2c92af1de24c0ed] that was reverted in
commit b13d4b9f16923b6146406ab2297a2c85bb7f1f8d [formerly bc2c7fcc2027aa94fe25f1dc35805870e5872eb6], but with a few
additional fixes to address flakiness exposed by tests.

Fixes: 322467516
Test: Included
Change-Id: Ibf5ea08f130d542c61b376e1ee7d530382fa9b32

Former-commit-id: 571982f4f6812fd5e51f6010c8f42302f2270b23