268575580: Allow suppressing SpUsage on the dimension element instead
Using a dp dimension for text-size related attributes is usually wrong,
and lint issues a warning, pointing primarily to the text attribute,
but also showing the relevant dimension attribute. Some users have
tried suppressing these attributes on the dimension declaration. This
isn't the ideal place to suppress it, since this will then allow *all*
text attribute usages for that dimension, but we'll allow it.

Test: Unit test included
Fixes: 268575580
Change-Id: Icd7dd2e2b614b12826c43ea2019531ab5301c15a

Former-commit-id: 70417d5a3d630356ed3e1cd5e2eb9253e16e4625