Lint api check: Don't flag calling super from overridden method
In code like this:

  override fun onPictureInPictureModeChanged(mode: Boolean, config: Configuration) {
    super.onPictureInPictureModeChanged(mode, config)
  }

we shouldn't flag the super.onPictureInPictureModeChanged method as
requiring API level N, since generally this method will only be called
on API level N as well.

We were already handling this scenario for methods in the SDK, but it
was not handled for API requirements supplied via @RequiresApi
annotations, such as the various AndroidX methods; this CL fixes that.

Test: Unit test included
Bug: Not filed
Change-Id: Id6af1cc9feab01cf66ca30985b1aded867ef871b

Former-commit-id: bfb013075abb862b2fe304457d6d9e29eca16bdb