[SDK Index] Add issue description as note
... for messages from developers.

Recent versions of the SDK Index snapshot contain a description for the
critical issues (messages from developers). Now we can display that
infromation in Android Studio. This change adds that description in the
genererateBlockingCriticalMessage and generateCriticalMessage methods so
users can see those descriptions without needing to navigate into the
SDK Index.

Bug: 336568007
Test: GooglePlaySdkIndexTest
Change-Id: Ifccc22ffdd3a0bc0ef19b21050c55e37fe6e4b9b

Former-commit-id: 14e312d19f43caeb80bc2e15b8287901e6504dc0