Assume That Tracked Database Are Open
This is true because `registerReference` only adds open database.
The exception is when a previously open database is closed, handleDatabaseSignal() will see it in databases[] before it calls unregisterReference. This will actually help us downstream when I refactor handleDatabaseSignal.

Note that the change to DatabaseLockingTest is caused by an artificial failure that is triggered because DatabaseRegistry is not updated properly while the test is running.

Bug: NA
Test: Existing
Change-Id: I97e78c240be7d1b922245f21146bf23f368c20e8

Former-commit-id: 3ad7ce73ee9ad9714383339b8aa722b29445a7b5