[SDK Index] Show warning only in 3rd party libraries
We currently display a message that says "These versions have not been
reviewed by Google Play. They could contain vulnerabilities or policy
violations. Carefully evaluate any third-party SDKs before integrating
them into your app" whenever we recommend a version in the SDK index
issues, regardless of the author of the library.

This change checks if the library is a third party before adding that
part in the issue message. The check is done using hasGroupId of
GoogleMavenRepository.

Bug: 338654964
Test: GooglePlaySdkIndexTest, GradleDetectorTest
Change-Id: I7d4c5a07e04baff73d64caf20d07bd1503bd93ce

Former-commit-id: 92e65a1451ddc08c3fdefa59072526724fdbd7c4